Cardano.Node.Configuration.LedgerDB

Documentation

newtype DeprecatedOptions

Some options that existed in the TopLevel were now moved to a subsection. We use this field to propagate the results from parsing those into the monadic part of the node so that we can emit warnings.

Constructors

DeprecatedOptions [String]

data LedgerDbConfiguration

Constructors

LedgerDbConfiguration SnapshotPolicyArgs QueryBatchSize LedgerDbSelectorFlag DeprecatedOptions

data LedgerDbSelectorFlag

Choose the LedgerDB Backend

As of UTxO-HD, the LedgerDB uses either an in-memory backend or an LSM-tree backend to keep track of differences in the UTxO set.

V2InMemory: uses more memory than the minimum requirements but is somewhat faster.

V2LSM: keeps the UTxO set on disk in an LSM tree, which uses less memory but is somewhat slower.

Constructors

V2InMemory

V2LSM

Fields

(Maybe FilePath)

Maybe a custom path to the LSM database. If not provided the default will be used (fast-storage/lsm).

(Maybe FilePath)

Maybe a path to which the LSM backend will export standalone snapshots on every snapshot. If not provided, no standalone snapshots are exported.

data Gigabytes

A number of gigabytes.

noDeprecatedOptions :: DeprecatedOptions

selectorToArgs :: (LedgerSupportsProtocol blk, LedgerDbSerialiseConstraints blk, CanUpgradeLedgerTables LedgerState blk) => LedgerDbSelectorFlag -> FilePath -> StdGen -> (LedgerDbBackendArgs IO blk, StdGen)