Cardano.Node.Configuration.NodeAddress

Node addresses

data NodeAddress' addr

IPv4 or IPv6 address with a port number.

Constructors

NodeAddress

Fields

naHostAddress :: !addr

naPort :: !PortNumber

type NodeIPAddress = NodeAddress' NodeHostIPAddress

nodeAddressToSockAddr :: NodeIPAddress -> SockAddr

type NodeIPv4Address = NodeAddress' NodeHostIPv4Address

type NodeIPv6Address = NodeAddress' NodeHostIPv6Address

type NodeDnsAddress = NodeAddress' NodeHostDnsAddress

nodeIPv4ToIPAddress :: NodeIPv4Address -> NodeIPAddress

nodeIPv6ToIPAddress :: NodeIPv6Address -> NodeIPAddress

nodeDnsAddressToRelayAccessPoint :: NodeDnsAddress -> RelayAccessPoint

newtype NodeHostIPAddress

Constructors

NodeHostIPAddress

Fields

unNodeHostIPAddress :: IP

nodeHostIPAddressToSockAddr :: NodeIPAddress -> SockAddr

newtype NodeHostIPv4Address

Constructors

NodeHostIPv4Address

Fields

unNodeHostIPv4Address :: IPv4

newtype NodeHostIPv6Address

Constructors

NodeHostIPv6Address

Fields

unNodeHostIPv6Address :: IPv6

nodeHostIPv4AddressToIPAddress :: NodeHostIPv4Address -> NodeHostIPAddress

nodeHostIPv6AddressToIPAddress :: NodeHostIPv6Address -> NodeHostIPAddress

newtype NodeHostDnsAddress

Domain name.

Constructors

NodeHostDnsAddress

Fields

unNodeHostDnsAddress :: Text

nodeHostDnsAddressToDomain :: NodeHostDnsAddress -> Domain

data PortNumber

Port number. Use the Num instance (i.e. use a literal) to create a PortNumber value.

>>> 1 :: PortNumber
1
>>> read "1" :: PortNumber
1
>>> show (12345 :: PortNumber)
"12345"
>>> 50000 < (51000 :: PortNumber)
True
>>> 50000 < (52000 :: PortNumber)
True
>>> 50000 + (10000 :: PortNumber)
60000

type SocketPath = File Socket 'InOut

newtype File content (direction :: FileDirection)

A file path with additional type information to indicate what the file is meant to contain and whether it is to be used for reading or writing.

Constructors

File

Fields

unFile :: FilePath