Cardano.Node.Configuration.POM

Documentation

data NodeConfiguration

Constructors

NodeConfiguration

Fields

ncNodeIPv4Addr ∷ !(Maybe NodeHostIPv4Address)

ncNodeIPv6Addr ∷ !(Maybe NodeHostIPv6Address)

ncNodePortNumber ∷ !(Maybe PortNumber)

ncConfigFile ∷ !ConfigYamlFilePath

Filepath of the configuration yaml file. This file determines all the configuration settings required for the cardano node (logging, tracing, protocol, slot length etc)

ncTopologyFile ∷ !TopologyFile

ncDatabaseFile ∷ !DbFile

ncProtocolFiles ∷ !ProtocolFilepaths

ncValidateDB ∷ !Bool

ncShutdownIPC ∷ !(Maybe Fd)

ncShutdownOnSlotSynced ∷ !MaxSlotNo

ncProtocolConfig ∷ !NodeProtocolConfiguration

ncSocketPath ∷ !(Maybe SocketPath)

ncDiffusionMode ∷ !DiffusionMode

ncMaxConcurrencyBulkSync ∷ !(Maybe MaxConcurrencyBulkSync)

ncMaxConcurrencyDeadline ∷ !(Maybe MaxConcurrencyDeadline)

ncLoggingSwitch ∷ !Bool

ncLogMetrics ∷ !Bool

ncTraceConfig ∷ !TraceOptions

data PartialNodeConfiguration

Constructors

PartialNodeConfiguration

Fields

pncNodeIPv4Addr ∷ !(Last NodeHostIPv4Address)

pncNodeIPv6Addr ∷ !(Last NodeHostIPv6Address)

pncNodePortNumber ∷ !(Last PortNumber)

pncConfigFile ∷ !(Last ConfigYamlFilePath)

Filepath of the configuration yaml file. This file determines all the configuration settings required for the cardano node (logging, tracing, protocol, slot length etc)

pncTopologyFile ∷ !(Last TopologyFile)

pncDatabaseFile ∷ !(Last DbFile)

pncProtocolFiles ∷ !(Last ProtocolFilepaths)

pncValidateDB ∷ !(Last Bool)

pncShutdownIPC ∷ !(Last (Maybe Fd))

pncShutdownOnSlotSynced ∷ !(Last MaxSlotNo)

pncProtocolConfig ∷ !(Last NodeProtocolConfiguration)

pncSocketPath ∷ !(Last SocketPath)

pncDiffusionMode ∷ !(Last DiffusionMode)

pncMaxConcurrencyBulkSync ∷ !(Last MaxConcurrencyBulkSync)

pncMaxConcurrencyDeadline ∷ !(Last MaxConcurrencyDeadline)

pncLoggingSwitch ∷ !(Last Bool)

pncLogMetrics ∷ !(Last Bool)

pncTraceConfig ∷ !(Last TraceOptions)

defaultPartialNodeConfiguration ∷ PartialNodeConfiguration

Default configuration is mainnet

lastOption ∷ Parser a → Parser (Last a)

makeNodeConfiguration ∷ PartialNodeConfiguration → Either String NodeConfiguration

parseNodeConfigurationFP ∷ Maybe ConfigYamlFilePath → IO PartialNodeConfiguration

pncProtocol ∷ PartialNodeConfiguration → Either Text Protocol

ncProtocol ∷ NodeConfiguration → Protocol