Cardano.Node.Configuration.Topology

Documentation

newtype TopologyError

Constructors

NodeIdNotFoundInToplogyFile FilePath

data NetworkTopology adr

Describes the non-P2P topology of a node. Whenever the node actually runs, the type parameter adr should be RemoteAddress. However, we might want to use and serialize this type with adr being NodeId, or another placeholder type, if we want the user to be able to edit the topology without knowing the actual addresses of the nodes: those might only be knowable at runtime.

Constructors

MockNodeTopology ![NodeSetup adr]

RealNodeTopology ![adr]

newtype NodeHostIPAddress

Constructors

NodeHostIPAddress

Fields

unNodeHostIPAddress :: IP

newtype NodeHostIPv4Address

Constructors

NodeHostIPv4Address

Fields

unNodeHostIPv4Address :: IPv4

newtype NodeHostIPv6Address

Constructors

NodeHostIPv6Address

Fields

unNodeHostIPv6Address :: IPv6

data NodeSetup adr

Constructors

NodeSetup

Fields

nodeId :: !Word64

nodeIPv4Address :: !(Maybe NodeIPv4Address)

nodeIPv6Address :: !(Maybe NodeIPv6Address)

producers :: ![adr]

data RemoteAddress

Domain name with port number

Constructors

RemoteAddress

Fields

raAddress :: !Text

Either a dns address or an ip address.

raPort :: !PortNumber

Port number of the destination.

raValency :: !Int

If a DNS address is given valency governs to how many resolved IP addresses should we maintain active (hot) connection; if an IP address is given valency is used as a Boolean value, 0 means to ignore the address;

nodeAddressToSockAddr :: NodeIPAddress -> SockAddr

readTopologyFile :: FromJSON adr => NodeConfiguration -> IO (Either Text (NetworkTopology adr))

Read the NetworkTopology configuration from the specified file. While running a real protocol, this gives your node its own address and other remote peers it will attempt to connect to.

readTopologyFileOrError :: FromJSON adr => NodeConfiguration -> IO (NetworkTopology adr)

remoteAddressToNodeAddress :: RemoteAddress -> Maybe (Either NodeIPAddress (NodeDnsAddress, Int))

Parse raAddress field as an IP address; if it parses and the valency is non zero return corresponding NodeAddress.