Cardano.Node.Configuration.TopologyP2P

Documentation

newtype TopologyError

Constructors

NodeIdNotFoundInToplogyFile FilePath

data NetworkTopology

Constructors

RealNodeTopology !LocalRootPeersGroups ![PublicRootPeers] !UseLedger

newtype PublicRootPeers

Constructors

PublicRootPeers

Fields

publicRoots ∷ RootConfig

data LocalRootPeersGroup

A local root peers group. Local roots are treated by the outbound governor in a special way. The node will make sure that a node has the requested number (valency) of connections to the local root peer group.

Constructors

LocalRootPeersGroup

Fields

localRoots ∷ RootConfig

valency ∷ Int

newtype LocalRootPeersGroups

Constructors

LocalRootPeersGroups

Fields

groups ∷ [LocalRootPeersGroup]

data RootConfig

Each root peer consists of a list of access points and a shared PeerAdvertise field.

Constructors

RootConfig

Fields

rootAccessPoints ∷ [RelayAccessPoint]

a list of relay access points, each of which is either an ip address or domain name and a port number.

rootAdvertise ∷ PeerAdvertise

advertise configures whether the root should be advertised through gossip.

newtype NodeHostIPAddress

Constructors

NodeHostIPAddress

Fields

unNodeHostIPAddress ∷ IP

newtype NodeHostIPv4Address

Constructors

NodeHostIPv4Address

Fields

unNodeHostIPv4Address ∷ IPv4

newtype NodeHostIPv6Address

Constructors

NodeHostIPv6Address

Fields

unNodeHostIPv6Address ∷ IPv6

data NodeSetup

Constructors

NodeSetup

Fields

nodeId ∷ !Word64

nodeIPv4Address ∷ !(Maybe NodeIPv4Address)

nodeIPv6Address ∷ !(Maybe NodeIPv6Address)

producers ∷ ![RootConfig]

useLedger ∷ !UseLedger

data PeerAdvertise

Constructors

DoNotAdvertisePeer

DoAdvertisePeer

newtype UseLedger

A newtype wrapper around UseLedgerAfter which provides FromJSON and ToJSON instances.

UseLedgerAfter is used to configure from which slot a p2p node can use on chain root peers.

Constructors

UseLedger UseLedgerAfter

nodeAddressToSockAddr ∷ NodeIPAddress → SockAddr

readTopologyFile ∷ Tracer IO (StartupTrace blk) → NodeConfiguration → IO (Either Text NetworkTopology)

Read the NetworkTopology configuration from the specified file.

readTopologyFileOrError ∷ Tracer IO (StartupTrace blk) → NodeConfiguration → IO NetworkTopology

rootConfigToRelayAccessPoint ∷ RootConfig → [(RelayAccessPoint, PeerAdvertise)]

Transforms a RootConfig into a pair of RelayAccessPoint and its corresponding PeerAdvertise value.