Cardano.Node.Handlers.Shutdown

Generalised shutdown handling

withShutdownHandling

Arguments

∷ Maybe Fd

→ Trace IO Text

→ IO ()
Action to potentially shutdown via file descriptor

→ IO ()

We provide an optional cross-platform method to politely request shut down. The parent process passes us the file descriptor number of the read end of a pipe, via the CLI with --shutdown-ipc FD

Watch ChainDB for passing a configured slot sync limit threshold,

maybeSpawnOnSlotSyncedShutdownHandler ∷ NodeConfiguration → Trace IO Text → ResourceRegistry IO → ChainDB IO blk → IO ()

Spawn a thread that would cause node to shutdown upon ChainDB reaching the configuration-defined slot.