Cardano.Node.Protocol.Alonzo

Documentation

data AlonzoProtocolInstantiationError

Constructors

InvalidCostModelError !FilePath

CostModelExtractionError !FilePath

AlonzoCostModelFileError !(FileError ())

AlonzoCostModelDecodeError !FilePath !Text

AlonzoGenesisFileError !(FileError ())

AlonzoGenesisDecodeError !FilePath !Text

GenesisFileNotFound !FilePath

renderAlonzoProtocolInstantiationError ∷ AlonzoProtocolInstantiationError → Text

Reusable parts

readAlonzoGenesis ∷ FilePath → ExceptT AlonzoProtocolInstantiationError IO AlonzoGenesis

In order to avoid introducing a separate Alonzo genesis file, we have added additional fields to the Shelley genesis that are required when hardforking to Alonzo. Unfortunately the ShelleyGenesis FromJSON instance exists in cardano-ledger-specs so we must duplicate code for now.