Cardano.Node.Protocol.Shelley

Protocol exposing the specific type

Use this when you need the specific instance

mkConsensusProtocolShelley ∷ NodeShelleyProtocolConfiguration → Maybe ProtocolFilepaths → ExceptT ShelleyProtocolInstantiationError IO (Protocol IO (ShelleyBlockHFC StandardShelley) ProtocolShelley)

Instantiate Protocol for Shelley specifically.

Use this when you need to run the consensus with this specific protocol.

Protocols hiding the specific type

Use this when you want to handle protocols generically

mkSomeConsensusProtocolShelley ∷ NodeShelleyProtocolConfiguration → Maybe ProtocolFilepaths → ExceptT ShelleyProtocolInstantiationError IO SomeConsensusProtocol

Make SomeConsensusProtocol using the Shelley instance.

This lets us handle multiple protocols in a generic way.

This also serves a purpose as a sanity check that we have all the necessary type class instances available.

Errors

data ShelleyProtocolInstantiationError

Constructors

GenesisReadError !FilePath !IOException

GenesisHashMismatch !GenesisHash !GenesisHash

GenesisDecodeError !FilePath !String

GenesisValidationFailure ![ValidationErr]

CredentialsReadError !FilePath !IOException

EnvelopeParseError !FilePath !String

FileError !(FileError TextEnvelopeError)

OCertNotSpecified

VRFKeyNotSpecified

KESKeyNotSpecified

renderShelleyProtocolInstantiationError ∷ ShelleyProtocolInstantiationError → Text

Reusable parts

readGenesis ∷ GenesisFile → Maybe GenesisHash → ExceptT ShelleyProtocolInstantiationError IO (ShelleyGenesis StandardShelley, GenesisHash)

readLeaderCredentials ∷ Maybe ProtocolFilepaths → ExceptT ShelleyProtocolInstantiationError IO [TPraosLeaderCredentials StandardCrypto]

genesisHashToPraosNonce ∷ GenesisHash → Nonce