Cardano.Node.Queries

Documentation

class ConvertTxId blk where

Convert a transaction ID to raw bytes.

Methods

txIdToRawBytes :: TxId (GenTx blk) -> ByteString

KES

newtype MaxKESEvolutions

The maximum number of evolutions that a KES key can undergo before it is considered expired.

Constructors

MaxKESEvolutions Word64

newtype OperationalCertStartKESPeriod

The start KES period of the configured operational certificate.

Constructors

OperationalCertStartKESPeriod Period

class GetKESInfo blk where

Minimal complete definition

Nothing

Methods

getKESInfoFromStateInfo :: Proxy blk -> ForgeStateInfo blk -> Maybe KESInfo

class HasKESInfo blk where

Minimal complete definition

Nothing

Methods

getKESInfo :: Proxy blk -> ForgeStateUpdateError blk -> Maybe KESInfo

data KESMetricsData

KES-related data to be traced as metrics.

Constructors

NoKESMetricsData
The current protocol does not support KES.

TPraosKESMetricsData

Fields

!Period

The current KES period of the hot key, relative to the start KES period of the operational certificate.

!MaxKESEvolutions

The configured max KES evolutions.

!OperationalCertStartKESPeriod

The start KES period of the configured operational certificate.

class HasKESMetricsData blk where

Minimal complete definition

Nothing

Methods

getKESMetricsData :: Proxy blk -> ForgeStateInfo blk -> KESMetricsData

General ledger

class LedgerQueries blk where

Methods

ledgerUtxoSize :: LedgerState blk -> Int

ledgerDelegMapSize :: LedgerState blk -> Int

Node kernel

newtype NodeKernelData blk

Constructors

NodeKernelData

Fields

unNodeKernelData :: IORef (StrictMaybe (NodeKernel IO RemoteAddress LocalConnectionId blk))

nkQueryChain :: (AnchoredFragment (Header blk) -> a) -> NodeKernel IO RemoteAddress LocalConnectionId blk -> IO a

nkQueryLedger :: IsLedger (LedgerState blk) => (ExtLedgerState blk -> a) -> NodeKernel IO RemoteAddress LocalConnectionId blk -> IO a

mapNodeKernelDataIO :: (NodeKernel IO RemoteAddress LocalConnectionId blk -> IO a) -> NodeKernelData blk -> IO (StrictMaybe a)

setNodeKernel :: NodeKernelData blk -> NodeKernel IO RemoteAddress LocalConnectionId blk -> IO ()

mkNodeKernelData :: IO (NodeKernelData blk)

Re-exports

data NodeKernel (m :: Type -> Type) addrNTN addrNTC blk

Interface against running relay node

Constructors

NodeKernel

Fields

getChainDB :: ChainDB m blk

The ChainDB of the node

getMempool :: Mempool m blk

The node's mempool

getTopLevelConfig :: TopLevelConfig blk

The node's top-level static configuration

getFetchClientRegistry :: FetchClientRegistry (ConnectionId addrNTN) (Header blk) blk m

The fetch client registry, used for the block fetch clients.

getFetchMode :: STM m FetchMode

The fetch mode, used by diffusion.

getNodeCandidates :: StrictTVar m (Map (ConnectionId addrNTN) (StrictTVar m (AnchoredFragment (Header blk))))

Read the current candidates

getPeerSharingRegistry :: PeerSharingRegistry addrNTN m

Read the current peer sharing registry, used for interacting with the PeerSharing protocol

getTracers :: Tracers m (ConnectionId addrNTN) addrNTC blk

The node's tracers

setBlockForging :: [BlockForging m blk] -> m ()

Set block forging

When set with the empty list '[]' block forging will be disabled.

type LocalConnectionId = ConnectionId LocalAddress

type RemoteConnectionId = ConnectionId RemoteAddress

data StrictMaybe a

Constructors

SNothing

SJust !a

fromSMaybe :: a -> StrictMaybe a -> a

Same as fromMaybe