Cardano.Node.Query

Description

Answer queries using the ledger state of the running node

This allows queries from within the running node as opposed to submitting queries via the LocalStateQuery protocol.

Documentation

answerQuery ∷ ∀ blk result. RunNode blk ⇒ Protocol IO blk (BlockProtocol blk) → NodeKernel IO RemoteConnectionId LocalConnectionId blk → HardForkCompatQuery blk result → IO result

Answer a general query about the current ledger state of the running node.

answerQueryWithLedgerState ∷ ∀ blk result. RunNode blk ⇒ Protocol IO blk (BlockProtocol blk) → ExtLedgerState blk → HardForkCompatQuery blk result → result

Answer a general query about the given ledger state.

type InterpreterQuery a = Qry a

To avoid confusing Qry with Query and 'HF.HardForkCompatQuery, we introduce a type synonym for the former which we can use in docstrings (and type signatures, of course).

Qry (and thus InterpreterQuery) is a query about slotsepochstime that a Interpreter can answer.

Query and HardForkCompatQuery are more general queries that can ask for anything related to the ledger state.

qryEpochStartTimeOfSlot ∷ SlotNo → InterpreterQuery RelativeTime

Return the start time of the epoch of the given slot.

getEpochStartTimeOfSlot ∷ ∀ blk. RunNode blk ⇒ Protocol IO blk (BlockProtocol blk) → NodeKernel IO RemoteConnectionId LocalConnectionId blk → SlotNo → IO (Either PastHorizonException UTCTime)

Execute the qryEpochStartTimeOfSlot InterpreterQuery against the current ledger state.

Will return a PastHorizonException when the current chain doesn't give us enough information about the given slot. For example, when syncing from scratch, we don't know yet when the transition to Shelley will happen, so we can't do conversions for slots from the Shelley era.