Cardano.Node.Run

Documentation

runNode :: PartialNodeConfiguration -> IO ()

checkVRFFilePermissions :: Tracer IO String -> File content direction -> ExceptT VRFPrivateKeyFilePermissionError IO ()

Make sure the VRF private key file is readable only by the current process owner the node is running under.