Cardano.Node.STM

Documentation

modifyReadTVar :: TVar a -> (a -> a) -> STM a

Mutate the contents of a TVar and return the new value of the TVar (non-strict).

modifyReadTVar' :: TVar a -> (a -> a) -> STM a

Mutate the contents of a TVar and return the new value of the TVar (strict).

modifyReadTVarIO :: TVar a -> (a -> a) -> IO a

Mutate the contents of a TVar and return the new value of the TVar (non-strict).

modifyReadTVarIO' :: TVar a -> (a -> a) -> IO a

Mutate the contents of a TVar and return the new value of the TVar (strict).