Cardano.Node.Startup

Documentation

data StartupTrace blk

Constructors

StartupInfo
Log startup information.

Fields

[SockAddr]

node-to-node addresses

(Maybe LocalSocketOrSocketInfo)

node-to-client socket path

(Map NodeToNodeVersion (BlockNodeToNodeVersion blk))

supported node-to-node versions

(Map NodeToClientVersion (BlockNodeToClientVersion blk))

supported node-to-client versions

StartupP2PInfo DiffusionMode
Log peer-to-peer diffusion mode

StartupTime UTCTime

StartupNetworkMagic NetworkMagic

StartupSocketConfigError SocketConfigError

StartupDBValidation

BlockForgingUpdate EnabledBlockForging
Log that the block forging is being updated

BlockForgingUpdateError ProtocolInstantiationError
Protocol instantiation error when updating block forging

BlockForgingBlockTypeMismatch
Mismatched block type

Fields

SomeBlockType

expected

SomeBlockType

provided

NetworkConfigUpdate
Log that the network configuration is being updated.

NetworkConfigUpdateUnsupported
Re-configuration of network config is not supported.

NetworkConfigUpdateError Text
Log network configuration update error.

NetworkConfigUpdateWarning Text
Log network configuration update warning.

NetworkConfigUpdateInfo Text
Log network configuration update info.

NetworkConfig [(HotValency, WarmValency, Map RelayAccessPoint CardanoLocalRootConfig)] (Map RelayAccessPoint PeerAdvertise) UseLedgerPeers (Maybe PeerSnapshotFile)
Log peer-to-peer network configuration, either on startup or when its updated.

NonP2PWarning
Warn when DisabledP2P is set.

WarningDevelopmentNodeToNodeVersions [NodeToNodeVersion]
Warn when ExperimentalProtocolsEnabled is set and affects node-to-node protocol.

WarningDevelopmentNodeToClientVersions [NodeToClientVersion]
Warn when ExperimentalProtocolsEnabled is set and affects node-to-client protocol.

BICommon BasicInfoCommon

BIShelley BasicInfoShelleyBased

BIByron BasicInfoByron

BINetwork BasicInfoNetwork

LedgerPeerSnapshotLoaded (WithOrigin SlotNo)

LedgerPeerSnapshotIgnored SlotNo SlotNo PeerSnapshotFile
Ledger peer snapshot ignored since the peer snapshot slot is older than UseLedgerPeers in the topology file. Arguments are: useLedgerPeersAfterSlot, peerSnapshotSlot, peerSnapshotFile.

LedgerPeerSnapshotError SlotNo SlotNo PeerSnapshotFile
Like above, but in GenesisMode it is an error to have an old snapshot.

MovedTopLevelOption String

data EnabledBlockForging

Constructors

EnabledBlockForging

DisabledBlockForging

NotEffective
one needs to send SIGHUP after consensus initialised itself (especially after replying all blocks).

data BasicInfoCommon

Constructors

BasicInfoCommon

Fields

biConfigPath :: FilePath

biNetworkMagic :: NetworkMagic

biProtocol :: Text

biVersion :: Text

biCommit :: Text

biNodeStartTime :: UTCTime

data BasicInfoShelleyBased

Constructors

BasicInfoShelleyBased

Fields

bisEra :: !Text

bisSystemStartTime :: !UTCTime

bisSlotLength :: !NominalDiffTime

bisEpochLength :: !Word64

bisSlotsPerKESPeriod :: !Word64

data BasicInfoByron

Constructors

BasicInfoByron

Fields

bibSystemStartTime :: UTCTime

bibSlotLength :: NominalDiffTime

bibEpochLength :: Word64

data BasicInfoNetwork

Constructors

BasicInfoNetwork

Fields

niAddresses :: [SocketOrSocketInfo]

niDiffusionMode :: DiffusionMode

data LedgerPeerSnapshotError

Constructors

LedgerPeerSnapshotTooOld SlotNo SlotNo PeerSnapshotFile

prepareNodeInfo :: NodeConfiguration -> SomeConsensusProtocol -> TraceConfig -> UTCTime -> IO NodeInfo

Prepare basic info about the node. This info will be sent to 'cardano-tracer'.

data NodeInfo

NodeInfo

Constructors

NodeInfo

Fields

niName :: Text

niProtocol :: Text

niVersion :: Text

niCommit :: Text

niStartTime :: UTCTime

niSystemStartTime :: UTCTime

data NodeStartupInfo

NodeStartupInfo

This information is taken from BasicInfoShelleyBased. It is required for 'cardano-tracer' service (particularly, for RTView).

Constructors

NodeStartupInfo

Fields

suiEra :: Text

suiSlotLength :: NominalDiffTime

suiEpochLength :: Word64

suiSlotsPerKESPeriod :: Word64