Cardano.Node.Tracing.Formatting

Orphan instances

LogFormatting Void
A bit of a weird one, but needed because some of the very general consensus interfaces are sometimes instantiated to Void, when there are no cases needed.

Instance details

Methods

forMachine ∷ DetailLevel → Void → Object

forHuman ∷ Void → Text

asMetrics ∷ Void → [Metric]

LogFormatting SlotNo

Instance details

Methods

forMachine ∷ DetailLevel → SlotNo → Object

forHuman ∷ SlotNo → Text

asMetrics ∷ SlotNo → [Metric]

LogFormatting ()

Instance details

Methods

forMachine ∷ DetailLevel → () → Object

forHuman ∷ () → Text

asMetrics ∷ () → [Metric]

ConvertRawHash blk ⇒ LogFormatting (RealPoint blk)

Instance details

Methods

forMachine ∷ DetailLevel → RealPoint blk → Object

forHuman ∷ RealPoint blk → Text

asMetrics ∷ RealPoint blk → [Metric]

ConvertRawHash blk ⇒ LogFormatting (Point blk)

Instance details

Methods

forMachine ∷ DetailLevel → Point blk → Object

forHuman ∷ Point blk → Text

asMetrics ∷ Point blk → [Metric]