Cardano.Node.Tracing.Tracers.ForgingThreadStats

Documentation

data ForgingStats

This structure stores counters of blockchain-related events, per individual thread in fsStats.

Constructors

ForgingStats

Fields

fsStats ∷ !(Map ThreadId ForgeThreadStats)

fsNodeCannotForgeNum ∷ !Int

fsNodeIsLeaderNum ∷ !Int

fsBlocksForgedNum ∷ !Int

fsSlotsMissedNum ∷ !Int

data ForgeThreadStats

Per-forging-thread statistics.

Constructors

ForgeThreadStats

Fields

ftsNodeCannotForgeNum ∷ !Int

ftsNodeIsLeaderNum ∷ !Int

ftsBlocksForgedNum ∷ !Int

ftsSlotsMissedNum ∷ !Int

Potentially missed slots. Note that this is not the same as the number of missed blocks, since this includes all occurrences of not reaching a leadership check decision, whether or not leadership was possible or not.

Also note that when the aggregate total for this metric is reported in the multi-pool case, it can be much larger than the actual number of slots occurring since node start, for it is a sum total for all threads.

ftsLastSlot ∷ !Int

forgeThreadStats ∷ Trace IO ForgingStats → IO (Trace IO (ForgeTracerType blk))