Cardano.Node.Types

Configuration

class AdjustFilePaths a where

Methods

adjustFilePaths ∷ (FilePath → FilePath) → a → a

newtype ConfigError

Errors for the cardano-config module.

Constructors

ConfigErrorFileNotFound FilePath

newtype ConfigYamlFilePath

Filepath of the configuration yaml file. This file determines all the configuration settings required for the cardano node (logging, tracing, protocol, slot length etc)

Constructors

ConfigYamlFilePath

Fields

unConfigPath ∷ FilePath

newtype DbFile

Constructors

DbFile

Fields

unDB ∷ FilePath

newtype GenesisFile

Constructors

GenesisFile

Fields

unGenesisFile ∷ FilePath

data ProtocolFilepaths

Constructors

ProtocolFilepaths

Fields

byronCertFile ∷ !(Maybe FilePath)

byronKeyFile ∷ !(Maybe FilePath)

shelleyKESFile ∷ !(Maybe FilePath)

shelleyVRFFile ∷ !(Maybe FilePath)

shelleyCertFile ∷ !(Maybe FilePath)

shelleyBulkCredsFile ∷ !(Maybe FilePath)

newtype GenesisHash

Constructors

GenesisHash (Hash Blake2b_256 ByteString)

newtype MaxConcurrencyBulkSync

Constructors

MaxConcurrencyBulkSync

Fields

unMaxConcurrencyBulkSync ∷ Word

newtype MaxConcurrencyDeadline

Constructors

MaxConcurrencyDeadline

Fields

unMaxConcurrencyDeadline ∷ Word

Node addresses

data NodeAddress' addr

IPv4 or IPv6 address with a port number.

Constructors

NodeAddress

Fields

naHostAddress ∷ !addr

naPort ∷ !PortNumber

type NodeIPAddress = NodeAddress' NodeHostIPAddress

nodeAddressToSockAddr ∷ NodeIPAddress → SockAddr

type NodeIPv4Address = NodeAddress' NodeHostIPv4Address

type NodeIPv6Address = NodeAddress' NodeHostIPv6Address

type NodeDnsAddress = NodeAddress' NodeHostDnsAddress

nodeIPv4ToIPAddress ∷ NodeIPv4Address → NodeIPAddress

nodeIPv6ToIPAddress ∷ NodeIPv6Address → NodeIPAddress

nodeDnsAddressToDomainAddress ∷ NodeDnsAddress → DomainAddress

newtype NodeHostIPAddress

Constructors

NodeHostIPAddress

Fields

unNodeHostIPAddress ∷ IP

nodeHostIPAddressToSockAddr ∷ NodeIPAddress → SockAddr

newtype NodeHostIPv4Address

Constructors

NodeHostIPv4Address

Fields

unNodeHostIPv4Address ∷ IPv4

newtype NodeHostIPv6Address

Constructors

NodeHostIPv6Address

Fields

unNodeHostIPv6Address ∷ IPv6

nodeHostIPv4AddressToIPAddress ∷ NodeHostIPv4Address → NodeHostIPAddress

nodeHostIPv6AddressToIPAddress ∷ NodeHostIPv6Address → NodeHostIPAddress

newtype NodeHostDnsAddress

Domain name.

Constructors

NodeHostDnsAddress

Fields

unNodeHostDnsAddress ∷ Text

nodeHostDnsAddressToDomain ∷ NodeHostDnsAddress → Domain

data PortNumber

newtype SocketPath

Constructors

SocketPath

Fields

unSocketPath ∷ FilePath

newtype TopologyFile

Constructors

TopologyFile

Fields

unTopology ∷ FilePath

newtype NodeDiffusionMode

Newtype wrapper which provides FromJSON instance for DiffusionMode.

Constructors

NodeDiffusionMode

Fields

getDiffusionMode ∷ DiffusionMode

Consensus protocol configuration

data NodeByronProtocolConfiguration

Constructors

NodeByronProtocolConfiguration

Fields

npcByronGenesisFile ∷ !GenesisFile

npcByronGenesisFileHash ∷ !(Maybe GenesisHash)

npcByronReqNetworkMagic ∷ !RequiresNetworkMagic

npcByronPbftSignatureThresh ∷ !(Maybe Double)

npcByronApplicationName ∷ !ApplicationName

Update application name.

npcByronApplicationVersion ∷ !NumSoftwareVersion

Application (ie software) version.

npcByronSupportedProtocolVersionMajor ∷ !Word16

These declare the version of the protocol that the node is prepared to run. This is usually the version of the protocol in use on the chain now, but during protocol updates this version will be the one that we declare that we are ready to move to. This is the endorsement mechanism for determining when enough block producers are ready to move to the next version.

npcByronSupportedProtocolVersionMinor ∷ !Word16

npcByronSupportedProtocolVersionAlt ∷ !Word8

data NodeHardForkProtocolConfiguration

Configuration relating to a hard forks themselves, not the specific eras.

Constructors

NodeHardForkProtocolConfiguration

Fields

npcTestShelleyHardForkAtEpoch ∷ Maybe EpochNo

For testing purposes we support specifying that the hard fork happens at an exact epoch number (ie the first epoch of the new era).

Obviously if this is used, all the nodes in the test cluster must be configured the same, or they will disagree.

npcTestShelleyHardForkAtVersion ∷ Maybe Word

For testing purposes we support specifying that the hard fork happens at a given major protocol version. For example this can be used to cause the Shelley hard fork to occur at the transition from protocol version 0 to version 1 (rather than the default of from 1 to 2) which can make the test setup simpler.

Obviously if this is used, all the nodes in the test cluster must be configured the same, or they will disagree.

npcTestAllegraHardForkAtEpoch ∷ Maybe EpochNo

For testing purposes we support specifying that the hard fork happens at an exact epoch number (ie the first epoch of the new era).

Obviously if this is used, all the nodes in the test cluster must be configured the same, or they will disagree.

npcTestAllegraHardForkAtVersion ∷ Maybe Word

For testing purposes we support specifying that the hard fork happens at a given major protocol version.

Obviously if this is used, all the nodes in the test cluster must be configured the same, or they will disagree.

npcTestMaryHardForkAtEpoch ∷ Maybe EpochNo

For testing purposes we support specifying that the hard fork happens at an exact epoch number (ie the first epoch of the new era).

Obviously if this is used, all the nodes in the test cluster must be configured the same, or they will disagree.

npcTestMaryHardForkAtVersion ∷ Maybe Word

For testing purposes we support specifying that the hard fork happens at an exact epoch number (ie the first epoch of the new era).

Obviously if this is used, all the nodes in the test cluster must be configured the same, or they will disagree.

data NodeProtocolConfiguration

Constructors

NodeProtocolConfigurationByron NodeByronProtocolConfiguration

NodeProtocolConfigurationShelley NodeShelleyProtocolConfiguration

NodeProtocolConfigurationCardano NodeByronProtocolConfiguration NodeShelleyProtocolConfiguration NodeHardForkProtocolConfiguration

data NodeShelleyProtocolConfiguration

Constructors

NodeShelleyProtocolConfiguration

Fields

npcShelleyGenesisFile ∷ !GenesisFile

npcShelleyGenesisFileHash ∷ !(Maybe GenesisHash)

data VRFPrivateKeyFilePermissionError

Constructors

OtherPermissionsExist FilePath

GroupPermissionsExist FilePath

GenericPermissionsExist FilePath

protocolName ∷ Protocol → String

A human readable name for the protocol

renderVRFPrivateKeyFilePermissionError ∷ VRFPrivateKeyFilePermissionError → Text