Cardano.Tracing.HasIssuer

Documentation

data BlockIssuerVerificationKeyHash

Block issuer verification key hash.

Constructors

BlockIssuerVerificationKeyHash !ByteString
Serialized block issuer verification key hash.

NoBlockIssuer
There is no block issuer.

For example, this could be relevant for epoch boundary blocks (EBBs), genesis blocks, etc.

class HasIssuer blk where

Get the block issuer verification key hash from a block header.

Methods

getIssuerVerificationKeyHash :: Header blk -> BlockIssuerVerificationKeyHash

Given a block header, return the serialized block issuer verification key hash.