Cardano.Tracing.MicroBenchmarking

Documentation

data MeasureTxs blk

Definition of the measurement datatype for the transactions.

Constructors

MeasureTxsTimeStart (GenTx blk) !Word !Word !Time

MeasureTxsTimeStop !SlotNo blk [GenTx blk] !Time

measureTxsStart ∷ ∀ blk. Trace IO Text → Tracer IO (TraceEventMempool blk)

Transformer for the start of the transaction, when the transaction was added to the mempool.

measureTxsEnd ∷ ∀ blk. Trace IO Text → Tracer IO (TraceForgeEvent blk)

Transformer for the end of the transaction, when the transaction was added to the block and the block was forged.

Re-exports so we localize the changes

class Monad m ⇒ Outcome (m ∷ Type → Type) a where

Associated Types

type IntermediateValue a

type OutcomeMetric a

Methods

classifyObservable ∷ a → m OutcomeProgressionStatus

captureObservableValue ∷ a → m (IntermediateValue a)

computeOutcomeMetric ∷ a → IntermediateValue a → IntermediateValue a → m (OutcomeMetric a)

type OutcomeEnhancedTracer (m ∷ Type → Type) a = Tracer m (Either a (OutcomeFidelity (OutcomeMetric a)))

mkOutcomeExtractor ∷ (MonadIO m, MonadMask m, Outcome m a) ⇒ m (OutcomeEnhancedTracer m a → Tracer m a)

Orphan instances

Transformable Text IO (Either (TraceForgeEvent blk) (OutcomeFidelity (Maybe (SlotNo, DiffTime, MempoolSize))))

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (Either (TraceForgeEvent blk) (OutcomeFidelity (Maybe (SlotNo, DiffTime, MempoolSize))))

(Monad m, MonadTime m) ⇒ Outcome m (TraceForgeEvent blk)

Instance details

Associated Types

type IntermediateValue (TraceForgeEvent blk)

type OutcomeMetric (TraceForgeEvent blk)

Methods

classifyObservable ∷ TraceForgeEvent blk → m OutcomeProgressionStatus

captureObservableValue ∷ TraceForgeEvent blk → m (IntermediateValue (TraceForgeEvent blk))

computeOutcomeMetric ∷ TraceForgeEvent blk → IntermediateValue (TraceForgeEvent blk) → IntermediateValue (TraceForgeEvent blk) → m (OutcomeMetric (TraceForgeEvent blk))

HasPrivacyAnnotation (Either (TraceForgeEvent blk) (OutcomeFidelity (Maybe (SlotNo, DiffTime, MempoolSize))))

Instance details

Methods

getPrivacyAnnotation ∷ Either (TraceForgeEvent blk) (OutcomeFidelity (Maybe (SlotNo, DiffTime, MempoolSize))) → PrivacyAnnotation

HasSeverityAnnotation (Either (TraceForgeEvent blk) (OutcomeFidelity (Maybe (SlotNo, DiffTime, MempoolSize))))

Instance details

Methods

getSeverityAnnotation ∷ Either (TraceForgeEvent blk) (OutcomeFidelity (Maybe (SlotNo, DiffTime, MempoolSize))) → Severity

ToObject (Either (TraceForgeEvent blk) (OutcomeFidelity (Maybe (SlotNo, DiffTime, MempoolSize))))

Instance details

Methods

toObject ∷ TracingVerbosity → Either (TraceForgeEvent blk) (OutcomeFidelity (Maybe (SlotNo, DiffTime, MempoolSize))) → Object

textTransformer ∷ Either (TraceForgeEvent blk) (OutcomeFidelity (Maybe (SlotNo, DiffTime, MempoolSize))) → Object → Text