Cardano.Tracing.OrphanInstances.Common

ToObject and helpers

class ToObject a where

Minimal complete definition

Nothing

Methods

toObject :: TracingVerbosity -> a -> Object

textTransformer :: a -> Object -> Text

data TracingVerbosity

Constructors

MinimalVerbosity

NormalVerbosity

MaximalVerbosity

mkObject :: ToObject a => [(Text, a)] -> HashMap Text a

emptyObject :: ToObject a => HashMap Text a

class ToJSON a

toJSON :: ToJSON a => a -> Value

(.=) :: (KeyValue e kv, ToJSON v) => Key -> v -> kv

Transformable and helpers

class (Monad m, HasPrivacyAnnotation b, HasSeverityAnnotation b) => Transformable a (m :: Type -> Type) b where

Minimal complete definition

Nothing

Methods

trTransformer :: TracingVerbosity -> Trace m a -> Tracer m b

trStructured :: forall b (m :: Type -> Type) a. (ToObject b, MonadIO m, HasPrivacyAnnotation b, HasSeverityAnnotation b) => TracingVerbosity -> Trace m a -> Tracer m b

trStructuredText :: forall b (m :: Type -> Type) a. (ToObject b, MonadIO m, HasTextFormatter b, HasPrivacyAnnotation b, HasSeverityAnnotation b) => TracingVerbosity -> Trace m a -> Tracer m b

class HasTextFormatter a where

Minimal complete definition

Nothing

Methods

formatText :: a -> Object -> Text

Severity and Privacy

class HasSeverityAnnotation a where

Minimal complete definition

Nothing

Methods

getSeverityAnnotation :: a -> Severity

data Severity

Constructors

Debug

Info

Notice

Warning

Error

Critical

Alert

Emergency

class HasPrivacyAnnotation a where

Minimal complete definition

Nothing

Methods

getPrivacyAnnotation :: a -> PrivacyAnnotation

data PrivacyAnnotation

Constructors

Confidential

Public

Tracer and related

data Tracer (m :: Type -> Type) a

example: simply output a message on the console

let logTrace = traceWith $ showTracing $ stdoutTracer
in  logTrace "hello world"

example: calling a function and passing in a Tracer

example1 :: IO ()
example1 = do
    let logTrace a = traceWith (showTracing (contramap ("Debug: " ++) stdoutTracer)) a
    void $ callFun1 logTrace

callFun1 :: (String -> IO ()) -> IO Int
callFun1 logTrace = do
    logTrace "in function 1"
    return 42

runTracer evaluates a Tracer (i.e. consumes its argument)

data LogObject a

Constructors

LogObject

Fields

loName :: LoggerName

loMeta :: !LOMeta

loContent :: !(LOContent a)

data LOContent a

Constructors

LogMessage a

LogError !Text

LogValue !Text !Measurable

LogStructuredText Object Text

LogStructured Object

ObserveOpen !CounterState

ObserveDiff !CounterState

ObserveClose !CounterState

AggregatedMessage [(Text, Aggregated)]

MonitoringEffect !MonitorAction

Command !CommandValue

KillPill

mkLOMeta :: MonadIO m => Severity -> PrivacyAnnotation -> m LOMeta

Orphan instances

FromJSON PortNumber

Instance details

Methods

parseJSON :: Value -> Parser PortNumber

parseJSONList :: Value -> Parser [PortNumber]

omittedField :: Maybe PortNumber

HasPrivacyAnnotation ResourceStats

Instance details

Methods

getPrivacyAnnotation :: ResourceStats -> PrivacyAnnotation

HasSeverityAnnotation ResourceStats

Instance details

Methods

getSeverityAnnotation :: ResourceStats -> Severity

ToObject Void
A bit of a weird one, but needed because some of the very general consensus interfaces are sometimes instantiated to Void, when there are no cases needed.

Instance details

Methods

toObject :: TracingVerbosity -> Void -> Object

textTransformer :: Void -> Object -> Text

ToObject ResourceStats

Instance details

Methods

toObject :: TracingVerbosity -> ResourceStats -> Object

textTransformer :: ResourceStats -> Object -> Text

Transformable Text IO ResourceStats

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO ResourceStats