Cardano.Tracing.OrphanInstances.Network

Orphan instances

HasPrivacyAnnotation AcceptConnectionsPolicyTrace

Instance details

Methods

getPrivacyAnnotation ∷ AcceptConnectionsPolicyTrace → PrivacyAnnotation

HasPrivacyAnnotation HandshakeTr

Instance details

Methods

getPrivacyAnnotation ∷ HandshakeTr → PrivacyAnnotation

HasPrivacyAnnotation HandshakeTr

Instance details

Methods

getPrivacyAnnotation ∷ HandshakeTr → PrivacyAnnotation

HasPrivacyAnnotation DiffusionInitializationTracer

Instance details

Methods

getPrivacyAnnotation ∷ DiffusionInitializationTracer → PrivacyAnnotation

HasSeverityAnnotation AcceptConnectionsPolicyTrace

Instance details

Methods

getSeverityAnnotation ∷ AcceptConnectionsPolicyTrace → Severity

HasSeverityAnnotation HandshakeTr

Instance details

Methods

getSeverityAnnotation ∷ HandshakeTr → Severity

HasSeverityAnnotation HandshakeTr

Instance details

Methods

getSeverityAnnotation ∷ HandshakeTr → Severity

HasSeverityAnnotation DiffusionInitializationTracer

Instance details

Methods

getSeverityAnnotation ∷ DiffusionInitializationTracer → Severity

HasTextFormatter AcceptConnectionsPolicyTrace

Instance details

Methods

formatText ∷ AcceptConnectionsPolicyTrace → Object → Text

HasTextFormatter HandshakeTr

Instance details

Methods

formatText ∷ HandshakeTr → Object → Text

HasTextFormatter HandshakeTr

Instance details

Methods

formatText ∷ HandshakeTr → Object → Text

HasTextFormatter DiffusionInitializationTracer

Instance details

Methods

formatText ∷ DiffusionInitializationTracer → Object → Text

ToObject SlotNo

Instance details

Methods

toObject ∷ TracingVerbosity → SlotNo → Object

textTransformer ∷ SlotNo → Object → Text

ToObject AcceptConnectionsPolicyTrace

Instance details

Methods

toObject ∷ TracingVerbosity → AcceptConnectionsPolicyTrace → Object

textTransformer ∷ AcceptConnectionsPolicyTrace → Object → Text

ToObject HandshakeTr

Instance details

Methods

toObject ∷ TracingVerbosity → HandshakeTr → Object

textTransformer ∷ HandshakeTr → Object → Text

ToObject HandshakeTr

Instance details

Methods

toObject ∷ TracingVerbosity → HandshakeTr → Object

textTransformer ∷ HandshakeTr → Object → Text

ToObject DiffusionInitializationTracer

Instance details

Methods

toObject ∷ TracingVerbosity → DiffusionInitializationTracer → Object

textTransformer ∷ DiffusionInitializationTracer → Object → Text

Transformable Text IO AcceptConnectionsPolicyTrace

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO AcceptConnectionsPolicyTrace

Transformable Text IO HandshakeTr

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO HandshakeTr

Transformable Text IO HandshakeTr

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO HandshakeTr

Transformable Text IO DiffusionInitializationTracer
instances of Transformable

NOTE: this list is sorted by the unqualified name of the outermost type.

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO DiffusionInitializationTracer

(StandardHash header, Show peer) ⇒ Transformable Text IO [TraceLabelPeer peer (FetchDecision [Point header])]

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO [TraceLabelPeer peer (FetchDecision [Point header])]

Transformable Text IO (Identity (SubscriptionTrace LocalAddress))

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (Identity (SubscriptionTrace LocalAddress))

Transformable Text IO (WithDomainName (SubscriptionTrace SockAddr))

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (WithDomainName (SubscriptionTrace SockAddr))

Transformable Text IO (WithDomainName DnsTrace)

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (WithDomainName DnsTrace)

Transformable Text IO (WithIPList (SubscriptionTrace SockAddr))

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (WithIPList (SubscriptionTrace SockAddr))

Show remotePeer ⇒ Transformable Text IO (TraceKeepAliveClient remotePeer)

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (TraceKeepAliveClient remotePeer)

Show peer ⇒ Transformable Text IO (WithMuxBearer peer MuxTrace)

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (WithMuxBearer peer MuxTrace)

Show addr ⇒ Transformable Text IO (WithAddr addr ErrorPolicyTrace)

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (WithAddr addr ErrorPolicyTrace)

Transformable Text IO (TraceTxSubmissionInbound txid tx)

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (TraceTxSubmissionInbound txid tx)

(Show peer, Show a, HasPrivacyAnnotation a, HasSeverityAnnotation a, ToObject a) ⇒ Transformable Text IO (TraceLabelPeer peer a)

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (TraceLabelPeer peer a)

(Show tx, Show txid) ⇒ Transformable Text IO (TraceTxSubmissionOutbound txid tx)

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (TraceTxSubmissionOutbound txid tx)

HasPrivacyAnnotation [TraceLabelPeer peer (FetchDecision [Point header])]

Instance details

Methods

getPrivacyAnnotation ∷ [TraceLabelPeer peer (FetchDecision [Point header])] → PrivacyAnnotation

HasPrivacyAnnotation (Identity (SubscriptionTrace LocalAddress))

Instance details

Methods

getPrivacyAnnotation ∷ Identity (SubscriptionTrace LocalAddress) → PrivacyAnnotation

HasPrivacyAnnotation (TraceSendRecv a)

Instance details

Methods

getPrivacyAnnotation ∷ TraceSendRecv a → PrivacyAnnotation

HasPrivacyAnnotation (WithDomainName (SubscriptionTrace SockAddr))

Instance details

Methods

getPrivacyAnnotation ∷ WithDomainName (SubscriptionTrace SockAddr) → PrivacyAnnotation

HasPrivacyAnnotation (WithDomainName DnsTrace)

Instance details

Methods

getPrivacyAnnotation ∷ WithDomainName DnsTrace → PrivacyAnnotation

HasPrivacyAnnotation (WithIPList (SubscriptionTrace SockAddr))

Instance details

Methods

getPrivacyAnnotation ∷ WithIPList (SubscriptionTrace SockAddr) → PrivacyAnnotation

HasPrivacyAnnotation (TraceFetchClientState header)

Instance details

Methods

getPrivacyAnnotation ∷ TraceFetchClientState header → PrivacyAnnotation

HasPrivacyAnnotation (TraceKeepAliveClient remotePeer)

Instance details

Methods

getPrivacyAnnotation ∷ TraceKeepAliveClient remotePeer → PrivacyAnnotation

HasSeverityAnnotation [TraceLabelPeer peer (FetchDecision [Point header])]

Instance details

Methods

getSeverityAnnotation ∷ [TraceLabelPeer peer (FetchDecision [Point header])] → Severity

HasSeverityAnnotation (Identity (SubscriptionTrace LocalAddress))

Instance details

Methods

getSeverityAnnotation ∷ Identity (SubscriptionTrace LocalAddress) → Severity

HasSeverityAnnotation (TraceSendRecv a)

Instance details

Methods

getSeverityAnnotation ∷ TraceSendRecv a → Severity

HasSeverityAnnotation (WithDomainName (SubscriptionTrace SockAddr))

Instance details

Methods

getSeverityAnnotation ∷ WithDomainName (SubscriptionTrace SockAddr) → Severity

HasSeverityAnnotation (WithDomainName DnsTrace)

Instance details

Methods

getSeverityAnnotation ∷ WithDomainName DnsTrace → Severity

HasSeverityAnnotation (WithIPList (SubscriptionTrace SockAddr))

Instance details

Methods

getSeverityAnnotation ∷ WithIPList (SubscriptionTrace SockAddr) → Severity

HasSeverityAnnotation (TraceFetchClientState header)

Instance details

Methods

getSeverityAnnotation ∷ TraceFetchClientState header → Severity

HasSeverityAnnotation (TraceKeepAliveClient remotePeer)

Instance details

Methods

getSeverityAnnotation ∷ TraceKeepAliveClient remotePeer → Severity

(StandardHash header, Show peer) ⇒ HasTextFormatter [TraceLabelPeer peer (FetchDecision [Point header])]

Instance details

Methods

formatText ∷ [TraceLabelPeer peer (FetchDecision [Point header])] → Object → Text

HasTextFormatter (Identity (SubscriptionTrace LocalAddress))

Instance details

Methods

formatText ∷ Identity (SubscriptionTrace LocalAddress) → Object → Text

HasTextFormatter (WithDomainName (SubscriptionTrace SockAddr))

Instance details

Methods

formatText ∷ WithDomainName (SubscriptionTrace SockAddr) → Object → Text

HasTextFormatter (WithDomainName DnsTrace)

Instance details

Methods

formatText ∷ WithDomainName DnsTrace → Object → Text

HasTextFormatter (WithIPList (SubscriptionTrace SockAddr))

Instance details

Methods

formatText ∷ WithIPList (SubscriptionTrace SockAddr) → Object → Text

Show addr ⇒ HasTextFormatter (TraceKeepAliveClient addr)

Instance details

Methods

formatText ∷ TraceKeepAliveClient addr → Object → Text

Show peer ⇒ ToObject [TraceLabelPeer peer (FetchDecision [Point header])]

Instance details

Methods

toObject ∷ TracingVerbosity → [TraceLabelPeer peer (FetchDecision [Point header])] → Object

textTransformer ∷ [TraceLabelPeer peer (FetchDecision [Point header])] → Object → Text

ToObject (Identity (SubscriptionTrace LocalAddress))

Instance details

Methods

toObject ∷ TracingVerbosity → Identity (SubscriptionTrace LocalAddress) → Object

textTransformer ∷ Identity (SubscriptionTrace LocalAddress) → Object → Text

ConvertRawHash blk ⇒ ToObject (Point blk)

Instance details

Methods

toObject ∷ TracingVerbosity → Point blk → Object

textTransformer ∷ Point blk → Object → Text

ToObject (AnyMessageAndAgency ps) ⇒ ToObject (TraceSendRecv ps)

Instance details

Methods

toObject ∷ TracingVerbosity → TraceSendRecv ps → Object

textTransformer ∷ TraceSendRecv ps → Object → Text

ToObject (WithDomainName (SubscriptionTrace SockAddr))

Instance details

Methods

toObject ∷ TracingVerbosity → WithDomainName (SubscriptionTrace SockAddr) → Object

textTransformer ∷ WithDomainName (SubscriptionTrace SockAddr) → Object → Text

ToObject (WithDomainName DnsTrace)

Instance details

Methods

toObject ∷ TracingVerbosity → WithDomainName DnsTrace → Object

textTransformer ∷ WithDomainName DnsTrace → Object → Text

ToObject (WithIPList (SubscriptionTrace SockAddr))

Instance details

Methods

toObject ∷ TracingVerbosity → WithIPList (SubscriptionTrace SockAddr) → Object

textTransformer ∷ WithIPList (SubscriptionTrace SockAddr) → Object → Text

ToObject (TraceFetchClientState header)

Instance details

Methods

toObject ∷ TracingVerbosity → TraceFetchClientState header → Object

textTransformer ∷ TraceFetchClientState header → Object → Text

ToObject (FetchDecision [Point header])

Instance details

Methods

toObject ∷ TracingVerbosity → FetchDecision [Point header] → Object

textTransformer ∷ FetchDecision [Point header] → Object → Text

Show remotePeer ⇒ ToObject (TraceKeepAliveClient remotePeer)

Instance details

Methods

toObject ∷ TracingVerbosity → TraceKeepAliveClient remotePeer → Object

textTransformer ∷ TraceKeepAliveClient remotePeer → Object → Text

(ConvertTxId blk, RunNode blk, HasTxs blk) ⇒ ToObject (AnyMessageAndAgency (BlockFetch blk (Point blk)))
instances of ToObject

NOTE: this list is sorted by the unqualified name of the outermost type.

Instance details

Methods

toObject ∷ TracingVerbosity → AnyMessageAndAgency (BlockFetch blk (Point blk)) → Object

textTransformer ∷ AnyMessageAndAgency (BlockFetch blk (Point blk)) → Object → Text

ToObject (AnyMessageAndAgency (ChainSync blk pt tip))

Instance details

Methods

toObject ∷ TracingVerbosity → AnyMessageAndAgency (ChainSync blk pt tip) → Object

textTransformer ∷ AnyMessageAndAgency (ChainSync blk pt tip) → Object → Text

(∀ result. Show (query result)) ⇒ ToObject (AnyMessageAndAgency (LocalStateQuery blk pt query))

Instance details

Methods

toObject ∷ TracingVerbosity → AnyMessageAndAgency (LocalStateQuery blk pt query) → Object

textTransformer ∷ AnyMessageAndAgency (LocalStateQuery blk pt query) → Object → Text

ToObject (AnyMessageAndAgency (LocalTxSubmission tx err))

Instance details

Methods

toObject ∷ TracingVerbosity → AnyMessageAndAgency (LocalTxSubmission tx err) → Object

textTransformer ∷ AnyMessageAndAgency (LocalTxSubmission tx err) → Object → Text

(Show txid, Show tx) ⇒ ToObject (AnyMessageAndAgency (TxSubmission txid tx))

Instance details

Methods

toObject ∷ TracingVerbosity → AnyMessageAndAgency (TxSubmission txid tx) → Object

textTransformer ∷ AnyMessageAndAgency (TxSubmission txid tx) → Object → Text

HasPrivacyAnnotation (WithMuxBearer peer MuxTrace)

Instance details

Methods

getPrivacyAnnotation ∷ WithMuxBearer peer MuxTrace → PrivacyAnnotation

HasPrivacyAnnotation (WithAddr addr ErrorPolicyTrace)

Instance details

Methods

getPrivacyAnnotation ∷ WithAddr addr ErrorPolicyTrace → PrivacyAnnotation

HasPrivacyAnnotation (TraceTxSubmissionInbound txid tx)

Instance details

Methods

getPrivacyAnnotation ∷ TraceTxSubmissionInbound txid tx → PrivacyAnnotation

HasPrivacyAnnotation a ⇒ HasPrivacyAnnotation (TraceLabelPeer peer a)

Instance details

Methods

getPrivacyAnnotation ∷ TraceLabelPeer peer a → PrivacyAnnotation

HasPrivacyAnnotation (TraceTxSubmissionOutbound txid tx)

Instance details

Methods

getPrivacyAnnotation ∷ TraceTxSubmissionOutbound txid tx → PrivacyAnnotation

HasSeverityAnnotation (WithMuxBearer peer MuxTrace)

Instance details

Methods

getSeverityAnnotation ∷ WithMuxBearer peer MuxTrace → Severity

HasSeverityAnnotation (WithAddr addr ErrorPolicyTrace)

Instance details

Methods

getSeverityAnnotation ∷ WithAddr addr ErrorPolicyTrace → Severity

HasSeverityAnnotation (TraceTxSubmissionInbound txid tx)

Instance details

Methods

getSeverityAnnotation ∷ TraceTxSubmissionInbound txid tx → Severity

HasSeverityAnnotation a ⇒ HasSeverityAnnotation (TraceLabelPeer peer a)

Instance details

Methods

getSeverityAnnotation ∷ TraceLabelPeer peer a → Severity

HasSeverityAnnotation (TraceTxSubmissionOutbound txid tx)

Instance details

Methods

getSeverityAnnotation ∷ TraceTxSubmissionOutbound txid tx → Severity

Show peer ⇒ HasTextFormatter (WithMuxBearer peer MuxTrace)

Instance details

Methods

formatText ∷ WithMuxBearer peer MuxTrace → Object → Text

Show addr ⇒ HasTextFormatter (WithAddr addr ErrorPolicyTrace)

Instance details

Methods

formatText ∷ WithAddr addr ErrorPolicyTrace → Object → Text

HasTextFormatter (TraceTxSubmissionInbound txid tx)

Instance details

Methods

formatText ∷ TraceTxSubmissionInbound txid tx → Object → Text

(Show peer, Show a) ⇒ HasTextFormatter (TraceLabelPeer peer a)

Instance details

Methods

formatText ∷ TraceLabelPeer peer a → Object → Text

(Show tx, Show txid) ⇒ HasTextFormatter (TraceTxSubmissionOutbound txid tx)

Instance details

Methods

formatText ∷ TraceTxSubmissionOutbound txid tx → Object → Text

Show peer ⇒ ToObject (WithMuxBearer peer MuxTrace)

Instance details

Methods

toObject ∷ TracingVerbosity → WithMuxBearer peer MuxTrace → Object

textTransformer ∷ WithMuxBearer peer MuxTrace → Object → Text

Show addr ⇒ ToObject (WithAddr addr ErrorPolicyTrace)

Instance details

Methods

toObject ∷ TracingVerbosity → WithAddr addr ErrorPolicyTrace → Object

textTransformer ∷ WithAddr addr ErrorPolicyTrace → Object → Text

ToObject (TraceTxSubmissionInbound txid tx)

Instance details

Methods

toObject ∷ TracingVerbosity → TraceTxSubmissionInbound txid tx → Object

textTransformer ∷ TraceTxSubmissionInbound txid tx → Object → Text

(Show peer, ToObject a) ⇒ ToObject (TraceLabelPeer peer a)

Instance details

Methods

toObject ∷ TracingVerbosity → TraceLabelPeer peer a → Object

textTransformer ∷ TraceLabelPeer peer a → Object → Text

(Show txid, Show tx) ⇒ ToObject (TraceTxSubmissionOutbound txid tx)

Instance details

Methods

toObject ∷ TracingVerbosity → TraceTxSubmissionOutbound txid tx → Object

textTransformer ∷ TraceTxSubmissionOutbound txid tx → Object → Text