Codec.CBOR.ByteArray

Description

A ByteArray with more instances than ByteArray. Some day when these instances are reliably available from primitive we can likely replace this with ByteArray.

Simple byte arrays

newtype ByteArray

Constructors

BA

Fields

unBA :: ByteArray

sizeofByteArray :: ByteArray -> Int

Conversions

fromShortByteString :: ShortByteString -> ByteArray

toShortByteString :: ByteArray -> ShortByteString

fromByteString :: ByteString -> ByteArray

toBuilder :: ByteArray -> Builder

toSliced :: ByteArray -> SlicedByteArray