Codec.CBOR.ByteArray.Sliced

Description

A ByteArray with more instances than ByteArray. Some day when these instances are reliably available from primitive we can likely replace this with ByteArray.

Documentation

data SlicedByteArray

Constructors

SBA

Fields

unSBA :: !ByteArray

offset :: !Int

length :: !Int

Conversions

sizeofSlicedByteArray :: SlicedByteArray -> Int

fromShortByteString :: ShortByteString -> SlicedByteArray

fromByteString :: ByteString -> SlicedByteArray

fromByteArray :: ByteArray -> SlicedByteArray

toByteString :: SlicedByteArray -> ByteString

Note that this may require a copy.

toBuilder :: SlicedByteArray -> Builder