Codec.Serialise

Description

This module provides functions to serialise and deserialise Haskell values for storage or transmission, to and from lazy ByteStrings. It also provides a type class and utilities to help you make your types serialisable.

For a full tutorial on using this module, see Codec.Serialise.Tutorial.

High level, one-shot API

The following API exposes a high level interface allowing you to quickly convert between arbitrary Haskell values (which are an instance of Serialise) and lazy ByteStrings.

serialise :: Serialise a => a -> ByteString

Serialise a Haskell value to an external binary representation.

The output is represented as a lazy ByteString and is constructed incrementally.

Since: 0.2.0.0

deserialise :: Serialise a => ByteString -> a

Deserialise a Haskell value from the external binary representation (which must have been made using serialise or related function).

Throws: DeserialiseFailure if the given external representation is invalid or does not correspond to a value of the expected type.

Since: 0.2.0.0

deserialiseOrFail :: Serialise a => ByteString -> Either DeserialiseFailure a

Deserialise a Haskell value from the external binary representation, or get back a DeserialiseFailure.

Since: 0.2.0.0

Deserialisation exceptions

data DeserialiseFailure

Constructors

DeserialiseFailure ByteOffset String

Incremental encoding interface

The following API allows you to encode or decode CBOR values incrementally, which is useful for large structures that require you to stream values in over time.

serialiseIncremental :: Serialise a => a -> Builder

Serialise a Haskell value to an external binary representation.

The output is represented as a Builder and is constructed incrementally. The representation as a Builder allows efficient concatenation with other data.

Since: 0.2.0.0

deserialiseIncremental :: Serialise a => ST s (IDecode s a)

Deserialise a Haskell value from the external binary representation.

This allows input data to be provided incrementally, rather than all in one go. It also gives an explicit representation of deserialisation errors.

Note that the incremental behaviour is only for the input data, not the output value: the final deserialised value is constructed and returned as a whole, not incrementally.

Since: 0.2.0.0

data IDecode s a

Constructors

Partial (Maybe ByteString -> ST s (IDecode s a))

Done !ByteString !ByteOffset a

Fail !ByteString !ByteOffset DeserialiseFailure

The Serialise class

class Serialise a where

Types that are instances of the Serialise class allow values to be quickly encoded or decoded directly to a CBOR representation, for object transmission or storage.

Since: 0.2.0.0

Minimal complete definition

Nothing

Methods

encode :: a -> Encoding

Definition for encoding a given type into a binary representation, using the Encoding Monoid.

Since: 0.2.0.0

default encode :: (Generic a, GSerialiseEncode (Rep a)) => a -> Encoding

decode :: Decoder s a

Definition of a given Decoder for a type.

Since: 0.2.0.0

default decode :: (Generic a, GSerialiseDecode (Rep a)) => Decoder s a

encodeList :: [a] -> Encoding

Utility to support specialised encoding for some list type - used for Char/String instances in this package.

Since: 0.2.0.0

decodeList :: Decoder s [a]

Utility to support specialised decoding for some list type - used for Char/String instances in this package.

Since: 0.2.0.0

IO operations

Convenient utilities for basic IO operations.

FilePath API

writeFileSerialise

Arguments

:: Serialise a

=> FilePath
The file to write to.

-> a
The value to be serialised and written.

-> IO ()

Serialise a ByteString and write it directly to the specified file.

Since: 0.2.0.0

readFileDeserialise

Arguments

:: Serialise a

=> FilePath
The file to read from.

-> IO a
The deserialised value.

Read the specified file (internally, by reading a ByteString) and attempt to decode it into a Haskell value using deserialise (the type of which is determined by the choice of the result type).

Throws: DeserialiseFailure if the file fails to deserialise properly.

Since: 0.2.0.0

Handle API

hPutSerialise

Arguments

:: Serialise a

=> Handle
The Handle to write to.

-> a
The value to be serialised and written.

-> IO ()

Serialise a ByteString (via serialise) and write it directly to the specified Handle.

Since: 0.2.0.0