Codec.Serialise.Class

Description

The Serialise class allows you to encode a given type into a CBOR object, or decode a CBOR object into the user-specified type.

The Serialise class

class Serialise a where

Types that are instances of the Serialise class allow values to be quickly encoded or decoded directly to a CBOR representation, for object transmission or storage.

Since: 0.2.0.0

Minimal complete definition

Nothing

Methods

encode :: a -> Encoding

Definition for encoding a given type into a binary representation, using the Encoding Monoid.

Since: 0.2.0.0

default encode :: (Generic a, GSerialiseEncode (Rep a)) => a -> Encoding

decode :: Decoder s a

Definition of a given Decoder for a type.

Since: 0.2.0.0

default decode :: (Generic a, GSerialiseDecode (Rep a)) => Decoder s a

encodeList :: [a] -> Encoding

Utility to support specialised encoding for some list type - used for Char/String instances in this package.

Since: 0.2.0.0

decodeList :: Decoder s [a]

Utility to support specialised decoding for some list type - used for Char/String instances in this package.

Since: 0.2.0.0

class GSerialiseEncode f where

Since: 0.2.0.0

Methods

gencode :: f a -> Encoding

Since: 0.2.0.0

class GSerialiseDecode f where

Since: 0.2.0.0

Methods

gdecode :: Decoder s (f a)

Since: 0.2.0.0

class GSerialiseProd f where

Serialization of product types

Methods

nFields :: Proxy f -> Word

Number of fields in product type

encodeSeq :: f a -> Encoding

Encode fields sequentially without writing header

gdecodeSeq :: Decoder s (f a)

Decode fields sequentially without reading header

class GSerialiseSum f where

Serialization of sum types

Since: 0.2.0.0

Methods

conNumber :: f a -> Word

Number of constructor of given value

numOfFields :: f a -> Word

Number of fields of given value

encodeSum :: f a -> Encoding

Encode field

decodeSum :: Word -> Decoder s (f a)

Decode field

nConstructors :: Proxy f -> Word

Number of constructors

fieldsForCon :: Proxy f -> Word -> Decoder s Word

Number of fields for given constructor number

encodeVector :: (Serialise a, Vector v a) => v a -> Encoding

Generic encoder for vectors. Its intended use is to allow easy definition of Serialise instances for custom vector

Since: 0.2.0.0

decodeVector :: (Serialise a, Vector v a) => Decoder s (v a)

Generic decoder for vectors. Its intended use is to allow easy definition of Serialise instances for custom vector

Since: 0.2.0.0

encodeContainerSkel

Arguments

:: (Word -> Encoding)
encoder of the length

-> (container -> Int)
length

-> (accumFunc -> Encoding -> container -> Encoding)
foldr

-> accumFunc

-> container

-> Encoding

Patch functions together to obtain an Encoding for a container.

encodeMapSkel

Arguments

:: (Serialise k, Serialise v)

=> (m -> Int)
obtain the length

-> ((k -> v -> Encoding -> Encoding) -> Encoding -> m -> Encoding)

-> m

-> Encoding

A helper function for encoding maps.

decodeMapSkel

Arguments

:: (Serialise k, Serialise v)

=> ([(k, v)] -> m)
fromList

-> Decoder s m

A utility function to construct a Decoder for maps.