Control.Exception.Annotated.UnliftIO

Description

This module presents the same interface as Control.Exception.Annotated, but uses MonadUnliftIO instead of MonadCatch or MonadThrow.

Since: 0.1.2.0

The Main Type

data AnnotatedException exception

The AnnotatedException type wraps an exception with a [Annotation]. This can provide a sort of a manual stack trace with programmer provided data.

Since: 0.1.0.0

Constructors

AnnotatedException

Fields

annotations :: [Annotation]

exception :: exception

exceptionWithCallStack :: (Exception e, HasCallStack) => e -> AnnotatedException e

Annotate the underlying exception with a CallStack.

Since: 0.2.0.0

throwWithCallStack :: forall e m a. (MonadIO m, Exception e, HasCallStack) => e -> m a

Like throwWithCallStack, but uses MonadIO instead of MonadThrow.

Since: 0.1.2.0

Annotating Exceptions

checkpoint :: forall m a. (MonadUnliftIO m, HasCallStack) => Annotation -> m a -> m a

Like checkpoint, but uses MonadUnliftIO instead of MonadCatch.

Since: 0.1.2.0

checkpointMany :: forall m a. (MonadUnliftIO m, HasCallStack) => [Annotation] -> m a -> m a

Like checkpointMany, but uses MonadUnliftIO instead of MonadCatch.

Since: 0.1.2.0

checkpointCallStack :: forall m a. (MonadUnliftIO m, HasCallStack) => m a -> m a

Like checkpointCallStack, but uses MonadUnliftIO instead of MonadCatch.

Since: 0.2.0.2

checkpointCallStackWith :: forall m a. (MonadUnliftIO m, HasCallStack) => [Annotation] -> m a -> m a

Deprecated: As of annotated-exception-0.2.0.0, this is an alias for checkpointMany

Like checkpointCallStackWith, but uses MonadUnliftIO instead of MonadCatch.

Deprecated in 0.2.0.0 as it is now an alias for checkpointMany.

Since: 0.1.2.0

Handling Exceptions

catch :: forall e m a. (MonadUnliftIO m, Exception e, HasCallStack) => m a -> (e -> m a) -> m a

Like catch, but uses MonadUnliftIO instead of MonadCatch.

Since: 0.1.2.0

catches :: forall m a. (MonadUnliftIO m, HasCallStack) => m a -> [Handler m a] -> m a

Like catches, bt uses MonadUnliftIO instead of MonadCatch.

Since: 0.1.2.0

tryAnnotated :: forall e m a. (MonadUnliftIO m, Exception e) => m a -> m (Either (AnnotatedException e) a)

Like tryAnnotated but uses MonadUnliftIO instead of MonadCatch.

Since: 0.1.2.0

try :: forall e m a. (MonadUnliftIO m, Exception e) => m a -> m (Either e a)

Like try but uses MonadUnliftIO instead of MonadCatch.

Since: 0.1.2.0

Manipulating Annotated Exceptions

check :: Exception e => AnnotatedException SomeException -> Maybe (AnnotatedException e)

Call fromException on the underlying Exception, attaching the annotations to the result.

Since: 0.1.0.0

hide :: Exception e => AnnotatedException e -> AnnotatedException SomeException

Call toException on the underlying Exception.

Since: 0.1.0.0

annotatedExceptionCallStack :: AnnotatedException exception -> Maybe CallStack

Retrieves the CallStack from an AnnotatedException if one is present.

The library maintains an internal check that a single CallStack is present in the list, so this only returns the first one found. If you have added a CallStack directly to the [Annotation], then this will likely break.

Since: 0.1.0.0

addCallStackToException :: CallStack -> AnnotatedException exception -> AnnotatedException exception

Adds a CallStack to the given AnnotatedException. This function will search through the existing annotations, and it will not add a second CallStack to the list. Instead, it will append the contents of the given CallStack to the existing one.

This mirrors the behavior of the way HasCallStack actually works.

Since: 0.1.0.0

Re-exports from Data.Annotation

data Annotation where

An Annotation is a wrapper around a value that includes a Typeable constraint so we can later unpack it. It is essentially a Dynamic, but we also include Show so that you can always fall back to simply showing the Annotation if it is otherwise unrecognized.

Since: 0.1.0.0

Constructors

Annotation :: AnnC a => a -> Annotation

newtype CallStackAnnotation

Deprecated: You can just use CallStack directly now.

A wrapper type for putting a CallStack into an Annotation. We need this because CallStack does not have an Eq instance.

Deprecated in 0.2.0.0 since you can just put a CallStack directly in an Annotation now that we have no need for an Eq constraint on it.

Since: 0.1.0.0

Constructors

CallStackAnnotation

Deprecated: You can just use CallStack directly now.

Fields

unCallStackAnnotation :: [(String, SrcLoc)]

Re-exports from Control.Exception.Safe

class (Typeable e, Show e) => Exception e where

Any type that you wish to throw or catch as an exception must be an instance of the Exception class. The simplest case is a new exception type directly below the root:

data MyException = ThisException | ThatException
    deriving Show

instance Exception MyException

The default method definitions in the Exception class do what we need in this case. You can now throw and catch ThisException and ThatException as exceptions:

*Main> throw ThisException `catch` \e -> putStrLn ("Caught " ++ show (e :: MyException))
Caught ThisException

In more complicated examples, you may wish to define a whole hierarchy of exceptions:

---------------------------------------------------------------------
-- Make the root exception type for all the exceptions in a compiler

data SomeCompilerException = forall e . Exception e => SomeCompilerException e

instance Show SomeCompilerException where
    show (SomeCompilerException e) = show e

instance Exception SomeCompilerException

compilerExceptionToException :: Exception e => e -> SomeException
compilerExceptionToException = toException . SomeCompilerException

compilerExceptionFromException :: Exception e => SomeException -> Maybe e
compilerExceptionFromException x = do
    SomeCompilerException a <- fromException x
    cast a

---------------------------------------------------------------------
-- Make a subhierarchy for exceptions in the frontend of the compiler

data SomeFrontendException = forall e . Exception e => SomeFrontendException e

instance Show SomeFrontendException where
    show (SomeFrontendException e) = show e

instance Exception SomeFrontendException where
    toException = compilerExceptionToException
    fromException = compilerExceptionFromException

frontendExceptionToException :: Exception e => e -> SomeException
frontendExceptionToException = toException . SomeFrontendException

frontendExceptionFromException :: Exception e => SomeException -> Maybe e
frontendExceptionFromException x = do
    SomeFrontendException a <- fromException x
    cast a

---------------------------------------------------------------------
-- Make an exception type for a particular frontend compiler exception

data MismatchedParentheses = MismatchedParentheses
    deriving Show

instance Exception MismatchedParentheses where
    toException   = frontendExceptionToException
    fromException = frontendExceptionFromException

We can now catch a MismatchedParentheses exception as MismatchedParentheses, SomeFrontendException or SomeCompilerException, but not other types, e.g. IOException:

*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: MismatchedParentheses))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: SomeFrontendException))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: SomeCompilerException))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: IOException))
*** Exception: MismatchedParentheses

Minimal complete definition

Nothing

Methods

toException :: e -> SomeException

fromException :: SomeException -> Maybe e

displayException :: e -> String

Render this exception value in a human-friendly manner.

Default implementation: show.

Since: base-4.8.0.0

data SomeException

The SomeException type is the root of the exception type hierarchy. When an exception of type e is thrown, behind the scenes it is encapsulated in a SomeException.

Constructors

Exception e => SomeException e

throw :: forall e m a. (MonadIO m, Exception e, HasCallStack) => e -> m a

Like throw, but uses MonadIO instead of MonadThrow.

Since: 0.1.2.0

data Handler (m :: Type -> Type) a

Generalized version of Handler

Constructors

Exception e => Handler (e -> m a)

class Monad m => MonadIO (m :: Type -> Type) where

Monads in which IO computations may be embedded. Any monad built by applying a sequence of monad transformers to the IO monad will be an instance of this class.

Instances should satisfy the following laws, which state that liftIO is a transformer of monads:

liftIO . return = return

liftIO (m >>= f) = liftIO m >>= (liftIO . f)

Methods

liftIO :: IO a -> m a

Lift a computation from the IO monad. This allows us to run IO computations in any monadic stack, so long as it supports these kinds of operations (i.e. IO is the base monad for the stack).

Example

Expand

import Control.Monad.Trans.State -- from the "transformers" library

printState :: Show s => StateT s IO ()
printState = do
  state <- get
  liftIO $ print state

Had we omitted liftIO, we would have ended up with this error:

• Couldn't match type ‘IO’ with ‘StateT s IO’
 Expected type: StateT s IO ()
   Actual type: IO ()

The important part here is the mismatch between StateT s IO () and IO ().

Luckily, we know of a function that takes an IO a and returns an (m a): liftIO, enabling us to run the program and see the expected results:

> evalStateT printState "hello"
"hello"

> evalStateT printState 3
3

class MonadIO m => MonadUnliftIO (m :: Type -> Type) where

Methods

withRunInIO :: ((forall a. m a -> IO a) -> IO b) -> m b