Control.Tracer

Description

Tracer is a contravariant functor to thread observable values through a number of transformers, possibly annotating them with additional information, or filtering them based on evaluating predicates.

Documentation

newtype Tracer m a

example: simply output a message on the console

let logTrace = traceWith $ showTracing $ stdoutTracer
in  logTrace "hello world"

example: calling a function and passing in a Tracer

example1 :: IO ()
example1 = do
    let logTrace a = traceWith (showTracing (contramap ("Debug: " ++) stdoutTracer)) a
    void $ callFun1 logTrace

callFun1 :: (String -> IO ()) -> IO Int
callFun1 logTrace = do
    logTrace "in function 1"
    return 42

runTracer evaluates a Tracer (i.e. consumes its argument)

Constructors

Tracer

Fields

runTracer :: a -> m ()

class Contravariant (f :: Type -> Type) where

The class of contravariant functors.

Whereas in Haskell, one can think of a Functor as containing or producing values, a contravariant functor is a functor that can be thought of as consuming values.

As an example, consider the type of predicate functions a -> Bool. One such predicate might be negative x = x < 0, which classifies integers as to whether they are negative. However, given this predicate, we can re-use it in other situations, providing we have a way to map values to integers. For instance, we can use the negative predicate on a person's bank balance to work out if they are currently overdrawn:

newtype Predicate a = Predicate { getPredicate :: a -> Bool }

instance Contravariant Predicate where
  contramap :: (a' -> a) -> (Predicate a -> Predicate a')
  contramap f (Predicate p) = Predicate (p . f)
                                         |   `- First, map the input...
                                         `----- then apply the predicate.

overdrawn :: Predicate Person
overdrawn = contramap personBankBalance negative

Any instance should be subject to the following laws:

Identity

contramap id = id

Composition

contramap (g . f) = contramap f . contramap g

Note, that the second law follows from the free theorem of the type of contramap and the first law, so you need only check that the former condition holds.

Minimal complete definition

contramap

Methods

contramap :: (a' -> a) -> f a -> f a'

(>$) :: b -> f b -> f a infixl 4

Replace all locations in the output with the same value. The default definition is contramap . const, but this may be overridden with a more efficient version.

tracing

traceWith :: Tracer m a -> a -> m ()

trace an observable value with a Tracer

tracers

nullTracer :: Applicative m => Tracer m a

this Tracer forgets about all arguments

stdoutTracer :: MonadIO m => Tracer m String

Output a traced String to stdout

debugTracer :: Applicative m => Tracer m String

Output a traced String using Trace

transformers

contramapM :: Monad m => (a -> m b) -> Tracer m b -> Tracer m a

Transform a tracer using a Kleisli map.

showTracing :: Show a => Tracer m String -> Tracer m a

transform a traced value to a showable instance.

condTracing :: Monad m => (a -> Bool) -> Tracer m a -> Tracer m a

conditionally trace an observable given the evaluation of a predicate.

condTracingM :: Monad m => m (a -> Bool) -> Tracer m a -> Tracer m a

conditionally trace an observable given the evaluation of a predicate in a monadic context.

natTracer :: (forall x. m x -> n x) -> Tracer m s -> Tracer n s

natural transformation from monad m to monad n.