Control.Tracer.Transformers

Documentation

newtype Counting a

Constructors

Counting Int

newtype Folding a f

Constructors

Folding f

counting :: forall m a. MonadIO m => Tracer m (Counting a) -> m (Tracer m a)

A stateful tracer transformer that substitutes messages with a monotonically incrementing occurence count.

fanning :: forall m a. (a -> Tracer m a) -> Tracer m a

A pure tracer combinator that allows to decide a further tracer to use, based on the message being processed.

folding :: forall m f a. MonadIO m => (f -> a -> f) -> f -> Tracer m (Folding a f) -> m (Tracer m a)

A generalised trace transformer that provides evolving state, defined as a strict left fold.