Control.Tracer.Transformers.ObserveOutcome

Description

Observing events with annotations of thread id and time.

transformer

class Monad m => Outcome m a where

Associated Types

type IntermediateValue a

type OutcomeMetric a

Methods

classifyObservable :: a -> m OutcomeProgressionStatus

captureObservableValue :: a -> m (IntermediateValue a)

computeOutcomeMetric :: a -> IntermediateValue a -> IntermediateValue a -> m (OutcomeMetric a)

type OutcomeEnhancedTracer m a = Tracer m (Either a (OutcomeFidelity (OutcomeMetric a)))

The Maybe (OutcomeMetric a) captures the 'DeltaQ-ness' of the nature of outcomes, may / may not complete.

data OutcomeFidelity a

Also need to know that observables happened in the "right way"

Constructors

EndsBeforeStarted

StartsBeforeEnds a

ProgressedNormally a

data OutcomeProgressionStatus

Constructors

OutcomeStarts

OutcomeOther

OutcomeEnds

mkOutcomeExtractor :: forall m a. (MonadIO m, MonadMask m, Outcome m a) => m (OutcomeEnhancedTracer m a -> Tracer m a)

Generic Trace transformer. It could be written to take an initial argument, but restricting the scope of that per-invocation state seems more appropriate (for the moment). That may be of use if/when explict management of timeout was required and/or non-termination of the outcome at the end of a run was of interest.