Criterion.Types

Description

Types for benchmarking.

The core type is Benchmarkable, which admits both pure functions and IO actions.

For a pure function of type a -> b, the benchmarking harness calls this function repeatedly, each time with a different Int64 argument (the number of times to run the function in a loop), and reduces the result the function returns to weak head normal form.

For an action of type IO a, the benchmarking harness calls the action repeatedly, but does not reduce the result.

Configuration

data Config

Top-level benchmarking configuration.

Constructors

Config

Fields

confInterval :: CL Double

Confidence interval for bootstrap estimation (greater than 0, less than 1).

timeLimit :: Double

Number of seconds to run a single benchmark. (In practice, execution time will very slightly exceed this limit.)

resamples :: Int

Number of resamples to perform when bootstrapping.

regressions :: [([String], String)]

Regressions to perform.

rawDataFile :: Maybe FilePath

File to write binary measurement and analysis data to. If not specified, this will be a temporary file.

reportFile :: Maybe FilePath

File to write report output to, with template expanded.

csvFile :: Maybe FilePath

File to write CSV summary to.

jsonFile :: Maybe FilePath

File to write JSON-formatted results to.

junitFile :: Maybe FilePath

File to write JUnit-compatible XML results to.

verbosity :: Verbosity

Verbosity level to use when running and analysing benchmarks.

template :: FilePath

Template file to use if writing a report.

data Verbosity

Control the amount of information displayed.

Constructors

Quiet

Normal

Verbose

Benchmark descriptions

data Benchmarkable

Constructors

NFData a => Benchmarkable

Fields

allocEnv :: Int64 -> IO a

cleanEnv :: Int64 -> a -> IO ()

runRepeatedly :: a -> Int64 -> IO ()

perRun :: Bool

data Benchmark where

Constructors

Environment :: forall env a. NFData env => IO env -> (env -> IO a) -> (env -> Benchmark) -> Benchmark

Benchmark :: String -> Benchmarkable -> Benchmark

BenchGroup :: String -> [Benchmark] -> Benchmark

Measurements

data Measured

Constructors

Measured

Fields

measTime :: !Double

measCpuTime :: !Double

measCycles :: !Int64

measIters :: !Int64

measAllocated :: !Int64

measPeakMbAllocated :: !Int64

measNumGcs :: !Int64

measBytesCopied :: !Int64

measMutatorWallSeconds :: !Double

measMutatorCpuSeconds :: !Double

measGcWallSeconds :: !Double

measGcCpuSeconds :: !Double

fromInt :: Int64 -> Maybe Int64

toInt :: Maybe Int64 -> Int64

fromDouble :: Double -> Maybe Double

toDouble :: Maybe Double -> Double

measureAccessors :: Map String (Measured -> Maybe Double, String)

measureKeys :: [String]

measure :: Unbox a => (Measured -> a) -> Vector Measured -> Vector a

rescale :: Measured -> Measured

Benchmark construction

env :: NFData env => IO env -> (env -> Benchmark) -> Benchmark

envWithCleanup :: NFData env => IO env -> (env -> IO a) -> (env -> Benchmark) -> Benchmark

perBatchEnv :: (NFData env, NFData b) => (Int64 -> IO env) -> (env -> IO b) -> Benchmarkable

perBatchEnvWithCleanup :: (NFData env, NFData b) => (Int64 -> IO env) -> (Int64 -> env -> IO ()) -> (env -> IO b) -> Benchmarkable

perRunEnv :: (NFData env, NFData b) => IO env -> (env -> IO b) -> Benchmarkable

perRunEnvWithCleanup :: (NFData env, NFData b) => IO env -> (env -> IO ()) -> (env -> IO b) -> Benchmarkable

toBenchmarkable :: (Int64 -> IO ()) -> Benchmarkable

bench :: String -> Benchmarkable -> Benchmark

bgroup :: String -> [Benchmark] -> Benchmark

addPrefix :: String -> String -> String

benchNames :: Benchmark -> [String]

Evaluation control

nf :: NFData b => (a -> b) -> a -> Benchmarkable

whnf :: (a -> b) -> a -> Benchmarkable

nfIO :: NFData a => IO a -> Benchmarkable

whnfIO :: IO a -> Benchmarkable

nfAppIO :: NFData b => (a -> IO b) -> a -> Benchmarkable

whnfAppIO :: (a -> IO b) -> a -> Benchmarkable

Result types

data Outliers

Outliers from sample data, calculated using the boxplot technique.

Constructors

Outliers

Fields

samplesSeen :: !Int64

lowSevere :: !Int64

More than 3 times the interquartile range (IQR) below the first quartile.

lowMild :: !Int64

Between 1.5 and 3 times the IQR below the first quartile.

highMild :: !Int64

Between 1.5 and 3 times the IQR above the third quartile.

highSevere :: !Int64

More than 3 times the IQR above the third quartile.

data OutlierEffect

A description of the extent to which outliers in the sample data affect the sample mean and standard deviation.

Constructors

Unaffected
Less than 1% effect.

Slight
Between 1% and 10%.

Moderate
Between 10% and 50%.

Severe
Above 50% (i.e. measurements are useless).

data OutlierVariance

Analysis of the extent to which outliers in a sample affect its standard deviation (and to some extent, its mean).

Constructors

OutlierVariance

Fields

ovEffect :: OutlierEffect

Qualitative description of effect.

ovDesc :: String

Brief textual description of effect.

ovFraction :: Double

Quantitative description of effect (a fraction between 0 and 1).

data Regression

Results of a linear regression.

Constructors

Regression

Fields

regResponder :: String

Name of the responding variable.

regCoeffs :: Map String (Estimate ConfInt Double)

Map from name to value of predictor coefficients.

regRSquare :: Estimate ConfInt Double

R² goodness-of-fit estimate.

data KDE

Data for a KDE chart of performance.

Constructors

KDE

Fields

kdeType :: String

kdeValues :: Vector Double

kdePDF :: Vector Double

data Report

Report of a sample analysis.

Constructors

Report

Fields

reportNumber :: Int

A simple index indicating that this is the nth report.

reportName :: String

The name of this report.

reportKeys :: [String]

See measureKeys.

reportMeasured :: Vector Measured

Raw measurements.

reportAnalysis :: SampleAnalysis

Report analysis.

reportOutliers :: Outliers

Analysis of outliers.

reportKDEs :: [KDE]

Data for a KDE of times.

data SampleAnalysis

Result of a bootstrap analysis of a non-parametric sample.

Constructors

SampleAnalysis

Fields

anRegress :: [Regression]

Estimates calculated via linear regression.

anMean :: Estimate ConfInt Double

Estimated mean.

anStdDev :: Estimate ConfInt Double

Estimated standard deviation.

anOutlierVar :: OutlierVariance

Description of the effects of outliers on the estimated variance.

data DataRecord

Constructors

Measurement Int String (Vector Measured)

Analysed Report