Data.GraphViz.Attributes.Colors

Description

This module defines the various colors, etc. for Graphviz. For information on colors in general, see: http://graphviz.org/doc/info/attrs.html#k:color For named colors, see: http://graphviz.org/doc/info/colors.html

Note that the ColorBrewer Color Schemes (shortened to just "Brewer" for the rest of this module) are covered by the following license (also available in the LICENSE file of this library): http://graphviz.org/doc/info/colors.html#brewer_license

Color schemes.

data ColorScheme

This represents the color schemes that Graphviz accepts.

Constructors

X11

SVG

Brewer BrewerScheme

Colors

data Color

Defining a color for use with Graphviz. Note that named colors have been split up into Colors and those based upon the Brewer color schemes.

Constructors

RGB

Fields

red :: Word8

green :: Word8

blue :: Word8

RGBA

Fields

red :: Word8

green :: Word8

blue :: Word8

alpha :: Word8

HSV
The hue, saturation and value values must all be 0 <= x <=1.

Fields

hue :: Double

saturation :: Double

value :: Double

X11Color X11Color

SVGColor SVGColor

BrewerColor BrewerColor

type ColorList = [WeightedColor]

The sum of the optional weightings must sum to at most 1.

data WeightedColor

A Color tagged with an optional weighting.

Constructors

WC

Fields

wColor :: Color

weighting :: Maybe Double

Must be in range 0 <= W <= 1.

toWC :: Color -> WeightedColor

For colors without weightings.

toColorList :: [Color] -> ColorList

For a list of colors without weightings.

class NamedColor nc where

More easily convert named colors to an overall Color value.

Minimal complete definition

colorScheme, toColor, printNC, parseNC'

Methods

toColor :: nc -> Color

toWColor :: NamedColor nc => nc -> WeightedColor

Conversion to/from Colour.

toColour :: Color -> Maybe (AlphaColour Double)

Attempt to convert a Color into a Colour value with an alpha channel. The use of Maybe is because the RGB values of the Colors haven't been stored here (primarily for licensing reasons).

fromColour :: Colour Double -> Color

Convert a Colour value to an RGB Color.

fromAColour :: AlphaColour Double -> Color

Convert an AlphaColour to an RGBA Color. The exception to this is for any AlphaColour which has alphaChannel ac == 0; these are converted to X11Color Transparent (note that the Show instance for such an AlphaColour is "transparent").