Data.GraphViz.Attributes.Colors.Brewer

Description

You almost definitely do not want to use this module. It is only defined for completeness when parsing existing Dot code.

Graphviz contains a list of colors known as the /Brewer color schemes/.

These colors are available under an Apache-style license: http://www.graphviz.org/doc/info/colors.html#brewer_license. As such, they are not recommended for general use, and have only been included in this package for completeness.

The complete list of Brewer colors can be found at http://www.graphviz.org/doc/info/colors.html#brewer.

Documentation

data BrewerScheme

Specify which colour palette and how many colours it has. Note the allowed values for the different BrewerNames.

Constructors

BScheme BrewerName Word8

data BrewerName

All of these have a minimum level value of 3, with a maximum of 9 unless otherwise specified.

Constructors

Accent
Maximum of 8.

Blues

Brbg
Maximum of 11.

Bugn

Bupu

Dark2
Maximum of 8.

Gnbu

Greens

Greys

Oranges

Orrd

Paired
Maximum of 12.

Pastel1

Pastel2
Maximum of 8.

Piyg
Maximum of 11.

Prgn
Maximum of 11.

Pubu

Pubugn

Puor
Maximum of 11; note that the last two are listed after the Purd values in the documentation.

Purd

Purples

Rdbu
Maximum of 11; note that the last two are listed first.

Rdgy
Maximum of 11; note that the last two are listed after the Rdpu values in the documentation.

Rdpu

Rdylbu
Maximum of 11.

Rdylgn
Maximum of 11.

Reds

Set1

Set2
Maximum of 8.

Set3
Maximum of 12.

Spectral
Maximum of 11.

Ylgn

Ylgnbu

Ylorbr

Ylorrd

data BrewerColor

This value should be between 1 and the level of the BrewerScheme being used.

Constructors

BC BrewerScheme Word8