Data.GraphViz.Attributes.HTML

Description

This module is written to be imported qualified. It defines the syntax for HTML-like values for use in Graphviz. Please note that these values are not really HTML, but the term "HTML" is used throughout as it is less cumbersome than "HTML-like". To be able to use this, the version of Graphviz must be at least 1.10. For more information, please see: http://graphviz.org/doc/info/shapes.html#html

The actual definition of the syntax specifies that these types must be valid XML syntax. As such, this assumed when printing and parsing, though the correct escape/descaping for ", &, < and > are automatically done when printing and parsing.

Differences from how Graphviz treats HTML-like values:

Graphviz only specifies the above-listed characters must be escaped; however, internally it also escapes -, ' and multiple sequences of spaces. This library attempts to match this behaviour. Please let me know if this behaviour (especially about escaping multiple spaces) is unwanted.

When parsing escaped HTML characters, numeric escapes are converted to the corresponding character as are the various characters listed above; all other escaped characters (apart from those listed above) are silently ignored and removed from the input (since technically these must be valid XML, which doesn't recognise as many named escape characters as does HTML).

All whitespace read in is kept (even if Graphviz would ignore multiple whitespace characters); when printing them, however, they are replaced with non-breaking spaces. As such, if multiple literal whitespace characters are used in a sequence, then the result of parsing and then printing some Dot code will not be the same as the initial Dot code. Furthermore, all whitespace characters are printed as spaces.

It is assumed that all parsed & values are the beginning of an XML escape sequence (which must finish with a ; character).

There should be no pre-escaped characters in values; when printing, the & will get escaped without considering if that is an escaped character.

Documentation

data Label

The overall type for HTML-like labels. Fundamentally, HTML-like values in Graphviz are either textual (i.e. a single element with formatting) or a table. Note that Label values can be nested via LabelCell.

Constructors

Text Text

Table Table

type Text = [TextItem]

Represents a textual component of an HTML-like label. It is assumed that a Label list is non-empty. It is preferable to "group" Str values together rather than have individual ones. Note that when printing, the individual values are concatenated together without spaces, and when parsing anything that isn't a tag is assumed to be a Str: that is, something like "<BR/> <BR/>" is parsed as:

[Newline [], Str " ", Newline []]

data TextItem

Textual items in HTML-like labels.

Constructors

Str Text

Newline Attributes
Only accepts an optional Attribute Attribute; defined this way for ease of printing/parsing.

Font Attributes Text

Format Format Text
Only available in Graphviz >= 2.28.0.

data Format

Constructors

Italics

Bold

Underline

Overline
Requires Graphviz >= 2.38.0.

Subscript

Superscript

data Table

A table in HTML-like labels. Tables are optionally wrapped in overall FONT tags.

Constructors

HTable

Fields

tableFontAttrs :: Maybe Attributes

Optional FONT attributes. Just [] denotes empty FONT tags; Nothing denotes no such tags.

tableAttrs :: Attributes

tableRows :: [Row]

This list is assumed to be non-empty.

data Row

A row in a Label. The list of Cell values is assumed to be non-empty.

Constructors

Cells [Cell]

HorizontalRule
Should be between Cells values, requires Graphviz >= 2.29.0

data Cell

Cells either recursively contain another Label or else a path to an image file.

Constructors

LabelCell Attributes Label

ImgCell Attributes Img

VerticalRule
Should be between LabelCell or ImgCell values, requires Graphviz >= 2.29.0

newtype Img

The path to an image; accepted Attributes are Attribute and Src.

Constructors

Img Attributes

type Attributes = [Attribute]

The various HTML-like label-specific attributes being used.

data Attribute

Note that not all Attribute values are valid everywhere: see the comments for each one on where it is valid.

Constructors

Align Align
Valid for: Label, Cell, Newline.

BAlign Align
Valid for: Cell.

BGColor Color
Valid for: Label (including tableFontAttrs), Cell, Font.

Border Word8
Valid for: Label, Cell. Default is 1; 0 represents no border.

CellBorder Word8
Valid for: Label. Default is 1; 0 represents no border.

CellPadding Word8
Valid for: Label, Cell. Default is 2.

CellSpacing Word8
Valid for: Label, Cell. Default is 2; maximum is 127.

Color Color
Valid for: Label, Cell.

ColSpan Word16
Valid for: Cell. Default is 1.

Columns CellFormat
Valid for: Label. Requires Graphviz >= 2.40.1

Face Text
Valid for: tableFontAttrs, Font.

FixedSize Bool
Valid for: Label, Cell. Default is False.

GradientAngle Int
Valid for: Label, Cell. Default is 0. Requires Graphviz >= 2.40.1

Height Word16
Valid for: Label, Cell.

HRef Text
Valid for: Label, Cell.

ID Text
Valid for: Label, Cell. Requires Graphviz >= 2.29.0

PointSize Double
Valid for: tableFontAttrs, Font.

Port PortName
Valid for: Label, Cell.

Rows CellFormat
Valid for: Label. Requires Graphviz >= 2.40.1

RowSpan Word16
Valid for: Cell.

Scale Scale
Valid for: Img.

Sides [Side]
Valid for: Label, Cell. Default is [LeftSide, TopSide, RightSide, BottomSide]. Requires Graphviz >= 2.40.1

Src FilePath
Valid for: Img.

Style Style
Valid for: Label, Cell. Requires Graphviz >= 2.40.1

Target Text
Valid for: Label, Cell.

Title Text
Valid for: Label, Cell. Has an alias of TOOLTIP.

VAlign VAlign
Valid for: Label, Cell.

Width Word16
Valid for: Label, Cell.

data Align

Specifies horizontal placement. When an object is allocated more space than required, this value determines where the extra space is placed left and right of the object.

Constructors

HLeft

HCenter
Default value.

HRight

HText
LabelCell values only; aligns lines of text using the full cell width. The alignment of a line is determined by its (possibly implicit) associated Newline element.

data VAlign

Specifies vertical placement. When an object is allocated more space than required, this value determines where the extra space is placed above and below the object.

Constructors

HTop

HMiddle
Default value.

HBottom

data CellFormat

Constructors

RuleBetween

data Scale

Specifies how an image will use any extra space available in its cell. If undefined, the image inherits the value of the ImageScale attribute.

Constructors

NaturalSize
Default value.

ScaleUniformly

ExpandWidth

ExpandHeight

ExpandBoth

data Side

Which sides of a border in a cell or table should be drawn, if a border is drawn.

Constructors

LeftSide

RightSide

TopSide

BottomSide

data Style

Constructors

Rounded
Valid for Label

Radial
Valid for Label, Cell.