Data.GraphViz.Parsing

Description

This module defines simple helper functions for use with Text.ParserCombinators.Poly.Lazy.

Note that the ParseDot instances for Bool, etc. match those specified for use with Graphviz (e.g. non-zero integers are equivalent to True).

You should not be using this module; rather, it is here for informative/documentative reasons. If you want to parse a DotRepr, you should use parseDotGraph rather than its ParseDot instance.

Re-exporting pertinent parts of Polyparse.

class Functor f => Applicative (f :: Type -> Type) where

A functor with application, providing operations to

embed pure expressions (pure), and

sequence computations and combine their results (<*> and liftA2).

A minimal complete definition must include implementations of pure and of either <*> or liftA2. If it defines both, then they must behave the same as their default definitions:

(<*>) = liftA2 id

liftA2 f x y = f <$> x <*> y

Further, any definition must satisfy the following:

Identity

pure id <*> v = v

Composition

pure (.) <*> u <*> v <*> w = u <*> (v <*> w)

Homomorphism

pure f <*> pure x = pure (f x)

Interchange

u <*> pure y = pure ($ y) <*> u

The other methods have the following default definitions, which may be overridden with equivalent specialized implementations:

u *> v = (id <$ u) <*> v

u <* v = liftA2 const u v

As a consequence of these laws, the Functor instance for f will satisfy

fmap f x = pure f <*> x

It may be useful to note that supposing

forall x y. p (q x y) = f x . g y

it follows from the above that

liftA2 p (liftA2 q u v) = liftA2 f u . liftA2 g v

If f is also a Monad, it should satisfy

pure = return

m1 <*> m2 = m1 >>= (\x1 -> m2 >>= (\x2 -> return (x1 x2)))

(*>) = (>>)

(which implies that pure and <*> satisfy the applicative functor laws).

Minimal complete definition

pure, ((<*>) | liftA2)

Methods

pure :: a -> f a

Lift a value.

(<*>) :: f (a -> b) -> f a -> f b infixl 4

Sequential application.

A few functors support an implementation of <*> that is more efficient than the default one.

Example

Expand

Used in combination with (<$>), (<*>) can be used to build a record.

>>> data MyState = MyState {arg1 :: Foo, arg2 :: Bar, arg3 :: Baz}

>>> produceFoo :: Applicative f => f Foo

>>> produceBar :: Applicative f => f Bar
>>> produceBaz :: Applicative f => f Baz

>>> mkState :: Applicative f => f MyState
>>> mkState = MyState <$> produceFoo <*> produceBar <*> produceBaz

liftA2 :: (a -> b -> c) -> f a -> f b -> f c

Lift a binary function to actions.

Some functors support an implementation of liftA2 that is more efficient than the default one. In particular, if fmap is an expensive operation, it is likely better to use liftA2 than to fmap over the structure and then use <*>.

This became a typeclass method in 4.10.0.0. Prior to that, it was a function defined in terms of <*> and fmap.

Example

Expand

>>> liftA2 (,) (Just 3) (Just 5)
Just (3,5)

(*>) :: f a -> f b -> f b infixl 4

Sequence actions, discarding the value of the first argument.

Examples

Expand

If used in conjunction with the Applicative instance for Maybe, you can chain Maybe computations, with a possible "early return" in case of Nothing.

>>> Just 2 *> Just 3
Just 3

>>> Nothing *> Just 3
Nothing

Of course a more interesting use case would be to have effectful computations instead of just returning pure values.

>>> import Data.Char
>>> import Text.ParserCombinators.ReadP
>>> let p = string "my name is " *> munch1 isAlpha <* eof
>>> readP_to_S p "my name is Simon"
[("Simon","")]

(<*) :: f a -> f b -> f a infixl 4

Sequence actions, discarding the value of the second argument.

class Applicative f => Alternative (f :: Type -> Type) where

A monoid on applicative functors.

If defined, some and many should be the least solutions of the equations:

some v = (:) <$> v <*> many v

many v = some v <|> pure []

Minimal complete definition

empty, (<|>)

Methods

(<|>) :: f a -> f a -> f a infixl 3

An associative binary operation

some :: f a -> f [a]

One or more.

many :: f a -> f [a]

Zero or more.

newtype Const a (b :: k)

The Const functor.

Constructors

Const

Fields

getConst :: a

newtype ZipList a

Lists, but with an Applicative functor based on zipping.

Constructors

ZipList

Fields

getZipList :: [a]

newtype WrappedArrow (a :: Type -> Type -> Type) b c

Constructors

WrapArrow

Fields

unwrapArrow :: a b c

newtype WrappedMonad (m :: Type -> Type) a

Constructors

WrapMonad

Fields

unwrapMonad :: m a

newtype Parser s a

Constructors

P (s -> Text -> Result (Text, s) a)

data Result z a

Constructors

Success z a

Failure z String

Committed (Result z a)

class Commitment (p :: Type -> Type) where

Methods

commit :: p a -> p a

adjustErr :: p a -> (String -> String) -> p a

oneOf' :: [(String, p a)] -> p a

class (Functor p, Monad p, MonadFail p, Applicative p, Alternative p, Commitment p) => PolyParse (p :: Type -> Type)

(<$>) :: Functor f => (a -> b) -> f a -> f b infixl 4

An infix synonym for fmap.

The name of this operator is an allusion to $. Note the similarities between their types:

 ($)  ::              (a -> b) ->   a ->   b
(<$>) :: Functor f => (a -> b) -> f a -> f b

Whereas $ is function application, <$> is function application lifted over a Functor.

Examples

Expand

Convert from a Maybe Int to a Maybe String using show:

>>> show <$> Nothing
Nothing
>>> show <$> Just 3
Just "3"

Convert from an Either Int Int to an Either Int String using show:

>>> show <$> Left 17
Left 17
>>> show <$> Right 17
Right "17"

Double each element of a list:

>>> (*2) <$> [1,2,3]
[2,4,6]

Apply even to the second element of a pair:

>>> even <$> (2,2)
(2,True)

bracket :: PolyParse p => p bra -> p ket -> p a -> p a

(<$) :: Functor f => a -> f b -> f a infixl 4

Replace all locations in the input with the same value. The default definition is fmap . const, but this may be overridden with a more efficient version.

Examples

Expand

Perform a computation with Maybe and replace the result with a constant value if it is Just:

>>> 'a' <$ Just 2
Just 'a'
>>> 'a' <$ Nothing
Nothing

(<**>) :: Applicative f => f a -> f (a -> b) -> f b infixl 4

A variant of <*> with the arguments reversed.

liftA :: Applicative f => (a -> b) -> f a -> f b

Lift a function to actions. Equivalent to Functor's fmap but implemented using only Applicative's methods: liftA f a = pure f <*> a

As such this function may be used to implement a Functor instance from an Applicative one.

Examples

Expand

Using the Applicative instance for Lists:

>>> liftA (+1) [1, 2]
[2,3]

Or the Applicative instance for Maybe

>>> liftA (+1) (Just 3)
Just 4

liftA3 :: Applicative f => (a -> b -> c -> d) -> f a -> f b -> f c -> f d

Lift a ternary function to actions.

satisfy :: (Char -> Bool) -> Parser s Char

eof :: Parser s ()

optional :: Alternative f => f a -> f (Maybe a)

One or none.

It is useful for modelling any computation that is allowed to fail.

Examples

Expand

Using the Alternative instance of Control.Monad.Except, the following functions:

>>> import Control.Monad.Except

>>> canFail = throwError "it failed" :: Except String Int
>>> final = return 42                :: Except String Int

Can be combined by allowing the first function to fail:

>>> runExcept $ canFail *> final
Left "it failed"
>>> runExcept $ optional canFail *> final
Right 42

many1 :: PolyParse p => p a -> p [a]

sepBy :: PolyParse p => p a -> p sep -> p [a]

sepBy1 :: PolyParse p => p a -> p sep -> p [a]

apply :: PolyParse p => p (a -> b) -> p a -> p b

asum :: (Foldable t, Alternative f) => t (f a) -> f a

The sum of a collection of actions using (<|>), generalizing concat.

asum is just like msum, but generalised to Alternative.

Examples

Expand

Basic usage:

>>> asum [Just "Hello", Nothing, Just "World"]
Just "Hello"

exactly :: PolyParse p => Int -> p a -> p [a]

stQuery :: (s -> a) -> Parser s a

stUpdate :: (s -> s) -> Parser s ()

adjustErrBad :: PolyParse p => p a -> (String -> String) -> p a

bracketSep :: PolyParse p => p bra -> p sep -> p ket -> p a -> p [a]

discard :: PolyParse p => p a -> p b -> p a

failBad :: PolyParse p => String -> p a

manyFinally :: PolyParse p => p a -> p z -> p [a]

manyFinally' :: (PolyParse p, Show a) => p a -> p z -> p [a]

oneOf :: PolyParse p => [p a] -> p a

upto :: PolyParse p => Int -> p a -> p [a]

many1Satisfy :: (Char -> Bool) -> Parser s Text

manySatisfy :: (Char -> Bool) -> Parser s Text

next :: Parser s Char

onFail :: Parser s a -> Parser s a -> Parser s a

reparse :: Text -> Parser s ()

stGet :: Parser s s

The ParseDot class.

type Parse a = Parser GraphvizState a

A ReadS-like type alias.

class ParseDot a where

Minimal complete definition

parseUnqt

Methods

parseUnqt :: Parse a

parse :: Parse a

parseUnqtList :: Parse [a]

parseList :: Parse [a]

parseIt :: ParseDot a => Text -> (a, Text)

Parse the required value, returning also the rest of the input Text that hasn't been parsed (for debugging purposes).

parseIt' :: ParseDot a => Text -> a

Parse the required value with the assumption that it will parse all of the input Text.

runParser :: Parse a -> Text -> (Either String a, Text)

runParser' :: Parse a -> Text -> a

A variant of runParser where it is assumed that the provided parsing function consumes all of the Text input (with the exception of whitespace at the end).

runParserWith :: (GraphvizState -> GraphvizState) -> Parse a -> Text -> (Either String a, Text)

parseLiberally :: GraphvizState -> GraphvizState

checkValidParse :: Either String a -> a

If unable to parse Dot code properly, throw a GraphvizException.

checkValidParseWithRest :: (Either String a, Text) -> a

If unable to parse Dot code properly, throw a GraphvizException, with the error containing the remaining unparsed code..

Convenience parsing combinators.

ignoreSep :: (a -> b -> c) -> Parse a -> Parse sep -> Parse b -> Parse c

The opposite of bracket.

onlyBool :: Parse Bool

Use this when you do not want numbers to be treated as Bool values.

quotelessString :: Parse Text

Parse a Text that doesn't need to be quoted.

stringBlock :: Parse Text

numString :: Bool -> Parse Text

isNumString :: Bool -> Text -> Bool

Determine if this String represents a number. Boolean parameter determines if exponents are considered part of numbers for this.

isIntString :: Text -> Bool

quotedString :: Parse Text

Used when quotes are explicitly required;

parseEscaped :: Bool -> [Char] -> [Char] -> Parse Text

Parse a Text where the provided Chars (as well as " and \) are escaped and the second list of Chars are those that are not permitted. Note: does not parse surrounding quotes. The Bool value indicates whether empty Texts are allowed or not.

parseAndSpace :: Parse a -> Parse a

string :: String -> Parse ()

strings :: [String] -> Parse ()

character :: Char -> Parse Char

Assumes that any letter is ASCII for case-insensitive comparisons.

parseStrictFloat :: Bool -> Parse Double

Parse a floating point number that actually contains decimals. Bool flag indicates whether values that need to be quoted are parsed.

parseSignedFloat :: Bool -> Parse Double

noneOf :: [Char] -> Parse Char

whitespace1 :: Parse ()

Parses at least one whitespace character.

whitespace :: Parse ()

Parses zero or more whitespace characters.

wrapWhitespace :: Parse a -> Parse a

Parse and discard optional surrounding whitespace.

optionalQuotedString :: String -> Parse ()

optionalQuoted :: Parse a -> Parse a

quotedParse :: Parse a -> Parse a

orQuote :: Parse Char -> Parse Char

quoteChar :: Char

newline :: Parse ()

Parses a newline.

newline' :: Parse ()

Consume all whitespace and newlines until a line with non-whitespace is reached. The whitespace on that line is not consumed.

parseComma :: Parse ()

parseEq :: Parse ()

tryParseList :: ParseDot a => Parse [a]

Try to parse a list of the specified type; returns an empty list if parsing fails.

tryParseList' :: Parse [a] -> Parse [a]

Return an empty list if parsing a list fails.

consumeLine :: Parse Text

Parses and returns all characters up till the end of the line, but does not touch the newline characters.

commaSep :: (ParseDot a, ParseDot b) => Parse (a, b)

commaSepUnqt :: (ParseDot a, ParseDot b) => Parse (a, b)

commaSep' :: Parse a -> Parse b -> Parse (a, b)

stringRep :: a -> String -> Parse a

stringReps :: a -> [String] -> Parse a

stringParse :: [(String, Parse a)] -> Parse a

stringValue :: [(String, a)] -> Parse a

parseAngled :: Parse a -> Parse a

parseBraced :: Parse a -> Parse a

parseColorScheme :: Bool -> Parse ColorScheme