Data.GraphViz.Printing

Description

This module defines simple helper functions for use with Text.PrettyPrint. It also re-exports all the pretty-printing combinators from that module.

Note that the PrintDot instances for Bool, etc. match those specified for use with Graphviz.

You should only be using this module if you are writing custom node types for use with Data.GraphViz.Types. For actual printing of code, use printDotGraph (which produces a Text value).

The Dot language specification specifies that any identifier is in one of four forms:

Any string of alphabetic ([a-zA-Z\200-\377]) characters, underscores ('_') or digits ([0-9]), not beginning with a digit;

a number [-]?(.[0-9]+ | [0-9]+(.[0-9]*)? );

any double-quoted string ("...") possibly containing escaped quotes (\");

an HTML string (<...>).

(Note that the first restriction is referring to a byte-by-byte comparison using octal values; when using UTF-8 this corresponds to all characters c where ord c >= 128.)

Due to these restrictions, you should only use text when you are sure that the Text in question is static and quotes are definitely needed/unneeded; it is better to use the Text instance for PrintDot. For more information, see the specification page: http://graphviz.org/doc/info/lang.html

Documentation

data Doc

text :: Applicative m => Text -> m Doc

group :: Functor m => m Doc -> m Doc

empty :: Applicative m => m Doc

fill :: Functor m => Int -> m Doc -> m Doc

char :: Applicative m => Char -> m Doc

parens :: Functor m => m Doc -> m Doc

list :: Functor m => m [Doc] -> m Doc

(<+>) :: Applicative m => m Doc -> m Doc -> m Doc

(</>) :: Applicative m => m Doc -> m Doc -> m Doc

semi :: Applicative m => m Doc

comma :: Applicative m => m Doc

colon :: Applicative m => m Doc

space :: Applicative m => m Doc

equals :: Applicative m => m Doc

lparen :: Applicative m => m Doc

rparen :: Applicative m => m Doc

lbrace :: Applicative m => m Doc

rbrace :: Applicative m => m Doc

int :: Applicative m => Int -> m Doc

integer :: Applicative m => Integer -> m Doc

float :: Applicative m => Float -> m Doc

double :: Applicative m => Double -> m Doc

rational :: Applicative m => Rational -> m Doc

brackets :: Functor m => m Doc -> m Doc

braces :: Functor m => m Doc -> m Doc

hcat :: Functor m => m [Doc] -> m Doc

hsep :: Functor m => m [Doc] -> m Doc

vcat :: Functor m => m [Doc] -> m Doc

nest :: Functor m => Int -> m Doc -> m Doc

hang :: Functor m => Int -> m Doc -> m Doc

punctuate :: Applicative m => m Doc -> m [Doc] -> m [Doc]

sep :: Functor m => m [Doc] -> m Doc

cat :: Functor m => m [Doc] -> m Doc

indent :: Functor m => Int -> m Doc -> m Doc

(<++>) :: Applicative m => m Doc -> m Doc -> m Doc

renderOneLine :: Doc -> SimpleDoc

beside :: Applicative m => m Doc -> m Doc -> m Doc

dquotes :: Functor m => m Doc -> m Doc

dot :: Applicative m => m Doc

langle :: Applicative m => m Doc

rangle :: Applicative m => m Doc

(<$$>) :: Applicative m => m Doc -> m Doc -> m Doc

(<//>) :: Applicative m => m Doc -> m Doc -> m Doc

align :: Functor m => m Doc -> m Doc

angles :: Functor m => m Doc -> m Doc

backslash :: Applicative m => m Doc

column :: Functor m => m (Int -> Doc) -> m Doc

displayB :: SimpleDoc -> Builder

displayTStrict :: SimpleDoc -> Text

dquote :: Applicative m => m Doc

enclose :: Applicative m => m Doc -> m Doc -> m Doc -> m Doc

encloseSep :: Applicative m => m Doc -> m Doc -> m Doc -> m [Doc] -> m Doc

fillBreak :: Functor m => Int -> m Doc -> m Doc

fillCat :: Functor m => m [Doc] -> m Doc

fillSep :: Functor m => m [Doc] -> m Doc

lbracket :: Applicative m => m Doc

line :: Applicative m => m Doc

linebreak :: Applicative m => m Doc

nesting :: Functor m => m (Int -> Doc) -> m Doc

rbracket :: Applicative m => m Doc

semiBraces :: Functor m => m [Doc] -> m Doc

softbreak :: Applicative m => m Doc

softline :: Applicative m => m Doc

spacebreak :: Applicative m => m Doc

squote :: Applicative m => m Doc

squotes :: Functor m => m Doc -> m Doc

stringStrict :: Monad m => Text -> m Doc

textStrict :: Monad m => Text -> m Doc

tupled :: Functor m => m [Doc] -> m Doc

vsep :: Functor m => m [Doc] -> m Doc

prettyM :: (Pretty a, Applicative m) => a -> m Doc

type DotCode = DotCodeM Doc

data DotCodeM a

A type alias to indicate what is being produced.

runDotCode :: DotCode -> Doc

renderDot :: DotCode -> Text

Correctly render Graphviz output.

class PrintDot a where

A class used to correctly print parts of the Graphviz Dot language. Minimal implementation is unqtDot.

Minimal complete definition

unqtDot

Methods

unqtDot :: a -> DotCode

The unquoted representation, for use when composing values to produce a larger printing value.

toDot :: a -> DotCode

The actual quoted representation; this should be quoted if it contains characters not permitted a plain ID String, a number or it is not an HTML string. Defaults to unqtDot.

unqtListToDot :: [a] -> DotCode

The correct way of representing a list of this value when printed; not all Dot values require this to be implemented. Defaults to Haskell-like list representation.

listToDot :: [a] -> DotCode

The quoted form of unqtListToDot; defaults to wrapping double quotes around the result of unqtListToDot (since the default implementation has characters that must be quoted).

unqtText :: Text -> DotCode

For use with OverloadedStrings to avoid ambiguous type variable errors.

dotText :: Text -> DotCode

For use with OverloadedStrings to avoid ambiguous type variable errors.

printIt :: PrintDot a => a -> Text

Convert to DotCode; note that this has no indentation, as we can only have one of indentation and (possibly) infinite line lengths.

addQuotes :: Text -> DotCode -> DotCode

unqtEscaped :: [Char] -> Text -> DotCode

Escape the specified chars as well as ".

printEscaped :: [Char] -> Text -> DotCode

Escape the specified chars as well as " and then wrap the result in quotes.

wrap :: DotCode -> DotCode -> DotCode -> DotCode

commaDel :: (PrintDot a, PrintDot b) => a -> b -> DotCode

printField :: PrintDot a => Text -> a -> DotCode

angled :: DotCode -> DotCode

fslash :: DotCode

printColorScheme :: Bool -> ColorScheme -> DotCode