Data.GraphViz.Types.Monadic

Description

This module is based upon the dotgen library by Andy Gill: http://hackage.haskell.org/package/dotgen

It provides a monadic interface for constructing generalised Dot graphs. Note that this does not have an instance for DotRepr (e.g. what would be the point of the fromCanonical function, as you can't do anything with the result): it is purely for construction purposes. Use the generalised Dot graph instance for printing, etc.

Note that the generalised Dot graph types are not re-exported, in case it causes a clash with other modules you may choose to import.

The example graph in Data.GraphViz.Types can be written as:

digraph (Str "G") $ do

   cluster (Int 0) $ do
       graphAttrs [style filled, color LightGray]
       nodeAttrs [style filled, color White]
       "a0" --> "a1"
       "a1" --> "a2"
       "a2" --> "a3"
       graphAttrs [textLabel "process #1"]

   cluster (Int 1) $ do
       nodeAttrs [style filled]
       "b0" --> "b1"
       "b1" --> "b2"
       "b2" --> "b3"
       graphAttrs [textLabel "process #2", color Blue]

   "start" --> "a0"
   "start" --> "b0"
   "a1" --> "b3"
   "b2" --> "a3"
   "a3" --> "end"
   "b3" --> "end"

   node "start" [shape MDiamond]
   node "end" [shape MSquare]

Documentation

type Dot n = DotM n ()

The monadic representation of a Dot graph.

data DotM n a

The actual monad; as with Dot but allows you to return a value within the do-block. The actual implementation is based upon the Writer monad.

data GraphID

A polymorphic type that covers all possible ID values allowed by Dot syntax. Note that whilst the ParseDot and PrintDot instances for String will properly take care of the special cases for numbers, they are treated differently here.

Constructors

Str Text

Num Number

Creating a generalised DotGraph.

digraph :: GraphID -> DotM n a -> DotGraph n

Create a directed dot graph with the specified graph ID.

digraph' :: DotM n a -> DotGraph n

Create a directed dot graph with no graph ID.

graph :: GraphID -> DotM n a -> DotGraph n

Create a undirected dot graph with the specified graph ID.

graph' :: DotM n a -> DotGraph n

Create a undirected dot graph with no graph ID.

Adding global attributes.

graphAttrs :: Attributes -> Dot n

Add graph/sub-graph/cluster attributes.

nodeAttrs :: Attributes -> Dot n

Add global node attributes.

edgeAttrs :: Attributes -> Dot n

Add global edge attributes

Adding items to the graph.

Subgraphs and clusters

subgraph :: GraphID -> DotM n a -> Dot n

Add a named subgraph to the graph.

anonSubgraph :: DotM n a -> Dot n

Add an anonymous subgraph to the graph.

It is highly recommended you use subgraph instead.

cluster :: GraphID -> DotM n a -> Dot n

Add a named cluster to the graph.

Nodes

node :: n -> Attributes -> Dot n

Add a node to the graph.

node' :: n -> Dot n

Add a node with no attributes to the graph.

Edges

If you wish to use something analogous to Dot's ability to write multiple edges with in-line subgraphs such as:

{a b c} -> {d e f}

Then you can use --> and <-> in combination with monadic traversal functions such as traverse_, for_, mapM_, forM_ and zipWithM_; for example:

("a" -->) `traverse_` ["d", "e", "f"]
["a", "b", "c"] `for_` (--> "d")
zipWithM_ (-->) ["a", "b", "c"] ["d", "e", "f"]

edge :: n -> n -> Attributes -> Dot n

Add an edge to the graph.

(-->) :: n -> n -> Dot n infixr 9

Add an edge with no attributes.

(<->) :: n -> n -> Dot n infixr 9

An alias for --> to make edges look more undirected.