Data.Monoid.Generic

Documentation

genericMappend :: (Generic a, MappendProduct (Rep a)) => a -> a -> a

A generic <> function which works for product types where each contained type is itself a Semigroup. It simply calls <> for each field.

If you don't want to use the deriving via mechanism, use this function to implement the Semigroup type class.

genericMempty :: (Generic a, MemptyProduct (Rep a)) => a

A generic mempty function which works for product types where each contained type is itself a Monoid. It simply calls mempty for each field.

If you don't want to use the deriving via mechanism, use this function to implement the Monoid type class.

newtype GenericSemigroup a

A newtype which allows you to using the DerivingVia extension to reduce boilerplate.

data X = X [Int] String
  deriving (Generic, Show)
  deriving Semigroup via GenericSemigroup X

Constructors

GenericSemigroup a

newtype GenericMonoid a

A newtype which allows you to using the DerivingVia extension to reduce boilerplate.

data X = X [Int] String
  deriving (Generic, Show)
  deriving Semigroup via GenericSemigroup X
  deriving Monoid    via GenericMonoid X

Note: Do NOT attempt to derive Semigroup via GenericMonoid. That will lead to infinite recursion.

Constructors

GenericMonoid a