Data.SOP.Counting

Description

Type-level counting

Intended for unqualified import.

Documentation

data AtMost xs a where

At most one value for each type level index

Constructors

AtMostNil :: AtMost xs a

AtMostCons :: !a -> !(AtMost xs a) -> AtMost (x ': xs) a

newtype Exactly xs a

Constructors

Exactly

Fields

getExactly :: NP (K a) xs

Bundled Patterns

pattern ExactlyNil :: () => xs ~ '[] => Exactly xs a

pattern ExactlyCons :: () => xs' ~ (x ': xs) => a -> Exactly xs a -> Exactly xs' a

Working with Exactly

exactlyHead :: Exactly (x ': xs) a -> a

Analogue of head

exactlyOne :: a -> Exactly '[x] a

Singleton

exactlyReplicate :: forall a r. Word -> a -> (forall xs. Exactly xs a -> r) -> r

Analogue of replicate

In CPS style because the xs type parameter is not statically known.

exactlyTail :: Exactly (x ': xs) a -> Exactly xs a

Analogue of tail

exactlyTwo :: a -> a -> Exactly '[x, y] a

From a pair

exactlyWeaken :: Exactly xs a -> AtMost xs a

exactlyWeakenNonEmpty :: Exactly (x ': xs) a -> NonEmpty (x ': xs) a

exactlyZip :: Exactly xs a -> Exactly xs b -> Exactly xs (a, b)

Analogue of zip

exactlyZipFoldable :: Foldable t => Exactly xs a -> t b -> AtMost xs (a, b)

Analogue of zip where the length of second argument is unknown

Working with AtMost

atMostFromNonEmpty :: NonEmpty xs a -> AtMost xs a

atMostHead :: AtMost xs a -> Maybe a

Analogue of head

atMostInit :: AtMost xs a -> Maybe (AtMost xs a, a)

Analogue of init

For simplicity we don't shrink the type-level index.

atMostLast :: AtMost xs a -> Maybe a

Analogue of last

atMostNonEmpty :: AtMost (x ': xs) a -> Maybe (NonEmpty (x ': xs) a)

atMostOne :: a -> AtMost (x ': xs) a

Singleton

atMostZipFoldable :: Foldable t => AtMost xs a -> t b -> AtMost xs (a, b)