Data.SOP.Lenses

Documentation

data Lens f xs a

Simple lens to access an element of an n-ary product.

Constructors

Lens

Fields

getter :: NP f xs -> f a

setter :: f a -> NP f xs -> NP f xs

lenses_NP :: SListI xs => NP (Lens f xs) xs

Generate all lenses to access the element of an n-ary product.