Data.SOP.Match

Description

Intended for qualified import

import           Data.SOP.Match (Mismatch(..))
import qualified Data.SOP.Match as Match

Documentation

data Mismatch f g xs where

We have a mismatch in the index between two NS

Constructors

ML :: f x -> NS g xs -> Mismatch f g (x ': xs)
The left is at the current x and the right is somewhere in the later xs

MR :: NS f xs -> g x -> Mismatch f g (x ': xs)
The right is at the current x and the left is somewhere in the later xs

MS :: Mismatch f g xs -> Mismatch f g (x ': xs)
There is a mismatch later on in the xs

flipMatch :: Mismatch f g xs -> Mismatch g f xs

matchNS :: NS f xs -> NS g xs -> Either (Mismatch f g xs) (NS (Product f g) xs)

matchTelescope :: NS h xs -> Telescope g f xs -> Either (Mismatch h f xs) (Telescope g (Product h f) xs)

telescopesMismatch :: Telescope a b xs -> Telescope g f xs -> Maybe (Mismatch b f xs)

Utilities

mismatchNotEmpty :: Mismatch f g xs -> (forall x xs'. xs ~ (x ': xs') => Mismatch f g (x ': xs') -> a) -> a

mismatchNotFirst :: Mismatch f g (x ': xs) -> Either (NS f xs) (NS g xs)

mismatchOne :: Mismatch f g '[x] -> Void

We cannot give a mismatch if we have only one type variable

mismatchToNS :: Mismatch f g xs -> (NS f xs, NS g xs)

Project two NS from a Mismatch

We should have the property that

uncurry matchNS (mismatchToNS m) == Left m

mismatchTwo :: Mismatch f g '[x, y] -> Either (f x, g y) (f y, g x)

If we only have two eras, only two possibilities for a mismatch

mkMismatchTwo :: Either (f x, g y) (f y, g x) -> Mismatch f g '[x, y]

mustMatchNS :: forall f g xs. HasCallStack => String -> NS f xs -> NS g xs -> NS (Product f g) xs

Variant of matchNS for when we know the two NSs must match. Otherwise an error, mentioning the given String, is thrown.

SOP operators

bihap :: NP (f -.-> f') xs -> NP (g -.-> g') xs -> Mismatch f g xs -> Mismatch f' g' xs

bihcmap :: All c xs => proxy c -> (forall x. c x => f x -> f' x) -> (forall x. c x => g x -> g' x) -> Mismatch f g xs -> Mismatch f' g' xs

Bifunctor analogue of hcmap

bihmap :: SListI xs => (forall x. f x -> f' x) -> (forall x. g x -> g' x) -> Mismatch f g xs -> Mismatch f' g' xs