Data.SOP.NonEmpty

Description

Type-level non-empty lists

Documentation

data NonEmpty xs a where

Non-empty variation on AtMost

Constructors

NonEmptyOne :: !a -> NonEmpty (x ': xs) a

NonEmptyCons :: !a -> !(NonEmpty xs a) -> NonEmpty (x ': xs) a

Proofs

class IsNonEmpty xs where

Methods

isNonEmpty :: proxy xs -> ProofNonEmpty xs

data ProofNonEmpty xs where

Constructors

ProofNonEmpty :: Proxy x -> Proxy xs -> ProofNonEmpty (x ': xs)

checkIsNonEmpty :: forall xs. SListI xs => Proxy xs -> Maybe (ProofNonEmpty xs)

Working with NonEmpty

nonEmptyFromList :: forall xs a. SListI xs => [a] -> Maybe (NonEmpty xs a)

Build a NonEmpty from a list. Returns Nothing when the list is empty or when it's longer than xs.

nonEmptyHead :: NonEmpty xs a -> a

Analogue of head

nonEmptyInit :: NonEmpty xs a -> (Maybe (NonEmpty xs a), a)

Analogue of init

nonEmptyLast :: NonEmpty xs a -> a

Analogue of last

nonEmptyMapOne :: forall m xs a. Alternative m => (a -> m a) -> NonEmpty xs a -> m (NonEmpty xs a)

Apply the specified function to exactly one element

nonEmptyMapTwo :: forall m xs a. Alternative m => (a -> m a) -> (a -> a -> m (a, a)) -> NonEmpty xs a -> m (NonEmpty xs a)

Variation on nonEmptyMapOne where we try to apply the function to pairs of elements

nonEmptyStrictPrefixes :: NonEmpty xs a -> [NonEmpty xs a]

A strict prefixes

   nonEmptyStrictPrefixes (fromJust (nonEmptyFromList [1..4]))
== [ NonEmptyOne  1
   , NonEmptyCons 1 $ NonEmptyOne  2
   , NonEmptyCons 1 $ NonEmptyCons 2 $ NonEmptyOne 3
   ]

nonEmptyToList :: forall xs a. NonEmpty xs a -> [a]

Convert a NonEmpty to a list.