Data.SOP.OptNP

Description

NP with optional values

Intended for qualified import

import           Data.SOP.OptNP (OptNP (..), ViewOptNP (..))
import qualified Data.SOP.OptNP as OptNP

Documentation

type NonEmptyOptNP = OptNP 'False

data OptNP empty f xs where

Like an NP, but with optional values

Constructors

OptNil :: OptNP 'True f '[]

OptCons :: !(f x) -> !(OptNP empty f xs) -> OptNP 'False f (x ': xs)

OptSkip :: !(OptNP empty f xs) -> OptNP empty f (x ': xs)

at :: SListI xs => f x -> Index xs x -> NonEmptyOptNP f xs

empty :: forall f xs. SListI xs => OptNP 'True f xs

fromNP :: (forall empty. OptNP empty f xs -> r) -> NP f xs -> r

fromNonEmptyNP :: forall f xs. IsNonEmpty xs => NP f xs -> NonEmptyOptNP f xs

fromSingleton :: NonEmptyOptNP f '[x] -> f x

If OptNP is not empty, it must contain at least one value

singleton :: f x -> NonEmptyOptNP f '[x]

toNP :: OptNP empty f xs -> NP (Maybe :.: f) xs

View

data ViewOptNP f xs where

Constructors

OptNP_ExactlyOne :: f x -> ViewOptNP f '[x]

OptNP_AtLeastTwo :: ViewOptNP f (x ': (y ': zs))

view :: forall f xs. NonEmptyOptNP f xs -> ViewOptNP f xs

Combining

combine :: forall (f :: Type -> Type) xs. (SListI xs, HasCallStack) => Maybe (NonEmptyOptNP f xs) -> Maybe (NonEmptyOptNP f xs) -> Maybe (NonEmptyOptNP f xs)

Precondition: there is no overlap between the two given lists: if there is a Just at a given position in one, it must be Nothing at the same position in the other.

combineWith :: SListI xs => (forall a. These1 f g a -> h a) -> Maybe (NonEmptyOptNP f xs) -> Maybe (NonEmptyOptNP g xs) -> Maybe (NonEmptyOptNP h xs)

zipWith :: forall f g h xs. (forall a. These1 f g a -> h a) -> NonEmptyOptNP f xs -> NonEmptyOptNP g xs -> NonEmptyOptNP h xs