Data.SOP.Telescope

Description

See Telescope

Intended for qualified import

import           Data.SOP.Telescope (Telescope(..))
import qualified Data.SOP.Telescope as Telescope

Telescope

data Telescope g f xs where

Telescope

A telescope is an extension of an NS, where every time we "go right" in the sum we have an additional value.

The Telescope API mostly follows sop-core conventions, supporting functor (hmap, hcmap), applicative (hap, hpure), foldable (hcollapse) and traversable (hsequence'). However, since Telescope is a bi-functor, it cannot reuse the sop-core classes. The naming scheme of the functions is adopted from sop-core though; for example:

bi h (c) zipWith
|  |  |    |
|  |  |    \ zipWith: the name from base
|  |  |
|  |  \ constrained: version of the function with a constraint parameter
|  |
|  \ higher order: 'Telescope' (like 'NS'/'NP') is a /higher order/ functor
|
\ bifunctor: 'Telescope' (unlike 'NS'/'NP') is a higher order /bifunctor/

In addition to the standard SOP operators, the new operators that make a Telescope a telescope are extend, retract and align; see their documentation for details.

Constructors

TZ :: !(f x) -> Telescope g f (x ': xs)

TS :: !(g x) -> !(Telescope g f xs) -> Telescope g f (x ': xs)

sequence :: forall m g f xs. Functor m => Telescope g (m :.: f) xs -> m (Telescope g f xs)

Specialization of hsequence' with weaker constraints (Functor rather than Applicative)

Utilities

fromTZ :: Telescope g f '[x] -> f x

fromTip :: NS f xs -> Telescope (K ()) f xs

tip :: Telescope g f xs -> NS f xs

toAtMost :: Telescope (K a) (K (Maybe a)) xs -> AtMost xs a

Bifunctor analogues of SOP functions

bihap :: NP (g -.-> g') xs -> NP (f -.-> f') xs -> Telescope g f xs -> Telescope g' f' xs

Bifunctor analogue of hap

bihczipWith :: All c xs => proxy c -> (forall x. c x => h x -> g x -> g' x) -> (forall x. c x => h x -> f x -> f' x) -> NP h xs -> Telescope g f xs -> Telescope g' f' xs

Bifunctor equivalent of hczipWith

bihmap :: SListI xs => (forall x. g x -> g' x) -> (forall x. f x -> f' x) -> Telescope g f xs -> Telescope g' f' xs

Bifunctor analogue of hmap

bihzipWith :: SListI xs => (forall x. h x -> g x -> g' x) -> (forall x. h x -> f x -> f' x) -> NP h xs -> Telescope g f xs -> Telescope g' f' xs

Bifunctor equivalent of hzipWith

Extension, retraction, alignment

newtype Extend m g f x y

Constructors

Extend

Fields

extendWith :: f x -> m (g x, f y)

newtype Retract m g f x y

Constructors

Retract

Fields

retractWith :: g x -> f y -> m (f x)

align

Arguments

:: forall m g' g f' f f'' xs. Monad m

=> InPairs (Requiring g' (Extend m g f)) xs
How to extend

-> Tails (Requiring f' (Retract m g f)) xs
How to retract

-> NP (f' -.-> (f -.-> f'')) xs
Function to apply at the tip

-> Telescope g' f' xs
Telescope we are aligning with

-> Telescope g f xs

-> m (Telescope g f'' xs)

Align a telescope with another, then apply a function to the tips

Aligning is a combination of extension and retraction, extending or retracting the telescope as required to match up with the other telescope.

extend

Arguments

:: forall m h g f xs. Monad m

=> InPairs (Requiring h (Extend m g f)) xs
How to extend

-> NP (f -.-> (Maybe :.: h)) xs
Where to extend from

-> Telescope g f xs

-> m (Telescope g f xs)

Extend the telescope

We will not attempt to extend the telescope past its final segment.

retract

Arguments

:: forall m h g f xs. Monad m

=> Tails (Requiring h (Retract m g f)) xs
How to retract

-> NP (g -.-> (Maybe :.: h)) xs
Where to retract to

-> Telescope g f xs

-> m (Telescope g f xs)

Retract a telescope

Simplified API

alignExtend

Arguments

:: (Monad m, HasCallStack)

=> InPairs (Requiring g' (Extend m g f)) xs
How to extend

-> NP (f' -.-> (f -.-> f'')) xs
Function to apply at the tip

-> Telescope g' f' xs
Telescope we are aligning with

-> Telescope g f xs

-> m (Telescope g f'' xs)

Version of align that never retracts, only extends

PRE: The telescope we are aligning with cannot be behind us.

alignExtendNS

Arguments

:: (Monad m, HasCallStack)

=> InPairs (Extend m g f) xs
How to extend

-> NP (f' -.-> (f -.-> f'')) xs
Function to apply at the tip

-> NS f' xs
NS we are aligning with

-> Telescope g f xs

-> m (Telescope g f'' xs)

Version of alignExtend that extends with an NS instead

extendIf

Arguments

:: Monad m

=> InPairs (Extend m g f) xs
How to extend

-> NP (f -.-> K Bool) xs
Where to extend from

-> Telescope g f xs

-> m (Telescope g f xs)

Version of extend where the evidence is a simple Bool

retractIf

Arguments

:: Monad m

=> Tails (Retract m g f) xs
How to retract

-> NP (g -.-> K Bool) xs
Where to retract to

-> Telescope g f xs

-> m (Telescope g f xs)

Version of retract where the evidence is a simple Bool

Additional API

newtype ScanNext h g x y

Constructors

ScanNext

Fields

getNext :: h x -> g x -> h y

newtype SimpleTelescope f xs

Telescope with both functors set to the same f

Constructors

SimpleTelescope

Fields

getSimpleTelescope :: Telescope f f xs

scanl :: InPairs (ScanNext h g) (x ': xs) -> h x -> Telescope g f (x ': xs) -> Telescope (Product h g) (Product h f) (x ': xs)

Telescope analogue of scanl on lists

This function is modelled on

scanl :: (b -> a -> b) -> b -> [a] -> [b]

but there are a few differences:

Since every seed has a different type, we must be given a function for each transition.

Unlike scanl, we preserve the length of the telescope (scanl prepends the initial seed)

Instead of generating a telescope containing only the seeds, we instead pair the seeds with the elements.