Data.Yaml

Description

Provides a high-level interface for processing YAML files.

This module reuses most of the infrastructure from the aeson package. This means that you can use all of the existing tools for JSON processing for processing YAML files. As a result, much of the documentation below mentions JSON; do not let that confuse you, it's intentional.

For the most part, YAML content translates directly into JSON, and therefore there is very little data loss. If you need to deal with YAML more directly (e.g., directly deal with aliases), you should use the Text.Libyaml module instead.

For documentation on the aeson types, functions, classes, and operators, please see the Data.Aeson module of the aeson package.

Look in the examples directory of the source repository for some initial pointers on how to use this library.

Encoding

encode :: ToJSON a => a -> ByteString

Encode a value into its YAML representation.

encodeWith :: ToJSON a => EncodeOptions -> a -> ByteString

Encode a value into its YAML representation with custom styling.

Since: 0.10.2.0

encodeFile :: ToJSON a => FilePath -> a -> IO ()

Encode a value into its YAML representation and save to the given file.

encodeFileWith :: ToJSON a => EncodeOptions -> FilePath -> a -> IO ()

Encode a value into its YAML representation with custom styling and save to the given file.

Since: 0.10.2.0

Decoding

decodeEither' :: FromJSON a => ByteString -> Either ParseException a

More helpful version of decodeEither which returns the YamlException.

Since: 0.8.3

decodeFileEither :: FromJSON a => FilePath -> IO (Either ParseException a)

A version of decodeFile which should not throw runtime exceptions.

Since: 0.8.4

decodeFileWithWarnings :: FromJSON a => FilePath -> IO (Either ParseException ([Warning], a))

A version of decodeFileEither that returns warnings along with the parse result.

Since: 0.10.0

decodeThrow :: (MonadThrow m, FromJSON a) => ByteString -> m a

A version of decodeEither' lifted to MonadThrow

Since: 0.8.31

decodeFileThrow :: (MonadIO m, FromJSON a) => FilePath -> m a

A version of decodeFileEither lifted to MonadIO

Since: 0.8.31

Decoding multiple documents

For situations where we need to be able to parse multiple documents separated by `---` in a YAML stream, these functions decode a list of values rather than a single value.

decodeAllEither' :: FromJSON a => ByteString -> Either ParseException [a]

Like decodeEither', but decode multiple documents.

Since: 0.11.5.0

decodeAllFileEither :: FromJSON a => FilePath -> IO (Either ParseException [a])

Like decodeFileEither, but decode multiple documents.

Since: 0.11.5.0

decodeAllFileWithWarnings :: FromJSON a => FilePath -> IO (Either ParseException ([Warning], [a]))

Like decodeFileWithWarnings, but decode multiple documents.

Since: 0.11.5.0

decodeAllThrow :: (MonadThrow m, FromJSON a) => ByteString -> m [a]

Like decodeThrow, but decode multiple documents.

Since: 0.11.5.0

decodeAllFileThrow :: (MonadIO m, FromJSON a) => FilePath -> m [a]

Like decodeFileThrow, but decode multiple documents.

Since: 0.11.5.0

More control over decoding

decodeHelper :: FromJSON a => ConduitM () Event Parse () -> IO (Either ParseException ([Warning], Either String a))

Types

data Value

Constructors

Object !Object

Array !Array

String !Text

Number !Scientific

Bool !Bool

Null

data Parser a

type Object = KeyMap Value

type Array = Vector Value

data ParseException

Constructors

NonScalarKey

UnknownAlias

Fields

_anchorName :: AnchorName

UnexpectedEvent

Fields

_received :: Maybe Event

_expected :: Maybe Event

InvalidYaml (Maybe YamlException)

MultipleDocuments

AesonException String

OtherParseException SomeException

NonStringKey JSONPath

NonStringKeyAlias AnchorName Value

CyclicIncludes

LoadSettingsException FilePath ParseException

prettyPrintParseException :: ParseException -> String

Alternative to show to display a ParseException on the screen. Instead of displaying the data constructors applied to their arguments, a more textual output is returned. For example, instead of printing:

InvalidYaml (Just (YamlParseException {yamlProblem = "did not find expected ',' or '}'", yamlContext = "while parsing a flow mapping", yamlProblemMark = YamlMark {yamlIndex = 42, yamlLine = 2, yamlColumn = 12}})))

It looks more pleasant to print:

YAML parse exception at line 2, column 12,
while parsing a flow mapping:
did not find expected ',' or '}'

Since 0.8.11

data YamlException

Constructors

YamlException String

YamlParseException

Fields

yamlProblem :: String

yamlContext :: String

yamlProblemMark :: YamlMark

data YamlMark

Constructors

YamlMark

Fields

yamlIndex :: Int

yamlLine :: Int

yamlColumn :: Int

Constructors and accessors

object :: [Pair] -> Value

array :: [Value] -> Value

Construct a new Value from a list of Values.

(.=) :: (KeyValue e kv, ToJSON v) => Key -> v -> kv

(.:) :: FromJSON a => Object -> Key -> Parser a

(.:?) :: FromJSON a => Object -> Key -> Parser (Maybe a)

(.!=) :: Parser (Maybe a) -> a -> Parser a

With helpers (since 0.8.23)

withObject :: String -> (Object -> Parser a) -> Value -> Parser a

withText :: String -> (Text -> Parser a) -> Value -> Parser a

withArray :: String -> (Array -> Parser a) -> Value -> Parser a

withScientific :: String -> (Scientific -> Parser a) -> Value -> Parser a

withBool :: String -> (Bool -> Parser a) -> Value -> Parser a

Parsing

parseMonad :: MonadFail m => (a -> Parser b) -> a -> m b

Deprecated: With the MonadFail split, this function is going to be removed in the future. Please migrate to parseEither.

parseEither :: (a -> Parser b) -> a -> Either String b

parseMaybe :: (a -> Parser b) -> a -> Maybe b

Classes

class ToJSON a where

Minimal complete definition

Nothing

Methods

toJSON :: a -> Value

toEncoding :: a -> Encoding

toJSONList :: [a] -> Value

toEncodingList :: [a] -> Encoding

omitField :: a -> Bool

class FromJSON a where

Minimal complete definition

Nothing

Methods

parseJSON :: Value -> Parser a

parseJSONList :: Value -> Parser [a]

omittedField :: Maybe a

Custom encoding

isSpecialString :: Text -> Bool

Determine whether a string must be quoted in YAML and can't appear as plain text. Useful if you want to use setStringStyle.

Since: 0.10.2.0

data EncodeOptions

Since: 0.10.2.0

defaultEncodeOptions :: EncodeOptions

Since: 0.10.2.0

defaultStringStyle :: StringStyle

Since: 0.11.2.0

setStringStyle :: (Text -> (Tag, Style)) -> EncodeOptions -> EncodeOptions

Set the string style in the encoded YAML. This is a function that decides for each string the type of YAML string to output.

WARNING: You must ensure that special strings (like "yes"/"no"/"null"/"1234") are not encoded with the Plain style, because then they will be decoded as boolean, null or numeric values. You can use isSpecialString to detect them.

By default, strings are encoded as follows:

Any string containing a newline character uses the Literal style

Otherwise, any special string (see isSpecialString) uses SingleQuoted

Otherwise, use Plain

Since: 0.10.2.0

setFormat :: FormatOptions -> EncodeOptions -> EncodeOptions

Set the encoding formatting for the encoded YAML. By default, this is defaultFormatOptions.

Since: 0.10.2.0

data FormatOptions

defaultFormatOptions :: FormatOptions

setWidth :: Maybe Int -> FormatOptions -> FormatOptions

Deprecated

decode :: FromJSON a => ByteString -> Maybe a

Deprecated: Please use decodeEither or decodeThrow, which provide information on how the decode failed

decodeFile :: FromJSON a => FilePath -> IO (Maybe a)

Deprecated: Please use decodeFileEither, which does not confused type-directed and runtime exceptions.

decodeEither :: FromJSON a => ByteString -> Either String a

Deprecated: Please use decodeEither' or decodeThrow, which provide more useful failures