Database.Sqlite.Easy.Internal

Description

The implementation of sqlite-easy.

This module is unstable and may change at any time.

Connection

newtype ConnectionString

A SQLite3 connection string

Constructors

ConnectionString

Fields

unConnectionString :: Text

createSqlitePool :: ConnectionString -> IO (Pool Database)

Create a pool of a sqlite3 db with a specific connection string. This also sets a few default pragmas.

createSqlitePoolWith :: ConnectionString -> [SQL] -> IO (Pool Database)

Create a pool of a sqlite3 db with a specific connection string. This will also run the supplied SQL statements after establishing each connection.

openWith :: ConnectionString -> [SQL] -> IO Database

Open a connection to a database, run the supplied statements, and return the connection.

defaultPragmas :: [SQL]

Default pragmas to be set when opening a connection from a pool.

withDb :: ConnectionString -> SQLite a -> IO a

Open a database, run some stuff, close the database.

withDatabase :: Database -> SQLite a -> IO a

Use an active database connection to run some stuff on a database.

withPool :: Pool Database -> SQLite a -> IO a

Use a resource pool to run some stuff on a database.

Execution

newtype SQL

A SQL statement

Constructors

SQL

Fields

unSQL :: Text

run :: SQL -> SQLite [[SQLData]]

Run a SQL statement on a database and fetch the results.

runWith :: SQL -> [SQLData] -> SQLite [[SQLData]]

Run a SQL statement with certain parameters on a database and fetch the results.

runWithMany :: SQL -> [[SQLData]] -> SQLite [[[SQLData]]]

Run a SQL statement binding it to all given parameter rows and fetch all of the data. This can significantly improve speed for cases like bulk INSERTs.

Since: 1.1.1.0

fetchAll :: Statement -> IO [[SQLData]]

Run a statement and fetch all of the data.

Transaction

newtype SQLite a

The type of actions to run on a SQLite database. In essence, it is almost the same as Database -> IO a.

SQLite actions can be created with the run and runWith functions, and can be composed using the type class instances.

SQLite actions can be run with the withDb, withDatabase, and withPool functions.

Constructors

SQLite

Fields

unSQLite :: SQLiteStuff -> IO a

data SQLiteStuff

Constructors

SQLiteStuff

Fields

dbConn :: Database

transactionNumber :: Maybe Int

getDB :: SQLite Database

runSQLite :: Database -> SQLite a -> IO a

transaction :: forall a. Typeable a => SQLite a -> SQLite a

Run operations as a transaction. If the action throws an error, the transaction is rolled back. For more information, visit: https://www.sqlite.org/lang_transaction.html

asTransaction' :: Database -> IO a -> IO a

rollback :: Typeable a => a -> SQLite a

Rollback the current (inner-most) transaction by supplying the return value. To be used inside transactions.

rollbackAll :: Typeable a => a -> SQLite a

Rollback all transaction structure by supplying the return value. To be used inside transactions.

data RollbackCurrent a

Constructors

RollbackCurrent a

data RollbackAll a

Constructors

RollbackAll a