Formatting.Formatters

Description

Formatting functions.

Text/string types

text :: Format r (Text -> r)

Output a lazy text.

stext :: Format r (Text -> r)

Output a strict text.

string :: Format r (String -> r)

Output a string.

shown :: Show a => Format r (a -> r)

Output a showable value (instance of Show) by turning it into Text:

>>> format ("Value number " % shown % " is " % shown % ".") 42 False
"Value number 42 is False."

char :: Format r (Char -> r)

Output a character.

builder :: Format r (Builder -> r)

Build a builder.

fconst :: Builder -> Format r (a -> r)

Like const but for formatters.

Numbers

int :: Integral a => Format r (a -> r)

Render an integral e.g. 123 -> "123", 0 -> "0".

float :: Real a => Format r (a -> r)

Render some floating point with the usual notation, e.g. 123.32 => "123.32"

fixed :: Real a => Int -> Format r (a -> r)

Render a floating point number using normal notation, with the given number of decimal places.

sci :: Format r (Scientific -> r)

Render a scientific number.

scifmt :: FPFormat -> Maybe Int -> Format r (Scientific -> r)

Render a scientific number with options.

shortest :: Real a => Format r (a -> r)

Render a floating point number using the smallest number of digits that correctly represent it. Note that in the case of whole numbers it will still add one decimal place, e.g. "1.0".

groupInt :: (Buildable n, Integral n) => Int -> Char -> Format r (n -> r)

Group integral numbers, e.g. groupInt 2 . on 123456 -> "12.34.56".

commas :: (Buildable n, Integral n) => Format r (n -> r)

Add commas to an integral, e.g 12000 -> "12,000".

ords :: Integral n => Format r (n -> r)

Add a suffix to an integral, e.g. 1st, 2nd, 3rd, 21st.

plural :: (Num a, Eq a) => Text -> Text -> Format r (a -> r)

English plural suffix for an integral.

For example:

>>> :set -XOverloadedStrings
>>> formatPeople = format (int % " " <> plural "person" "people" % ".") :: Int -> Data.Text.Lazy.Text
>>> formatPeople 1
"1 person."
>>> formatPeople 3
"3 people."

asInt :: Enum a => Format r (a -> r)

Shows the Int value of Enum instances using fromEnum.

>>> format ("Got: " % char % " (" % asInt % ")") 'a' 'a'
"Got: a (97)"

Padding

left :: Buildable a => Int -> Char -> Format r (a -> r)

Pad the left hand side of a string until it reaches k characters wide, if necessary filling with character c.

right :: Buildable a => Int -> Char -> Format r (a -> r)

Pad the right hand side of a string until it reaches k characters wide, if necessary filling with character c.

center :: Buildable a => Int -> Char -> Format r (a -> r)

Pad the left & right hand side of a string until it reaches k characters wide, if necessary filling with character c.

fitLeft :: Buildable a => Int -> Format r (a -> r)

Fit in the given length, truncating on the left.

fitRight :: Buildable a => Int -> Format r (a -> r)

Fit in the given length, truncating on the right.

Bases

base :: Integral a => Int -> Format r (a -> r)

Render an integral at base n.

bin :: Integral a => Format r (a -> r)

Render an integer using binary notation. (No leading 0b is added.) Defined as bin = base 2.

oct :: Integral a => Format r (a -> r)

Render an integer using octal notation. (No leading 0o is added.) Defined as oct = base 8.

hex :: Integral a => Format r (a -> r)

Render an integer using hexadecimal notation. (No leading 0x is added.) Has a specialized implementation.

prefixBin :: Integral a => Format r (a -> r)

Render an integer using binary notation with a leading 0b.

See also binPrefix for fixed-width formatting.

prefixOct :: Integral a => Format r (a -> r)

Render an integer using octal notation with a leading 0o.

See also octPrefix for fixed-width formatting.

prefixHex :: Integral a => Format r (a -> r)

Render an integer using hexadecimal notation with a leading 0x.

See also hexPrefix for fixed-width formatting.

bytes

Arguments

:: (Ord f, Integral a, Fractional f)

=> Format Builder (f -> Builder)
formatter for the decimal part

-> Format r (a -> r)

Renders a given byte count using an appropiate decimal binary suffix:

>>> format (bytes shortest) 1024
"1KB"

>>> format (bytes (fixed 2 % " ")) (1024*1024*5)
"5.00 MB"

Buildables

build :: Buildable a => Format r (a -> r)

Build anything that implements the Buildable class.

class Buildable p

The class of types that can be rendered to a Builder.

Minimal complete definition

build