Formatting.Internal.Raw

Description

Reexports of things that were previously in the text-format package.

Format control

left :: Buildable a => Int -> Char -> a -> Builder

Pad the left hand side of a string until it reaches k characters wide, if necessary filling with character c.

right :: Buildable a => Int -> Char -> a -> Builder

Pad the right hand side of a string until it reaches k characters wide, if necessary filling with character c.

Integers

hex :: Integral a => a -> Builder

Render an integer using hexadecimal notation. (No leading "0x" is added.)

Floating point numbers

fixed

Arguments

:: Real a

=> Int
Number of digits of precision after the decimal.

-> a

-> Builder

Render a floating point number using normal notation, with the given number of decimal places.

shortest :: Real a => a -> Builder

Render a floating point number using the smallest number of digits that correctly represent it.

(<>) :: Builder -> Builder -> Builder infixr 4

The normal mappend function with right associativity instead of left.

i2d :: Int -> Char

Unsafe conversion for decimal digits.

module Data.Text.Lazy.Builder.Int

newtype Shown a

Render a value using its Show instance.

Constructors

Shown

Fields

shown :: a

Integer format control

newtype Hex a

Render an integral type in hexadecimal.

Constructors

Hex a