Formatting.ShortFormatters

Description

Single letters for short formatting.

Documentation

t :: Format r (Text -> r)

Output a lazy text.

d :: Integral a => Format r (a -> r)

Render an integral e.g. 123 -> "123", 0 -> "0".

b :: Integral a => Format r (a -> r)

Render an integer using binary notation. (No leading 0b is added.)

o :: Integral a => Format r (a -> r)

Render an integer using octal notation. (No leading 0o is added.)

x :: Integral a => Format r (a -> r)

Render an integer using hexadecimal notation. (No leading 0x is added.)

st :: Format r (Text -> r)

Output a strict text.

s :: Format r (String -> r)

Output a string.

sh :: Show a => Format r (a -> r)

Output a showable value (instance of Show) by turning it into Text.

c :: Format r (Char -> r)

Output a character.

f :: Real a => Int -> Format r (a -> r)

Render a floating point number using normal notation, with the given number of decimal places.

sf :: Real a => Format r (a -> r)

Render a floating point number using the smallest number of digits that correctly represent it.

l :: Buildable a => Int -> Char -> Format r (a -> r)

Pad the left hand side of a string until it reaches k characters wide, if necessary filling with character ch.

r :: Buildable a => Int -> Char -> Format r (a -> r)

Pad the right hand side of a string until it reaches k characters wide, if necessary filling with character ch.