Generic.Data.Internal.Compat

Description

Shim for backwards compatibility.

Warning

This is an internal module: it is not subject to any versioning policy, breaking changes can happen at any time.

If something here seems useful, please report it or create a pull request to export it from an external module.

Documentation

readPrec1 :: (Read1 f, Read a) => ReadPrec (f a)

Lift the standard readPrec and readListPrec functions through the type constructor.

Since: base-4.10.0.0

type family Div (a :: Natural) (b :: Natural) :: Natural where ... infixl 7

Division (round down) of natural numbers. Div x 0 is undefined (i.e., it cannot be reduced).

Since: base-4.11.0.0