Generic.Data.Internal.Data

Description

Generic representations as data types.

Warning

This is an internal module: it is not subject to any versioning policy, breaking changes can happen at any time.

If something here seems useful, please report it or create a pull request to export it from an external module.

Documentation

newtype Data r p

Synthetic data type.

A wrapper to view a generic Rep as the datatype it's supposed to represent, without needing a declaration.

Constructors

Data

Fields

unData :: r p

toData :: Generic a => a -> Data (Rep a) p

Conversion between a generic type and the synthetic type made using its representation. Inverse of fromData.

fromData :: Generic a => Data (Rep a) p -> a

Inverse of toData.