Generic.Data.Internal.Functions

Description

Type-level functions on generic representations.

Warning

This is an internal module: it is not subject to any versioning policy, breaking changes can happen at any time.

If something here seems useful, please report it or create a pull request to export it from an external module.

Documentation

type family NConstructors (r :: k -> Type) :: Nat

Number of constructors of a data type.

nconstructors :: forall r. KnownNat (NConstructors r) => Integer

Number of constructors of a data type.

type family NFields (r :: k -> Type) :: Nat

Arity of a constructor.

nfields :: forall r. KnownNat (NFields r) => Integer

Arity of a constructor.