Generic.Data.Internal.Generically

Description

Newtypes with instances implemented using generic combinators.

Warning

This is an internal module: it is not subject to any versioning policy, breaking changes can happen at any time.

If something here seems useful, please report it or create a pull request to export it from an external module.

Documentation

newtype Generically a

A datatype whose instances are defined generically, using the Generic representation. Generically1 is a higher-kinded version of Generically that uses Generic1.

Generic instances can be derived via Generically A using -XDerivingVia.

{-# LANGUAGE DeriveGeneric      #-}
{-# LANGUAGE DerivingStrategies #-}
{-# LANGUAGE DerivingVia        #-}

import GHC.Generics (Generic)

data V4 a = V4 a a a a
  deriving stock Generic

  deriving (Semigroup, Monoid)
  via Generically (V4 a)

This corresponds to Semigroup and Monoid instances defined by pointwise lifting:

instance Semigroup a => Semigroup (V4 a) where
  (<>) :: V4 a -> V4 a -> V4 a
  V4 a1 b1 c1 d1 <> V4 a2 b2 c2 d2 =
    V4 (a1 <> a2) (b1 <> b2) (c1 <> c2) (d1 <> d2)

instance Monoid a => Monoid (V4 a) where
  mempty :: V4 a
  mempty = V4 mempty mempty mempty mempty

Historically this required modifying the type class to include generic method definitions (-XDefaultSignatures) and deriving it with the anyclass strategy (-XDeriveAnyClass). Having a /via type/ like Generically decouples the instance from the type class.

Since: base-4.17.0.0

Constructors

Generically a

newtype Generically1 (f :: k -> Type) (a :: k) where

A type whose instances are defined generically, using the Generic1 representation. Generically1 is a higher-kinded version of Generically that uses Generic.

Generic instances can be derived for type constructors via Generically1 F using -XDerivingVia.

{-# LANGUAGE DeriveGeneric      #-}
{-# LANGUAGE DerivingStrategies #-}
{-# LANGUAGE DerivingVia        #-}

import GHC.Generics (Generic)

data V4 a = V4 a a a a
  deriving stock (Functor, Generic1)

  deriving Applicative
  via Generically1 V4

This corresponds to Applicative instances defined by pointwise lifting:

instance Applicative V4 where
  pure :: a -> V4 a
  pure a = V4 a a a a

  liftA2 :: (a -> b -> c) -> (V4 a -> V4 b -> V4 c)
  liftA2 (·) (V4 a1 b1 c1 d1) (V4 a2 b2 c2 d2) =
    V4 (a1 · a2) (b1 · b2) (c1 · c2) (d1 · d2)

Historically this required modifying the type class to include generic method definitions (-XDefaultSignatures) and deriving it with the anyclass strategy (-XDeriveAnyClass). Having a /via type/ like Generically1 decouples the instance from the type class.

Since: base-4.17.0.0

Constructors

Generically1 :: forall {k} (f :: k -> Type) (a :: k). f a -> Generically1 f a

newtype FiniteEnumeration a

Type with Enum instance derived via Generic with FiniteEnum option. This allows deriving Enum for types whose constructors have fields.

Some caution is advised; see details in FiniteEnum.

Example

Expand

>>> :{
data Booool = Booool Bool Bool
  deriving Generic
  deriving (Enum, Bounded) via (FiniteEnumeration Booool)
:}

Constructors

FiniteEnumeration a

newtype GenericProduct a

Product type with generic instances of Semigroup and Monoid.

This is similar to Generically in most cases, but GenericProduct also works for types T with deriving via GenericProduct U, where U is a generic product type coercible to, but distinct from T. In particular, U may not have an instance of Semigroup, which Generically requires.

Example

Expand

>>> import Data.Monoid (Sum(..))
>>> data Point a = Point a a deriving Generic
>>> :{
  newtype Vector a = Vector (Point a)
    deriving (Semigroup, Monoid)
      via GenericProduct (Point (Sum a))
:}

If it were via Generically (Point (Sum a)) instead, then Vector's mappend (the Monoid method) would be defined as Point's (<>) (the Semigroup method), which might not exist, or might not be equivalent to Vector's generic Semigroup instance, which would be unlawful.

Constructors

GenericProduct a

Orphan instances

Generic1 f => Generic1 (Generically1 f :: Type -> Type)
This is a hack to implicitly wrap/unwrap in the instances of Generically1.

Instance details

Associated Types

type Rep1 (Generically1 f) :: k -> Type

Methods

from1 :: forall (a :: k). Generically1 f a -> Rep1 (Generically1 f) a

to1 :: forall (a :: k). Rep1 (Generically1 f) a -> Generically1 f a

(Generic a, GBounded (Rep a)) => Bounded (Generically a)

Instance details

Methods

minBound :: Generically a

maxBound :: Generically a

(Generic a, GEnum StandardEnum (Rep a)) => Enum (Generically a)

Instance details

Methods

succ :: Generically a -> Generically a

pred :: Generically a -> Generically a

toEnum :: Int -> Generically a

fromEnum :: Generically a -> Int

enumFrom :: Generically a -> [Generically a]

enumFromThen :: Generically a -> Generically a -> [Generically a]

enumFromTo :: Generically a -> Generically a -> [Generically a]

enumFromThenTo :: Generically a -> Generically a -> Generically a -> [Generically a]

Generic a => Generic (Generically a)
This is a hack to implicitly wrap/unwrap in the instances of Generically.

Instance details

Associated Types

type Rep (Generically a) :: Type -> Type

Methods

from :: Generically a -> Rep (Generically a) x

to :: Rep (Generically a) x -> Generically a

(Generic a, Ord (Rep a ()), GIx (Rep a)) => Ix (Generically a)

Instance details

Methods

range :: (Generically a, Generically a) -> [Generically a]

index :: (Generically a, Generically a) -> Generically a -> Int

unsafeIndex :: (Generically a, Generically a) -> Generically a -> Int

inRange :: (Generically a, Generically a) -> Generically a -> Bool

rangeSize :: (Generically a, Generically a) -> Int

unsafeRangeSize :: (Generically a, Generically a) -> Int

(Generic a, GRead0 (Rep a)) => Read (Generically a)

Instance details

Methods

readsPrec :: Int -> ReadS (Generically a)

readList :: ReadS [Generically a]

readPrec :: ReadPrec (Generically a)

readListPrec :: ReadPrec [Generically a]

(Generic a, GShow0 (Rep a)) => Show (Generically a)

Instance details

Methods

showsPrec :: Int -> Generically a -> ShowS

show :: Generically a -> String

showList :: [Generically a] -> ShowS

(Generic a, Eq (Rep a ())) => Eq (Generically a)

Instance details

Methods

(==) :: Generically a -> Generically a -> Bool

(/=) :: Generically a -> Generically a -> Bool

(Generic a, Ord (Rep a ())) => Ord (Generically a)

Instance details

Methods

compare :: Generically a -> Generically a -> Ordering

(<) :: Generically a -> Generically a -> Bool

(<=) :: Generically a -> Generically a -> Bool

(>) :: Generically a -> Generically a -> Bool

(>=) :: Generically a -> Generically a -> Bool

max :: Generically a -> Generically a -> Generically a

min :: Generically a -> Generically a -> Generically a

(Generic1 f, GFoldable (Rep1 f)) => Foldable (Generically1 f)

Instance details

Methods

fold :: Monoid m => Generically1 f m -> m

foldMap :: Monoid m => (a -> m) -> Generically1 f a -> m

foldMap' :: Monoid m => (a -> m) -> Generically1 f a -> m

foldr :: (a -> b -> b) -> b -> Generically1 f a -> b

foldr' :: (a -> b -> b) -> b -> Generically1 f a -> b

foldl :: (b -> a -> b) -> b -> Generically1 f a -> b

foldl' :: (b -> a -> b) -> b -> Generically1 f a -> b

foldr1 :: (a -> a -> a) -> Generically1 f a -> a

foldl1 :: (a -> a -> a) -> Generically1 f a -> a

toList :: Generically1 f a -> [a]

null :: Generically1 f a -> Bool

length :: Generically1 f a -> Int

elem :: Eq a => a -> Generically1 f a -> Bool

maximum :: Ord a => Generically1 f a -> a

minimum :: Ord a => Generically1 f a -> a

sum :: Num a => Generically1 f a -> a

product :: Num a => Generically1 f a -> a

(Generic1 f, GRead1 (Rep1 f)) => Read1 (Generically1 f)

Instance details

Methods

liftReadsPrec :: (Int -> ReadS a) -> ReadS [a] -> Int -> ReadS (Generically1 f a)

liftReadList :: (Int -> ReadS a) -> ReadS [a] -> ReadS [Generically1 f a]

liftReadPrec :: ReadPrec a -> ReadPrec [a] -> ReadPrec (Generically1 f a)

liftReadListPrec :: ReadPrec a -> ReadPrec [a] -> ReadPrec [Generically1 f a]

(Generic1 f, GShow1 (Rep1 f)) => Show1 (Generically1 f)

Instance details

Methods

liftShowsPrec :: (Int -> a -> ShowS) -> ([a] -> ShowS) -> Int -> Generically1 f a -> ShowS

liftShowList :: (Int -> a -> ShowS) -> ([a] -> ShowS) -> [Generically1 f a] -> ShowS

(Generic1 f, Functor (Rep1 f), GFoldable (Rep1 f), GTraversable (Rep1 f)) => Traversable (Generically1 f)

Instance details

Methods

traverse :: Applicative f0 => (a -> f0 b) -> Generically1 f a -> f0 (Generically1 f b)

sequenceA :: Applicative f0 => Generically1 f (f0 a) -> f0 (Generically1 f a)

mapM :: Monad m => (a -> m b) -> Generically1 f a -> m (Generically1 f b)

sequence :: Monad m => Generically1 f (m a) -> m (Generically1 f a)

Generic (f a) => Generic (Generically1 f a)
This is a hack to implicitly wrap/unwrap in the instances of Generically1.

Instance details

Associated Types

type Rep (Generically1 f a) :: Type -> Type

Methods

from :: Generically1 f a -> Rep (Generically1 f a) x

to :: Rep (Generically1 f a) x -> Generically1 f a

(Generic1 f, GRead1 (Rep1 f), Read a) => Read (Generically1 f a)

Instance details

Methods

readsPrec :: Int -> ReadS (Generically1 f a)

readList :: ReadS [Generically1 f a]

readPrec :: ReadPrec (Generically1 f a)

readListPrec :: ReadPrec [Generically1 f a]

(Generic1 f, GShow1 (Rep1 f), Show a) => Show (Generically1 f a)

Instance details

Methods

showsPrec :: Int -> Generically1 f a -> ShowS

show :: Generically1 f a -> String

showList :: [Generically1 f a] -> ShowS