Generic.Data.Internal.Microsurgery

Description

Surgeries that are just coerce.

Warning

This is an internal module: it is not subject to any versioning policy, breaking changes can happen at any time.

If something here seems useful, please report it or create a pull request to export it from an external module.

Surgery

type Surgery (s :: Type) (a :: Type) = Generically (Surgery' s a)

Apply a microsurgery s to a type a for DerivingVia.

For the Monoid class, see ProductSurgery.

Example

{-# LANGUAGE DerivingVia #-}

-- The constructors must be visible.
import Generic.Data.Microsurgery
  (Surgery, Surgery'(..), Generically(..), Derecordify)

data T = T { unT :: Int }
  deriving Show via (Surgery Derecordify T)

-- T won't be shown as a record:
--   show (T {unT = 3}) == "T 3"

type ProductSurgery (s :: Type) (a :: Type) = GenericProduct (Surgery' s a)

Apply a microsurgery s to a type a for DerivingVia for the Monoid class.

type Surgeries (s :: [Type]) (a :: Type) = Surgery (Cat s) a

Plural of Surgery. Apply a list of microsurgeries.

type ProductSurgeries (s :: [Type]) (a :: Type) = ProductSurgery (Cat s) a

Plural of ProductSurgery. Apply a list of microsurgeries.

newtype Surgery' (s :: Type) (a :: Type)

See Surgery.

Constructors

Surgery'

Fields

unSurgery' :: a

type family GSurgery (s :: Type) (f :: k -> Type) :: k -> Type

Apply a microsurgery represented by a symbol s (declared as a dummy data type) to a generic representation f.

Derecordify

derecordify :: Coercible (GSurgery Derecordify f) f => Data f p -> Data (GSurgery Derecordify f) p

underecordify :: Coercible f (GSurgery Derecordify f) => Data (GSurgery Derecordify f) p -> Data f p

data Derecordify :: Type

Forget that a type was declared using record syntax.

data Foo = Bar { baz :: Zap }

-- becomes --

data Foo = Bar Zap

Concretely, set the last field of MetaCons to False and forget field names.

This is a defunctionalized symbol, applied using GSurgery or Surgery.

type family GDerecordify (f :: k -> Type) :: k -> Type

Type aging ("denewtypify")

typeage :: Coercible (GSurgery Typeage f) f => Data f p -> Data (GSurgery Typeage f) p

untypeage :: Coercible f (GSurgery Typeage f) => Data (GSurgery Typeage f) p -> Data f p

data Typeage :: Type

Forget that a type is a newtype. (The pun is that "aging" a type makes it no longer "new".)

newtype Foo = Bar Baz

-- becomes --

data Foo = Bar Baz

This is a defunctionalized symbol, applied using GSurgery or Surgery.

Renaming

renameFields :: forall rnm f p. Coercible (GSurgery (RenameFields rnm) f) f => Data f p -> Data (GSurgery (RenameFields rnm) f) p

unrenameFields :: forall rnm f p. Coercible (GSurgery (RenameFields rnm) f) f => Data f p -> Data (GSurgery (RenameFields rnm) f) p

renameConstrs :: forall rnm f p. Coercible (GSurgery (RenameConstrs rnm) f) f => Data f p -> Data (GSurgery (RenameConstrs rnm) f) p

unrenameConstrs :: forall rnm f p. Coercible (GSurgery (RenameConstrs rnm) f) f => Data f p -> Data (GSurgery (RenameConstrs rnm) f) p

data RenameFields (rnm :: Type) :: Type

Rename fields using the function rnm given as a parameter.

data Foo = Bar { baz :: Zap }

-- becomes, renaming "baz" to "bag" --

data Foo = Bar { bag :: Zap }

This is a defunctionalized symbol, applied using GSurgery or Surgery.

type family GRenameFields (rnm :: Type) (f :: k -> Type) :: k -> Type

data RenameConstrs (rnm :: Type) :: Type

Rename constructors using the function rnm given as a parameter.

data Foo = Bar { baz :: Zap }

-- becomes, renaming "Bar" to "Car" --

data Foo = Car { baz :: Zap }

This is a defunctionalized symbol, applied using GSurgery or Surgery.

type family GRenameConstrs (rnm :: Type) (f :: k -> Type) :: k -> Type

Defining symbol functions

type family (f :: Type) @@ (s :: Symbol) :: Symbol

f @@ s is the application of a type-level function symbolized by f to a s :: Symbol.

A function FooToBar can be defined as follows:

data FooToBar
type instance FooToBar @@ "foo" = "bar"

data SId

Identity function Symbol -> Symbol.

data SError

Empty function (compile-time error when applied).

data SConst (s :: Symbol)

Constant function.

data SRename (xs :: [(Symbol, Symbol)]) (f :: Type)

Define a function for a fixed set of strings, and fall back to f for the others.

type family SRename' (xs :: [(Symbol, Symbol)]) (f :: Type) (s :: Symbol) where ...

Closed type family for SRename.

Equations

SRename' '[] f s = f @@ s

SRename' ('(s, t) ': _xs) _f s = t

SRename' ('(_r, _t) ': xs) f s = SRename' xs f s

Other

class UnifyRep (f :: k -> Type) (g :: k -> Type)

Unify the "spines" of two generic representations (the "spine" is everything except the field types).

onData :: (UnifyRep r s, UnifyRep s r) => p (Data r x) (Data s y) -> p (Data r x) (Data s y)

onData :: _ => (Data r x -> Data s y) -> (Data r x -> Data s y)  -- possible specialization

Can be used with generic-lens for type-changing field updates with field_ (and possibly other generic optics).

A specialization of the identity function to be used to fix types of functions on Data, unifying the "spines" of input and output generic representations (the "spine" is everything except field types, which may thus change).

data OnFields (f :: Type -> Type) :: Type

Apply a type constructor f to every field type of a generic representation r.

data Color = RGB
  { r :: Int
  , g :: Int
  , b :: Int }

-- becomes --

data Color f = RGB
  { r :: f Int
  , g :: f Int
  , b :: f Int }

This is a defunctionalized symbol, applied using GSurgery or Surgery.

type family GOnFields (f :: Type -> Type) (g :: k -> Type) :: k -> Type

type DOnFields (f :: Type -> Type) (a :: Type) = Data (GSurgery (OnFields f) (Rep a)) ()

Apply a type constructor f to every field type of a type a to make a synthetic type.

data OnField (s :: Symbol) (f :: Type -> Type) :: Type

Apply a type constructor f to the field named s in a generic record r.

data Vec a = Vec
  { len :: Int
  , contents :: [a] }

-- with (OnField "len" Sum) becomes --

data Vec a = Vec
  { len :: Sum Int
  , contents :: [a] }

This is a defunctionalized symbol, applied using GSurgery or Surgery. See also the synonym (%~).

type family GOnField (x :: Symbol) (f :: Type -> Type) (g :: k -> Type) :: k -> Type where ...

Equations

GOnField x f (M1 S ('MetaSel ('Just x) a b c) (K1 i t)) = M1 S ('MetaSel ('Just x) a b c) (K1 i (f t))

GOnField x f (M1 S m r) = M1 S m r

GOnField x f (M1 C m r) = M1 C m (GOnField x f r)

GOnField x f (M1 D m r) = M1 D m (GOnField x f r)

GOnField x f (r :+: s) = GOnField x f r :+: GOnField x f s

GOnField x f (r :*: s) = GOnField x f r :*: GOnField x f s

GOnField x f (K1 i a) = K1 i (f a)

GOnField x f U1 = U1

GOnField x f V1 = V1

type (%~) = OnField infixr 4

Infix name for OnField. To be used with Surgeries or Cat.

Examples

Expand

Transform one Int field into Sum Int for deriving Monoid:

data Vec a = Vec
  { len :: Int
  , contents :: [a] }
  deriving Generic
  deriving (Eq, Show) via Generically (Vec a)
  deriving (Semigroup, Monoid) via ProductSurgeries '["len" %~ Sum] (Vec a)

Wrap unshowable fields in Opaque for deriving Show:

data Unshowable = Unshowable
  { fun :: Int -> Int
  , io :: IO Bool
  , int :: Int }
  deriving Generic
  deriving Show via Surgeries '["fun" %~ Opaque, "io" %~ Opaque] Unshowable

-- show (Unshowable id (pure True) 42) = "Unshowable _ _ 42"

data Cat (ss :: [Type]) :: Type

Compose surgeries together.

type DCat (ss :: [Type]) (a :: Type) = Data (GSurgery (Cat ss) (Rep a)) ()

Make a synthetic type (Data) by chaining multiple surgeries.

data CopyRep (a :: Type) :: Type

Change the generic representation to that of another type a.

copyRep :: forall a f p. Coercible (GSurgery (CopyRep a) f) f => Data f p -> Data (GSurgery (CopyRep a) f) p

uncopyRep :: forall a f p. Coercible f (GSurgery (CopyRep a) f) => Data (GSurgery (CopyRep a) f) p -> Data f p