Generic.Data.Internal.Read

Description

Generic implementation of Read

Warning

This is an internal module: it is not subject to any versioning policy, breaking changes can happen at any time.

If something here seems useful, please report it or create a pull request to export it from an external module.

Documentation

greadPrec :: (Generic a, GRead0 (Rep a)) => ReadPrec a

Generic readPrec.

instance Read MyType where
  readPrec = greadPrec
  readListPrec = readListPrecDefault

type GRead0 = GRead Proxy

Generic representation of Read types.

gliftReadPrec :: (Generic1 f, GRead1 (Rep1 f)) => ReadPrec a -> ReadPrec [a] -> ReadPrec (f a)

Generic liftReadPrec.

type GRead1 = GRead Identity

Generic representation of Read1 types.

class GRead p f where

Methods

gPrecRead :: p (ReadPrec a, ReadPrec [a]) -> ReadPrec (f a)

class IsNullaryDataType f where

Methods

isNullaryDataType :: f a -> Bool

class GReadC p c f where

Methods

gPrecReadC :: p (ReadPrec a, ReadPrec [a]) -> String -> Fixity -> ReadPrec (M1 C c f a)

class GReadFields p f where

Methods

gPrecReadFields :: p (ReadPrec a, ReadPrec [a]) -> ReadPrecTree (f a)

class GReadNamed p f where

Methods

gPrecReadNamed :: p (ReadPrec a, ReadPrec [a]) -> ReadPrec (f a)

class GReadSingle p f where

Methods

gPrecReadSingle :: p (ReadPrec a, ReadPrec [a]) -> ReadPrec (f a)

coerceM1 :: ReadPrec (f p) -> ReadPrec (M1 i c f p)

liftReadPrecCompat :: Read1 f => (ReadPrec a, ReadPrec [a]) -> ReadPrec (f a)

A backwards-compatible version of liftReadPrec. This is needed for compatibility with base-4.9, where Read1 only offers liftReadsPrec, not liftReadPrec.

data ReadPrecTree a where

Constructors

U1Leaf :: ReadPrecTree (U1 a)

M1Leaf :: ReadPrec (f a) -> ReadPrecTree (M1 i c f a)

Branch :: ReadPrecTree (f a) -> ReadPrecTree (g a) -> ReadPrecTree ((f :*: g) a)

toReadPrec :: ReadPrecTree a -> ReadPrec a

identHLexemes :: String -> [Lexeme]

readPrefixCon :: String -> ReadPrec ()

readSurround :: Char -> ReadPrec a -> Char -> ReadPrec a

snocView :: [a] -> Maybe ([a], a)