Generic.Data.Internal.Resolvers

Description

Newtypes with special instances for deriving.

Warning

This is an internal module: it is not subject to any versioning policy, breaking changes can happen at any time.

If something here seems useful, please report it or create a pull request to export it from an external module.

Documentation

newtype Id1 f a

A newtype whose instances for simple classes (Eq, Ord, Read, Show) use higher-kinded class instances for f (Eq1, Ord1, Read1, Show1).

Constructors

Id1

Fields

unId1 :: f a

newtype Opaque a

A newtype with trivial instances, that considers every value equivalent to every other one, and shows as just "_".

Constructors

Opaque

Fields

unOpaque :: a

newtype Opaque1 f a

A higher-kinded version of Opaque.

Constructors

Opaque1

Fields

unOpaque1 :: f a