Generic.Data.Internal.Show

Description

Generic implementation of Show

Warning

This is an internal module: it is not subject to any versioning policy, breaking changes can happen at any time.

If something here seems useful, please report it or create a pull request to export it from an external module.

Documentation

gshowsPrec :: (Generic a, GShow0 (Rep a)) => Int -> a -> ShowS

Generic showsPrec.

instance Show MyType where
  showsPrec = gshowsPrec

gprecShows :: (Generic a, GShow0 (Rep a)) => a -> PrecShowS

type GShow0 = GShow Proxy

Generic representation of Show types.

gliftShowsPrec :: (Generic1 f, GShow1 (Rep1 f)) => (Int -> a -> ShowS) -> ([a] -> ShowS) -> Int -> f a -> ShowS

Generic liftShowsPrec.

gLiftPrecShows :: GShow1 f => (Int -> a -> ShowS) -> ([a] -> ShowS) -> f a -> PrecShowS

type ShowsPrec a = (Int -> a -> ShowS, [a] -> ShowS)

type GShow1 = GShow Identity

Generic representation of Show1 types.

class GShow p f where

Methods

gPrecShows :: p (ShowsPrec a) -> f a -> PrecShowS

class GShowC p c f where

Methods

gPrecShowsC :: p (ShowsPrec a) -> String -> Fixity -> M1 C c f a -> PrecShowS

class GShowFields p f where

Methods

gPrecShowsFields :: p (ShowsPrec a) -> f a -> [PrecShowS]

class GShowNamed p f where

Methods

gPrecShowsNamed :: p (ShowsPrec a) -> f a -> ShowFields

class GShowSingle p f where

Methods

gPrecShowsSingle :: p (ShowsPrec a) -> f a -> PrecShowS

surroundConName :: Fixity -> String -> String