Generic.Data.Internal.Traversable

Description

Generic implementation of Foldable and Traversable.

There is already a naive implementation using the generic Rep's own instances of Foldable and Traversable. However, deriving then generates a lot of code that may not be simplified away by GHC, that results in unnecessary run-time overhead.

In contrast, this implementation guarantees that the generated code is identical to stock-derived instances of Foldable and Traversable, which have the following syntactic properties:

constructors with zero fields use pure once;

constructors with one field use fmap once;

constructors with n >= 2 fields use liftA2 once and (<*>) n-2 times.

The heavy lifting is actually done by the ap-normalize library.

Library

gfoldMap :: (Generic1 f, GFoldable (Rep1 f), Monoid m) => (a -> m) -> f a -> m

Generic foldMap.

instance Foldable MyTypeF where
  foldMap = gfoldMap

gtraverse :: (Generic1 f, GTraversable (Rep1 f), Applicative m) => (a -> m b) -> f a -> m (f b)

Generic traverse.

instance Traversable MyTypeF where
  traverse = gtraverse

gsequenceA :: (Generic1 f, GTraversable (Rep1 f), Applicative m) => f (m a) -> m (f a)

Generic sequenceA.

instance Traversable MyTypeF where
  sequenceA = gsequenceA

See also gtraverse.

class GFoldable_ t => GFoldable t

Class of generic representations for which Foldable can be derived.

class GTraversable_ t => GTraversable t

Class of generic representations for which Traversable can be derived.

class (GFoldMap t, Foldable t) => GFoldable_ t

Internal definition of GFoldable.

class (GTraverse Kleisli t, GTraverse Equal t) => GTraversable_ t

Internal definition of GTraversable.

Foldable

data Maybe' m

Isomorphic to Maybe m, but we need to micromanage the use of Monoid vs Semigroup to match exactly the output of stock deriving, for inspection testing.

Constructors

Nothing'

Just' m

type EndoM m = Endo (Maybe' m)

liftEndoM :: Monoid m => m -> EndoM m

lowerEndoM :: Monoid m => EndoM m -> m

lowerMaybe :: Monoid m => Maybe' m -> m

class GFoldMap t where

Methods

gfoldMap_ :: Monoid m => (a -> m) -> t a -> EndoM m

Traversable

data Equal (f :: Type -> Type) a b where

Constructors

Refl :: Equal f (f b) b

newtype Kleisli f a b

Constructors

Kleisli (a -> f b)

class GTraverse arr t where

Methods

gtraverse_ :: Applicative f => arr f a b -> t a -> Aps f (t b)