Hedgehog

Description

This module includes almost everything you need to get started writing property tests with Hedgehog.

It is designed to be used alongside Hedgehog.Gen and Hedgehog.Range, which should be imported qualified. You also need to enable Template Haskell so the Hedgehog test runner can find your properties.

{-# LANGUAGE TemplateHaskell #-}

import           Hedgehog
import qualified Hedgehog.Gen as Gen
import qualified Hedgehog.Range as Range

Once you have your imports set up, you can write a simple property:

prop_reverse :: Property
prop_reverse =
  property $ do
    xs <- forAll $ Gen.list (Range.linear 0 100) Gen.alpha
    reverse (reverse xs) === xs

And add the Template Haskell splice which will discover your properties:

tests :: IO Bool
tests =
  checkParallel $$(discover)

If you prefer to avoid macros, you can specify the group of properties to run manually instead:

{-# LANGUAGE OverloadedStrings #-}

tests :: IO Bool
tests =
  checkParallel $ Group "Test.Example" [
      ("prop_reverse", prop_reverse)
    ]

You can then load the module in GHCi, and run it:

λ tests
━━━ Test.Example ━━━
  ✓ prop_reverse passed 100 tests.

Properties

data Property

A property test, along with some configurable limits like how many times to run the test.

data PropertyT m a

The property monad transformer allows both the generation of test inputs and the assertion of expectations.

data Group

A named collection of property tests.

Constructors

Group

Fields

groupName :: !GroupName

groupProperties :: ![(PropertyName, Property)]

data PropertyName

The name of a property.

Should be constructed using OverloadedStrings:

  "apples" :: PropertyName

data GroupName

The name of a group of properties.

Should be constructed using OverloadedStrings:

  "fruit" :: GroupName

property :: HasCallStack => PropertyT IO () -> Property

Creates a property with the default configuration.

test :: Monad m => TestT m a -> PropertyT m a

Lift a test in to a property.

Because both TestT and PropertyT have MonadTest instances, this function is not often required. It can however be useful for writing functions directly in TestT and thus gaining a MonadTransControl instance at the expense of not being able to generate additional inputs using forAll.

An example where this is useful is parallel state machine testing, as executeParallel requires MonadBaseControl IO in order to be able to spawn threads in MonadTest.

forAll :: (Monad m, Show a, HasCallStack) => Gen a -> PropertyT m a

Generates a random input for the test by running the provided generator.

forAllWith :: (Monad m, HasCallStack) => (a -> String) -> Gen a -> PropertyT m a

Generates a random input for the test by running the provided generator.

This is a the same as forAll but allows the user to provide a custom rendering function. This is useful for values which don't have a Show instance.

discard :: Monad m => PropertyT m a

Discards the current test entirely.

check :: MonadIO m => Property -> m Bool

Check a property.

recheck :: MonadIO m => Size -> Seed -> Property -> m ()

Check a property using a specific size and seed.

recheckAt :: MonadIO m => Seed -> Skip -> Property -> m ()

discover :: TExpQ Group

Discover all the properties in a module.

Functions starting with prop_ are assumed to be properties.

discoverPrefix :: String -> TExpQ Group

checkParallel :: MonadIO m => Group -> m Bool

Check a group of properties in parallel.

Warning: although this check function runs tests faster than checkSequential, it should be noted that it may cause problems with properties that are not self-contained. For example, if you have a group of tests which all use the same database table, you may find that they interfere with each other when being run in parallel.

Using Template Haskell for property discovery:

tests :: IO Bool
tests =
  checkParallel $$(discover)

With manually specified properties:

tests :: IO Bool
tests =
  checkParallel $ Group "Test.Example" [
      ("prop_reverse", prop_reverse)
    ]

checkSequential :: MonadIO m => Group -> m Bool

Check a group of properties sequentially.

Using Template Haskell for property discovery:

tests :: IO Bool
tests =
  checkSequential $$(discover)

With manually specified properties:

tests :: IO Bool
tests =
  checkSequential $ Group "Test.Example" [
      ("prop_reverse", prop_reverse)
    ]

data Confidence

The acceptable occurrence of false positives

Example, Confidence 10^9 would mean that you'd accept a false positive for 1 in 10^9 tests.

verifiedTermination :: Property -> Property

withConfidence :: Confidence -> Property -> Property

Make sure that the result is statistically significant in accordance to the passed Confidence

withTests :: TestLimit -> Property -> Property

Set the number of times a property should be executed before it is considered successful.

If you have a test that does not involve any generators and thus does not need to run repeatedly, you can use withTests 1 to define a property that will only be checked once.

data TestLimit

The number of successful tests that need to be run before a property test is considered successful.

Can be constructed using numeric literals:

  200 :: TestLimit

withDiscards :: DiscardLimit -> Property -> Property

Set the number of times a property is allowed to discard before the test runner gives up.

data DiscardLimit

The number of discards to allow before giving up.

Can be constructed using numeric literals:

  10000 :: DiscardLimit

withShrinks :: ShrinkLimit -> Property -> Property

Set the number of times a property is allowed to shrink before the test runner gives up and prints the counterexample.

data ShrinkLimit

The number of shrinks to try before giving up on shrinking.

Can be constructed using numeric literals:

  1000 :: ShrinkLimit

withRetries :: ShrinkRetries -> Property -> Property

Set the number of times a property will be executed for each shrink before the test runner gives up and tries a different shrink. See ShrinkRetries for more information.

data ShrinkRetries

The number of times to re-run a test during shrinking. This is useful if you are testing something which fails non-deterministically and you want to increase the change of getting a good shrink.

If you are doing parallel state machine testing, you should probably set shrink retries to something like 10. This will mean that during shrinking, a parallel test case requires 10 successful runs before it is passes and we try a different shrink.

Can be constructed using numeric literals:

  0 :: ShrinkRetries

withSkip :: Skip -> Property -> Property

Set the target that a property will skip to before it starts to run.

data Skip

Where to start running a property's tests.

Generating Test Data

type Gen = GenT Identity

Generator for random values of a.

data GenT m a

Monad transformer which can generate random values of a.

class (Monad m, Monad (GenBase m)) => MonadGen m where

Class of monads which can generate input data for tests.

Associated Types

type GenBase m :: Type -> Type

Methods

toGenT :: m a -> GenT (GenBase m) a

Extract a GenT from a MonadGen.

fromGenT :: GenT (GenBase m) a -> m a

Lift a GenT in to a MonadGen.

data Range a

A range describes the bounds of a number to generate, which may or may not be dependent on a Size.

The constructor takes an origin between the lower and upper bound, and a function from Size to bounds. As the size goes towards 0, the values go towards the origin.

newtype Size

Tests are parameterized by the size of the randomly-generated data. The meaning of a Size value depends on the particular generator used, but it must always be a number between 0 and 99 inclusive.

Constructors

Size

Fields

unSize :: Int

data Seed

A splittable random number generator.

Constructors

Seed

Fields

seedValue :: !Word64

seedGamma :: !Word64

must be an odd number

Tests

type Test = TestT Identity

A test monad allows the assertion of expectations.

data TestT m a

A test monad transformer allows the assertion of expectations.

class Monad m => MonadTest m where

Methods

liftTest :: Test a -> m a

annotate :: (MonadTest m, HasCallStack) => String -> m ()

Annotates the source code with a message that might be useful for debugging a test failure.

annotateShow :: (MonadTest m, Show a, HasCallStack) => a -> m ()

Annotates the source code with a value that might be useful for debugging a test failure.

footnote :: MonadTest m => String -> m ()

Logs a message to be displayed as additional information in the footer of the failure report.

footnoteShow :: (MonadTest m, Show a) => a -> m ()

Logs a value to be displayed as additional information in the footer of the failure report.

success :: MonadTest m => m ()

Another name for pure ().

failure :: (MonadTest m, HasCallStack) => m a

Causes a test to fail.

assert :: (MonadTest m, HasCallStack) => Bool -> m ()

Fails the test if the condition provided is False.

diff :: (MonadTest m, Show a, Show b, HasCallStack) => a -> (a -> b -> Bool) -> b -> m ()

Fails the test and shows a git-like diff if the comparison operation evaluates to False when applied to its arguments.

The comparison function is the second argument, which may be counter-intuitive to Haskell programmers. However, it allows operators to be written infix for easy reading:

  diff y (<) 87
  diff x (<=) r

This function behaves like the unix diff tool, which gives a 0 exit code if the compared files are identical, or a 1 exit code code otherwise. Like unix diff, if the arguments fail the comparison, a /diff is shown.

(===) :: (MonadTest m, Eq a, Show a, HasCallStack) => a -> a -> m () infix 4

Fails the test if the two arguments provided are not equal.

(/==) :: (MonadTest m, Eq a, Show a, HasCallStack) => a -> a -> m () infix 4

Fails the test if the two arguments provided are equal.

tripping :: (MonadTest m, Applicative f, Show b, Show (f a), Eq (f a), HasCallStack) => a -> (a -> b) -> (b -> f a) -> m ()

Test that a pair of encode / decode functions are compatible.

Given a printer from some type a -> b, and a parser with a potential failure case b -> f a. Ensure that a valid a round trips through the "print" and "parse" to yield the same a.

For example, types should have tripping Read and Show instances:

trippingShowRead :: (Show a, Read a, Eq a, MonadTest m) => a -> m ()
trippingShowRead a = tripping a show readEither

eval :: (MonadTest m, HasCallStack) => a -> m a

Fails the test if the value throws an exception when evaluated to weak head normal form (WHNF).

evalNF :: (MonadTest m, NFData a, HasCallStack) => a -> m a

Fails the test if the value throws an exception when evaluated to normal form (NF).

evalM :: (MonadTest m, MonadCatch m, HasCallStack) => m a -> m a

Fails the test if the action throws an exception.

The benefit of using this over simply letting the exception bubble up is that the location of the closest evalM will be shown in the output.

evalIO :: (MonadTest m, MonadIO m, HasCallStack) => IO a -> m a

Fails the test if the IO action throws an exception.

The benefit of using this over liftIO is that the location of the exception will be shown in the output.

evalEither :: (MonadTest m, Show x, HasCallStack) => Either x a -> m a

Fails the test if the Either is Left, otherwise returns the value in the Right.

evalEitherM :: (MonadTest m, Show x, MonadCatch m, HasCallStack) => m (Either x a) -> m a

Fails the test if the action throws an exception, or if the Either is Left, otherwise returns the value in the Right.

evalExceptT :: (MonadTest m, Show x, HasCallStack) => ExceptT x m a -> m a

Fails the test if the ExceptT is Left, otherwise returns the value in the Right.

evalMaybe :: (MonadTest m, HasCallStack) => Maybe a -> m a

Fails the test if the Maybe is Nothing, otherwise returns the value in the Just.

evalMaybeM :: (MonadTest m, MonadCatch m, HasCallStack) => m (Maybe a) -> m a

Fails the test if the action throws an exception, or if the Maybe is Nothing, otherwise returns the value in the Just.

Coverage

data LabelName

The name of a classifier.

Should be constructed using OverloadedStrings:

  "apples" :: LabelName

classify :: (MonadTest m, HasCallStack) => LabelName -> Bool -> m ()

Records the proportion of tests which satisfy a given condition.

   prop_with_classifier :: Property
   prop_with_classifier =
     property $ do
       xs <- forAll $ Gen.list (Range.linear 0 100) Gen.alpha
       for_ xs $ \x -> do
         classify "newborns" $ x == 0
         classify "children" $ x > 0 && x < 13
         classify "teens" $ x > 12 && x < 20

cover :: (MonadTest m, HasCallStack) => CoverPercentage -> LabelName -> Bool -> m ()

Require a certain percentage of the tests to be covered by the classifier.

   prop_with_coverage :: Property
   prop_with_coverage =
     property $ do
       match <- forAll Gen.bool
       cover 30 "True" $ match
       cover 30 "False" $ not match

The example above requires a minimum of 30% coverage for both classifiers. If these requirements are not met, it will fail the test.

label :: (MonadTest m, HasCallStack) => LabelName -> m ()

Add a label for each test run. It produces a table showing the percentage of test runs that produced each label.

collect :: (MonadTest m, Show a, HasCallStack) => a -> m ()

Like label, but uses Show to render its argument for display.

State Machine Tests

data Command gen m (state :: (Type -> Type) -> Type)

The specification for the expected behaviour of an Action. These are used to generate sequences of actions to test.

This is the main type you will use when writing state machine tests. gen is usually an instance of MonadGen, and m is usually an instance of MonadTest. These constraints appear when you pass your Command list to sequential or parallel.

Constructors

forall input output.(TraversableB input, Show (input Symbolic), Show output, Typeable output) => Command

Fields

commandGen :: state Symbolic -> Maybe (gen (input Symbolic))

A generator which provides random arguments for a command. If the command cannot be executed in the current state, it should return Nothing.

commandExecute :: input Concrete -> m output

Executes a command using the arguments generated by commandGen.

commandCallbacks :: [Callback input output state]

A set of callbacks which provide optional command configuration such as pre-condtions, post-conditions and state updates.

data Callback input output state

Optional command configuration.

Constructors

Require (state Symbolic -> input Symbolic -> Bool)
A pre-condition for a command that must be verified before the command can be executed. This is mainly used during shrinking to ensure that it is still OK to run a command despite the fact that some previously executed commands may have been removed from the sequence.

Update (forall v. Ord1 v => state v -> input v -> Var output v -> state v)
Updates the model state, given the input and output of the command. Note that this function is polymorphic in the type of values. This is because it must work over Symbolic values when we are generating actions, and Concrete values when we are executing them.

Ensure (state Concrete -> state Concrete -> input Concrete -> output -> Test ())
A post-condition for a command that must be verified for the command to be considered a success.

This callback receives the state prior to execution as the first argument, and the state after execution as the second argument.

data Action m (state :: (Type -> Type) -> Type)

An instantiation of a Command which can be executed, and its effect evaluated.

newtype Sequential m state

A sequence of actions to execute.

Constructors

Sequential

Fields

sequentialActions :: [Action m state]

The sequence of actions.

data Parallel m state

A sequential prefix of actions to execute, with two branches to execute in parallel.

Constructors

Parallel

Fields

parallelPrefix :: [Action m state]

The sequential prefix.

parallelBranch1 :: [Action m state]

The first branch.

parallelBranch2 :: [Action m state]

The second branch.

executeSequential :: (MonadTest m, MonadCatch m, HasCallStack) => (forall v. state v) -> Sequential m state -> m ()

Executes a list of actions sequentially, verifying that all post-conditions are met and no exceptions are thrown.

To generate a sequence of actions to execute, see the sequential combinator in the Hedgehog.Gen module.

executeParallel :: (MonadTest m, MonadCatch m, MonadBaseControl IO m, HasCallStack) => (forall v. state v) -> Parallel m state -> m ()

Executes the prefix actions sequentially, then executes the two branches in parallel, verifying that no exceptions are thrown and that there is at least one sequential interleaving where all the post-conditions are met.

To generate parallel actions to execute, see the parallel combinator in the Hedgehog.Gen module.

newtype Var a v

Variables are the potential or actual result of executing an action. They are parameterised by either Symbolic or Concrete depending on the phase of the test.

Symbolic variables are the potential results of actions. These are used when generating the sequence of actions to execute. They allow actions which occur later in the sequence to make use of the result of an action which came earlier in the sequence.

Concrete variables are the actual results of actions. These are used during test execution. They provide access to the actual runtime value of a variable.

The state update Callback for a command needs to be polymorphic in the type of variable because it is used in both the generation and the execution phase.

The order of arguments makes Var FunctorB and TraversableB, which is how Symbolic values are turned into Concrete ones.

Constructors

Var (v a)

concrete :: Var a Concrete -> a

Take the value from a concrete variable.

opaque :: Var (Opaque a) Concrete -> a

Take the value from an opaque concrete variable.

data Symbolic a

Symbolic values: Because hedgehog generates actions in a separate phase before execution, you will sometimes need to refer to the result of a previous action in a generator without knowing the value of the result (e.g., to get the ID of a previously-created user).

Symbolic variables provide a token to stand in for the actual variables at generation time (and in Require/Update callbacks). At execution time, real values are available, so your execute actions work on Concrete variables.

See also: Command, Var

newtype Concrete a where

Concrete values: At test-execution time, Symbolic values from generation are replaced with Concrete values from performing actions. This type gives us something of the same kind as Symbolic to pass as a type argument to Var.

Constructors

Concrete :: a -> Concrete a

newtype Opaque a

Opaque values.

Useful if you want to put something without a Show instance inside something which you'd like to be able to display.

For example:

  data State v =
    State {
        stateRefs :: [Var (Opaque (IORef Int)) v]
      } deriving (Eq, Show)

Constructors

Opaque

Fields

unOpaque :: a

Transformers

distributeT :: (MonadTransDistributive g, Transformer f g m) => g (f m) a -> f (g m) a

Distribute one monad transformer over another.

Functors

FunctorB and TraversableB must be implemented for all Command input types.

This is most easily achieved using DeriveGeneric:

data Register v =
  Register Name (Var Pid v)
  deriving (Eq, Show, Generic)

instance FunctorB Register
instance TraversableB Register

newtype Unregister (v :: * -> *) =
  Unregister Name
  deriving (Eq, Show, Generic)

instance FunctorB Unregister
instance TraversableB Unregister

DeriveAnyClass and DerivingStrategies allow a more compact syntax:

data Register v =
  Register Name (Var Pid v)
  deriving (Eq, Show, Generic, FunctorB, TraversableB)

newtype Unregister (v :: * -> *) =
  Unregister Name
  deriving (Eq, Show, Generic)
  deriving anyclass (FunctorB, TraversableB)

class FunctorB (b :: (k -> Type) -> Type) where

Minimal complete definition

Nothing

Methods

bmap :: (forall (a :: k). f a -> g a) -> b f -> b g

class FunctorB b => TraversableB (b :: (k -> Type) -> Type) where

Minimal complete definition

Nothing

Methods

btraverse :: Applicative e => (forall (a :: k). f a -> e (g a)) -> b f -> e (b g)

newtype Rec p a (x :: k)

Constructors

Rec

Fields

unRec :: K1 R a x

class (forall a. Eq a => Eq (f a)) => Eq1 (f :: Type -> Type)

Lifting of the Eq class to unary type constructors.

Any instance should be subject to the following law that canonicity is preserved:

liftEq (==) = (==)

This class therefore represents the generalization of Eq by decomposing its main method into a canonical lifting on a canonical inner method, so that the lifting can be reused for other arguments than the canonical one.

Since: base-4.9.0.0

eq1 :: (Eq1 f, Eq a) => f a -> f a -> Bool

Lift the standard (==) function through the type constructor.

Since: base-4.9.0.0

class (Eq1 f, forall a. Ord a => Ord (f a)) => Ord1 (f :: Type -> Type)

Lifting of the Ord class to unary type constructors.

Any instance should be subject to the following law that canonicity is preserved:

liftCompare compare = compare

This class therefore represents the generalization of Ord by decomposing its main method into a canonical lifting on a canonical inner method, so that the lifting can be reused for other arguments than the canonical one.

Since: base-4.9.0.0

compare1 :: (Ord1 f, Ord a) => f a -> f a -> Ordering

Lift the standard compare function through the type constructor.

Since: base-4.9.0.0

class (forall a. Show a => Show (f a)) => Show1 (f :: Type -> Type)

Lifting of the Show class to unary type constructors.

Any instance should be subject to the following laws that canonicity is preserved:

liftShowsPrec showsPrec showList = showsPrec

liftShowList showsPrec showList = showList

This class therefore represents the generalization of Show by decomposing it's methods into a canonical lifting on a canonical inner method, so that the lifting can be reused for other arguments than the canonical one.

Since: base-4.9.0.0

showsPrec1 :: (Show1 f, Show a) => Int -> f a -> ShowS

Lift the standard showsPrec and showList functions through the type constructor.

Since: base-4.9.0.0

Deprecated

class HTraversable t where

Deprecated: Replace with Hedgehog.TraversableB (defined in Data.Functor.Barbie) which can be derived automatically using GHC.Generics

Higher-order traversable functors.

Deprecated in favor of TraversableB which can be derived using GHC.Generics

Methods

htraverse :: Applicative f => (forall a. g a -> f (h a)) -> t g -> f (t h)