Hedgehog.Extras.Stock.IO.Network.Port

Documentation

randomPort :: MonadIO m => MonadFail m => HostAddress -> m PortNumber

Return a random available port on a specified host address

reserveRandomPort :: MonadFail m => MonadResource m => HostAddress -> m (ReleaseKey, PortNumber)

portInUse :: MonadIO m => HostAddress -> PortNumber -> m Bool

Check if a port is in use on a specified host address