Hedgehog.Extras.Stock.IO.Network.Socket

Documentation

doesSocketExist :: FilePath -> IO Bool

isPortOpen :: Int -> IO Bool

Check if a TCP port is open

canConnect :: SockAddr -> IO ()

Check if it is possible to connect to a socket address

listenOn :: Int -> IO Socket

Open a socket at the specified port for listening

allocateRandomPorts :: Int -> IO [Int]

Allocate the specified number of random ports