Hedgehog.Extras.Stock.String

Documentation

strip :: String -> String

Strip whitespace from the beginning and end of the string.

lastLine :: String -> String

Get the last line in the string

firstLine :: String -> String

Get the first line in the string

readNoteM :: (Read a, Show a, MonadTest m, MonadCatch m, HasCallStack) => String -> m a

Trim leading and trailing whitespace and read the string into a value. Report the read value in the test annotations.