Hedgehog.Extras.Test.MonadAssertion

Documentation

class Monad m => MonadAssertion m where

Methods

throwAssertion :: Failure -> m a

catchAssertion :: m a -> (Failure -> m a) -> m a

assertFailure :: HasCallStack => Show a => MonadAssertion m => MonadTest m => m a -> m Failure

Run the given action and succeed if the action fails, but fail if it succeeds.

assertFailure_ :: HasCallStack => Show a => MonadAssertion m => MonadTest m => m a -> m ()

Run the given action and succeed if the action fails, but fail if it succeeds.

tryAssertion :: MonadAssertion m => m a -> m (Either Failure a)

Attempt to run a function that may assert, returning either a failure or the result of the assertion.