Hedgehog.Extras.Test.Tripwire

Description

This modules provides a tripwire abstraction. You can use tripwire as a detection mechanism if the code path was executed. Trip a tripwire with trip in the place where you'd like to detect if it was reached. The tripwire can then be checked in the other place in the code using for example isTripped or assertNotTripped.

Create a tripwire

data Tripwire

Represents a tripwire which can be tripped only once. It can be used to detect if a particular code path was reached.

makeTripwire :: MonadIO m => m Tripwire

Creates a new tripwire

makeTripwireWithLabel :: MonadIO m => String -> m Tripwire

Creates a new tripwire with a label, which is visible when showed: Tripwire mylabel

Tripwire operations

trip :: HasCallStack => MonadIO m => MonadTest m => Tripwire -> m ()

Triggers the tripwire and registers the place of the first trigger. Idempotent. Prints the information in the test log about tripping the tripwire.

trip_ :: HasCallStack => MonadIO m => Tripwire -> m ()

Triggers the tripwire and registers the place of the first trigger. Idempotent. A silent variant of trip which does not require MonadTest, but also does not log the information about tripping.

isTripped :: MonadIO m => Tripwire -> m Bool

Check if the tripwire was tripped.

getTripSite :: MonadIO m => Tripwire -> m (Maybe CallStack)

Return the call stack, where the tripwire was tripped - if it was tripped.

resetTripwire :: MonadIO m => Tripwire -> m ()

Restore tripwire to initial non triggered state

Assertions

assertNotTripped :: HasCallStack => MonadTest m => MonadIO m => Tripwire -> m ()

Fails the test if the tripwire was triggered. Prints the call stack where the tripwire was triggered.

assertTripped :: HasCallStack => MonadTest m => MonadIO m => Tripwire -> m ()

Fails the test if the tripwire was not triggered yet.