Hedgehog.Internal.Config

Documentation

data UseColor

Whether to render output using ANSI colors or not.

Constructors

DisableColor
Disable ANSI colors in report output.

EnableColor
Enable ANSI colors in report output.

resolveColor :: MonadIO m => Maybe UseColor -> m UseColor

data Seed

A splittable random number generator.

Constructors

Seed

Fields

seedValue :: !Word64

seedGamma :: !Word64

must be an odd number

resolveSeed :: MonadIO m => Maybe Seed -> m Seed

data Verbosity

How verbose should the report output be.

Constructors

Quiet
Only display the summary of the test run.

Normal
Display each property as it is running, as well as the summary.

resolveVerbosity :: MonadIO m => Maybe Verbosity -> m Verbosity

newtype WorkerCount

The number of workers to use when running properties in parallel.

Constructors

WorkerCount Int

resolveWorkers :: MonadIO m => Maybe WorkerCount -> m WorkerCount

data Skip

Where to start running a property's tests.

Constructors

SkipNothing
Don't skip anything.

SkipToTest TestCount DiscardCount
Skip to a specific test number. If it fails, shrink as normal. If it passes, move on to the next test. Coverage checks are disabled.

We also need to count discards, since failing "after 7 tests" points at a different generated value than failing "after 7 tests and 5 discards".

SkipToShrink TestCount DiscardCount ShrinkPath
Skip to a specific test number and shrink state. If it fails, stop without shrinking further. If it passes, the property will pass without running any more tests.

Due to implementation details, all intermediate shrink states - those on the direct path from the original test input to the target state - will be tested too, and their results discarded.

resolveSkip :: MonadIO m => Maybe Skip -> m Skip

detectMark :: MonadIO m => m Bool

detectColor :: MonadIO m => m UseColor

detectSeed :: MonadIO m => m Seed

detectVerbosity :: MonadIO m => m Verbosity

detectWorkers :: MonadIO m => m WorkerCount

detectSkip :: MonadIO m => m Skip