Hedgehog.Internal.Gen

Transformer

type Gen = GenT Identity

Generator for random values of a.

newtype GenT m a

Monad transformer which can generate random values of a.

Constructors

GenT

Fields

unGenT :: Size -> Seed -> TreeT (MaybeT m) a

class (Monad m, Monad (GenBase m)) => MonadGen m where

Class of monads which can generate input data for tests.

Associated Types

type GenBase m :: Type -> Type

Methods

toGenT :: m a -> GenT (GenBase m) a

Extract a GenT from a MonadGen.

fromGenT :: GenT (GenBase m) a -> m a

Lift a GenT in to a MonadGen.

Combinators

generalize :: Monad m => Gen a -> GenT m a

Lift a Gen / GenT Identity in to a Monad m => GenT m

Shrinking

shrink :: MonadGen m => (a -> [a]) -> m a -> m a

Apply a shrinking function to a generator.

This will give the generator additional shrinking options, while keeping the existing shrinks intact.

prune :: MonadGen m => m a -> m a

Throw away a generator's shrink tree.

Size

small :: MonadGen m => m a -> m a

Make a generator smaller by scaling its size parameter.

scale :: (HasCallStack, MonadGen m) => (Size -> Size) -> m a -> m a

Adjust the size parameter by transforming it with the given function.

resize :: MonadGen m => Size -> m a -> m a

Override the size parameter. Returns a generator which uses the given size instead of the runtime-size parameter.

sized :: MonadGen m => (Size -> m a) -> m a

Construct a generator that depends on the size parameter.

Integral

integral :: forall m a. (MonadGen m, Integral a) => Range a -> m a

Generates a random integral number in the given [inclusive,inclusive] range.

When the generator tries to shrink, it will shrink towards the origin of the specified Range.

For example, the following generator will produce a number between 1970 and 2100, but will shrink towards 2000:

integral (Range.constantFrom 2000 1970 2100) :: Gen Int

Some sample outputs from this generator might look like:

=== Outcome ===
1973
=== Shrinks ===
2000
1987
1980
1976
1974

=== Outcome ===
2061
=== Shrinks ===
2000
2031
2046
2054
2058
2060

integral_ :: (MonadGen m, Integral a) => Range a -> m a

Generates a random integral number in the [inclusive,inclusive] range.

This generator does not shrink.

int :: MonadGen m => Range Int -> m Int

Generates a random machine integer in the given [inclusive,inclusive] range.

This is a specialization of integral, offered for convenience.

int8 :: MonadGen m => Range Int8 -> m Int8

Generates a random 8-bit integer in the given [inclusive,inclusive] range.

This is a specialization of integral, offered for convenience.

int16 :: MonadGen m => Range Int16 -> m Int16

Generates a random 16-bit integer in the given [inclusive,inclusive] range.

This is a specialization of integral, offered for convenience.

int32 :: MonadGen m => Range Int32 -> m Int32

Generates a random 32-bit integer in the given [inclusive,inclusive] range.

This is a specialization of integral, offered for convenience.

int64 :: MonadGen m => Range Int64 -> m Int64

Generates a random 64-bit integer in the given [inclusive,inclusive] range.

This is a specialization of integral, offered for convenience.

word :: MonadGen m => Range Word -> m Word

Generates a random machine word in the given [inclusive,inclusive] range.

This is a specialization of integral, offered for convenience.

word8 :: MonadGen m => Range Word8 -> m Word8

Generates a random byte in the given [inclusive,inclusive] range.

This is a specialization of integral, offered for convenience.

word16 :: MonadGen m => Range Word16 -> m Word16

Generates a random 16-bit word in the given [inclusive,inclusive] range.

This is a specialization of integral, offered for convenience.

word32 :: MonadGen m => Range Word32 -> m Word32

Generates a random 32-bit word in the given [inclusive,inclusive] range.

This is a specialization of integral, offered for convenience.

word64 :: MonadGen m => Range Word64 -> m Word64

Generates a random 64-bit word in the given [inclusive,inclusive] range.

This is a specialization of integral, offered for convenience.

Floating-point

realFloat :: (MonadGen m, RealFloat a) => Range a -> m a

Generates a random floating-point number in the [inclusive,exclusive) range.

This generator works the same as integral, but for floating point numbers.

realFrac_ :: (MonadGen m, RealFrac a) => Range a -> m a

Generates a random fractional number in the [inclusive,exclusive) range.

This generator does not shrink.

float :: MonadGen m => Range Float -> m Float

Generates a random floating-point number in the [inclusive,exclusive) range.

This is a specialization of realFloat, offered for convenience.

double :: MonadGen m => Range Double -> m Double

Generates a random floating-point number in the [inclusive,exclusive) range.

This is a specialization of realFloat, offered for convenience.

Enumeration

enum :: (MonadGen m, Enum a) => a -> a -> m a

Generates an element from an enumeration.

This generator shrinks towards the first argument.

For example:

enum 'a' 'z' :: Gen Char

enumBounded :: (MonadGen m, Enum a, Bounded a) => m a

Generates a random value from a bounded enumeration.

This generator shrinks towards minBound.

For example:

enumBounded :: Gen Bool

This is implemented in terms of the Enum class, and thus may be partial for integral types larger than Int, e.g. Word64.

bool :: MonadGen m => m Bool

Generates a random boolean.

This generator shrinks to False.

This is a specialization of enumBounded, offered for convenience.

bool_ :: MonadGen m => m Bool

Generates a random boolean.

This generator does not shrink.

Characters

binit :: MonadGen m => m Char

Generates an ASCII binit: '0'..'1'

octit :: MonadGen m => m Char

Generates an ASCII octit: '0'..'7'

digit :: MonadGen m => m Char

Generates an ASCII digit: '0'..'9'

hexit :: MonadGen m => m Char

Generates an ASCII hexit: '0'..'9', 'a'..'f', 'A'..'F'

lower :: MonadGen m => m Char

Generates an ASCII lowercase letter: 'a'..'z'

upper :: MonadGen m => m Char

Generates an ASCII uppercase letter: 'A'..'Z'

alpha :: MonadGen m => m Char

Generates an ASCII letter: 'a'..'z', 'A'..'Z'

alphaNum :: MonadGen m => m Char

Generates an ASCII letter or digit: 'a'..'z', 'A'..'Z', '0'..'9'

ascii :: MonadGen m => m Char

Generates an ASCII character: '0'..'127'

latin1 :: MonadGen m => m Char

Generates a Latin-1 character: '0'..'255'

unicode :: MonadGen m => m Char

Generates a Unicode character, excluding noncharacters and invalid standalone surrogates: '0'..'1114111' (excluding '55296'..'57343', '65534', '65535')

unicodeAll :: MonadGen m => m Char

Generates a Unicode character, including noncharacters and invalid standalone surrogates: '0'..'1114111'

Strings

string :: MonadGen m => Range Int -> m Char -> m String

Generates a string using Range to determine the length.

This is a specialization of list, offered for convenience.

text :: MonadGen m => Range Int -> m Char -> m Text

Generates a string using Range to determine the length.

utf8 :: MonadGen m => Range Int -> m Char -> m ByteString

Generates a UTF-8 encoded string, using Range to determine the length.

bytes :: MonadGen m => Range Int -> m ByteString

Generates a random ByteString, using Range to determine the length.

Choice

constant :: MonadGen m => a -> m a

Trivial generator that always produces the same element.

This is another name for pure / return.

element :: (HasCallStack, Foldable f, MonadGen m) => f a -> m a

Randomly selects one of the elements in the list.

This generator shrinks towards the first element in the list.

The input list must be non-empty.

element_ :: (HasCallStack, MonadGen m) => [a] -> m a

Randomly selects one of the elements in the list.

This generator does not shrink the choice of element.

The input list must be non-empty.

choice :: (HasCallStack, MonadGen m) => [m a] -> m a

Randomly selects one of the generators in the list.

This generator shrinks towards the first generator in the list.

The input list must be non-empty.

frequency :: (HasCallStack, MonadGen m) => [(Int, m a)] -> m a

Uses a weighted distribution to randomly select one of the generators in the list.

This generator shrinks towards the first generator in the list.

The input list must be non-empty.

recursive :: MonadGen m => ([m a] -> m a) -> [m a] -> [m a] -> m a

Modifies combinators which choose from a list of generators, like choice or frequency, so that they can be used in recursive scenarios.

This combinator modifies its target to select one of the generators in either the non-recursive or the recursive list. When a selection is made from the recursive list, the Size is halved. When the Size gets to one or less, selections are no longer made from the recursive list, this ensures termination.

A good example of where this might be useful is abstract syntax trees:

data Expr =
    Var String
  | Lam String Expr
  | App Expr Expr

-- Assuming we have a name generator
genName :: MonadGen m => m String

-- We can write a generator for expressions
genExpr :: MonadGen m => m Expr
genExpr =
  Gen.recursive Gen.choice [
      -- non-recursive generators
      Var <$> genName
    ] [
      -- recursive generators
      Gen.subtermM genExpr (x -> Lam <$> genName <*> pure x)
    , Gen.subterm2 genExpr genExpr App
    ]

If we wrote the above example using only choice, it is likely that it would fail to terminate. This is because for every call to genExpr, there is a 2 in 3 chance that we will recurse again.

Conditional

discard :: MonadGen m => m a

Discards the whole generator.

ensure :: MonadGen m => (a -> Bool) -> m a -> m a

Discards the generator if the generated value does not satisfy the predicate.

filter :: (MonadGen m, GenBase m ~ Identity) => (a -> Bool) -> m a -> m a

Generates a value that satisfies a predicate.

Shrinks of the generated value will also satisfy the predicate. From the original generator's shrink tree, any values that fail the predicate will be removed, but any subsequent shrinks that satisfy it will be retained. Compared to filter, shrinking may be slower but will be optimal.

It's possible that the predicate will never pass, or will only pass at a larger size than we're currently running at. To avoid looping forever, we limit the number of retries, and grow the size with each retry. If we retry too many times then the whole generator is discarded.

mapMaybe :: (MonadGen m, GenBase m ~ Identity) => (a -> Maybe b) -> m a -> m b

Generates a value which is the result of the given function returning a Just.

The original generator's shrink tree will be retained, with values returning Nothing removed. Subsequent shrinks of those values will be retained. Compared to mapMaybeT, shrinking may be slower but will be optimal.

It's possible that the function will never return Just, or will only do so a larger size than we're currently running at. To avoid looping forever, we limit the number of retries, and grow the size with each retry. If we retry too many times then the whole generator is discarded.

filterT :: MonadGen m => (a -> Bool) -> m a -> m a

Generates a value that satisfies a predicate.

Shrinks of the generated value will also satisfy the predicate. From the original generator's shrink tree, any values that fail the predicate will be removed, along with their subsequent shrinks. Compared to filter, shrinking may be faster but may also be less optimal.

The type is also more general, because the shrink behavior from filter would force the entire shrink tree to be evaluated when applied to an impure tree.

This is essentially:

  filterT p gen = mfilter p gen <|> filterT p gen

But that could loop forever, if the predicate will never pass or will only pass at a larger size than we're currently running at. We differ from the above in keeping some state to avoid that. We limit the number of retries, and grow the size with each retry. If we retry too many times then the whole generator is discarded.

mapMaybeT :: MonadGen m => (a -> Maybe b) -> m a -> m b

Generates a value which is the result of the given function returning a Just.

The original generator's shrink tree will be retained, with values returning Nothing removed. Subsequent shrinks of those values will be retained. Compared to mapMaybeT, shrinking may be slower but will be optimal.

The type is also more general, because the shrink behavior from mapMaybe would force the entire shrink tree to be evaluated when applied to an impure tree.

It's possible that the function will never return Just, or will only do so a larger size than we're currently running at. To avoid looping forever, we limit the number of retries, and grow the size with each retry. If we retry too many times then the whole generator is discarded.

just :: (MonadGen m, GenBase m ~ Identity) => m (Maybe a) -> m a

Runs a Maybe generator until it produces a Just.

This is implemented using filter and has the same caveats.

justT :: MonadGen m => m (Maybe a) -> m a

Runs a Maybe generator until it produces a Just.

This is implemented using filter and has the same caveats.

Collections

maybe :: MonadGen m => m a -> m (Maybe a)

Generates a Nothing some of the time.

either :: MonadGen m => m a -> m b -> m (Either a b)

Generates either an a or a b.

As the size grows, this generator generates Rights more often than Lefts.

either_ :: MonadGen m => m a -> m b -> m (Either a b)

Generates either an a or a b, without bias.

This generator generates as many Rights as it does Lefts.

list :: MonadGen m => Range Int -> m a -> m [a]

Generates a list using a Range to determine the length.

seq :: MonadGen m => Range Int -> m a -> m (Seq a)

Generates a seq using a Range to determine the length.

nonEmpty :: MonadGen m => Range Int -> m a -> m (NonEmpty a)

Generates a non-empty list using a Range to determine the length.

set :: (MonadGen m, Ord a) => Range Int -> m a -> m (Set a)

Generates a set using a Range to determine the length.

This may fail to generate anything if the element generator cannot produce a large enough number of unique items to satify the required set size.

map :: (MonadGen m, Ord k) => Range Int -> m (k, v) -> m (Map k v)

Generates a map using a Range to determine the length.

This may fail to generate anything if the keys produced by the generator do not account for a large enough number of unique items to satify the required map size.

Subterms

freeze :: MonadGen m => m a -> m (a, m a)

Freeze the size and seed used by a generator, so we can inspect the value which it will produce.

This is used for implementing list and subtermMVec. It allows us to shrink the list itself before trying to shrink the values inside the list.

subterm :: MonadGen m => m a -> (a -> a) -> m a

Constructs a generator from a sub-term generator.

Shrinks to the sub-term if possible.

subtermM :: MonadGen m => m a -> (a -> m a) -> m a

Constructs a generator from a sub-term generator.

Shrinks to the sub-term if possible.

subterm2 :: MonadGen m => m a -> m a -> (a -> a -> a) -> m a

Constructs a generator from two sub-term generators.

Shrinks to one of the sub-terms if possible.

subtermM2 :: MonadGen m => m a -> m a -> (a -> a -> m a) -> m a

Constructs a generator from two sub-term generators.

Shrinks to one of the sub-terms if possible.

subterm3 :: MonadGen m => m a -> m a -> m a -> (a -> a -> a -> a) -> m a

Constructs a generator from three sub-term generators.

Shrinks to one of the sub-terms if possible.

subtermM3 :: MonadGen m => m a -> m a -> m a -> (a -> a -> a -> m a) -> m a

Constructs a generator from three sub-term generators.

Shrinks to one of the sub-terms if possible.

Combinations & Permutations

subsequence :: MonadGen m => [a] -> m [a]

Generates a random subsequence of a list.

For example:

Gen.print (Gen.subsequence [1..5])

=== Outcome ===
[1,2,4]
=== Shrinks ===
[]
[2,4]
[1,4]
[1,2]

subset :: MonadGen m => Set a -> m (Set a)

Generates a random subset of a set.

This shrinks towards the empty set.

shuffle :: MonadGen m => [a] -> m [a]

Generates a random permutation of a list.

This shrinks towards the order of the list being identical to the input list.

shuffleSeq :: MonadGen m => Seq a -> m (Seq a)

Generates a random permutation of a sequence.

This shrinks towards the order of the sequence being identical to the input sequence.

Sampling Generators

sample :: (HasCallStack, MonadIO m) => Gen a -> m a

Generate a sample from a generator.

This function is useful for examining a Gen in GHCi or other contexts. It is not appropriate for use in a test suite directly. You will only get a single sample from this function, and it will not give you a property test. The seed is random, so the test is not deterministic.

If you only want a single test to run, then use withTests 1:

prop_OnlyRunOnce :: Property
prop_OnlyRunOnce =
  withTests 1 $ property $ do
    i <- Gen.int
    i /== 0

print :: (MonadIO m, Show a) => Gen a -> m ()

Run a generator with a random seed and print the outcome, and the first level of shrinks.

Gen.print (Gen.enum 'a' 'f')

=== Outcome ===
'd'
=== Shrinks ===
'a'
'b'
'c'

printTree :: (MonadIO m, Show a) => Gen a -> m ()

Run a generator with a random seed and print the resulting shrink tree.

Gen.printTree (Gen.enum 'a' 'f')

'd'
 ├╼'a'
 ├╼'b'
 │  └╼'a'
 └╼'c'
    ├╼'a'
    └╼'b'
       └╼'a'

This may not terminate when the tree is very large.

printWith :: (MonadIO m, Show a) => Size -> Seed -> Gen a -> m ()

Print the value produced by a generator, and the first level of shrinks, for the given size and seed.

Use print to generate a value from a random seed.

printTreeWith :: (MonadIO m, Show a) => Size -> Seed -> Gen a -> m ()

Print the shrink tree produced by a generator, for the given size and seed.

Use printTree to generate a value from a random seed.

renderTree :: Show a => Size -> Seed -> Gen a -> String

Render the shrink tree produced by a generator, for the given size and seed.

Internal

These functions are exported in case you need them in a pinch, but are not part of the public API and may change at any time, even as part of a minor update.

Transfomer

runGenT :: Size -> Seed -> GenT m a -> TreeT (MaybeT m) a

Runs a generator, producing its shrink tree.

evalGen :: Size -> Seed -> Gen a -> Maybe (Tree a)

Run a generator, producing its shrink tree.

Nothing means discarded, Just means we have a value.

evalGenT :: Monad m => Size -> Seed -> GenT m a -> TreeT m (Maybe a)

Runs a generator, producing its shrink tree.

mapGenT :: (TreeT (MaybeT m) a -> TreeT (MaybeT n) b) -> GenT m a -> GenT n b

Map over a generator's shrink tree.

generate :: MonadGen m => (Size -> Seed -> a) -> m a

Generate a value with no shrinks from a Size and a Seed.

toTree :: forall m a. (MonadGen m, GenBase m ~ Identity) => m a -> m (Tree a)

Observe a generator's shrink tree.

toTreeMaybeT :: MonadGen m => m a -> m (TreeT (MaybeT (GenBase m)) a)

Lift a predefined shrink tree in to a generator, ignoring the seed and the size.

fromTree :: MonadGen m => Tree a -> m a

Lift a predefined shrink tree in to a generator, ignoring the seed and the size.

fromTreeT :: MonadGen m => TreeT (GenBase m) a -> m a

Lift a predefined shrink tree in to a generator, ignoring the seed and the size.

fromTreeMaybeT :: MonadGen m => TreeT (MaybeT (GenBase m)) a -> m a

Lift a predefined shrink tree in to a generator, ignoring the seed and the size.

runDiscardEffect :: TreeT (MaybeT Identity) a -> Maybe (Tree a)

Lazily run the discard effects through the tree and reify it a Maybe (Tree a).

Nothing means discarded, Just means we have a value.

Discards in the child nodes of the tree are simply removed.

runDiscardEffectT :: Monad m => TreeT (MaybeT m) a -> TreeT m (Maybe a)

Run the discard effects through the tree and reify them as Maybe values at the nodes.

Nothing means discarded, Just means we have a value.

Size

golden :: Size -> Size

Scale a size using the golden ratio.

golden x = x / φ
golden x = x / 1.61803..

Shrinking

atLeast :: Int -> [a] -> Bool

Check that list contains at least a certain number of elements.

Characters

isSurrogate :: Char -> Bool

Check if a character is in the surrogate category.

isNoncharacter :: Char -> Bool

Check if a character is one of the noncharacters '65534', '65535'.

Subterms

data Vec n a where

Constructors

Nil :: Vec 'Z a

(:.) :: a -> Vec n a -> Vec ('S n) a infixr 5

data Nat

Constructors

Z

S Nat

subtermMVec :: MonadGen m => Vec n (m a) -> (Vec n a -> m a) -> m a

Constructs a generator from a number of sub-term generators.

Shrinks to one of the sub-terms if possible.