Hedgehog.Internal.Property

Property

data Property

A property test, along with some configurable limits like how many times to run the test.

Constructors

Property

Fields

propertyConfig :: !PropertyConfig

propertyTest :: PropertyT IO ()

newtype PropertyT m a

The property monad transformer allows both the generation of test inputs and the assertion of expectations.

Constructors

PropertyT

Fields

unPropertyT :: TestT (GenT m) a

newtype PropertyName

The name of a property.

Should be constructed using OverloadedStrings:

  "apples" :: PropertyName

Constructors

PropertyName

Fields

unPropertyName :: String

data PropertyConfig

Configuration for a property test.

Constructors

PropertyConfig

Fields

propertyDiscardLimit :: !DiscardLimit

propertyShrinkLimit :: !ShrinkLimit

propertyShrinkRetries :: !ShrinkRetries

propertyTerminationCriteria :: !TerminationCriteria

propertySkip :: Maybe Skip

If this is Nothing, we take the Skip from the environment variable HEDGEHOG_SKIP.

newtype TestLimit

The number of successful tests that need to be run before a property test is considered successful.

Can be constructed using numeric literals:

  200 :: TestLimit

Constructors

TestLimit Int

newtype TestCount

The number of tests a property ran successfully.

Constructors

TestCount Int

newtype DiscardLimit

The number of discards to allow before giving up.

Can be constructed using numeric literals:

  10000 :: DiscardLimit

Constructors

DiscardLimit Int

newtype DiscardCount

The number of tests a property had to discard.

Constructors

DiscardCount Int

newtype ShrinkLimit

The number of shrinks to try before giving up on shrinking.

Can be constructed using numeric literals:

  1000 :: ShrinkLimit

Constructors

ShrinkLimit Int

newtype ShrinkCount

The numbers of times a property was able to shrink after a failing test.

Constructors

ShrinkCount Int

data Skip

Where to start running a property's tests.

Constructors

SkipNothing
Don't skip anything.

SkipToTest TestCount DiscardCount
Skip to a specific test number. If it fails, shrink as normal. If it passes, move on to the next test. Coverage checks are disabled.

We also need to count discards, since failing "after 7 tests" points at a different generated value than failing "after 7 tests and 5 discards".

SkipToShrink TestCount DiscardCount ShrinkPath
Skip to a specific test number and shrink state. If it fails, stop without shrinking further. If it passes, the property will pass without running any more tests.

Due to implementation details, all intermediate shrink states - those on the direct path from the original test input to the target state - will be tested too, and their results discarded.

newtype ShrinkPath

The path taken to reach a shrink state.

Constructors

ShrinkPath [Int]

newtype ShrinkRetries

The number of times to re-run a test during shrinking. This is useful if you are testing something which fails non-deterministically and you want to increase the change of getting a good shrink.

If you are doing parallel state machine testing, you should probably set shrink retries to something like 10. This will mean that during shrinking, a parallel test case requires 10 successful runs before it is passes and we try a different shrink.

Can be constructed using numeric literals:

  0 :: ShrinkRetries

Constructors

ShrinkRetries Int

withTests :: TestLimit -> Property -> Property

Set the number of times a property should be executed before it is considered successful.

If you have a test that does not involve any generators and thus does not need to run repeatedly, you can use withTests 1 to define a property that will only be checked once.

withDiscards :: DiscardLimit -> Property -> Property

Set the number of times a property is allowed to discard before the test runner gives up.

withShrinks :: ShrinkLimit -> Property -> Property

Set the number of times a property is allowed to shrink before the test runner gives up and prints the counterexample.

withRetries :: ShrinkRetries -> Property -> Property

Set the number of times a property will be executed for each shrink before the test runner gives up and tries a different shrink. See ShrinkRetries for more information.

withSkip :: Skip -> Property -> Property

Set the target that a property will skip to before it starts to run.

property :: HasCallStack => PropertyT IO () -> Property

Creates a property with the default configuration.

test :: Monad m => TestT m a -> PropertyT m a

Lift a test in to a property.

Because both TestT and PropertyT have MonadTest instances, this function is not often required. It can however be useful for writing functions directly in TestT and thus gaining a MonadTransControl instance at the expense of not being able to generate additional inputs using forAll.

An example where this is useful is parallel state machine testing, as executeParallel requires MonadBaseControl IO in order to be able to spawn threads in MonadTest.

forAll :: (Monad m, Show a, HasCallStack) => Gen a -> PropertyT m a

Generates a random input for the test by running the provided generator.

forAllT :: (Monad m, Show a, HasCallStack) => GenT m a -> PropertyT m a

Generates a random input for the test by running the provided generator.

forAllWith :: (Monad m, HasCallStack) => (a -> String) -> Gen a -> PropertyT m a

Generates a random input for the test by running the provided generator.

This is a the same as forAll but allows the user to provide a custom rendering function. This is useful for values which don't have a Show instance.

forAllWithT :: (Monad m, HasCallStack) => (a -> String) -> GenT m a -> PropertyT m a

Generates a random input for the test by running the provided generator.

This is a the same as forAllT but allows the user to provide a custom rendering function. This is useful for values which don't have a Show instance.

defaultMinTests :: TestLimit

The minimum amount of tests to run for a Property

discard :: Monad m => PropertyT m a

Discards the current test entirely.

skipCompress :: Skip -> String

Compress a Skip into a hopefully-short alphanumeric string.

The bit that might be long is the ShrinkPath in SkipToShrink. For that, we encode the path components in base 26, alternating between uppercase and lowercase alphabets to distinguish list elements. Additionally when we have runs of equal components, we use the normal base 10 encoding to indicate the length.

This gives something which is hopefully quite short, but a human can roughly interpret it by eyeball.

shrinkPathCompress :: ShrinkPath -> String

Compress a ShrinkPath into a hopefully-short alphanumeric string.

We encode the path components in base 26, alternating between uppercase and lowercase alphabets to distinguish list elements. Additionally when we have runs of equal components, we use the normal base 10 encoding to indicate the length.

skipDecompress :: String -> Maybe Skip

Decompress a Skip.

This satisfies

  skipDecompress (skipCompress a) == Just a

shrinkPathDecompress :: String -> Maybe ShrinkPath

Decompress a ShrinkPath.

This satisfies

  shrinkPathDecompress (shrinkPathCompress a) == Just a

Group

data Group

A named collection of property tests.

Constructors

Group

Fields

groupName :: !GroupName

groupProperties :: ![(PropertyName, Property)]

newtype GroupName

The name of a group of properties.

Should be constructed using OverloadedStrings:

  "fruit" :: GroupName

Constructors

GroupName

Fields

unGroupName :: String

newtype PropertyCount

The number of properties in a group.

Constructors

PropertyCount Int

TestT

class Monad m => MonadTest m where

Methods

liftTest :: Test a -> m a

type Test = TestT Identity

A test monad allows the assertion of expectations.

newtype TestT m a

A test monad transformer allows the assertion of expectations.

Constructors

TestT

Fields

unTest :: ExceptT Failure (WriterT Journal m) a

data Log

Log messages which are recorded during a test run.

Constructors

Annotation (Maybe Span) String

Footnote String

Label (Label Cover)

newtype Journal

A record containing the details of a test run.

Constructors

Journal

Fields

journalLogs :: [Log]

data Failure

Details on where and why a test failed.

Constructors

Failure (Maybe Span) String (Maybe Diff)

data Diff

The difference between some expected and actual value.

Constructors

Diff

Fields

diffPrefix :: String

diffRemoved :: String

diffInfix :: String

diffAdded :: String

diffSuffix :: String

diffValue :: ValueDiff

annotate :: (MonadTest m, HasCallStack) => String -> m ()

Annotates the source code with a message that might be useful for debugging a test failure.

annotateShow :: (MonadTest m, Show a, HasCallStack) => a -> m ()

Annotates the source code with a value that might be useful for debugging a test failure.

footnote :: MonadTest m => String -> m ()

Logs a message to be displayed as additional information in the footer of the failure report.

footnoteShow :: (MonadTest m, Show a) => a -> m ()

Logs a value to be displayed as additional information in the footer of the failure report.

failure :: (MonadTest m, HasCallStack) => m a

Causes a test to fail.

success :: MonadTest m => m ()

Another name for pure ().

assert :: (MonadTest m, HasCallStack) => Bool -> m ()

Fails the test if the condition provided is False.

diff :: (MonadTest m, Show a, Show b, HasCallStack) => a -> (a -> b -> Bool) -> b -> m ()

Fails the test and shows a git-like diff if the comparison operation evaluates to False when applied to its arguments.

The comparison function is the second argument, which may be counter-intuitive to Haskell programmers. However, it allows operators to be written infix for easy reading:

  diff y (<) 87
  diff x (<=) r

This function behaves like the unix diff tool, which gives a 0 exit code if the compared files are identical, or a 1 exit code code otherwise. Like unix diff, if the arguments fail the comparison, a /diff is shown.

(===) :: (MonadTest m, Eq a, Show a, HasCallStack) => a -> a -> m () infix 4

Fails the test if the two arguments provided are not equal.

(/==) :: (MonadTest m, Eq a, Show a, HasCallStack) => a -> a -> m () infix 4

Fails the test if the two arguments provided are equal.

eval :: (MonadTest m, HasCallStack) => a -> m a

Fails the test if the value throws an exception when evaluated to weak head normal form (WHNF).

evalNF :: (MonadTest m, NFData a, HasCallStack) => a -> m a

Fails the test if the value throws an exception when evaluated to normal form (NF).

evalM :: (MonadTest m, MonadCatch m, HasCallStack) => m a -> m a

Fails the test if the action throws an exception.

The benefit of using this over simply letting the exception bubble up is that the location of the closest evalM will be shown in the output.

evalIO :: (MonadTest m, MonadIO m, HasCallStack) => IO a -> m a

Fails the test if the IO action throws an exception.

The benefit of using this over liftIO is that the location of the exception will be shown in the output.

evalEither :: (MonadTest m, Show x, HasCallStack) => Either x a -> m a

Fails the test if the Either is Left, otherwise returns the value in the Right.

evalEitherM :: (MonadTest m, Show x, MonadCatch m, HasCallStack) => m (Either x a) -> m a

Fails the test if the action throws an exception, or if the Either is Left, otherwise returns the value in the Right.

evalExceptT :: (MonadTest m, Show x, HasCallStack) => ExceptT x m a -> m a

Fails the test if the ExceptT is Left, otherwise returns the value in the Right.

evalMaybe :: (MonadTest m, HasCallStack) => Maybe a -> m a

Fails the test if the Maybe is Nothing, otherwise returns the value in the Just.

evalMaybeM :: (MonadTest m, MonadCatch m, HasCallStack) => m (Maybe a) -> m a

Fails the test if the action throws an exception, or if the Maybe is Nothing, otherwise returns the value in the Just.

Coverage

newtype Coverage a

The extent to which all classifiers cover a test.

When a given classification's coverage does not exceed the required minimum, the test will be failed.

Constructors

Coverage

Fields

coverageLabels :: Map LabelName (Label a)

data Label a

The extent to which a test is covered by a classifier.

When a classifier's coverage does not exceed the required minimum, the test will be failed.

Constructors

MkLabel

Fields

labelName :: !LabelName

labelLocation :: !(Maybe Span)

labelMinimum :: !CoverPercentage

labelAnnotation :: !a

newtype LabelName

The name of a classifier.

Should be constructed using OverloadedStrings:

  "apples" :: LabelName

Constructors

LabelName

Fields

unLabelName :: String

cover :: (MonadTest m, HasCallStack) => CoverPercentage -> LabelName -> Bool -> m ()

Require a certain percentage of the tests to be covered by the classifier.

   prop_with_coverage :: Property
   prop_with_coverage =
     property $ do
       match <- forAll Gen.bool
       cover 30 "True" $ match
       cover 30 "False" $ not match

The example above requires a minimum of 30% coverage for both classifiers. If these requirements are not met, it will fail the test.

classify :: (MonadTest m, HasCallStack) => LabelName -> Bool -> m ()

Records the proportion of tests which satisfy a given condition.

   prop_with_classifier :: Property
   prop_with_classifier =
     property $ do
       xs <- forAll $ Gen.list (Range.linear 0 100) Gen.alpha
       for_ xs $ \x -> do
         classify "newborns" $ x == 0
         classify "children" $ x > 0 && x < 13
         classify "teens" $ x > 12 && x < 20

label :: (MonadTest m, HasCallStack) => LabelName -> m ()

Add a label for each test run. It produces a table showing the percentage of test runs that produced each label.

collect :: (MonadTest m, Show a, HasCallStack) => a -> m ()

Like label, but uses Show to render its argument for display.

coverPercentage :: TestCount -> CoverCount -> CoverPercentage

labelCovered :: TestCount -> Label CoverCount -> Bool

coverageSuccess :: TestCount -> Coverage CoverCount -> Bool

All labels are covered

coverageFailures :: TestCount -> Coverage CoverCount -> [Label CoverCount]

journalCoverage :: Journal -> Coverage CoverCount

data Cover

Whether a test is covered by a classifier, and therefore belongs to a Class.

Constructors

NoCover

Cover

newtype CoverCount

The total number of tests which are covered by a classifier.

Can be constructed using numeric literals:

  30 :: CoverCount

Constructors

CoverCount

Fields

unCoverCount :: Int

newtype CoverPercentage

The relative number of tests which are covered by a classifier.

Can be constructed using numeric literals:

  30 :: CoverPercentage

Constructors

CoverPercentage

Fields

unCoverPercentage :: Double

toCoverCount :: Cover -> CoverCount

Confidence

newtype Confidence

The acceptable occurrence of false positives

Example, Confidence 10^9 would mean that you'd accept a false positive for 1 in 10^9 tests.

Constructors

Confidence

Fields

unConfidence :: Int64

data TerminationCriteria

Constructors

EarlyTermination Confidence TestLimit

NoEarlyTermination Confidence TestLimit

NoConfidenceTermination TestLimit

confidenceSuccess :: TestCount -> Confidence -> Coverage CoverCount -> Bool

Is true when the test coverage satisfies the specified Confidence contstraint for all 'Coverage CoverCount's

confidenceFailure :: TestCount -> Confidence -> Coverage CoverCount -> Bool

Is true when there exists a label that is sure to have failed according to the Confidence constraint

withConfidence :: Confidence -> Property -> Property

Make sure that the result is statistically significant in accordance to the passed Confidence

verifiedTermination :: Property -> Property

defaultConfidence :: Confidence

The default confidence allows one false positive in 10^9 tests

Internal

These functions are exported in case you need them in a pinch, but are not part of the public API and may change at any time, even as part of a minor update.

defaultConfig :: PropertyConfig

The default configuration for a property test.

mapConfig :: (PropertyConfig -> PropertyConfig) -> Property -> Property

Map a config modification function over a property.

failDiff :: (MonadTest m, Show a, Show b, HasCallStack) => a -> b -> m ()

Fails with an error that shows the difference between two values.

failException :: (MonadTest m, HasCallStack) => SomeException -> m a

Fails with an error which renders the type of an exception and its error message.

failWith :: (MonadTest m, HasCallStack) => Maybe Diff -> String -> m a

Fail the test with an error message, useful for building other failure combinators.

writeLog :: MonadTest m => Log -> m ()

Log some information which might be relevant to a potential test failure.

mkTest :: (Either Failure a, Journal) -> Test a

mkTestT :: m (Either Failure a, Journal) -> TestT m a

runTest :: Test a -> (Either Failure a, Journal)

runTestT :: TestT m a -> m (Either Failure a, Journal)

wilsonBounds :: Integer -> Integer -> Double -> (Double, Double)