Hedgehog.Internal.Report

Report

data Summary

A summary of all the properties executed.

Constructors

Summary

Fields

summaryWaiting :: !PropertyCount

summaryRunning :: !PropertyCount

summaryFailed :: !PropertyCount

summaryGaveUp :: !PropertyCount

summaryOK :: !PropertyCount

data Report a

A report on a running or completed property test.

Constructors

Report

Fields

reportTests :: !TestCount

reportDiscards :: !DiscardCount

reportCoverage :: !(Coverage CoverCount)

reportSeed :: !Seed

reportStatus :: !a

data Progress

The status of a running property test.

Constructors

Running

Shrinking !FailureReport

data Result

The status of a completed property test.

In the case of a failure it provides the seed used for the test, the number of shrinks, and the execution log.

Constructors

Failed !FailureReport

GaveUp

OK

data FailureReport

Constructors

FailureReport

Fields

failureShrinks :: !ShrinkCount

failureShrinkPath :: !ShrinkPath

failureCoverage :: !(Maybe (Coverage CoverCount))

failureAnnotations :: ![FailedAnnotation]

failureLocation :: !(Maybe Span)

failureMessage :: !String

failureDiff :: !(Maybe Diff)

failureFootnotes :: ![String]

data FailedAnnotation

Constructors

FailedAnnotation

Fields

failedSpan :: !(Maybe Span)

failedValue :: !String

data Style

Constructors

StyleDefault

StyleAnnotation

StyleFailure

data Markup

Constructors

WaitingIcon

WaitingHeader

RunningIcon

RunningHeader

ShrinkingIcon

ShrinkingHeader

FailedIcon

FailedText

GaveUpIcon

GaveUpText

SuccessIcon

SuccessText

CoverageIcon

CoverageText

CoverageFill

DeclarationLocation

StyledLineNo !Style

StyledBorder !Style

StyledSource !Style

AnnotationGutter

AnnotationValue

FailureArrows

FailureGutter

FailureMessage

DiffPrefix

DiffInfix

DiffSuffix

DiffSame

DiffRemoved

DiffAdded

ReproduceHeader

ReproduceGutter

ReproduceSource

data Config

Constructors

Config

Fields

configContext :: Context

configPrintFailedAtLocation :: Bool

configPrintShrinkPath :: Bool

configPrintReproduceMessage :: Bool

configPrintPrefixIcons :: PrintPrefixIcons

defaultConfig :: Config

data Context

Constructors

FullContext

Context Lines

data Lines

data PrintPrefixIcons

Whether to add icons to the start of important output lines or not.

Constructors

DisablePrefixIcons
Do not add icons to the start of important output lines.

EnablePrefixIcons
Add icons to the start of important output lines.

renderProgress :: MonadIO m => UseColor -> Maybe PropertyName -> Report Progress -> m String

renderResult :: MonadIO m => UseColor -> Maybe PropertyName -> Report Result -> m String

renderResultWith :: MonadIO m => Config -> UseColor -> Maybe PropertyName -> Report Result -> m String

renderSummary :: MonadIO m => UseColor -> Summary -> m String

renderDoc :: MonadIO m => UseColor -> Doc Markup -> m String

ppProgress :: MonadIO m => Maybe PropertyName -> Report Progress -> m (Doc Markup)

ppResult :: MonadIO m => Maybe PropertyName -> Report Result -> m (Doc Markup)

ppResultWith :: MonadIO m => Config -> Maybe PropertyName -> Report Result -> m (Doc Markup)

ppSummary :: MonadIO m => Summary -> m (Doc Markup)

fromResult :: Result -> Summary

Construct a summary from a single result.

mkFailure :: ShrinkCount -> ShrinkPath -> Maybe (Coverage CoverCount) -> Maybe Span -> String -> Maybe Diff -> [Log] -> FailureReport