Network.DNS.Resolver

Description

Resolver related data types.

Configuration for resolver

data ResolvConf

defaultResolvConf :: ResolvConf

Accessors

resolvInfo :: ResolvConf -> FileOrNumericHost

resolvTimeout :: ResolvConf -> Int

resolvRetry :: ResolvConf -> Int

resolvConcurrent :: ResolvConf -> Bool

resolvCache :: ResolvConf -> Maybe CacheConf

resolvQueryControls :: ResolvConf -> QueryControls

Specifying DNS servers

data FileOrNumericHost

Constructors

RCFilePath FilePath

RCHostName HostName

RCHostNames [HostName]

RCHostPort HostName PortNumber

Configuring cache

data CacheConf

defaultCacheConf :: CacheConf

maximumTTL :: CacheConf -> TTL

pruningDelay :: CacheConf -> Int

Intermediate data type for resolver

data ResolvSeed

makeResolvSeed :: ResolvConf -> IO ResolvSeed

Make a ResolvSeed from a ResolvConf.

Examples:

>>> rs <- makeResolvSeed defaultResolvConf

Type and function for resolver

data Resolver

withResolver :: ResolvSeed -> (Resolver -> IO a) -> IO a

Giving a thread-safe Resolver to the function of the second argument.

withResolvers :: [ResolvSeed] -> ([Resolver] -> IO a) -> IO a

Deprecated: Use withResolver with resolvConcurrent set to True

Giving thread-safe Resolvers to the function of the second argument. For each Resolver, multiple lookups must be done sequentially. Resolvers can be used concurrently.