Network.HTTP.Client.Conduit

Description

A new, experimental API to replace Network.HTTP.Conduit.

For most users, Network.HTTP.Simple is probably a better choice. For more information, see:

https://haskell-lang.org/library/http-client

For more information on using this module, please be sure to read the documentation in the Network.HTTP.Client module.

Conduit-specific interface

withResponse :: (MonadUnliftIO m, MonadIO n, MonadReader env m, HasHttpManager env) => Request -> (Response (ConduitM i ByteString n ()) -> m a) -> m a

Conduit powered version of withResponse. Differences are:

Response body is represented as a Producer.

Generalized to any instance of MonadUnliftIO, not just IO.

The Manager is contained by a MonadReader context.

Since 2.1.0

responseOpen :: (MonadIO m, MonadIO n, MonadReader env m, HasHttpManager env) => Request -> m (Response (ConduitM i ByteString n ()))

Conduit-powered version of responseOpen.

See withResponse for the differences with responseOpen.

Since 2.1.0

responseClose :: MonadIO m => Response body -> m ()

Generalized version of responseClose.

Since 2.1.0

acquireResponse :: (MonadIO n, MonadReader env m, HasHttpManager env) => Request -> m (Acquire (Response (ConduitM i ByteString n ())))

An Acquire for getting a Response.

Since 2.1.0

httpSource :: (MonadResource m, MonadIO n, MonadReader env m, HasHttpManager env) => Request -> (Response (ConduitM () ByteString n ()) -> ConduitM () r m ()) -> ConduitM () r m ()

Same as httpSource, but uses Manager from Reader environment instead of the global one.

Since 2.3.6

Manager helpers

defaultManagerSettings :: ManagerSettings

TLS-powered manager settings.

Since 2.1.0

newManager :: MonadIO m => m Manager

Get a new manager using defaultManagerSettings.

Since 2.1.0

newManagerSettings :: MonadIO m => ManagerSettings -> m Manager

Get a new manager using the given settings.

Since 2.1.0

General HTTP client interface

data Proxy

Constructors

Proxy

Fields

proxyHost :: ByteString

proxyPort :: Int

data RequestBody

Constructors

RequestBodyLBS ByteString

RequestBodyBS ByteString

RequestBodyBuilder Int64 Builder

RequestBodyStream Int64 (GivesPopper ())

RequestBodyStreamChunked (GivesPopper ())

RequestBodyIO (IO RequestBody)

data Request

data Response body

data Manager

data ManagerSettings

data ResponseTimeout

data Cookie

Constructors

Cookie

Fields

cookie_name :: ByteString

cookie_value :: ByteString

cookie_expiry_time :: UTCTime

cookie_domain :: ByteString

cookie_path :: ByteString

cookie_creation_time :: UTCTime

cookie_last_access_time :: UTCTime

cookie_persistent :: Bool

cookie_host_only :: Bool

cookie_secure_only :: Bool

cookie_http_only :: Bool

data CookieJar

data HttpException

Constructors

HttpExceptionRequest Request HttpExceptionContent

InvalidUrlException String String

data HttpExceptionContent

Constructors

StatusCodeException (Response ()) ByteString

TooManyRedirects [Response ByteString]

OverlongHeaders

TooManyHeaderFields

ResponseTimeout

ConnectionTimeout

ConnectionFailure SomeException

InvalidStatusLine ByteString

InvalidHeader ByteString

InvalidRequestHeader ByteString

InternalException SomeException

ProxyConnectException ByteString Int Status

NoResponseDataReceived

TlsNotSupported

WrongRequestBodyStreamSize Word64 Word64

ResponseBodyTooShort Word64 Word64

InvalidChunkHeaders

IncompleteHeaders

InvalidDestinationHost ByteString

HttpZlibException ZlibException

InvalidProxyEnvironmentVariable Text Text

ConnectionClosed

InvalidProxySettings Text

class HasHttpManager a where

Methods

getHttpManager :: a -> Manager

type GivesPopper a = NeedsPopper a -> IO a

type NeedsPopper a = Popper -> IO a

type Popper = IO ByteString

data ProxyOverride

data StreamFileStatus

Constructors

StreamFileStatus

Fields

fileSize :: Int64

readSoFar :: Int64

thisChunkSize :: Int

data HistoriedResponse body

setQueryString :: [(ByteString, Maybe ByteString)] -> Request -> Request

responseTimeoutMicro :: Int -> ResponseTimeout

responseTimeoutNone :: ResponseTimeout

responseTimeoutDefault :: ResponseTimeout

createCookieJar :: [Cookie] -> CookieJar

destroyCookieJar :: CookieJar -> [Cookie]

parseUrl :: MonadThrow m => String -> m Request

parseUrlThrow :: MonadThrow m => String -> m Request

parseRequest :: MonadThrow m => String -> m Request

parseRequest_ :: String -> Request

defaultRequest :: Request

applyBasicAuth :: ByteString -> ByteString -> Request -> Request

urlEncodedBody :: [(ByteString, ByteString)] -> Request -> Request

setRequestIgnoreStatus :: Request -> Request

setRequestCheckStatus :: Request -> Request

brReadSome :: BodyReader -> Int -> IO ByteString

makeConnection :: IO ByteString -> (ByteString -> IO ()) -> IO () -> IO Connection

socketConnection :: Socket -> Int -> IO Connection

strippedHostName :: String -> String

computeCookieString :: Request -> CookieJar -> UTCTime -> Bool -> (ByteString, CookieJar)

defaultPath :: Request -> ByteString

domainMatches :: ByteString -> ByteString -> Bool

evictExpiredCookies :: CookieJar -> UTCTime -> CookieJar

generateCookie :: SetCookie -> Request -> UTCTime -> Bool -> Maybe Cookie

insertCheckedCookie :: Cookie -> CookieJar -> Bool -> CookieJar

insertCookiesIntoRequest :: Request -> CookieJar -> UTCTime -> (Request, CookieJar)

isIpAddress :: ByteString -> Bool

isPotentiallyTrustworthyOrigin :: Bool -> ByteString -> Bool

pathMatches :: ByteString -> ByteString -> Bool

receiveSetCookie :: SetCookie -> Request -> UTCTime -> Bool -> CookieJar -> CookieJar

removeExistingCookieFromCookieJar :: Cookie -> CookieJar -> (Maybe Cookie, CookieJar)

updateCookieJar :: Response a -> Request -> UTCTime -> CookieJar -> (CookieJar, Response a)

withConnection :: Request -> Manager -> (Connection -> IO a) -> IO a

defaultProxy :: ProxyOverride

noProxy :: ProxyOverride

proxyEnvironment :: Maybe Proxy -> ProxyOverride

proxyEnvironmentNamed :: Text -> Maybe Proxy -> ProxyOverride

proxyFromRequest :: ProxyOverride

rawConnectionModifySocket :: (Socket -> IO ()) -> IO (Maybe HostAddress -> String -> Int -> IO Connection)

rawConnectionModifySocketSize :: (Socket -> IO ()) -> IO (Int -> Maybe HostAddress -> String -> Int -> IO Connection)

useProxy :: Proxy -> ProxyOverride

useProxySecureWithoutConnect :: Proxy -> ProxyOverride

withManager :: ManagerSettings -> (Manager -> IO a) -> IO a

applyBasicProxyAuth :: ByteString -> ByteString -> Request -> Request

applyBearerAuth :: ByteString -> Request -> Request

getUri :: Request -> URI

observedStreamFile :: (StreamFileStatus -> IO ()) -> FilePath -> IO RequestBody

requestFromURI :: MonadThrow m => URI -> m Request

requestFromURI_ :: URI -> Request

setQueryStringPartialEscape :: [(ByteString, [EscapeItem])] -> Request -> Request

streamFile :: FilePath -> IO RequestBody

throwErrorStatusCodes :: MonadIO m => Request -> Response BodyReader -> m ()

getOriginalRequest :: Response a -> Request

compareCookies :: Cookie -> Cookie -> Ordering

equalCookie :: Cookie -> Cookie -> Bool

equalCookieJar :: CookieJar -> CookieJar -> Bool

equivCookie :: Cookie -> Cookie -> Bool

equivCookieJar :: CookieJar -> CookieJar -> Bool

managerSetInsecureProxy :: ProxyOverride -> ManagerSettings -> ManagerSettings

managerSetMaxHeaderLength :: Int -> ManagerSettings -> ManagerSettings

managerSetMaxNumberHeaders :: Int -> ManagerSettings -> ManagerSettings

managerSetProxy :: ProxyOverride -> ManagerSettings -> ManagerSettings

managerSetSecureProxy :: ProxyOverride -> ManagerSettings -> ManagerSettings

responseOpenHistory :: Request -> Manager -> IO (HistoriedResponse BodyReader)

withResponseHistory :: Request -> Manager -> (HistoriedResponse BodyReader -> IO a) -> IO a

httpLbs :: (MonadIO m, HasHttpManager env, MonadReader env m) => Request -> m (Response ByteString)

Same as httpLbs, except it uses the Manager in the reader environment.

Since 2.1.1

httpNoBody :: (MonadIO m, HasHttpManager env, MonadReader env m) => Request -> m (Response ())

Same as httpNoBody, except it uses the Manager in the reader environment.

This can be more convenient that using withManager as it avoids the need to specify the base monad for the response body.

Since 2.1.2

Lower-level conduit functions

requestBodySource :: Int64 -> ConduitM () ByteString IO () -> RequestBody

requestBodySourceChunked :: ConduitM () ByteString IO () -> RequestBody

bodyReaderSource :: MonadIO m => BodyReader -> ConduitM i ByteString m ()