Network.HTTP.Simple

Description

Simplified interface for common HTTP client interactions. Tutorial available at https://haskell-lang.org/library/http-client

Important note: Request is an instance of IsString, and therefore recommended usage is to turn on OverloadedStrings, e.g.

{-# LANGUAGE OverloadedStrings #-}
import Network.HTTP.Simple
import qualified Data.ByteString.Char8 as B8

main :: IO ()
main = httpBS "http://example.com" >>= B8.putStrLn . getResponseBody

The IsString instance uses parseRequest behind the scenes and inherits its behavior.

Perform requests

httpBS :: MonadIO m => Request -> m (Response ByteString)

Perform an HTTP request and return the body as a ByteString.

Since: 2.2.4

httpLBS :: MonadIO m => Request -> m (Response ByteString)

Perform an HTTP request and return the body as a lazy ByteString. Note that the entire value will be read into memory at once (no lazy I/O will be performed). The advantage of a lazy ByteString here (versus using httpBS) is--if needed--a better in-memory representation.

Since: 2.1.10

httpNoBody :: MonadIO m => Request -> m (Response ())

Perform an HTTP request and ignore the response body.

Since: 2.2.2

httpJSON :: (MonadIO m, FromJSON a) => Request -> m (Response a)

Perform an HTTP request and parse the body as JSON. In the event of an JSON parse errors, a JSONException runtime exception will be thrown.

NOTE: Depends on the aeson cabal flag being enabled

Since: 2.1.10

httpJSONEither :: (MonadIO m, FromJSON a) => Request -> m (Response (Either JSONException a))

Perform an HTTP request and parse the body as JSON. In the event of an JSON parse errors, a Left value will be returned.

NOTE: Depends on the aeson cabal flag being enabled

Since: 2.1.10

httpSink :: MonadUnliftIO m => Request -> (Response () -> ConduitM ByteString Void m a) -> m a

Perform an HTTP request and consume the body with the given Sink

Since: 2.1.10

httpSource :: (MonadResource m, MonadIO n) => Request -> (Response (ConduitM i ByteString n ()) -> ConduitM i o m r) -> ConduitM i o m r

Perform an HTTP request, and get the response body as a Source.

The second argument to this function tells us how to make the Source from the Response itself. This allows you to perform actions with the status or headers, for example, in addition to the raw bytes themselves. If you just care about the response body, you can use getResponseBody as the second argument here.

{-# LANGUAGE OverloadedStrings #-}
import           Control.Monad.IO.Class       (liftIO)
import           Control.Monad.Trans.Resource (runResourceT)
import           Data.Conduit                 (($$))
import qualified Data.Conduit.Binary          as CB
import qualified Data.Conduit.List            as CL
import           Network.HTTP.Simple
import           System.IO                    (stdout)

main :: IO ()
main =
    runResourceT
        $ httpSource "http://httpbin.org/robots.txt" getSrc
       $$ CB.sinkHandle stdout
  where
    getSrc res = do
        liftIO $ print (getResponseStatus res, getResponseHeaders res)
        getResponseBody res

Since: 2.2.1

withResponse :: (MonadUnliftIO m, MonadIO n) => Request -> (Response (ConduitM i ByteString n ()) -> m a) -> m a

Perform an action with the given request. This employes the bracket pattern.

This is similar to httpSource, but does not require MonadResource and allows the result to not contain a ConduitM value.

Since: 2.2.3

Types

type Header = (HeaderName, ByteString)

type Query = [QueryItem]

type QueryItem = (ByteString, Maybe ByteString)

data Request

type RequestHeaders = [Header]

data Response body

type ResponseHeaders = [Header]

data JSONException

An exception that can occur when parsing JSON

NOTE: Depends on the aeson cabal flag being enabled

Since: 2.1.10

Constructors

JSONParseException Request (Response ()) ParseError

JSONConversionException Request (Response Value) String

data HttpException

Constructors

HttpExceptionRequest Request HttpExceptionContent

InvalidUrlException String String

data Proxy

Constructors

Proxy

Fields

proxyHost :: ByteString

proxyPort :: Int

Request constructions

defaultRequest :: Request

parseRequest :: MonadThrow m => String -> m Request

parseRequest_ :: String -> Request

parseRequestThrow :: MonadThrow m => String -> m Request

Same as parseRequest, except will throw an HttpException in the event of a non-2XX response. This uses throwErrorStatusCodes to implement checkResponse.

Exactly the same as parseUrlThrow, but has a name that is more consistent with the other parseRequest functions.

Since: 2.3.2

parseRequestThrow_ :: String -> Request

Same as parseRequestThrow, but parse errors cause an impure exception. Mostly useful for static strings which are known to be correctly formatted.

Since: 2.3.2

Request lenses

Basics

setRequestMethod :: ByteString -> Request -> Request

Set the request method

Since: 2.1.10

setRequestSecure :: Bool -> Request -> Request

Set whether this is a secureHTTPS (True) or insecureHTTP (False) request

Since: 2.1.10

setRequestHost :: ByteString -> Request -> Request

Set the destination host of the request

Since: 2.1.10

setRequestPort :: Int -> Request -> Request

Set the destination port of the request

Since: 2.1.10

setRequestPath :: ByteString -> Request -> Request

Lens for the requested path info of the request

Since: 2.1.10

addRequestHeader :: HeaderName -> ByteString -> Request -> Request

Add a request header name/value combination

Since: 2.1.10

getRequestHeader :: HeaderName -> Request -> [ByteString]

Get all request header values for the given name

Since: 2.1.10

setRequestHeader :: HeaderName -> [ByteString] -> Request -> Request

Set the given request header to the given list of values. Removes any previously set header values with the same name.

Since: 2.1.10

setRequestHeaders :: RequestHeaders -> Request -> Request

Set the request headers, wiping out all previously set headers. This means if you use setRequestHeaders to set some headers and also use one of the other setters that modifies the content-type header (such as setRequestBodyJSON), be sure that setRequestHeaders is evaluated first.

Since: 2.1.10

setRequestQueryString :: Query -> Request -> Request

Set the query string parameters

Since: 2.1.10

getRequestQueryString :: Request -> Query

Get the query string parameters

Since: 2.1.10

addToRequestQueryString :: Query -> Request -> Request

Add to the existing query string parameters.

Since: 2.3.5

Request body

setRequestBody :: RequestBody -> Request -> Request

Set the request body to the given RequestBody. You may want to consider using one of the convenience functions in the modules, e.g. requestBodyJSON.

Note: This will not modify the request method. For that, please use requestMethod. You likely don't want the default of GET.

Since: 2.1.10

setRequestBodyJSON :: ToJSON a => a -> Request -> Request

Set the request body as a JSON value

Note: This will not modify the request method. For that, please use requestMethod. You likely don't want the default of GET.

This also sets the Content-Type to application/json; charset=utf-8

NOTE: Depends on the aeson cabal flag being enabled

Since: 2.1.10

setRequestBodyLBS :: ByteString -> Request -> Request

Set the request body as a lazy ByteString

Note: This will not modify the request method. For that, please use requestMethod. You likely don't want the default of GET.

Since: 2.1.10

setRequestBodySource

Arguments

:: Int64
length of source

-> ConduitM () ByteString IO ()

-> Request

-> Request

Set the request body as a Source

Note: This will not modify the request method. For that, please use requestMethod. You likely don't want the default of GET.

Since: 2.1.10

setRequestBodyFile :: FilePath -> Request -> Request

Set the request body as a file

Note: This will not modify the request method. For that, please use requestMethod. You likely don't want the default of GET.

Since: 2.1.10

setRequestBodyURLEncoded :: [(ByteString, ByteString)] -> Request -> Request

Set the request body as URL encoded data

Note: This will change the request method to POST and set the content-type to application/x-www-form-urlencoded

Since: 2.1.10

Special fields

setRequestIgnoreStatus :: Request -> Request

setRequestCheckStatus :: Request -> Request

setRequestBasicAuth

Arguments

:: ByteString
username

-> ByteString
password

-> Request

-> Request

Set basic auth with the given username and password

Since: 2.1.10

setRequestBearerAuth

Arguments

:: ByteString
token

-> Request

-> Request

Set bearer auth with the given token

Since: 2.3.8

setRequestManager :: Manager -> Request -> Request

Instead of using the default global Manager, use the supplied Manager.

Since: 2.1.10

setRequestProxy :: Maybe Proxy -> Request -> Request

Override the default proxy server settings

Since: 2.1.10

setRequestResponseTimeout :: ResponseTimeout -> Request -> Request

Set the maximum time to wait for a response

Since: 2.3.8

Response lenses

getResponseStatus :: Response a -> Status

Get the status of the response

Since: 2.1.10

getResponseStatusCode :: Response a -> Int

Get the integral status code of the response

Since: 2.1.10

getResponseHeader :: HeaderName -> Response a -> [ByteString]

Get all response header values with the given name

Since: 2.1.10

getResponseHeaders :: Response a -> [(HeaderName, ByteString)]

Get all response headers

Since: 2.1.10

getResponseBody :: Response a -> a

Get the response body

Since: 2.1.10

Alternate spellings

httpLbs :: MonadIO m => Request -> m (Response ByteString)

Alternate spelling of httpLBS

Since: 2.1.10