Options.Applicative.Help.Pretty

Documentation

class Pretty a where

Minimal complete definition

pretty

Methods

pretty :: a -> Doc ann

prettyList :: [a] -> Doc ann

data SimpleDocStream ann

Constructors

SFail

SEmpty

SChar !Char (SimpleDocStream ann)

SText !Int !Text (SimpleDocStream ann)

SLine !Int (SimpleDocStream ann)

SAnnPush ann (SimpleDocStream ann)

SAnnPop (SimpleDocStream ann)

newtype LayoutOptions

Constructors

LayoutOptions

Fields

layoutPageWidth :: PageWidth

data FusionDepth

Constructors

Shallow

Deep

data PageWidth

Constructors

AvailablePerLine !Int !Double

Unbounded

group :: Doc ann -> Doc ann

(<>) :: Semigroup a => a -> a -> a infixr 6

An associative operation.

Examples

Expand

>>> [1,2,3] <> [4,5,6]
[1,2,3,4,5,6]

>>> Just [1, 2, 3] <> Just [4, 5, 6]
Just [1,2,3,4,5,6]

>>> putStr "Hello, " <> putStrLn "World!"
Hello, World!

fill :: Int -> Doc ann -> Doc ann

parens :: Doc ann -> Doc ann

list :: [Doc ann] -> Doc ann

(<+>) :: Doc ann -> Doc ann -> Doc ann

annotate :: ann -> Doc ann -> Doc ann

semi :: Doc ann

comma :: Doc ann

colon :: Doc ann

space :: Doc ann

equals :: Doc ann

lparen :: Doc ann

rparen :: Doc ann

lbrace :: Doc ann

rbrace :: Doc ann

brackets :: Doc ann -> Doc ann

braces :: Doc ann -> Doc ann

hcat :: [Doc ann] -> Doc ann

hsep :: [Doc ann] -> Doc ann

vcat :: [Doc ann] -> Doc ann

nest :: Int -> Doc ann -> Doc ann

hang :: Int -> Doc ann -> Doc ann

punctuate :: Doc ann -> [Doc ann] -> [Doc ann]

sep :: [Doc ann] -> Doc ann

cat :: [Doc ann] -> Doc ann

indent :: Int -> Doc ann -> Doc ann

flatAlt :: Doc ann -> Doc ann -> Doc ann

line :: Doc ann

softline :: Doc ann

nesting :: (Int -> Doc ann) -> Doc ann

column :: (Int -> Doc ann) -> Doc ann

align :: Doc ann -> Doc ann

layoutPretty :: LayoutOptions -> Doc ann -> SimpleDocStream ann

alterAnnotations :: (ann -> [ann']) -> Doc ann -> Doc ann'

alterAnnotationsS :: (ann -> Maybe ann') -> SimpleDocStream ann -> SimpleDocStream ann'

concatWith :: Foldable t => (Doc ann -> Doc ann -> Doc ann) -> t (Doc ann) -> Doc ann

defaultLayoutOptions :: LayoutOptions

emptyDoc :: Doc ann

enclose :: Doc ann -> Doc ann -> Doc ann -> Doc ann

encloseSep :: Doc ann -> Doc ann -> Doc ann -> [Doc ann] -> Doc ann

fillBreak :: Int -> Doc ann -> Doc ann

fillCat :: [Doc ann] -> Doc ann

fillSep :: [Doc ann] -> Doc ann

fuse :: FusionDepth -> Doc ann -> Doc ann

hardline :: Doc ann

layoutCompact :: Doc ann1 -> SimpleDocStream ann2

layoutSmart :: LayoutOptions -> Doc ann -> SimpleDocStream ann

line' :: Doc ann

pageWidth :: (PageWidth -> Doc ann) -> Doc ann

plural :: (Num amount, Eq amount) => doc -> doc -> amount -> doc

reAnnotate :: (ann -> ann') -> Doc ann -> Doc ann'

reAnnotateS :: (ann -> ann') -> SimpleDocStream ann -> SimpleDocStream ann'

removeTrailingWhitespace :: SimpleDocStream ann -> SimpleDocStream ann

softline' :: Doc ann

surround :: Doc ann -> Doc ann -> Doc ann -> Doc ann

tupled :: [Doc ann] -> Doc ann

unAnnotate :: Doc ann -> Doc xxx

unAnnotateS :: SimpleDocStream ann -> SimpleDocStream xxx

unsafeViaShow :: Show a => a -> Doc ann

viaShow :: Show a => a -> Doc ann

vsep :: [Doc ann] -> Doc ann

width :: Doc ann -> (Int -> Doc ann) -> Doc ann

angles :: Doc ann -> Doc ann

backslash :: Doc ann

dot :: Doc ann

dquote :: Doc ann

dquotes :: Doc ann -> Doc ann

langle :: Doc ann

lbracket :: Doc ann

pipe :: Doc ann

rangle :: Doc ann

rbracket :: Doc ann

slash :: Doc ann

squote :: Doc ann

squotes :: Doc ann -> Doc ann

type Doc = Doc AnsiStyle

type SimpleDoc = SimpleDocStream AnsiStyle

(.$.) :: Doc -> Doc -> Doc

(</>) :: Doc -> Doc -> Doc

groupOrNestLine :: Doc -> Doc

Render flattened text on this line, or start a new line before rendering any text.

This will also nest subsequent lines in the group.

altSep :: Doc -> Doc -> Doc

Separate items in an alternative with a pipe.

If the first document and the pipe don't fit on the line, then mandatorily flow the next entry onto the following line.

The (//) softbreak ensures that if the document does fit on the line, there is at least a space, but it's possible for y to still appear on the next line.

hangAtIfOver :: Int -> Int -> Doc -> Doc

Printer hacks to get nice indentation for long commands and subcommands.

If we're starting this section over the desired width (usually 1/3 of the ribbon), then we will make a line break, indent all of the usage, and go.

The ifAtRoot is an interesting clause. If this whole operation is put under a group then the linebreak will disappear; then item d will therefore not be at the starting column, and it won't be indented more.

prettyString :: Double -> Int -> Doc -> String