Options.Applicative.Help.Style

Documentation

data SetStyle

Constructors

SetStyle

Fields

ansiReset :: Bool

ansiForeground :: Maybe (ColorIntensity, Color)

Set the foreground color, or keep the old one.

ansiBackground :: Maybe (ColorIntensity, Color)

Set the background color, or keep the old one.

ansiConsoleIntensity :: Maybe ConsoleIntensity

Adjust boldness

ansiItalics :: Maybe Italicized

Adjust italics

ansiUnderlining :: Maybe Underlining

Adjust underlining

data ColorIntensity

Constructors

Dull

Vivid

data Layer

Constructors

Foreground

Background

data ConsoleIntensity

Constructors

BoldIntensity

FaintIntensity

NormalIntensity

data Underlining

Constructors

SingleUnderline

DoubleUnderline

CurlyUnderline

DottedUnderline

DashedUnderline

NoUnderline

data Italicized

Constructors

Italicized

NoItalics

data Color

Constructors

Black

Red

Green

Yellow

Blue

Magenta

Cyan

White

color :: Color -> SetStyle

Style the foreground with a vivid color.

bgColor :: Color -> SetStyle

Style the background with a vivid color.

colorDull :: Color -> SetStyle

Style the foreground with a dull color.

bgColorDull :: Color -> SetStyle

Style the background with a dull color.

bold :: SetStyle

Render in bold.

underlined :: SetStyle

Render underlined.

italicized :: SetStyle

Render in italics.

styleToRawText :: SetStyle -> String

defaultStyle :: SetStyle