RIO.ByteString.Lazy

Description

Lazy ByteString. Import as:

import qualified RIO.ByteString.Lazy as BL

This module does not export any partial functions. For those, see RIO.ByteString.Lazy.Partial

The ByteString type

data ByteString

A space-efficient representation of a Word8 vector, supporting many efficient operations.

A lazy ByteString contains 8-bit bytes, or by using the operations from Data.ByteString.Lazy.Char8 it can be interpreted as containing 8-bit characters.

Introducing and eliminating ByteStrings

empty :: ByteString

O(1) The empty ByteString

singleton :: Word8 -> ByteString

O(1) Convert a Word8 into a ByteString

pack :: [Word8] -> ByteString

O(n) Convert a '[Word8]' into a ByteString.

unpack :: ByteString -> [Word8]

O(n) Converts a ByteString to a '[Word8]'.

fromStrict :: ByteString -> ByteString

O(1) Convert a strict ByteString into a lazy ByteString.

toStrict :: ByteString -> ByteString

O(n) Convert a lazy ByteString into a strict ByteString.

Note that this is an expensive operation that forces the whole lazy ByteString into memory and then copies all the data. If possible, try to avoid converting back and forth between strict and lazy bytestrings.

fromChunks :: [ByteString] -> ByteString

O(c) Convert a list of strict ByteString into a lazy ByteString

toChunks :: ByteString -> [ByteString]

O(c) Convert a lazy ByteString into a list of strict ByteString

foldrChunks :: (ByteString -> a -> a) -> a -> ByteString -> a

Consume the chunks of a lazy ByteString with a natural right fold.

foldlChunks :: (a -> ByteString -> a) -> a -> ByteString -> a

Consume the chunks of a lazy ByteString with a strict, tail-recursive, accumulating left fold.

Basic interface

cons :: Word8 -> ByteString -> ByteString infixr 5

O(1) cons is analogous to (:) for lists.

cons' :: Word8 -> ByteString -> ByteString infixr 5

O(1) Unlike cons, cons' is strict in the ByteString that we are consing onto. More precisely, it forces the head and the first chunk. It does this because, for space efficiency, it may coalesce the new byte onto the first 'chunk' rather than starting a new 'chunk'.

So that means you can't use a lazy recursive contruction like this:

let xs = cons' c xs in xs

You can however use cons, as well as repeat and cycle, to build infinite lazy ByteStrings.

snoc :: ByteString -> Word8 -> ByteString infixl 5

O(n/c) Append a byte to the end of a ByteString

append :: ByteString -> ByteString -> ByteString

O(n/c) Append two ByteStrings

uncons :: ByteString -> Maybe (Word8, ByteString)

O(1) Extract the head and tail of a ByteString, returning Nothing if it is empty.

unsnoc :: ByteString -> Maybe (ByteString, Word8)

O(n/c) Extract the init and last of a ByteString, returning Nothing if it is empty.

It is no faster than using init and last

null :: ByteString -> Bool

O(1) Test whether a ByteString is empty.

length :: ByteString -> Int64

O(c) length returns the length of a ByteString as an Int64

Transforming ByteStrings

map :: (Word8 -> Word8) -> ByteString -> ByteString

O(n) map f xs is the ByteString obtained by applying f to each element of xs.

reverse :: ByteString -> ByteString

O(n) reverse xs returns the elements of xs in reverse order.

intersperse :: Word8 -> ByteString -> ByteString

The intersperse function takes a Word8 and a ByteString and `intersperses' that byte between the elements of the ByteString. It is analogous to the intersperse function on Lists.

intercalate :: ByteString -> [ByteString] -> ByteString

O(n) The intercalate function takes a ByteString and a list of ByteStrings and concatenates the list after interspersing the first argument between each element of the list.

transpose :: [ByteString] -> [ByteString]

The transpose function transposes the rows and columns of its ByteString argument.

Reducing ByteStrings (folds)

foldl :: (a -> Word8 -> a) -> a -> ByteString -> a

foldl, applied to a binary operator, a starting value (typically the left-identity of the operator), and a ByteString, reduces the ByteString using the binary operator, from left to right.

foldl' :: (a -> Word8 -> a) -> a -> ByteString -> a

foldl' is like foldl, but strict in the accumulator.

foldr :: (Word8 -> a -> a) -> a -> ByteString -> a

foldr, applied to a binary operator, a starting value (typically the right-identity of the operator), and a ByteString, reduces the ByteString using the binary operator, from right to left.

Special folds

concat :: [ByteString] -> ByteString

O(n) Concatenate a list of ByteStrings.

concatMap :: (Word8 -> ByteString) -> ByteString -> ByteString

Map a function over a ByteString and concatenate the results

any :: (Word8 -> Bool) -> ByteString -> Bool

O(n) Applied to a predicate and a ByteString, any determines if any element of the ByteString satisfies the predicate.

all :: (Word8 -> Bool) -> ByteString -> Bool

O(n) Applied to a predicate and a ByteString, all determines if all elements of the ByteString satisfy the predicate.

Building ByteStrings

Scans

scanl

Arguments

:: (Word8 -> Word8 -> Word8)
accumulator -> element -> new accumulator

-> Word8
starting value of accumulator

-> ByteString
input of length n

-> ByteString
output of length n+1

scanl is similar to foldl, but returns a list of successive reduced values from the left.

scanl f z [x1, x2, ...] == [z, z `f` x1, (z `f` x1) `f` x2, ...]

Note that

head (scanl f z xs) == z
last (scanl f z xs) == foldl f z xs

Accumulating maps

mapAccumL :: (acc -> Word8 -> (acc, Word8)) -> acc -> ByteString -> (acc, ByteString)

The mapAccumL function behaves like a combination of map and foldl; it applies a function to each element of a ByteString, passing an accumulating parameter from left to right, and returning a final value of this accumulator together with the new ByteString.

mapAccumR :: (acc -> Word8 -> (acc, Word8)) -> acc -> ByteString -> (acc, ByteString)

The mapAccumR function behaves like a combination of map and foldr; it applies a function to each element of a ByteString, passing an accumulating parameter from right to left, and returning a final value of this accumulator together with the new ByteString.

Infinite ByteStrings

repeat :: Word8 -> ByteString

repeat x is an infinite ByteString, with x the value of every element.

replicate :: Int64 -> Word8 -> ByteString

O(n) replicate n x is a ByteString of length n with x the value of every element.

cycle :: HasCallStack => ByteString -> ByteString

cycle ties a finite ByteString into a circular one, or equivalently, the infinite repetition of the original ByteString.

iterate :: (Word8 -> Word8) -> Word8 -> ByteString

iterate f x returns an infinite ByteString of repeated applications of f to x:

iterate f x == [x, f x, f (f x), ...]

Unfolding ByteStrings

unfoldr :: (a -> Maybe (Word8, a)) -> a -> ByteString

O(n) The unfoldr function is analogous to the List 'unfoldr'. unfoldr builds a ByteString from a seed value. The function takes the element and returns Nothing if it is done producing the ByteString or returns Just (a,b), in which case, a is a prepending to the ByteString and b is used as the next element in a recursive call.

Substrings

Breaking strings

take :: Int64 -> ByteString -> ByteString

O(n/c) take n, applied to a ByteString xs, returns the prefix of xs of length n, or xs itself if n > length xs.

drop :: Int64 -> ByteString -> ByteString

O(n/c) drop n xs returns the suffix of xs after the first n elements, or empty if n > length xs.

splitAt :: Int64 -> ByteString -> (ByteString, ByteString)

O(n/c) splitAt n xs is equivalent to (take n xs, drop n xs).

takeWhile :: (Word8 -> Bool) -> ByteString -> ByteString

Similar to takeWhile, returns the longest (possibly empty) prefix of elements satisfying the predicate.

dropWhile :: (Word8 -> Bool) -> ByteString -> ByteString

Similar to dropWhile, drops the longest (possibly empty) prefix of elements satisfying the predicate and returns the remainder.

span :: (Word8 -> Bool) -> ByteString -> (ByteString, ByteString)

Similar to span, returns the longest (possibly empty) prefix of elements satisfying the predicate and the remainder of the string.

span p is equivalent to break (not . p) and to (takeWhile p &&& dropWhile p).

break :: (Word8 -> Bool) -> ByteString -> (ByteString, ByteString)

Similar to break, returns the longest (possibly empty) prefix of elements which do not satisfy the predicate and the remainder of the string.

break p is equivalent to span (not . p) and to (takeWhile (not . p) &&& dropWhile (not . p)).

group :: ByteString -> [ByteString]

The group function takes a ByteString and returns a list of ByteStrings such that the concatenation of the result is equal to the argument. Moreover, each string in the result contains only equal elements. For example,

group "Mississippi" = ["M","i","ss","i","ss","i","pp","i"]

It is a special case of groupBy, which allows the programmer to supply their own equality test.

groupBy :: (Word8 -> Word8 -> Bool) -> ByteString -> [ByteString]

The groupBy function is the non-overloaded version of group.

inits :: ByteString -> [ByteString]

Returns all initial segments of the given ByteString, shortest first.

tails :: ByteString -> [ByteString]

O(n) Returns all final segments of the given ByteString, longest first.

stripPrefix :: ByteString -> ByteString -> Maybe ByteString

O(n) The stripPrefix function takes two ByteStrings and returns Just the remainder of the second iff the first is its prefix, and otherwise Nothing.

Since: bytestring-0.10.8.0

stripSuffix :: ByteString -> ByteString -> Maybe ByteString

O(n) The stripSuffix function takes two ByteStrings and returns Just the remainder of the second iff the first is its suffix, and otherwise Nothing.

Breaking into many substrings

split :: Word8 -> ByteString -> [ByteString]

O(n) Break a ByteString into pieces separated by the byte argument, consuming the delimiter. I.e.

split 10  "a\nb\nd\ne" == ["a","b","d","e"]   -- fromEnum '\n' == 10
split 97  "aXaXaXa"    == ["","X","X","X",""] -- fromEnum 'a' == 97
split 120 "x"          == ["",""]             -- fromEnum 'x' == 120
split undefined ""     == []                  -- and not [""]

and

intercalate [c] . split c == id
split == splitWith . (==)

As for all splitting functions in this library, this function does not copy the substrings, it just constructs new ByteStrings that are slices of the original.

splitWith :: (Word8 -> Bool) -> ByteString -> [ByteString]

O(n) Splits a ByteString into components delimited by separators, where the predicate returns True for a separator element. The resulting components do not contain the separators. Two adjacent separators result in an empty component in the output. eg.

splitWith (==97) "aabbaca" == ["","","bb","c",""] -- fromEnum 'a' == 97
splitWith undefined ""     == []                  -- and not [""]

Predicates

isPrefixOf :: ByteString -> ByteString -> Bool

O(n) The isPrefixOf function takes two ByteStrings and returns True iff the first is a prefix of the second.

isSuffixOf :: ByteString -> ByteString -> Bool

O(n) The isSuffixOf function takes two ByteStrings and returns True iff the first is a suffix of the second.

The following holds:

isSuffixOf x y == reverse x `isPrefixOf` reverse y

Search ByteStrings

Searching by equality

elem :: Word8 -> ByteString -> Bool

O(n) elem is the ByteString membership predicate.

notElem :: Word8 -> ByteString -> Bool

O(n) notElem is the inverse of elem

Searching with a predicate

find :: (Word8 -> Bool) -> ByteString -> Maybe Word8

O(n) The find function takes a predicate and a ByteString, and returns the first element in matching the predicate, or Nothing if there is no such element.

find f p = case findIndex f p of Just n -> Just (p ! n) ; _ -> Nothing

filter :: (Word8 -> Bool) -> ByteString -> ByteString

O(n) filter, applied to a predicate and a ByteString, returns a ByteString containing those characters that satisfy the predicate.

partition :: (Word8 -> Bool) -> ByteString -> (ByteString, ByteString)

O(n) The partition function takes a predicate a ByteString and returns the pair of ByteStrings with elements which do and do not satisfy the predicate, respectively; i.e.,

partition p bs == (filter p xs, filter (not . p) xs)

Indexing ByteStrings

index :: HasCallStack => ByteString -> Int64 -> Word8

O(c) ByteString index (subscript) operator, starting from 0.

This is a partial function, consider using indexMaybe instead.

elemIndex :: Word8 -> ByteString -> Maybe Int64

O(n) The elemIndex function returns the index of the first element in the given ByteString which is equal to the query element, or Nothing if there is no such element. This implementation uses memchr(3).

elemIndexEnd :: Word8 -> ByteString -> Maybe Int64

O(n) The elemIndexEnd function returns the last index of the element in the given ByteString which is equal to the query element, or Nothing if there is no such element. The following holds:

elemIndexEnd c xs = case elemIndex c (reverse xs) of
  Nothing -> Nothing
  Just i  -> Just (length xs - 1 - i)

Since: bytestring-0.10.6.0

elemIndices :: Word8 -> ByteString -> [Int64]

O(n) The elemIndices function extends elemIndex, by returning the indices of all elements equal to the query element, in ascending order. This implementation uses memchr(3).

findIndex :: (Word8 -> Bool) -> ByteString -> Maybe Int64

The findIndex function takes a predicate and a ByteString and returns the index of the first element in the ByteString satisfying the predicate.

findIndices :: (Word8 -> Bool) -> ByteString -> [Int64]

The findIndices function extends findIndex, by returning the indices of all elements satisfying the predicate, in ascending order.

count :: Word8 -> ByteString -> Int64

count returns the number of times its argument appears in the ByteString

count = length . elemIndices

But more efficiently than using length on the intermediate list.

Zipping and unzipping ByteStrings

zip :: ByteString -> ByteString -> [(Word8, Word8)]

O(n) zip takes two ByteStrings and returns a list of corresponding pairs of bytes. If one input ByteString is short, excess elements of the longer ByteString are discarded. This is equivalent to a pair of unpack operations.

zipWith :: (Word8 -> Word8 -> a) -> ByteString -> ByteString -> [a]

zipWith generalises zip by zipping with the function given as the first argument, instead of a tupling function. For example, zipWith (+) is applied to two ByteStrings to produce the list of corresponding sums.

unzip :: [(Word8, Word8)] -> (ByteString, ByteString)

O(n) unzip transforms a list of pairs of bytes into a pair of ByteStrings. Note that this performs two pack operations.

Low level conversions

Copying ByteStrings

copy :: ByteString -> ByteString

O(n) Make a copy of the ByteString with its own storage. This is mainly useful to allow the rest of the data pointed to by the ByteString to be garbage collected, for example if a large string has been read in, and only a small part of it is needed in the rest of the program.

I/O with ByteStrings

Standard input and output

getContents :: MonadIO m => m LByteString

Lifted getContents

putStr :: MonadIO m => LByteString -> m ()

Lifted putStr

putStrLn :: MonadIO m => LByteString -> m ()

Lifted putStrLn

interact :: MonadIO m => (LByteString -> LByteString) -> m ()

Lifted interact

Files

readFile :: MonadIO m => FilePath -> m LByteString

Lifted readFile

writeFile :: MonadIO m => FilePath -> LByteString -> m ()

Lifted writeFile

appendFile :: MonadIO m => FilePath -> LByteString -> m ()

Lifted appendFile

I/O with Handles

hGetContents :: MonadIO m => Handle -> m LByteString

Lifted hGetContents

hGet :: MonadIO m => Handle -> Int -> m LByteString

Lifted hGet

hGetNonBlocking :: MonadIO m => Handle -> Int -> m LByteString

Lifted hGetNonBlocking

hPut :: MonadIO m => Handle -> LByteString -> m ()

Lifted hPut

hPutNonBlocking :: MonadIO m => Handle -> LByteString -> m LByteString

Lifted hPutNonBlocking

hPutStr :: MonadIO m => Handle -> LByteString -> m ()

Lifted hPutStr