RIO.NonEmpty.Partial

Description

NonEmpty list partial functions. Import as:

import qualified RIO.NonEmpty.Partial as NE'

Indexing streams

(!!) :: HasCallStack => NonEmpty a -> Int -> a infixl 9

xs !! n returns the element of the stream xs at index n. Note that the head of the stream has index 0.

Beware: a negative or out-of-bounds index will cause an error.

Converting to and from a list

fromList :: HasCallStack => [a] -> NonEmpty a

Converts a normal list to a NonEmpty stream.

Raises an error if given an empty list.