RIO.Orphans

Description

Orphan instances for the RIO data type.

Documentation

class HasResourceMap env where

An environment with a ResourceMap

Since: 0.1.0.0

Methods

resourceMapL :: Lens' env ResourceMap

type ResourceMap = IORef ReleaseMap

A collection of all of the registered resource cleanup actions.

Since: 0.1.0.0

withResourceMap :: MonadUnliftIO m => (ResourceMap -> m a) -> m a

Perform an action with a ResourceMap

Since: 0.1.0.0

Orphan instances

Display LogStr
Since: 0.1.1.0

Instance details

Methods

display :: LogStr -> Utf8Builder

textDisplay :: LogStr -> Text

MonadBaseControl IO (RIO env)
Since: 0.1.0.0

Instance details

Associated Types

type StM (RIO env) a

Methods

liftBaseWith :: (RunInBase (RIO env) IO -> IO a) -> RIO env a

restoreM :: StM (RIO env) a -> RIO env a

MonadBase IO (RIO env)
Since: 0.1.0.0

Instance details

Methods

liftBase :: IO α -> RIO env α

MonadCatch (RIO env)
Since: 0.1.0.0

Instance details

Methods

catch :: (HasCallStack, Exception e) => RIO env a -> (e -> RIO env a) -> RIO env a

MonadMask (RIO env)
Since: 0.1.0.0

Instance details

Methods

mask :: HasCallStack => ((forall a. RIO env a -> RIO env a) -> RIO env b) -> RIO env b

uninterruptibleMask :: HasCallStack => ((forall a. RIO env a -> RIO env a) -> RIO env b) -> RIO env b

generalBracket :: HasCallStack => RIO env a -> (a -> ExitCase b -> RIO env c) -> (a -> RIO env b) -> RIO env (b, c)

HasLogFunc env => MonadLogger (RIO env)
Since: 0.1.1.0

Instance details

Methods

monadLoggerLog :: ToLogStr msg => Loc -> LogSource -> LogLevel -> msg -> RIO env ()

HasLogFunc env => MonadLoggerIO (RIO env)
Do not let the generated function escape its RIO context. This may lead to log-related cleanup running before the function is called.

Since: 0.1.2.0

Instance details

Methods

askLoggerIO :: RIO env (Loc -> LogSource -> LogLevel -> LogStr -> IO ())

HasResourceMap env => MonadResource (RIO env)

Instance details

Methods

liftResourceT :: ResourceT IO a -> RIO env a