RIO.Text.Lazy

Description

Lazy Text. Import as:

import qualified RIO.Text.Lazy as TL

This module does not export any partial functions. For those, see RIO.Text.Lazy.Partial

Types

data Text

Creation and elimination

pack :: String -> Text

O(n) Convert a String into a Text.

Performs replacement on invalid scalar values, so unpack . pack is not id:

>>> Data.Text.Lazy.unpack (Data.Text.Lazy.pack "\55555")
"\65533"

unpack :: Text -> String

O(n) Convert a Text into a String.

singleton :: Char -> Text

O(1) Convert a character into a Text. Performs replacement on invalid scalar values.

empty :: Text

Smart constructor for Empty.

fromChunks :: [Text] -> Text

O(c) Convert a list of strict Texts into a lazy Text.

toChunks :: Text -> [Text]

O(n) Convert a lazy Text into a list of strict Texts.

toStrict :: Text -> Text

O(n) Convert a lazy Text into a strict Text.

fromStrict :: Text -> Text

O(c) Convert a strict Text into a lazy Text.

foldrChunks :: (Text -> a -> a) -> a -> Text -> a

Consume the chunks of a lazy Text with a natural right fold.

foldlChunks :: (a -> Text -> a) -> a -> Text -> a

Consume the chunks of a lazy Text with a strict, tail-recursive, accumulating left fold.

Basic interface

cons :: Char -> Text -> Text infixr 5

O(1) Adds a character to the front of a Text.

snoc :: Text -> Char -> Text

O(n) Adds a character to the end of a Text. This copies the entire array in the process.

append :: Text -> Text -> Text

O(n/c) Appends one Text to another.

uncons :: Text -> Maybe (Char, Text)

O(1) Returns the first character and rest of a Text, or Nothing if empty.

null :: Text -> Bool

O(1) Tests whether a Text is empty or not.

length :: Text -> Int64

O(n) Returns the number of characters in a Text.

compareLength :: Text -> Int64 -> Ordering

O(min(n,c)) Compare the count of characters in a Text to a number.

compareLength t c = compare (length t) c

This function gives the same answer as comparing against the result of length, but can short circuit if the count of characters is greater than the number, and hence be more efficient.

Transformations

map :: (Char -> Char) -> Text -> Text

O(n) map f t is the Text obtained by applying f to each element of t. Performs replacement on invalid scalar values.

intercalate :: Text -> [Text] -> Text

O(n) The intercalate function takes a Text and a list of Texts and concatenates the list after interspersing the first argument between each element of the list.

intersperse :: Char -> Text -> Text

O(n) The intersperse function takes a character and places it between the characters of a Text. Performs replacement on invalid scalar values.

transpose :: [Text] -> [Text]

O(n) The transpose function transposes the rows and columns of its Text argument. Note that this function uses pack, unpack, and the list version of transpose, and is thus not very efficient.

reverse :: Text -> Text

O(n) reverse t returns the elements of t in reverse order.

Case conversion

toCaseFold :: Text -> Text

O(n) Convert a string to folded case.

This function is mainly useful for performing caseless (or case insensitive) string comparisons.

A string x is a caseless match for a string y if and only if:

toCaseFold x == toCaseFold y

The result string may be longer than the input string, and may differ from applying toLower to the input string. For instance, the Armenian small ligature men now (U+FB13) is case folded to the bigram men now (U+0574 U+0576), while the micro sign (U+00B5) is case folded to the Greek small letter letter mu (U+03BC) instead of itself.

toLower :: Text -> Text

O(n) Convert a string to lower case, using simple case conversion.

The result string may be longer than the input string. For instance, the Latin capital letter I with dot above (U+0130) maps to the sequence Latin small letter i (U+0069) followed by combining dot above (U+0307).

toUpper :: Text -> Text

O(n) Convert a string to upper case, using simple case conversion.

The result string may be longer than the input string. For instance, the German eszett (U+00DF) maps to the two-letter sequence SS.

toTitle :: Text -> Text

O(n) Convert a string to title case, using simple case conversion.

The first letter (as determined by isLetter) of the input is converted to title case, as is every subsequent letter that immediately follows a non-letter. Every letter that immediately follows another letter is converted to lower case.

The result string may be longer than the input string. For example, the Latin small ligature ﬂ (U+FB02) is converted to the sequence Latin capital letter F (U+0046) followed by Latin small letter l (U+006C).

This function is not idempotent. Consider lower-case letter ŉ (U+0149 LATIN SMALL LETTER N PRECEDED BY APOSTROPHE). Then toTitle "ŉ" = "ʼN": the first (and the only) letter of the input is converted to title case, becoming two letters. Now ʼ (U+02BC MODIFIER LETTER APOSTROPHE) is a modifier letter and as such is recognised as a letter by isLetter, so toTitle "ʼN" = "'n".

Note: this function does not take language or culture specific rules into account. For instance, in English, different style guides disagree on whether the book name "The Hill of the Red Fox" is correctly title cased—but this function will capitalize every word.

Since: text-1.0.0.0

Justification

justifyLeft :: Int64 -> Char -> Text -> Text

O(n) Left-justify a string to the given length, using the specified fill character on the right. Performs replacement on invalid scalar values.

Examples:

justifyLeft 7 'x' "foo"    == "fooxxxx"
justifyLeft 3 'x' "foobar" == "foobar"

justifyRight :: Int64 -> Char -> Text -> Text

O(n) Right-justify a string to the given length, using the specified fill character on the left. Performs replacement on invalid scalar values.

Examples:

justifyRight 7 'x' "bar"    == "xxxxbar"
justifyRight 3 'x' "foobar" == "foobar"

center :: Int64 -> Char -> Text -> Text

O(n) Center a string to the given length, using the specified fill character on either side. Performs replacement on invalid scalar values.

Examples:

center 8 'x' "HS" = "xxxHSxxx"

Folds

foldl :: (a -> Char -> a) -> a -> Text -> a

O(n) foldl, applied to a binary operator, a starting value (typically the left-identity of the operator), and a Text, reduces the Text using the binary operator, from left to right.

foldl' :: (a -> Char -> a) -> a -> Text -> a

O(n) A strict version of foldl.

foldr :: (Char -> a -> a) -> a -> Text -> a

O(n) foldr, applied to a binary operator, a starting value (typically the right-identity of the operator), and a Text, reduces the Text using the binary operator, from right to left.

foldr is lazy like foldr for lists: evaluation actually traverses the Text from left to right, only as far as it needs to.

For example, head can be defined with O(1) complexity using foldr:

head :: Text -> Char
head = foldr const (error "head empty")

Special folds

concat :: [Text] -> Text

O(n) Concatenate a list of Texts.

concatMap :: (Char -> Text) -> Text -> Text

O(n) Map a function over a Text that results in a Text, and concatenate the results.

any :: (Char -> Bool) -> Text -> Bool

O(n) any p t determines whether any character in the Text t satisfies the predicate p.

all :: (Char -> Bool) -> Text -> Bool

O(n) all p t determines whether all characters in the Text t satisfy the predicate p.

Construction

Scans

scanl :: (Char -> Char -> Char) -> Char -> Text -> Text

O(n) scanl is similar to foldl, but returns a list of successive reduced values from the left. Performs replacement on invalid scalar values.

scanl f z [x1, x2, ...] == [z, z `f` x1, (z `f` x1) `f` x2, ...]

Note that

last (scanl f z xs) == foldl f z xs.

scanl1 :: (Char -> Char -> Char) -> Text -> Text

O(n) scanl1 is a variant of scanl that has no starting value argument. Performs replacement on invalid scalar values.

scanl1 f [x1, x2, ...] == [x1, x1 `f` x2, ...]

scanr :: (Char -> Char -> Char) -> Char -> Text -> Text

O(n) scanr is the right-to-left dual of scanl. Performs replacement on invalid scalar values.

scanr f v == reverse . scanl (flip f) v . reverse

scanr1 :: (Char -> Char -> Char) -> Text -> Text

O(n) scanr1 is a variant of scanr that has no starting value argument. Performs replacement on invalid scalar values.

Accumulating maps

mapAccumL :: (a -> Char -> (a, Char)) -> a -> Text -> (a, Text)

O(n) Like a combination of map and foldl'. Applies a function to each element of a Text, passing an accumulating parameter from left to right, and returns a final Text. Performs replacement on invalid scalar values.

mapAccumR :: (a -> Char -> (a, Char)) -> a -> Text -> (a, Text)

The mapAccumR function behaves like a combination of map and a strict foldr; it applies a function to each element of a Text, passing an accumulating parameter from right to left, and returning a final value of this accumulator together with the new Text. Performs replacement on invalid scalar values.

Generation and unfolding

repeat :: Char -> Text

repeat x is an infinite Text, with x the value of every element.

Since: text-1.2.0.5

replicate :: Int64 -> Text -> Text

O(n*m) replicate n t is a Text consisting of the input t repeated n times.

cycle :: HasCallStack => Text -> Text

cycle ties a finite, non-empty Text into a circular one, or equivalently, the infinite repetition of the original Text.

Since: text-1.2.0.5

iterate :: (Char -> Char) -> Char -> Text

iterate f x returns an infinite Text of repeated applications of f to x:

iterate f x == [x, f x, f (f x), ...]

Since: text-1.2.0.5

unfoldr :: (a -> Maybe (Char, a)) -> a -> Text

O(n), where n is the length of the result. The unfoldr function is analogous to the List unfoldr. unfoldr builds a Text from a seed value. The function takes the element and returns Nothing if it is done producing the Text, otherwise Just (a,b). In this case, a is the next Char in the string, and b is the seed value for further production. Performs replacement on invalid scalar values.

unfoldrN :: Int64 -> (a -> Maybe (Char, a)) -> a -> Text

O(n) Like unfoldr, unfoldrN builds a Text from a seed value. However, the length of the result should be limited by the first argument to unfoldrN. This function is more efficient than unfoldr when the maximum length of the result is known and correct, otherwise its performance is similar to unfoldr. Performs replacement on invalid scalar values.

Substrings

Breaking strings

take :: Int64 -> Text -> Text

O(n) take n, applied to a Text, returns the prefix of the Text of length n, or the Text itself if n is greater than the length of the Text.

takeEnd :: Int64 -> Text -> Text

O(n) takeEnd n t returns the suffix remaining after taking n characters from the end of t.

Examples:

takeEnd 3 "foobar" == "bar"

Since: text-1.1.1.0

drop :: Int64 -> Text -> Text

O(n) drop n, applied to a Text, returns the suffix of the Text after the first n characters, or the empty Text if n is greater than the length of the Text.

dropEnd :: Int64 -> Text -> Text

O(n) dropEnd n t returns the prefix remaining after dropping n characters from the end of t.

Examples:

dropEnd 3 "foobar" == "foo"

Since: text-1.1.1.0

takeWhile :: (Char -> Bool) -> Text -> Text

O(n) takeWhile, applied to a predicate p and a Text, returns the longest prefix (possibly empty) of elements that satisfy p.

takeWhileEnd :: (Char -> Bool) -> Text -> Text

O(n) takeWhileEnd, applied to a predicate p and a Text, returns the longest suffix (possibly empty) of elements that satisfy p. Examples:

takeWhileEnd (=='o') "foo" == "oo"

Since: text-1.2.2.0

dropWhile :: (Char -> Bool) -> Text -> Text

O(n) dropWhile p t returns the suffix remaining after takeWhile p t.

dropWhileEnd :: (Char -> Bool) -> Text -> Text

O(n) dropWhileEnd p t returns the prefix remaining after dropping characters that satisfy the predicate p from the end of t.

Examples:

dropWhileEnd (=='.') "foo..." == "foo"

dropAround :: (Char -> Bool) -> Text -> Text

O(n) dropAround p t returns the substring remaining after dropping characters that satisfy the predicate p from both the beginning and end of t.

strip :: Text -> Text

O(n) Remove leading and trailing white space from a string. Equivalent to:

dropAround isSpace

stripStart :: Text -> Text

O(n) Remove leading white space from a string. Equivalent to:

dropWhile isSpace

stripEnd :: Text -> Text

O(n) Remove trailing white space from a string. Equivalent to:

dropWhileEnd isSpace

splitAt :: Int64 -> Text -> (Text, Text)

O(n) splitAt n t returns a pair whose first element is a prefix of t of length n, and whose second is the remainder of the string. It is equivalent to (take n t, drop n t).

span :: (Char -> Bool) -> Text -> (Text, Text)

O(n) span, applied to a predicate p and text t, returns a pair whose first element is the longest prefix (possibly empty) of t of elements that satisfy p, and whose second is the remainder of the text.

>>> T.span (=='0') "000AB"
("000","AB")

break :: (Char -> Bool) -> Text -> (Text, Text)

O(n) break is like span, but the prefix returned is over elements that fail the predicate p.

>>> T.break (=='c') "180cm"
("180","cm")

group :: Text -> [Text]

The group function takes a Text and returns a list of Texts such that the concatenation of the result is equal to the argument. Moreover, each sublist in the result contains only equal elements. For example,

group "Mississippi" = ["M","i","ss","i","ss","i","pp","i"]

It is a special case of groupBy, which allows the programmer to supply their own equality test.

groupBy :: (Char -> Char -> Bool) -> Text -> [Text]

The groupBy function is the non-overloaded version of group.

inits :: Text -> [Text]

O(n) Return all initial segments of the given Text, shortest first.

tails :: Text -> [Text]

O(n) Return all final segments of the given Text, longest first.

Breaking into many substrings

split :: (Char -> Bool) -> Text -> [Text]

O(n) Splits a Text into components delimited by separators, where the predicate returns True for a separator element. The resulting components do not contain the separators. Two adjacent separators result in an empty component in the output. eg.

split (=='a') "aabbaca" == ["","","bb","c",""]
split (=='a') []        == [""]

chunksOf :: Int64 -> Text -> [Text]

O(n) Splits a Text into components of length k. The last element may be shorter than the other chunks, depending on the length of the input. Examples:

chunksOf 3 "foobarbaz"   == ["foo","bar","baz"]
chunksOf 4 "haskell.org" == ["hask","ell.","org"]

Breaking into lines and words

lines :: Text -> [Text]

O(n) Breaks a Text up into a list of Texts at newline characters '\n' (LF, line feed). The resulting strings do not contain newlines.

lines does not treat '\r' (CR, carriage return) as a newline character.

words :: Text -> [Text]

O(n) Breaks a Text up into a list of words, delimited by Chars representing white space.

unlines :: [Text] -> Text

O(n) Joins lines, after appending a terminating newline to each.

unwords :: [Text] -> Text

O(n) Joins words using single space characters.

Predicates

isPrefixOf :: Text -> Text -> Bool

O(n) The isPrefixOf function takes two Texts and returns True if and only if the first is a prefix of the second.

isSuffixOf :: Text -> Text -> Bool

O(n) The isSuffixOf function takes two Texts and returns True if and only if the first is a suffix of the second.

isInfixOf :: Text -> Text -> Bool

O(n+m) The isInfixOf function takes two Texts and returns True if and only if the first is contained, wholly and intact, anywhere within the second.

This function is strict in its first argument, and lazy in its second.

In (unlikely) bad cases, this function's time complexity degrades towards O(n*m).

View patterns

stripPrefix :: Text -> Text -> Maybe Text

O(n) Return the suffix of the second string if its prefix matches the entire first string.

Examples:

stripPrefix "foo" "foobar" == Just "bar"
stripPrefix ""    "baz"    == Just "baz"
stripPrefix "foo" "quux"   == Nothing

This is particularly useful with the ViewPatterns extension to GHC, as follows:

{-# LANGUAGE ViewPatterns #-}
import Data.Text.Lazy as T

fnordLength :: Text -> Int
fnordLength (stripPrefix "fnord" -> Just suf) = T.length suf
fnordLength _                                 = -1

stripSuffix :: Text -> Text -> Maybe Text

O(n) Return the prefix of the second string if its suffix matches the entire first string.

Examples:

stripSuffix "bar" "foobar" == Just "foo"
stripSuffix ""    "baz"    == Just "baz"
stripSuffix "foo" "quux"   == Nothing

This is particularly useful with the ViewPatterns extension to GHC, as follows:

{-# LANGUAGE ViewPatterns #-}
import Data.Text.Lazy as T

quuxLength :: Text -> Int
quuxLength (stripSuffix "quux" -> Just pre) = T.length pre
quuxLength _                                = -1

commonPrefixes :: Text -> Text -> Maybe (Text, Text, Text)

O(n) Find the longest non-empty common prefix of two strings and return it, along with the suffixes of each string at which they no longer match.

If the strings do not have a common prefix or either one is empty, this function returns Nothing.

Examples:

commonPrefixes "foobar" "fooquux" == Just ("foo","bar","quux")
commonPrefixes "veeble" "fetzer"  == Nothing
commonPrefixes "" "baz"           == Nothing

Searching

filter :: (Char -> Bool) -> Text -> Text

O(n) filter, applied to a predicate and a Text, returns a Text containing those characters that satisfy the predicate.

find :: (Char -> Bool) -> Text -> Maybe Char

O(n) The find function takes a predicate and a Text, and returns the first element in matching the predicate, or Nothing if there is no such element.

partition :: (Char -> Bool) -> Text -> (Text, Text)

O(n) The partition function takes a predicate and a Text, and returns the pair of Texts with elements which do and do not satisfy the predicate, respectively; i.e.

partition p t == (filter p t, filter (not . p) t)

Indexing

index :: HasCallStack => Text -> Int64 -> Char

O(n) Text index (subscript) operator, starting from 0.

count :: HasCallStack => Text -> Text -> Int64

O(n+m) The count function returns the number of times the query string appears in the given Text. An empty query string is invalid, and will cause an error to be raised.

In (unlikely) bad cases, this function's time complexity degrades towards O(n*m).

Zipping and unzipping

zip :: Text -> Text -> [(Char, Char)]

O(n) zip takes two Texts and returns a list of corresponding pairs of bytes. If one input Text is short, excess elements of the longer Text are discarded. This is equivalent to a pair of unpack operations.

zipWith :: (Char -> Char -> Char) -> Text -> Text -> Text

O(n) zipWith generalises zip by zipping with the function given as the first argument, instead of a tupling function. Performs replacement on invalid scalar values.