RIO.Vector

Description

Generic Vector interface. Import as:

import qualified RIO.Vector as V

This module does not export any partial or unsafe functions. For those, see RIO.Vector.Partial and RIO.Vector.Unsafe

Immutable vectors

class MVector (Mutable v) a => Vector (v :: Type -> Type) a

Minimal complete definition

basicUnsafeFreeze, basicUnsafeThaw, basicLength, basicUnsafeSlice, basicUnsafeIndexM

Accessors

Length information

length :: Vector v a => v a -> Int

null :: Vector v a => v a -> Bool

Indexing

(!?) :: Vector v a => v a -> Int -> Maybe a

Extracting subvectors

slice :: (HasCallStack, Vector v a) => Int -> Int -> v a -> v a

take :: Vector v a => Int -> v a -> v a

drop :: Vector v a => Int -> v a -> v a

splitAt :: Vector v a => Int -> v a -> (v a, v a)

Construction

Initialisation

empty :: Vector v a => v a

singleton :: Vector v a => a -> v a

replicate :: Vector v a => Int -> a -> v a

generate :: Vector v a => Int -> (Int -> a) -> v a

iterateN :: Vector v a => Int -> (a -> a) -> a -> v a

Monadic initialisation

replicateM :: (Monad m, Vector v a) => Int -> m a -> m (v a)

generateM :: (Monad m, Vector v a) => Int -> (Int -> m a) -> m (v a)

iterateNM :: (Monad m, Vector v a) => Int -> (a -> m a) -> a -> m (v a)

create :: Vector v a => (forall s. ST s (Mutable v s a)) -> v a

createT :: (Traversable f, Vector v a) => (forall s. ST s (f (Mutable v s a))) -> f (v a)

Unfolding

unfoldr :: Vector v a => (b -> Maybe (a, b)) -> b -> v a

unfoldrN :: Vector v a => Int -> (b -> Maybe (a, b)) -> b -> v a

unfoldrM :: (Monad m, Vector v a) => (b -> m (Maybe (a, b))) -> b -> m (v a)

unfoldrNM :: (Monad m, Vector v a) => Int -> (b -> m (Maybe (a, b))) -> b -> m (v a)

constructN :: Vector v a => Int -> (v a -> a) -> v a

constructrN :: Vector v a => Int -> (v a -> a) -> v a

Enumeration

enumFromN :: (Vector v a, Num a) => a -> Int -> v a

enumFromStepN :: (Vector v a, Num a) => a -> a -> Int -> v a

enumFromTo :: (Vector v a, Enum a) => a -> a -> v a

enumFromThenTo :: (Vector v a, Enum a) => a -> a -> a -> v a

Concatenation

cons :: Vector v a => a -> v a -> v a

snoc :: Vector v a => v a -> a -> v a

(++) :: Vector v a => v a -> v a -> v a

concat :: Vector v a => [v a] -> v a

concatNE :: Vector v a => NonEmpty (v a) -> v a

Restricting memory usage

force :: Vector v a => v a -> v a

Modifying vectors

Permutations

reverse :: Vector v a => v a -> v a

Safe destructive update

modify :: Vector v a => (forall s. Mutable v s a -> ST s ()) -> v a -> v a

Elementwise operations

Indexing

indexed :: (Vector v a, Vector v (Int, a)) => v a -> v (Int, a)

Mapping

map :: (Vector v a, Vector v b) => (a -> b) -> v a -> v b

imap :: (Vector v a, Vector v b) => (Int -> a -> b) -> v a -> v b

concatMap :: (Vector v a, Vector v b) => (a -> v b) -> v a -> v b

Monadic mapping

mapM :: (Monad m, Vector v a, Vector v b) => (a -> m b) -> v a -> m (v b)

imapM :: (Monad m, Vector v a, Vector v b) => (Int -> a -> m b) -> v a -> m (v b)

mapM_ :: (Monad m, Vector v a) => (a -> m b) -> v a -> m ()

imapM_ :: (Monad m, Vector v a) => (Int -> a -> m b) -> v a -> m ()

forM :: (Monad m, Vector v a, Vector v b) => v a -> (a -> m b) -> m (v b)

forM_ :: (Monad m, Vector v a) => v a -> (a -> m b) -> m ()

Zipping

zipWith :: (Vector v a, Vector v b, Vector v c) => (a -> b -> c) -> v a -> v b -> v c

zipWith3 :: (Vector v a, Vector v b, Vector v c, Vector v d) => (a -> b -> c -> d) -> v a -> v b -> v c -> v d

zipWith4 :: (Vector v a, Vector v b, Vector v c, Vector v d, Vector v e) => (a -> b -> c -> d -> e) -> v a -> v b -> v c -> v d -> v e

zipWith5 :: (Vector v a, Vector v b, Vector v c, Vector v d, Vector v e, Vector v f) => (a -> b -> c -> d -> e -> f) -> v a -> v b -> v c -> v d -> v e -> v f

zipWith6 :: (Vector v a, Vector v b, Vector v c, Vector v d, Vector v e, Vector v f, Vector v g) => (a -> b -> c -> d -> e -> f -> g) -> v a -> v b -> v c -> v d -> v e -> v f -> v g

izipWith :: (Vector v a, Vector v b, Vector v c) => (Int -> a -> b -> c) -> v a -> v b -> v c

izipWith3 :: (Vector v a, Vector v b, Vector v c, Vector v d) => (Int -> a -> b -> c -> d) -> v a -> v b -> v c -> v d

izipWith4 :: (Vector v a, Vector v b, Vector v c, Vector v d, Vector v e) => (Int -> a -> b -> c -> d -> e) -> v a -> v b -> v c -> v d -> v e

izipWith5 :: (Vector v a, Vector v b, Vector v c, Vector v d, Vector v e, Vector v f) => (Int -> a -> b -> c -> d -> e -> f) -> v a -> v b -> v c -> v d -> v e -> v f

izipWith6 :: (Vector v a, Vector v b, Vector v c, Vector v d, Vector v e, Vector v f, Vector v g) => (Int -> a -> b -> c -> d -> e -> f -> g) -> v a -> v b -> v c -> v d -> v e -> v f -> v g

zip :: (Vector v a, Vector v b, Vector v (a, b)) => v a -> v b -> v (a, b)

zip3 :: (Vector v a, Vector v b, Vector v c, Vector v (a, b, c)) => v a -> v b -> v c -> v (a, b, c)

zip4 :: (Vector v a, Vector v b, Vector v c, Vector v d, Vector v (a, b, c, d)) => v a -> v b -> v c -> v d -> v (a, b, c, d)

zip5 :: (Vector v a, Vector v b, Vector v c, Vector v d, Vector v e, Vector v (a, b, c, d, e)) => v a -> v b -> v c -> v d -> v e -> v (a, b, c, d, e)

zip6 :: (Vector v a, Vector v b, Vector v c, Vector v d, Vector v e, Vector v f, Vector v (a, b, c, d, e, f)) => v a -> v b -> v c -> v d -> v e -> v f -> v (a, b, c, d, e, f)

Monadic zipping

zipWithM :: (Monad m, Vector v a, Vector v b, Vector v c) => (a -> b -> m c) -> v a -> v b -> m (v c)

izipWithM :: (Monad m, Vector v a, Vector v b, Vector v c) => (Int -> a -> b -> m c) -> v a -> v b -> m (v c)

zipWithM_ :: (Monad m, Vector v a, Vector v b) => (a -> b -> m c) -> v a -> v b -> m ()

izipWithM_ :: (Monad m, Vector v a, Vector v b) => (Int -> a -> b -> m c) -> v a -> v b -> m ()

Unzipping

unzip :: (Vector v a, Vector v b, Vector v (a, b)) => v (a, b) -> (v a, v b)

unzip3 :: (Vector v a, Vector v b, Vector v c, Vector v (a, b, c)) => v (a, b, c) -> (v a, v b, v c)

unzip4 :: (Vector v a, Vector v b, Vector v c, Vector v d, Vector v (a, b, c, d)) => v (a, b, c, d) -> (v a, v b, v c, v d)

unzip5 :: (Vector v a, Vector v b, Vector v c, Vector v d, Vector v e, Vector v (a, b, c, d, e)) => v (a, b, c, d, e) -> (v a, v b, v c, v d, v e)

unzip6 :: (Vector v a, Vector v b, Vector v c, Vector v d, Vector v e, Vector v f, Vector v (a, b, c, d, e, f)) => v (a, b, c, d, e, f) -> (v a, v b, v c, v d, v e, v f)

Working with predicates

Filtering

filter :: Vector v a => (a -> Bool) -> v a -> v a

ifilter :: Vector v a => (Int -> a -> Bool) -> v a -> v a

uniq :: (Vector v a, Eq a) => v a -> v a

mapMaybe :: (Vector v a, Vector v b) => (a -> Maybe b) -> v a -> v b

imapMaybe :: (Vector v a, Vector v b) => (Int -> a -> Maybe b) -> v a -> v b

filterM :: (Monad m, Vector v a) => (a -> m Bool) -> v a -> m (v a)

takeWhile :: Vector v a => (a -> Bool) -> v a -> v a

dropWhile :: Vector v a => (a -> Bool) -> v a -> v a

Partitioning

partition :: Vector v a => (a -> Bool) -> v a -> (v a, v a)

unstablePartition :: Vector v a => (a -> Bool) -> v a -> (v a, v a)

span :: Vector v a => (a -> Bool) -> v a -> (v a, v a)

break :: Vector v a => (a -> Bool) -> v a -> (v a, v a)

Searching

elem :: (Vector v a, Eq a) => a -> v a -> Bool

notElem :: (Vector v a, Eq a) => a -> v a -> Bool

find :: Vector v a => (a -> Bool) -> v a -> Maybe a

findIndex :: Vector v a => (a -> Bool) -> v a -> Maybe Int

findIndices :: (Vector v a, Vector v Int) => (a -> Bool) -> v a -> v Int

elemIndex :: (Vector v a, Eq a) => a -> v a -> Maybe Int

elemIndices :: (Vector v a, Vector v Int, Eq a) => a -> v a -> v Int

Folding

foldl :: Vector v b => (a -> b -> a) -> a -> v b -> a

foldl' :: Vector v b => (a -> b -> a) -> a -> v b -> a

foldr :: Vector v a => (a -> b -> b) -> b -> v a -> b

foldr' :: Vector v a => (a -> b -> b) -> b -> v a -> b

ifoldl :: Vector v b => (a -> Int -> b -> a) -> a -> v b -> a

ifoldl' :: Vector v b => (a -> Int -> b -> a) -> a -> v b -> a

ifoldr :: Vector v a => (Int -> a -> b -> b) -> b -> v a -> b

ifoldr' :: Vector v a => (Int -> a -> b -> b) -> b -> v a -> b

Specialised folds

all :: Vector v a => (a -> Bool) -> v a -> Bool

any :: Vector v a => (a -> Bool) -> v a -> Bool

and :: Vector v Bool => v Bool -> Bool

or :: Vector v Bool => v Bool -> Bool

sum :: (Vector v a, Num a) => v a -> a

product :: (Vector v a, Num a) => v a -> a

Monadic folds

foldM :: (Monad m, Vector v b) => (a -> b -> m a) -> a -> v b -> m a

ifoldM :: (Monad m, Vector v b) => (a -> Int -> b -> m a) -> a -> v b -> m a

foldM' :: (Monad m, Vector v b) => (a -> b -> m a) -> a -> v b -> m a

ifoldM' :: (Monad m, Vector v b) => (a -> Int -> b -> m a) -> a -> v b -> m a

foldM_ :: (Monad m, Vector v b) => (a -> b -> m a) -> a -> v b -> m ()

ifoldM_ :: (Monad m, Vector v b) => (a -> Int -> b -> m a) -> a -> v b -> m ()

foldM'_ :: (Monad m, Vector v b) => (a -> b -> m a) -> a -> v b -> m ()

ifoldM'_ :: (Monad m, Vector v b) => (a -> Int -> b -> m a) -> a -> v b -> m ()

Monadic sequencing

sequence :: (Monad m, Vector v a, Vector v (m a)) => v (m a) -> m (v a)

sequence_ :: (Monad m, Vector v (m a)) => v (m a) -> m ()

Prefix sums (scans)

prescanl :: (Vector v a, Vector v b) => (a -> b -> a) -> a -> v b -> v a

prescanl' :: (Vector v a, Vector v b) => (a -> b -> a) -> a -> v b -> v a

postscanl :: (Vector v a, Vector v b) => (a -> b -> a) -> a -> v b -> v a

postscanl' :: (Vector v a, Vector v b) => (a -> b -> a) -> a -> v b -> v a

scanl :: (Vector v a, Vector v b) => (a -> b -> a) -> a -> v b -> v a

scanl' :: (Vector v a, Vector v b) => (a -> b -> a) -> a -> v b -> v a

iscanl :: (Vector v a, Vector v b) => (Int -> a -> b -> a) -> a -> v b -> v a

iscanl' :: (Vector v a, Vector v b) => (Int -> a -> b -> a) -> a -> v b -> v a

prescanr :: (Vector v a, Vector v b) => (a -> b -> b) -> b -> v a -> v b

prescanr' :: (Vector v a, Vector v b) => (a -> b -> b) -> b -> v a -> v b

postscanr :: (Vector v a, Vector v b) => (a -> b -> b) -> b -> v a -> v b

postscanr' :: (Vector v a, Vector v b) => (a -> b -> b) -> b -> v a -> v b

scanr :: (Vector v a, Vector v b) => (a -> b -> b) -> b -> v a -> v b

scanr' :: (Vector v a, Vector v b) => (a -> b -> b) -> b -> v a -> v b

iscanr :: (Vector v a, Vector v b) => (Int -> a -> b -> b) -> b -> v a -> v b

iscanr' :: (Vector v a, Vector v b) => (Int -> a -> b -> b) -> b -> v a -> v b

Conversions

Lists

toList :: Vector v a => v a -> [a]

fromList :: Vector v a => [a] -> v a

fromListN :: Vector v a => Int -> [a] -> v a

Different vector types

convert :: (Vector v a, Vector w a) => v a -> w a

Mutable vectors

freeze :: (PrimMonad m, Vector v a) => Mutable v (PrimState m) a -> m (v a)

thaw :: (PrimMonad m, Vector v a) => v a -> m (Mutable v (PrimState m) a)

copy :: (HasCallStack, PrimMonad m, Vector v a) => Mutable v (PrimState m) a -> v a -> m ()

Fusion support

Conversion to/from Bundles

stream :: Vector v a => v a -> Bundle v a

unstream :: Vector v a => Bundle v a -> v a

streamR :: forall v a (u :: Type -> Type). Vector v a => v a -> Bundle u a

unstreamR :: Vector v a => Bundle v a -> v a

Recycling support

new :: Vector v a => New v a -> v a

clone :: Vector v a => v a -> New v a

Utilities

Comparisons

eq :: (Vector v a, Eq a) => v a -> v a -> Bool

cmp :: (Vector v a, Ord a) => v a -> v a -> Ordering

eqBy :: (Vector v a, Vector v b) => (a -> b -> Bool) -> v a -> v b -> Bool

cmpBy :: (Vector v a, Vector v b) => (a -> b -> Ordering) -> v a -> v b -> Ordering

Show and Read

showsPrec :: (Vector v a, Show a) => Int -> v a -> ShowS

readPrec :: (Vector v a, Read a) => ReadPrec (v a)

liftShowsPrec :: Vector v a => (Int -> a -> ShowS) -> ([a] -> ShowS) -> Int -> v a -> ShowS

liftReadsPrec :: Vector v a => (Int -> ReadS a) -> ReadS [a] -> Int -> ReadS (v a)

Data and Typeable

gfoldl :: (Vector v a, Data a) => (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> v a -> c (v a)

dataCast :: (Vector v a, Data a, Typeable v, Typeable t) => (forall d. Data d => c (t d)) -> Maybe (c (v a))

mkType :: String -> DataType