RIO.Vector.Storable

Description

Storable Vector. Import as:

import qualified RIO.Vector.Storable as VS

This module does not export any partial or unsafe functions. For those, see RIO.Vector.Storable.Partial and RIO.Vector.Storable.Unsafe

Storable vectors

data Vector a

data MVector s a

Constructors

MVector !Int !(ForeignPtr a)

class Storable a

The member functions of this class facilitate writing values of primitive types to raw memory (which may have been allocated with the above mentioned routines) and reading values from blocks of raw memory. The class, furthermore, includes support for computing the storage requirements and alignment restrictions of storable types.

Memory addresses are represented as values of type Ptr a, for some a which is an instance of class Storable. The type argument to Ptr helps provide some valuable type safety in FFI code (you can't mix pointers of different types without an explicit cast), while helping the Haskell type system figure out which marshalling method is needed for a given pointer.

All marshalling between Haskell and a foreign language ultimately boils down to translating Haskell data structures into the binary representation of a corresponding data structure of the foreign language and vice versa. To code this marshalling in Haskell, it is necessary to manipulate primitive data types stored in unstructured memory blocks. The class Storable facilitates this manipulation on all types for which it is instantiated, which are the standard basic types of Haskell, the fixed size Int types (Int8, Int16, Int32, Int64), the fixed size Word types (Word8, Word16, Word32, Word64), StablePtr, all types from Foreign.C.Types, as well as Ptr.

Minimal complete definition

sizeOf, alignment, (peek | peekElemOff | peekByteOff), (poke | pokeElemOff | pokeByteOff)

Accessors

Length information

length :: Storable a => Vector a -> Int

null :: Storable a => Vector a -> Bool

Indexing

(!?) :: Storable a => Vector a -> Int -> Maybe a

Extracting subvectors

slice :: Storable a => Int -> Int -> Vector a -> Vector a

take :: Storable a => Int -> Vector a -> Vector a

drop :: Storable a => Int -> Vector a -> Vector a

splitAt :: Storable a => Int -> Vector a -> (Vector a, Vector a)

Construction

Initialisation

empty :: Storable a => Vector a

singleton :: Storable a => a -> Vector a

replicate :: Storable a => Int -> a -> Vector a

generate :: Storable a => Int -> (Int -> a) -> Vector a

iterateN :: Storable a => Int -> (a -> a) -> a -> Vector a

Monadic initialisation

replicateM :: (Monad m, Storable a) => Int -> m a -> m (Vector a)

generateM :: (Monad m, Storable a) => Int -> (Int -> m a) -> m (Vector a)

iterateNM :: (Monad m, Storable a) => Int -> (a -> m a) -> a -> m (Vector a)

create :: Storable a => (forall s. ST s (MVector s a)) -> Vector a

createT :: (Traversable f, Storable a) => (forall s. ST s (f (MVector s a))) -> f (Vector a)

Unfolding

unfoldr :: Storable a => (b -> Maybe (a, b)) -> b -> Vector a

unfoldrN :: Storable a => Int -> (b -> Maybe (a, b)) -> b -> Vector a

unfoldrM :: (Monad m, Storable a) => (b -> m (Maybe (a, b))) -> b -> m (Vector a)

unfoldrNM :: (Monad m, Storable a) => Int -> (b -> m (Maybe (a, b))) -> b -> m (Vector a)

constructN :: Storable a => Int -> (Vector a -> a) -> Vector a

constructrN :: Storable a => Int -> (Vector a -> a) -> Vector a

Enumeration

enumFromN :: (Storable a, Num a) => a -> Int -> Vector a

enumFromStepN :: (Storable a, Num a) => a -> a -> Int -> Vector a

enumFromTo :: (Storable a, Enum a) => a -> a -> Vector a

enumFromThenTo :: (Storable a, Enum a) => a -> a -> a -> Vector a

Concatenation

cons :: Storable a => a -> Vector a -> Vector a

snoc :: Storable a => Vector a -> a -> Vector a

(++) :: Storable a => Vector a -> Vector a -> Vector a

concat :: Storable a => [Vector a] -> Vector a

Restricting memory usage

force :: Storable a => Vector a -> Vector a

Modifying vectors

Permutations

reverse :: Storable a => Vector a -> Vector a

Safe destructive update

modify :: Storable a => (forall s. MVector s a -> ST s ()) -> Vector a -> Vector a

Elementwise operations

Mapping

map :: (Storable a, Storable b) => (a -> b) -> Vector a -> Vector b

imap :: (Storable a, Storable b) => (Int -> a -> b) -> Vector a -> Vector b

concatMap :: (Storable a, Storable b) => (a -> Vector b) -> Vector a -> Vector b

Monadic mapping

mapM :: (Monad m, Storable a, Storable b) => (a -> m b) -> Vector a -> m (Vector b)

mapM_ :: (Monad m, Storable a) => (a -> m b) -> Vector a -> m ()

forM :: (Monad m, Storable a, Storable b) => Vector a -> (a -> m b) -> m (Vector b)

forM_ :: (Monad m, Storable a) => Vector a -> (a -> m b) -> m ()

Zipping

zipWith :: (Storable a, Storable b, Storable c) => (a -> b -> c) -> Vector a -> Vector b -> Vector c

zipWith3 :: (Storable a, Storable b, Storable c, Storable d) => (a -> b -> c -> d) -> Vector a -> Vector b -> Vector c -> Vector d

zipWith4 :: (Storable a, Storable b, Storable c, Storable d, Storable e) => (a -> b -> c -> d -> e) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e

zipWith5 :: (Storable a, Storable b, Storable c, Storable d, Storable e, Storable f) => (a -> b -> c -> d -> e -> f) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f

zipWith6 :: (Storable a, Storable b, Storable c, Storable d, Storable e, Storable f, Storable g) => (a -> b -> c -> d -> e -> f -> g) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f -> Vector g

izipWith :: (Storable a, Storable b, Storable c) => (Int -> a -> b -> c) -> Vector a -> Vector b -> Vector c

izipWith3 :: (Storable a, Storable b, Storable c, Storable d) => (Int -> a -> b -> c -> d) -> Vector a -> Vector b -> Vector c -> Vector d

izipWith4 :: (Storable a, Storable b, Storable c, Storable d, Storable e) => (Int -> a -> b -> c -> d -> e) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e

izipWith5 :: (Storable a, Storable b, Storable c, Storable d, Storable e, Storable f) => (Int -> a -> b -> c -> d -> e -> f) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f

izipWith6 :: (Storable a, Storable b, Storable c, Storable d, Storable e, Storable f, Storable g) => (Int -> a -> b -> c -> d -> e -> f -> g) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f -> Vector g

Monadic zipping

zipWithM :: (Monad m, Storable a, Storable b, Storable c) => (a -> b -> m c) -> Vector a -> Vector b -> m (Vector c)

zipWithM_ :: (Monad m, Storable a, Storable b) => (a -> b -> m c) -> Vector a -> Vector b -> m ()

Working with predicates

Filtering

filter :: Storable a => (a -> Bool) -> Vector a -> Vector a

ifilter :: Storable a => (Int -> a -> Bool) -> Vector a -> Vector a

uniq :: (Storable a, Eq a) => Vector a -> Vector a

mapMaybe :: (Storable a, Storable b) => (a -> Maybe b) -> Vector a -> Vector b

imapMaybe :: (Storable a, Storable b) => (Int -> a -> Maybe b) -> Vector a -> Vector b

filterM :: (Monad m, Storable a) => (a -> m Bool) -> Vector a -> m (Vector a)

takeWhile :: Storable a => (a -> Bool) -> Vector a -> Vector a

dropWhile :: Storable a => (a -> Bool) -> Vector a -> Vector a

Partitioning

partition :: Storable a => (a -> Bool) -> Vector a -> (Vector a, Vector a)

unstablePartition :: Storable a => (a -> Bool) -> Vector a -> (Vector a, Vector a)

span :: Storable a => (a -> Bool) -> Vector a -> (Vector a, Vector a)

break :: Storable a => (a -> Bool) -> Vector a -> (Vector a, Vector a)

Searching

elem :: (Storable a, Eq a) => a -> Vector a -> Bool

notElem :: (Storable a, Eq a) => a -> Vector a -> Bool

find :: Storable a => (a -> Bool) -> Vector a -> Maybe a

findIndex :: Storable a => (a -> Bool) -> Vector a -> Maybe Int

findIndices :: Storable a => (a -> Bool) -> Vector a -> Vector Int

elemIndex :: (Storable a, Eq a) => a -> Vector a -> Maybe Int

elemIndices :: (Storable a, Eq a) => a -> Vector a -> Vector Int

Folding

foldl :: Storable b => (a -> b -> a) -> a -> Vector b -> a

foldl' :: Storable b => (a -> b -> a) -> a -> Vector b -> a

foldr :: Storable a => (a -> b -> b) -> b -> Vector a -> b

foldr' :: Storable a => (a -> b -> b) -> b -> Vector a -> b

ifoldl :: Storable b => (a -> Int -> b -> a) -> a -> Vector b -> a

ifoldl' :: Storable b => (a -> Int -> b -> a) -> a -> Vector b -> a

ifoldr :: Storable a => (Int -> a -> b -> b) -> b -> Vector a -> b

ifoldr' :: Storable a => (Int -> a -> b -> b) -> b -> Vector a -> b

Specialised folds

all :: Storable a => (a -> Bool) -> Vector a -> Bool

any :: Storable a => (a -> Bool) -> Vector a -> Bool

and :: Vector Bool -> Bool

or :: Vector Bool -> Bool

sum :: (Storable a, Num a) => Vector a -> a

product :: (Storable a, Num a) => Vector a -> a

Monadic folds

foldM :: (Monad m, Storable b) => (a -> b -> m a) -> a -> Vector b -> m a

foldM' :: (Monad m, Storable b) => (a -> b -> m a) -> a -> Vector b -> m a

foldM_ :: (Monad m, Storable b) => (a -> b -> m a) -> a -> Vector b -> m ()

foldM'_ :: (Monad m, Storable b) => (a -> b -> m a) -> a -> Vector b -> m ()

Prefix sums (scans)

prescanl :: (Storable a, Storable b) => (a -> b -> a) -> a -> Vector b -> Vector a

prescanl' :: (Storable a, Storable b) => (a -> b -> a) -> a -> Vector b -> Vector a

postscanl :: (Storable a, Storable b) => (a -> b -> a) -> a -> Vector b -> Vector a

postscanl' :: (Storable a, Storable b) => (a -> b -> a) -> a -> Vector b -> Vector a

scanl :: (Storable a, Storable b) => (a -> b -> a) -> a -> Vector b -> Vector a

scanl' :: (Storable a, Storable b) => (a -> b -> a) -> a -> Vector b -> Vector a

prescanr :: (Storable a, Storable b) => (a -> b -> b) -> b -> Vector a -> Vector b

prescanr' :: (Storable a, Storable b) => (a -> b -> b) -> b -> Vector a -> Vector b

postscanr :: (Storable a, Storable b) => (a -> b -> b) -> b -> Vector a -> Vector b

postscanr' :: (Storable a, Storable b) => (a -> b -> b) -> b -> Vector a -> Vector b

scanr :: (Storable a, Storable b) => (a -> b -> b) -> b -> Vector a -> Vector b

scanr' :: (Storable a, Storable b) => (a -> b -> b) -> b -> Vector a -> Vector b

Conversions

Lists

toList :: Storable a => Vector a -> [a]

fromList :: Storable a => [a] -> Vector a

fromListN :: Storable a => Int -> [a] -> Vector a

Different vector types

convert :: (Vector v a, Vector w a) => v a -> w a

Mutable vectors

freeze :: (Storable a, PrimMonad m) => MVector (PrimState m) a -> m (Vector a)

thaw :: (Storable a, PrimMonad m) => Vector a -> m (MVector (PrimState m) a)

copy :: (Storable a, PrimMonad m) => MVector (PrimState m) a -> Vector a -> m ()