RIO.Vector.Unboxed

Description

Unboxed Vector. Import as:

import qualified RIO.Vector.Unboxed as VU

This module does not export any partial or unsafe functions. For those, see RIO.Vector.Unboxed.Partial and RIO.Vector.Unboxed.Unsafe

Unboxed vectors

data family Vector a

data family MVector s a

class (Vector Vector a, MVector MVector a) => Unbox a

Accessors

Length information

length :: Unbox a => Vector a -> Int

null :: Unbox a => Vector a -> Bool

Indexing

(!?) :: Unbox a => Vector a -> Int -> Maybe a

Extracting subvectors

slice :: Unbox a => Int -> Int -> Vector a -> Vector a

take :: Unbox a => Int -> Vector a -> Vector a

drop :: Unbox a => Int -> Vector a -> Vector a

splitAt :: Unbox a => Int -> Vector a -> (Vector a, Vector a)

Construction

Initialisation

empty :: Unbox a => Vector a

singleton :: Unbox a => a -> Vector a

replicate :: Unbox a => Int -> a -> Vector a

generate :: Unbox a => Int -> (Int -> a) -> Vector a

iterateN :: Unbox a => Int -> (a -> a) -> a -> Vector a

Monadic initialisation

replicateM :: (Monad m, Unbox a) => Int -> m a -> m (Vector a)

generateM :: (Monad m, Unbox a) => Int -> (Int -> m a) -> m (Vector a)

iterateNM :: (Monad m, Unbox a) => Int -> (a -> m a) -> a -> m (Vector a)

create :: Unbox a => (forall s. ST s (MVector s a)) -> Vector a

createT :: (Traversable f, Unbox a) => (forall s. ST s (f (MVector s a))) -> f (Vector a)

Unfolding

unfoldr :: Unbox a => (b -> Maybe (a, b)) -> b -> Vector a

unfoldrN :: Unbox a => Int -> (b -> Maybe (a, b)) -> b -> Vector a

unfoldrM :: (Monad m, Unbox a) => (b -> m (Maybe (a, b))) -> b -> m (Vector a)

unfoldrNM :: (Monad m, Unbox a) => Int -> (b -> m (Maybe (a, b))) -> b -> m (Vector a)

constructN :: Unbox a => Int -> (Vector a -> a) -> Vector a

constructrN :: Unbox a => Int -> (Vector a -> a) -> Vector a

Enumeration

enumFromN :: (Unbox a, Num a) => a -> Int -> Vector a

enumFromStepN :: (Unbox a, Num a) => a -> a -> Int -> Vector a

enumFromTo :: (Unbox a, Enum a) => a -> a -> Vector a

enumFromThenTo :: (Unbox a, Enum a) => a -> a -> a -> Vector a

Concatenation

cons :: Unbox a => a -> Vector a -> Vector a

snoc :: Unbox a => Vector a -> a -> Vector a

(++) :: Unbox a => Vector a -> Vector a -> Vector a

concat :: Unbox a => [Vector a] -> Vector a

Restricting memory usage

force :: Unbox a => Vector a -> Vector a

Modifying vectors

Permutations

reverse :: Unbox a => Vector a -> Vector a

Safe destructive update

modify :: Unbox a => (forall s. MVector s a -> ST s ()) -> Vector a -> Vector a

Elementwise operations

Indexing

indexed :: Unbox a => Vector a -> Vector (Int, a)

Mapping

map :: (Unbox a, Unbox b) => (a -> b) -> Vector a -> Vector b

imap :: (Unbox a, Unbox b) => (Int -> a -> b) -> Vector a -> Vector b

concatMap :: (Unbox a, Unbox b) => (a -> Vector b) -> Vector a -> Vector b

Monadic mapping

mapM :: (Monad m, Unbox a, Unbox b) => (a -> m b) -> Vector a -> m (Vector b)

imapM :: (Monad m, Unbox a, Unbox b) => (Int -> a -> m b) -> Vector a -> m (Vector b)

mapM_ :: (Monad m, Unbox a) => (a -> m b) -> Vector a -> m ()

imapM_ :: (Monad m, Unbox a) => (Int -> a -> m b) -> Vector a -> m ()

forM :: (Monad m, Unbox a, Unbox b) => Vector a -> (a -> m b) -> m (Vector b)

forM_ :: (Monad m, Unbox a) => Vector a -> (a -> m b) -> m ()

Zipping

zipWith :: (Unbox a, Unbox b, Unbox c) => (a -> b -> c) -> Vector a -> Vector b -> Vector c

zipWith3 :: (Unbox a, Unbox b, Unbox c, Unbox d) => (a -> b -> c -> d) -> Vector a -> Vector b -> Vector c -> Vector d

zipWith4 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e) => (a -> b -> c -> d -> e) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e

zipWith5 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e, Unbox f) => (a -> b -> c -> d -> e -> f) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f

zipWith6 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e, Unbox f, Unbox g) => (a -> b -> c -> d -> e -> f -> g) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f -> Vector g

izipWith :: (Unbox a, Unbox b, Unbox c) => (Int -> a -> b -> c) -> Vector a -> Vector b -> Vector c

izipWith3 :: (Unbox a, Unbox b, Unbox c, Unbox d) => (Int -> a -> b -> c -> d) -> Vector a -> Vector b -> Vector c -> Vector d

izipWith4 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e) => (Int -> a -> b -> c -> d -> e) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e

izipWith5 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e, Unbox f) => (Int -> a -> b -> c -> d -> e -> f) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f

izipWith6 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e, Unbox f, Unbox g) => (Int -> a -> b -> c -> d -> e -> f -> g) -> Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f -> Vector g

zip :: (Unbox a, Unbox b) => Vector a -> Vector b -> Vector (a, b)

zip3 :: (Unbox a, Unbox b, Unbox c) => Vector a -> Vector b -> Vector c -> Vector (a, b, c)

zip4 :: (Unbox a, Unbox b, Unbox c, Unbox d) => Vector a -> Vector b -> Vector c -> Vector d -> Vector (a, b, c, d)

zip5 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e) => Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector (a, b, c, d, e)

zip6 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e, Unbox f) => Vector a -> Vector b -> Vector c -> Vector d -> Vector e -> Vector f -> Vector (a, b, c, d, e, f)

Monadic zipping

zipWithM :: (Monad m, Unbox a, Unbox b, Unbox c) => (a -> b -> m c) -> Vector a -> Vector b -> m (Vector c)

izipWithM :: (Monad m, Unbox a, Unbox b, Unbox c) => (Int -> a -> b -> m c) -> Vector a -> Vector b -> m (Vector c)

zipWithM_ :: (Monad m, Unbox a, Unbox b) => (a -> b -> m c) -> Vector a -> Vector b -> m ()

izipWithM_ :: (Monad m, Unbox a, Unbox b) => (Int -> a -> b -> m c) -> Vector a -> Vector b -> m ()

Unzipping

unzip :: (Unbox a, Unbox b) => Vector (a, b) -> (Vector a, Vector b)

unzip3 :: (Unbox a, Unbox b, Unbox c) => Vector (a, b, c) -> (Vector a, Vector b, Vector c)

unzip4 :: (Unbox a, Unbox b, Unbox c, Unbox d) => Vector (a, b, c, d) -> (Vector a, Vector b, Vector c, Vector d)

unzip5 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e) => Vector (a, b, c, d, e) -> (Vector a, Vector b, Vector c, Vector d, Vector e)

unzip6 :: (Unbox a, Unbox b, Unbox c, Unbox d, Unbox e, Unbox f) => Vector (a, b, c, d, e, f) -> (Vector a, Vector b, Vector c, Vector d, Vector e, Vector f)

Working with predicates

Filtering

filter :: Unbox a => (a -> Bool) -> Vector a -> Vector a

ifilter :: Unbox a => (Int -> a -> Bool) -> Vector a -> Vector a

uniq :: (Unbox a, Eq a) => Vector a -> Vector a

mapMaybe :: (Unbox a, Unbox b) => (a -> Maybe b) -> Vector a -> Vector b

imapMaybe :: (Unbox a, Unbox b) => (Int -> a -> Maybe b) -> Vector a -> Vector b

filterM :: (Monad m, Unbox a) => (a -> m Bool) -> Vector a -> m (Vector a)

takeWhile :: Unbox a => (a -> Bool) -> Vector a -> Vector a

dropWhile :: Unbox a => (a -> Bool) -> Vector a -> Vector a

Partitioning

partition :: Unbox a => (a -> Bool) -> Vector a -> (Vector a, Vector a)

unstablePartition :: Unbox a => (a -> Bool) -> Vector a -> (Vector a, Vector a)

span :: Unbox a => (a -> Bool) -> Vector a -> (Vector a, Vector a)

break :: Unbox a => (a -> Bool) -> Vector a -> (Vector a, Vector a)

Searching

elem :: (Unbox a, Eq a) => a -> Vector a -> Bool

notElem :: (Unbox a, Eq a) => a -> Vector a -> Bool

find :: Unbox a => (a -> Bool) -> Vector a -> Maybe a

findIndex :: Unbox a => (a -> Bool) -> Vector a -> Maybe Int

findIndices :: Unbox a => (a -> Bool) -> Vector a -> Vector Int

elemIndex :: (Unbox a, Eq a) => a -> Vector a -> Maybe Int

elemIndices :: (Unbox a, Eq a) => a -> Vector a -> Vector Int

Folding

foldl :: Unbox b => (a -> b -> a) -> a -> Vector b -> a

foldl' :: Unbox b => (a -> b -> a) -> a -> Vector b -> a

foldr :: Unbox a => (a -> b -> b) -> b -> Vector a -> b

foldr' :: Unbox a => (a -> b -> b) -> b -> Vector a -> b

ifoldl :: Unbox b => (a -> Int -> b -> a) -> a -> Vector b -> a

ifoldl' :: Unbox b => (a -> Int -> b -> a) -> a -> Vector b -> a

ifoldr :: Unbox a => (Int -> a -> b -> b) -> b -> Vector a -> b

ifoldr' :: Unbox a => (Int -> a -> b -> b) -> b -> Vector a -> b

Specialised folds

all :: Unbox a => (a -> Bool) -> Vector a -> Bool

any :: Unbox a => (a -> Bool) -> Vector a -> Bool

and :: Vector Bool -> Bool

or :: Vector Bool -> Bool

sum :: (Unbox a, Num a) => Vector a -> a

product :: (Unbox a, Num a) => Vector a -> a

Monadic folds

foldM :: (Monad m, Unbox b) => (a -> b -> m a) -> a -> Vector b -> m a

ifoldM :: (Monad m, Unbox b) => (a -> Int -> b -> m a) -> a -> Vector b -> m a

foldM' :: (Monad m, Unbox b) => (a -> b -> m a) -> a -> Vector b -> m a

ifoldM' :: (Monad m, Unbox b) => (a -> Int -> b -> m a) -> a -> Vector b -> m a

foldM_ :: (Monad m, Unbox b) => (a -> b -> m a) -> a -> Vector b -> m ()

ifoldM_ :: (Monad m, Unbox b) => (a -> Int -> b -> m a) -> a -> Vector b -> m ()

foldM'_ :: (Monad m, Unbox b) => (a -> b -> m a) -> a -> Vector b -> m ()

ifoldM'_ :: (Monad m, Unbox b) => (a -> Int -> b -> m a) -> a -> Vector b -> m ()

Prefix sums (scans)

prescanl :: (Unbox a, Unbox b) => (a -> b -> a) -> a -> Vector b -> Vector a

prescanl' :: (Unbox a, Unbox b) => (a -> b -> a) -> a -> Vector b -> Vector a

postscanl :: (Unbox a, Unbox b) => (a -> b -> a) -> a -> Vector b -> Vector a

postscanl' :: (Unbox a, Unbox b) => (a -> b -> a) -> a -> Vector b -> Vector a

scanl :: (Unbox a, Unbox b) => (a -> b -> a) -> a -> Vector b -> Vector a

scanl' :: (Unbox a, Unbox b) => (a -> b -> a) -> a -> Vector b -> Vector a

prescanr :: (Unbox a, Unbox b) => (a -> b -> b) -> b -> Vector a -> Vector b

prescanr' :: (Unbox a, Unbox b) => (a -> b -> b) -> b -> Vector a -> Vector b

postscanr :: (Unbox a, Unbox b) => (a -> b -> b) -> b -> Vector a -> Vector b

postscanr' :: (Unbox a, Unbox b) => (a -> b -> b) -> b -> Vector a -> Vector b

scanr :: (Unbox a, Unbox b) => (a -> b -> b) -> b -> Vector a -> Vector b

scanr' :: (Unbox a, Unbox b) => (a -> b -> b) -> b -> Vector a -> Vector b

Conversions

Lists

toList :: Unbox a => Vector a -> [a]

fromList :: Unbox a => [a] -> Vector a

fromListN :: Unbox a => Int -> [a] -> Vector a

Different vector types

convert :: (Vector v a, Vector w a) => v a -> w a

Mutable vectors

freeze :: (Unbox a, PrimMonad m) => MVector (PrimState m) a -> m (Vector a)

thaw :: (Unbox a, PrimMonad m) => Vector a -> m (MVector (PrimState m) a)

copy :: (Unbox a, PrimMonad m) => MVector (PrimState m) a -> Vector a -> m ()