RIO.Vector.Unsafe

Description

Generic Vector interface unsafe functions. These perform no bounds checking, and may cause segmentation faults etc.! Import as:

import qualified RIO.Vector.Unsafe as V'

Immutable vectors

class MVector (Mutable v) a => Vector (v :: Type -> Type) a where

Minimal complete definition

basicUnsafeFreeze, basicUnsafeThaw, basicLength, basicUnsafeSlice, basicUnsafeIndexM

Methods

basicUnsafeFreeze :: Mutable v s a -> ST s (v a)

basicUnsafeThaw :: v a -> ST s (Mutable v s a)

basicLength :: v a -> Int

basicUnsafeSlice :: Int -> Int -> v a -> v a

basicUnsafeIndexM :: v a -> Int -> Box a

basicUnsafeCopy :: Mutable v s a -> v a -> ST s ()

elemseq :: v a -> a -> b -> b

Accessors

Indexing

unsafeIndex :: Vector v a => v a -> Int -> a

unsafeHead :: Vector v a => v a -> a

unsafeLast :: Vector v a => v a -> a

Monadic indexing

unsafeIndexM :: (Vector v a, Monad m) => v a -> Int -> m a

unsafeHeadM :: (Vector v a, Monad m) => v a -> m a

unsafeLastM :: (Vector v a, Monad m) => v a -> m a

Extracting subvectors

unsafeSlice :: Vector v a => Int -> Int -> v a -> v a

unsafeInit :: Vector v a => v a -> v a

unsafeTail :: Vector v a => v a -> v a

unsafeTake :: Vector v a => Int -> v a -> v a

unsafeDrop :: Vector v a => Int -> v a -> v a

Modifying vectors

Bulk updates

unsafeUpd :: Vector v a => v a -> [(Int, a)] -> v a

unsafeUpdate :: (Vector v a, Vector v (Int, a)) => v a -> v (Int, a) -> v a

unsafeUpdate_ :: (Vector v a, Vector v Int) => v a -> v Int -> v a -> v a

Accumulations

unsafeAccum :: Vector v a => (a -> b -> a) -> v a -> [(Int, b)] -> v a

unsafeAccumulate :: (Vector v a, Vector v (Int, b)) => (a -> b -> a) -> v a -> v (Int, b) -> v a

unsafeAccumulate_ :: (Vector v a, Vector v Int, Vector v b) => (a -> b -> a) -> v a -> v Int -> v b -> v a

Permutations

unsafeBackpermute :: (Vector v a, Vector v Int) => v a -> v Int -> v a

Conversions

Mutable vectors

unsafeFreeze :: (PrimMonad m, Vector v a) => Mutable v (PrimState m) a -> m (v a)

unsafeThaw :: (PrimMonad m, Vector v a) => v a -> m (Mutable v (PrimState m) a)

unsafeCopy :: (PrimMonad m, Vector v a) => Mutable v (PrimState m) a -> v a -> m ()