Servant.Server.UVerb

Documentation

respond :: forall (x :: Type) (xs :: [Type]) (f :: Type -> Type). (Applicative f, HasStatus x, IsMember x xs) => x -> f (Union xs)

return for UVerb handlers. Takes a value of any of the members of the open union, and will construct a union value in an Applicative (eg. Server).

class IsServerResource (cts :: [Type]) a

Minimal complete definition

resourceResponse, resourceHeaders

Orphan instances

(ReflectMethod method, AllMime contentTypes, All (IsServerResourceWithStatus contentTypes) as, Unique (Statuses as)) => HasServer (UVerb method contentTypes as :: Type) context

Instance details

Associated Types

type ServerT (UVerb method contentTypes as) m

Methods

route :: Proxy (UVerb method contentTypes as) -> Context context -> Delayed env (Server (UVerb method contentTypes as)) -> Router env

hoistServerWithContext :: Proxy (UVerb method contentTypes as) -> Proxy context -> (forall x. m x -> n x) -> ServerT (UVerb method contentTypes as) m -> ServerT (UVerb method contentTypes as) n