Statistics.Autocorrelation

Description

Functions for computing autocovariance and autocorrelation of a sample.

Documentation

autocovariance :: (Vector v Double, Vector v Int) => v Double -> v Double

Compute the autocovariance of a sample, i.e. the covariance of the sample against a shifted version of itself.

autocorrelation :: (Vector v Double, Vector v Int) => v Double -> (v Double, v Double, v Double)

Compute the autocorrelation function of a sample, and the upper and lower bounds of confidence intervals for each element.

Note: The calculation of the 95% confidence interval assumes a stationary Gaussian process.