Statistics.Distribution

Description

Type classes for probability distributions

Type classes

class Distribution d where

Type class common to all distributions. Only c.d.f. could be defined for both discrete and continuous distributions.

Minimal complete definition

(cumulative | complCumulative)

Methods

cumulative :: d -> Double -> Double

Cumulative distribution function. The probability that a random variable X is less or equal than x, i.e. P(X≤x). Cumulative should be defined for infinities as well:

cumulative d +∞ = 1
cumulative d -∞ = 0

complCumulative :: d -> Double -> Double

One's complement of cumulative distribution:

complCumulative d x = 1 - cumulative d x

It's useful when one is interested in P(X>x) and expression on the right side begin to lose precision. This function have default implementation but implementors are encouraged to provide more precise implementation.

class Distribution d => DiscreteDistr d where

Discrete probability distribution.

Minimal complete definition

(probability | logProbability)

Methods

probability :: d -> Int -> Double

Probability of n-th outcome.

logProbability :: d -> Int -> Double

Logarithm of probability of n-th outcome

class Distribution d => ContDistr d where

Continuous probability distribution.

Minimal complete definition is quantile and either density or logDensity.

Minimal complete definition

(density | logDensity), (quantile | complQuantile)

Methods

density :: d -> Double -> Double

Probability density function. Probability that random variable X lies in the infinitesimal interval [x,x+δx) equal to density(x)⋅δx

logDensity :: d -> Double -> Double

Natural logarithm of density.

quantile :: d -> Double -> Double

Inverse of the cumulative distribution function. The value x for which P(X≤x) = p. If probability is outside of [0,1] range function should call error

complQuantile :: d -> Double -> Double

1-complement of quantile:

complQuantile x ≡ quantile (1 - x)

Distribution statistics

class Distribution d => MaybeMean d where

Type class for distributions with mean. maybeMean should return Nothing if it's undefined for current value of data

Methods

maybeMean :: d -> Maybe Double

class MaybeMean d => Mean d where

Type class for distributions with mean. If a distribution has finite mean for all valid values of parameters it should be instance of this type class.

Methods

mean :: d -> Double

class MaybeMean d => MaybeVariance d where

Type class for distributions with variance. If variance is undefined for some parameter values both maybeVariance and maybeStdDev should return Nothing.

Minimal complete definition is maybeVariance or maybeStdDev

Minimal complete definition

(maybeVariance | maybeStdDev)

Methods

maybeVariance :: d -> Maybe Double

maybeStdDev :: d -> Maybe Double

class (Mean d, MaybeVariance d) => Variance d where

Type class for distributions with variance. If distribution have finite variance for all valid parameter values it should be instance of this type class.

Minimal complete definition is variance or stdDev

Minimal complete definition

(variance | stdDev)

Methods

variance :: d -> Double

stdDev :: d -> Double

class Distribution d => MaybeEntropy d where

Type class for distributions with entropy, meaning Shannon entropy in the case of a discrete distribution, or differential entropy in the case of a continuous one. maybeEntropy should return Nothing if entropy is undefined for the chosen parameter values.

Methods

maybeEntropy :: d -> Maybe Double

Returns the entropy of a distribution, in nats, if such is defined.

class MaybeEntropy d => Entropy d where

Type class for distributions with entropy, meaning Shannon entropy in the case of a discrete distribution, or differential entropy in the case of a continuous one. If the distribution has well-defined entropy for all valid parameter values then it should be an instance of this type class.

Methods

entropy :: d -> Double

Returns the entropy of a distribution, in nats.

class FromSample d a where

Estimate distribution from sample. First parameter in sample is distribution type and second is element type.

Methods

fromSample :: Vector v a => v a -> Maybe d

Estimate distribution from sample. Returns Nothing if there is not enough data, or if no usable fit results from the method used, e.g., the estimated distribution parameters would be invalid or inaccurate.

Random number generation

class Distribution d => ContGen d where

Generate discrete random variates which have given distribution.

Methods

genContVar :: StatefulGen g m => d -> g -> m Double

class (DiscreteDistr d, ContGen d) => DiscreteGen d where

Generate discrete random variates which have given distribution. ContGen is superclass because it's always possible to generate real-valued variates from integer values

Methods

genDiscreteVar :: StatefulGen g m => d -> g -> m Int

genContinuous :: (ContDistr d, StatefulGen g m) => d -> g -> m Double

Generate variates from continuous distribution using inverse transform rule.

Helper functions

findRoot

Arguments

:: ContDistr d

=> d
Distribution

-> Double
Probability p

-> Double
Initial guess

-> Double
Lower bound on interval

-> Double
Upper bound on interval

-> Double

Approximate the value of X for which P(x>X)=p.

This method uses a combination of Newton-Raphson iteration and bisection with the given guess as a starting point. The upper and lower bounds specify the interval in which the probability distribution reaches the value p.

sumProbabilities :: DiscreteDistr d => d -> Int -> Int -> Double

Sum probabilities in inclusive interval.