Statistics.Distribution.Beta

Description

Documentation

data BetaDistribution

The beta distribution

Constructor

betaDistr

Arguments

:: Double
Shape parameter alpha

-> Double
Shape parameter beta

-> BetaDistribution

Create beta distribution. Both shape parameters must be positive.

betaDistrE

Arguments

:: Double
Shape parameter alpha

-> Double
Shape parameter beta

-> Maybe BetaDistribution

Create beta distribution. Both shape parameters must be positive.

improperBetaDistr

Arguments

:: Double
Shape parameter alpha

-> Double
Shape parameter beta

-> BetaDistribution

Create beta distribution. Both shape parameters must be non-negative. So it allows to construct improper beta distribution which could be used as improper prior.

improperBetaDistrE

Arguments

:: Double
Shape parameter alpha

-> Double
Shape parameter beta

-> Maybe BetaDistribution

Create beta distribution. Both shape parameters must be non-negative. So it allows to construct improper beta distribution which could be used as improper prior.

Accessors

bdAlpha :: BetaDistribution -> Double

Alpha shape parameter

bdBeta :: BetaDistribution -> Double

Beta shape parameter