Statistics.Distribution.Geometric

Description

The Geometric distribution. There are two variants of distribution. First is the probability distribution of the number of Bernoulli trials needed to get one success, supported on the set 1,2... Sometimes it's referred to as the shifted geometric distribution to distinguish from another one.

Second variant is probability distribution of the number of failures before first success, defined over the set 0,1...

Documentation

data GeometricDistribution

Distribution over [1..]

data GeometricDistribution0

Distribution over [0..]

Constructors

geometric

Arguments

:: Double
Success rate

-> GeometricDistribution

Create geometric distribution.

geometricE

Arguments

:: Double
Success rate

-> Maybe GeometricDistribution

Create geometric distribution.

geometric0

Arguments

:: Double
Success rate

-> GeometricDistribution0

Create geometric distribution.

geometric0E

Arguments

:: Double
Success rate

-> Maybe GeometricDistribution0

Create geometric distribution.

Accessors

gdSuccess :: GeometricDistribution -> Double

gdSuccess0 :: GeometricDistribution0 -> Double