Statistics.Quantile

Description

Functions for approximating quantiles, i.e. points taken at regular intervals from the cumulative distribution function of a random variable.

The number of quantiles is described below by the variable q, so with q=4, a 4-quantile (also known as a quartile) has 4 intervals, and contains 5 points. The parameter k describes the desired point, where 0 ≤ k ≤ q.

Quantile estimation functions

Below is family of functions which use same algorithm for estimation of sample quantiles. It approximates empirical CDF as continuous piecewise function which interpolates linearly between points \((X_k,p_k)\) where \(X_k\) is k-th order statistics (k-th smallest element) and \(p_k\) is probability corresponding to it. ContParam determines how \(p_k\) is chosen. For more detailed explanation see [Hyndman1996].

This is the method used by most statistical software, such as R, Mathematica, SPSS, and S.

data ContParam

Parameters α and β to the continuousBy function. Exact meaning of parameters is described in [Hyndman1996] in section "Piecewise linear functions"

Constructors

ContParam !Double !Double

class Default a where

Minimal complete definition

Nothing

Methods

def :: a

quantile

Arguments

:: Vector v Double

=> ContParam
Parameters α and β.

-> Int
k, the desired quantile.

-> Int
q, the number of quantiles.

-> v Double
x, the sample data.

-> Double

O(n·log n). Estimate the kth q-quantile of a sample x, using the continuous sample method with the given parameters.

The following properties should hold, otherwise an error will be thrown.

input sample must be nonempty

the input does not contain NaN

0 ≤ k ≤ q

quantiles :: (Vector v Double, Foldable f, Functor f) => ContParam -> f Int -> Int -> v Double -> f Double

O(k·n·log n). Estimate set of the kth q-quantile of a sample x, using the continuous sample method with the given parameters. This is faster than calling quantile repeatedly since sample should be sorted only once

The following properties should hold, otherwise an error will be thrown.

input sample must be nonempty

the input does not contain NaN

for every k in set of quantiles 0 ≤ k ≤ q

quantilesVec :: (Vector v Double, Vector v Int) => ContParam -> v Int -> Int -> v Double -> v Double

O(k·n·log n). Same as quantiles but uses Vector container instead of Foldable one.

Parameters for the continuous sample method

cadpw :: ContParam

California Department of Public Works definition, α=0, β=1. Gives a linear interpolation of the empirical CDF. This corresponds to method 4 in R and Mathematica.

hazen :: ContParam

Hazen's definition, α=0.5, β=0.5. This is claimed to be popular among hydrologists. This corresponds to method 5 in R and Mathematica.

spss :: ContParam

Definition used by the SPSS statistics application, with α=0, β=0 (also known as Weibull's definition). This corresponds to method 6 in R and Mathematica.

s :: ContParam

Definition used by the S statistics application, with α=1, β=1. The interpolation points divide the sample range into n-1 intervals. This corresponds to method 7 in R and Mathematica and is default in R.

medianUnbiased :: ContParam

Median unbiased definition, α=1/3, β=1/3. The resulting quantile estimates are approximately median unbiased regardless of the distribution of x. This corresponds to method 8 in R and Mathematica.

normalUnbiased :: ContParam

Normal unbiased definition, α=3/8, β=3/8. An approximately unbiased estimate if the empirical distribution approximates the normal distribution. This corresponds to method 9 in R and Mathematica.

Other algorithms

weightedAvg

Arguments

:: Vector v Double

=> Int
k, the desired quantile.

-> Int
q, the number of quantiles.

-> v Double
x, the sample data.

-> Double

O(n·log n). Estimate the kth q-quantile of a sample, using the weighted average method. Up to rounding errors it's same as quantile s.

The following properties should hold otherwise an error will be thrown.

the length of the input is greater than 0

the input does not contain NaN

k ≥ 0 and k ≤ q

Median & other specializations

median

Arguments

:: Vector v Double

=> ContParam
Parameters α and β.

-> v Double
x, the sample data.

-> Double

O(n·log n) Estimate median of sample

mad

Arguments

:: Vector v Double

=> ContParam
Parameters α and β.

-> v Double
x, the sample data.

-> Double

O(n·log n). Estimate the median absolute deviation (MAD) of a sample x using continuousBy. It's robust estimate of variability in sample and defined as:

\[ MAD = \operatorname{median}(| X_i - \operatorname{median}(X) |) \]

midspread

Arguments

:: Vector v Double

=> ContParam
Parameters α and β.

-> Int
q, the number of quantiles.

-> v Double
x, the sample data.

-> Double

O(n·log n). Estimate the range between q-quantiles 1 and q-1 of a sample x, using the continuous sample method with the given parameters.

For instance, the interquartile range (IQR) can be estimated as follows:

midspread medianUnbiased 4 (U.fromList [1,1,2,2,3])
==> 1.333333

Deprecated

continuousBy

Arguments

:: Vector v Double

=> ContParam
Parameters α and β.

-> Int
k, the desired quantile.

-> Int
q, the number of quantiles.

-> v Double
x, the sample data.

-> Double

Deprecated: Use quantile instead

References

Weisstein, E.W. Quantile. MathWorld. http://mathworld.wolfram.com/Quantile.html

[Hyndman1996] Hyndman, R.J.; Fan, Y. (1996) Sample quantiles in statistical packages. American Statistician 50(4):361–365. http://www.jstor.org/stable/2684934