Statistics.Resampling

Description

Resampling statistics.

Data types

newtype Resample

A resample drawn randomly, with replacement, from a set of data points. Distinct from a normal array to make it harder for your humble author's brain to go wrong.

Constructors

Resample

Fields

fromResample :: Vector Double

data Bootstrap v a

Constructors

Bootstrap

Fields

fullSample :: !a

resamples :: v a

data Estimator

An estimator of a property of a sample, such as its mean.

The use of an algebraic data type here allows functions such as jackknife and bootstrapBCA to use more efficient algorithms when possible.

Constructors

Mean

Variance

VarianceUnbiased

StdDev

Function (Sample -> Double)

estimate :: Estimator -> Sample -> Double

Run an Estimator over a sample.

Resampling

resampleST

Arguments

:: PrimMonad m

=> Gen (PrimState m)

-> [Estimator]
Estimation functions.

-> Int
Number of resamples to compute.

-> Vector Double
Original sample.

-> m [Bootstrap Vector Double]

Single threaded and deterministic version of resample.

resample

Arguments

:: GenIO

-> [Estimator]
Estimation functions.

-> Int
Number of resamples to compute.

-> Vector Double
Original sample.

-> IO [(Estimator, Bootstrap Vector Double)]

O(e*r*s) Resample a data set repeatedly, with replacement, computing each estimate over the resampled data.

This function is expensive; it has to do work proportional to e*r*s, where e is the number of estimation functions, r is the number of resamples to compute, and s is the number of original samples.

To improve performance, this function will make use of all available CPUs. At least with GHC 7.0, parallel performance seems best if the parallel garbage collector is disabled (RTS option -qg).

resampleVector :: (PrimMonad m, Vector v a) => Gen (PrimState m) -> v a -> m (v a)

Create vector using resamples

Jackknife

jackknife :: Estimator -> Sample -> Vector Double

O(n) or O(n^2) Compute a statistical estimate repeatedly over a sample, each time omitting a successive element.

jackknifeMean :: Sample -> Vector Double

O(n) Compute the jackknife mean of a sample.

jackknifeVariance :: Sample -> Vector Double

O(n) Compute the jackknife variance of a sample.

jackknifeVarianceUnb :: Sample -> Vector Double

O(n) Compute the unbiased jackknife variance of a sample.

jackknifeStdDev :: Sample -> Vector Double

O(n) Compute the jackknife standard deviation of a sample.

Helper functions

splitGen :: Int -> GenIO -> IO [GenIO]

Split a generator into several that can run independently.