Statistics.Resampling.Bootstrap

Description

The bootstrap method for statistical inference.

Documentation

bootstrapBCA

Arguments

:: CL Double
Confidence level

-> Sample
Full data sample

-> [(Estimator, Bootstrap Vector Double)]
Estimates obtained from resampled data and estimator used for this.

-> [Estimate ConfInt Double]

Bias-corrected accelerated (BCA) bootstrap. This adjusts for both bias and skewness in the resampled distribution.

BCA algorithm is described in ch. 5 of Davison, Hinkley "Confidence intervals" in section 5.3 "Percentile method"

basicBootstrap

Arguments

:: (Vector v a, Ord a, Num a)

=> CL Double
Confidence vector

-> Bootstrap v a
Estimate from full sample and vector of estimates obtained from resamples

-> Estimate ConfInt a

Basic bootstrap. This method simply uses empirical quantiles for confidence interval.

References

Davison, A.C; Hinkley, D.V. (1997) Bootstrap methods and their application. http://statwww.epfl.ch/davison/BMA/