Statistics.Test.ChiSquared

Description

Pearson's chi squared test.

Documentation

chi2test

Arguments

:: (Vector v (Int, Double), Vector v Double)

=> Int
Number of additional degrees of freedom. One degree of freedom is due to the fact that the are N observation in total and accounted for automatically.

-> v (Int, Double)
Observation and expectation.

-> Maybe (Test ChiSquared)

Generic form of Pearson chi squared tests for binned data. Data sample is supplied in form of tuples (observed quantity, expected number of events). Both must be positive.

This test should be used only if all bins have expected values of at least 5.

chi2testCont

Arguments

:: (Vector v (Estimate NormalErr Double, Double), Vector v Double)

=> Int
Number of additional degrees of freedom.

-> v (Estimate NormalErr Double, Double)
Observation and expectation.

-> Maybe (Test ChiSquared)

Chi squared test for data with normal errors. Data is supplied in form of pair (observation with error, and expectation).

module Statistics.Test.Types