Statistics.Test.KruskalWallis

Description

Kruskal-Wallis test

kruskalWallisTest :: (Ord a, Unbox a) => [Vector a] -> Maybe (Test ())

Perform Kruskal-Wallis Test for the given samples and required significance. For additional information check kruskalWallis. This is just a helper function.

It uses Chi-Squared distribution for approximation as long as the sizes are larger than 5. Otherwise the test returns Nothing.

Building blocks

kruskalWallisRank :: (Unbox a, Ord a) => [Vector a] -> [Vector Double]

Kruskal-Wallis ranking.

All values are replaced by the absolute rank in the combined samples.

The samples and values need not to be ordered but the values in the result are ordered. Assigned ranks (ties are given their average rank).

kruskalWallis :: (Unbox a, Ord a) => [Vector a] -> Double

The Kruskal-Wallis Test.

In textbooks the output value is usually represented by K or H. This function already does the ranking.

module Statistics.Test.Types