Statistics.Test.StudentT

Description

Student's T-test is for assessing whether two samples have different mean. This module contain several variations of T-test. It's a parametric tests and assumes that samples are normally distributed.

Documentation

studentTTest

Arguments

:: Vector v Double

=> PositionTest
one- or two-tailed test

-> v Double
Sample A

-> v Double
Sample B

-> Maybe (Test StudentT)

Two-sample Student's t-test. It assumes that both samples are normally distributed and have same variance. Returns Nothing if sample sizes are not sufficient.

welchTTest

Arguments

:: Vector v Double

=> PositionTest
one- or two-tailed test

-> v Double
Sample A

-> v Double
Sample B

-> Maybe (Test StudentT)

Two-sample Welch's t-test. It assumes that both samples are normally distributed but doesn't assume that they have same variance. Returns Nothing if sample sizes are not sufficient.

pairedTTest

Arguments

:: forall v. Vector v (Double, Double)

=> PositionTest
one- or two-tailed test

-> v (Double, Double)
paired samples

-> Maybe (Test StudentT)

Paired two-sample t-test. Two samples are paired in a within-subject design. Returns Nothing if sample size is not sufficient.

module Statistics.Test.Types