Statistics.Test.Types

Documentation

data Test distr

Result of statistical test.

Constructors

Test

Fields

testSignificance :: !(PValue Double)

Probability of getting value of test statistics at least as extreme as measured.

testStatistics :: !Double

Statistic used for test.

testDistribution :: distr

Distribution of test statistics if null hypothesis is correct.

isSignificant :: PValue Double -> Test d -> TestResult

Check whether test is significant for given p-value.

data TestResult

Result of hypothesis testing

Constructors

Significant
Null hypothesis should be rejected

NotSignificant
Data is compatible with hypothesis

significant :: Bool -> TestResult

significant if parameter is True, not significant otherwise

data PositionTest

Test type for test which compare positional (mean,median etc.) information of samples.

Constructors

SamplesDiffer
Test whether samples differ in position. Null hypothesis is samples are not different

AGreater
Test if first sample (A) is larger than second (B). Null hypothesis is first sample is not larger than second.

BGreater
Test if second sample is larger than first.