System.Metrics.Prometheus.MetricId

Documentation

newtype Name

Construct with makeName to ensure that names use only valid characters

Constructors

Name

Fields

unName :: Text

newtype Labels

Constructors

Labels

Fields

unLabels :: Map Text Text

data MetricId

Constructors

MetricId

Fields

name :: Name

labels :: Labels

addLabel :: Text -> Text -> Labels -> Labels

fromList :: [(Text, Text)] -> Labels

toList :: Labels -> [(Text, Text)]

null :: Labels -> Bool

makeValid :: Text -> Text

Make the input match the regex [a-zA-Z_][a-zA-Z0-9_] which defines valid metric and label names, according to https://prometheus.io/docs/concepts/data_model/#metric-names-and-labels Replace invalid characters with _ and add a leading _ if the first character is only valid as a later character.

makeName :: Text -> Name

Construct a Name, replacing disallowed characters.