Test.Golden.Shelley.Key.ConvertCardanoAddressKey

Documentation

golden_convertCardanoAddressByronSigningKey ∷ Property

Test that converting a cardano-address Byron signing key yields the expected result.

golden_convertCardanoAddressIcarusSigningKey ∷ Property

Test that converting a cardano-address Icarus signing key yields the expected result.

golden_convertCardanoAddressShelleyPaymentSigningKey ∷ Property

Test that converting a cardano-address Shelley payment signing key yields the expected result.

golden_convertCardanoAddressShelleyStakeSigningKey ∷ Property

Test that converting a cardano-address Shelley stake signing key yields the expected result.