Test.OptParse

Documentation

checkTextEnvelopeFormat ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ TextEnvelopeType → FilePath → FilePath → m ()

Checks that the tvType and tvDescription are equivalent between two files.

equivalence ∷ (MonadTest m, Eq a, Show a, HasCallStack) ⇒ a → a → m ()

Check for equivalence between two types and perform a file cleanup on failure.

execCardanoCLI

Arguments

∷ (MonadTest m, MonadCatch m, MonadIO m, HasCallStack)

⇒ [String]
Arguments to the CLI command

→ m String
Captured stdout

Execute cardano-cli via the command line.

Waits for the process to finish and returns the stdout.

propertyOnce ∷ PropertyT IO () → Property

withSnd ∷ (a → b) → a → (a, b)

Return the supply value with the result of the supplied function as a tuple

noteInputFile ∷ (MonadTest m, HasCallStack) ⇒ FilePath → m FilePath

Return the input file path after annotating it relative to the project root directory

noteTempFile ∷ (MonadTest m, HasCallStack) ⇒ FilePath → FilePath → m FilePath

Return the test file path after annotating it relative to the project root directory