Test.Tasty.Hedgehog

Description

This package lets you test Hedgehog properties with tasty.

Typical usage would look like this:

testGroup "tasty-hedgehog tests" [
   testPropertyNamed "reverse involutive" "prop_reverse_involutive" prop_reverse_involutive
 , testPropertyNamed "sort idempotent"    "prop_sort_idempotent"    prop_sort_idempotent
 ]

Documentation

testProperty :: TestName -> Property -> TestTree

Create a TestTree from a Hedgehog Property.

testPropertyNamed :: TestName -> PropertyName -> Property -> TestTree

testPropertyNamed testName propertyName property creates a TestTree from property using testName as the displayed description for the property. The propertyName is used by Hedgehog when a failure occurs to provide instructions for how to re-run the property and should normally be set to a string representation of the property argument.

testPropertyNamed
 "reverse is involutive"
 "prop_reverse_involutive"
 prop_reverse_involutive

Since: 1.2.0.0

fromGroup :: Group -> TestTree

Create a TestTree from a Hedgehog Group.

Options you can pass in via tasty

newtype HedgehogReplay

The replay token to use for replaying a previous test run

Constructors

HedgehogReplay (Maybe (Skip, Seed))

newtype HedgehogShowReplay

If a test case fails, show a replay token for replaying tests

Constructors

HedgehogShowReplay Bool

newtype HedgehogTestLimit

The number of successful test cases required before Hedgehog will pass a test

Constructors

HedgehogTestLimit (Maybe TestLimit)

newtype HedgehogDiscardLimit

The number of discarded cases allowed before Hedgehog will fail a test

Constructors

HedgehogDiscardLimit (Maybe DiscardLimit)

newtype HedgehogShrinkLimit

The number of shrinks allowed before Hedgehog will fail a test

Constructors

HedgehogShrinkLimit (Maybe ShrinkLimit)

newtype HedgehogShrinkRetries

The number of times to re-run a test during shrinking

Constructors

HedgehogShrinkRetries (Maybe ShrinkRetries)