Test.Tasty.Ingredients.Basic

Description

This module exports the basic ingredients defined in the tasty packages.

Note that if defaultIngredients from Test.Tasty suits your needs, use that instead of importing this module.

Since: 0.8

Console test reporter

consoleTestReporter :: Ingredient

A simple console UI.

Since: 0.4

newtype Quiet

Do not print test results (see README for details).

Since: 0.8

Constructors

Quiet Bool

newtype HideSuccesses

Report only failed tests.

At the moment, this option only works globally. As an argument to localOption, it does nothing.

Since: 0.8

Constructors

HideSuccesses Bool

newtype AnsiTricks

By default, when the option --hide-successes is given and the output goes to an ANSI-capable terminal, we employ some ANSI terminal tricks to display the name of the currently running test and then erase it if it succeeds.

These tricks sometimes fail, however—in particular, when the test names happen to be longer than the width of the terminal window. See

https://github.com/UnkindPartition/tasty/issues/152

https://github.com/UnkindPartition/tasty/issues/250

When that happens, this option can be used to disable the tricks. In that case, the test name will be printed only once the test fails.

Since: 1.3

Constructors

AnsiTricks

Fields

getAnsiTricks :: Bool

Listing tests

listingTests :: Ingredient

The ingredient that provides the test listing functionality.

Since: 0.4

newtype ListTests

This option, when set to True, specifies that we should run in the «list tests» mode.

Since: 0.4

Constructors

ListTests Bool

testsNames :: OptionSet -> TestTree -> [TestName]

Obtain the list of all tests in the suite.

Since: 0.4

Adding options

includingOptions :: [OptionDescription] -> Ingredient

This ingredient doesn't do anything apart from registering additional options.

The option values can be accessed using askOption.

Since: 0.6