Test.Tasty.QuickCheck

Description

This module allows to use QuickCheck properties in tasty.

Documentation

testProperty :: Testable a => TestName -> a -> TestTree

Create a TestTree for a QuickCheck Testable property

testProperties :: TestName -> [(String, Property)] -> TestTree

Create a test from a list of QuickCheck properties. To be used with allProperties. E.g.

tests :: TestTree
tests = testProperties "Foo" $allProperties

newtype QuickCheckTests

Number of test cases for QuickCheck to generate

Constructors

QuickCheckTests Int

data QuickCheckReplay

Replay seed

Constructors

QuickCheckReplayNone
No seed

Since: 0.11

QuickCheckReplayLegacy Int
Legacy integer seed

Since: 0.11

QuickCheckReplay (QCGen, Int)
(qcgen, intSize) holds both the seed and the size to run QuickCheck tests

Since: 0.11

newtype QuickCheckShowReplay

If a test case fails unexpectedly, show the replay token

Constructors

QuickCheckShowReplay Bool

newtype QuickCheckMaxSize

Size of the biggest test cases

Constructors

QuickCheckMaxSize Int

newtype QuickCheckMaxRatio

Maximum number of of discarded tests per successful test before giving up.

Constructors

QuickCheckMaxRatio Int

newtype QuickCheckVerbose

Show the test cases that QuickCheck generates

Constructors

QuickCheckVerbose Bool

newtype QuickCheckMaxShrinks

Number of shrinks allowed before QuickCheck will fail a test.

Since: 0.10.2

Constructors

QuickCheckMaxShrinks Int

newtype QuickCheckTimeout

Timeout for individual tests within a property.

Since: 0.11.1

Constructors

QuickCheckTimeout Timeout

Re-export of Test.QuickCheck

data Fun a b

Constructors

Fun (a :-> b, b, Shrunk) (a -> b)

newtype Fixed a

Constructors

Fixed

Fields

getFixed :: a

data Property

class Arbitrary a where

Minimal complete definition

arbitrary

Methods

arbitrary :: Gen a

shrink :: a -> [a]

class Arbitrary1 (f :: Type -> Type) where

Minimal complete definition

liftArbitrary

Methods

liftArbitrary :: Gen a -> Gen (f a)

liftShrink :: (a -> [a]) -> f a -> [f a]

class Arbitrary2 (f :: Type -> Type -> Type) where

Minimal complete definition

liftArbitrary2

Methods

liftArbitrary2 :: Gen a -> Gen b -> Gen (f a b)

liftShrink2 :: (a -> [a]) -> (b -> [b]) -> f a b -> [f a b]

class CoArbitrary a where

Minimal complete definition

Nothing

Methods

coarbitrary :: a -> Gen b -> Gen b

class Function a where

Minimal complete definition

Nothing

Methods

function :: (a -> b) -> a :-> b

data Gen a

newtype ASCIIString

Constructors

ASCIIString

Fields

getASCIIString :: String

newtype Blind a

Constructors

Blind

Fields

getBlind :: a

data InfiniteList a

Constructors

InfiniteList

Fields

getInfiniteList :: [a]

infiniteListInternalData :: InfiniteListInternalData a

newtype Large a

Constructors

Large

Fields

getLarge :: a

newtype Negative a

Constructors

Negative

Fields

getNegative :: a

newtype NonEmptyList a

Constructors

NonEmpty

Fields

getNonEmpty :: [a]

newtype NonNegative a

Constructors

NonNegative

Fields

getNonNegative :: a

newtype NonPositive a

Constructors

NonPositive

Fields

getNonPositive :: a

newtype NonZero a

Constructors

NonZero

Fields

getNonZero :: a

newtype OrderedList a

Constructors

Ordered

Fields

getOrdered :: [a]

newtype Positive a

Constructors

Positive

Fields

getPositive :: a

newtype PrintableString

Constructors

PrintableString

Fields

getPrintableString :: String

newtype Shrink2 a

Constructors

Shrink2

Fields

getShrink2 :: a

class ShrinkState s a where

Methods

shrinkInit :: a -> s

shrinkState :: a -> s -> [(a, s)]

data Shrinking s a

Constructors

Shrinking s a

newtype Small a

Constructors

Small

Fields

getSmall :: a

data Smart a

Constructors

Smart Int a

newtype SortedList a

Constructors

Sorted

Fields

getSorted :: [a]

newtype UnicodeString

Constructors

UnicodeString

Fields

getUnicodeString :: String

data Discard

Constructors

Discard

class Testable prop where

Minimal complete definition

property

Methods

property :: prop -> Property

propertyForAllShrinkShow :: Gen a -> (a -> [a]) -> (a -> [String]) -> (a -> prop) -> Property

data Witness

Constructors

(Typeable a, Show a) => Wit a

data Confidence

Constructors

Confidence

Fields

certainty :: Integer

tolerance :: Double

pattern Success :: Int -> Int -> !(Map [String] Int) -> !(Map String Int) -> !(Map String (Map String Int)) -> String -> Result

pattern Failure :: Int -> Int -> Int -> Int -> Int -> QCGen -> Int -> String -> Maybe AnException -> String -> [String] -> [String] -> Set String -> [Witness] -> Result

pattern GaveUp :: Int -> Int -> !(Map [String] Int) -> !(Map String Int) -> !(Map String (Map String Int)) -> String -> Result

pattern NoExpectedFailure :: Int -> Int -> !(Map [String] Int) -> !(Map String Int) -> !(Map String (Map String Int)) -> String -> Result

classes :: Result -> Map String Int

failingClasses :: Result -> Set String

failingLabels :: Result -> [String]

failingTestCase :: Result -> [String]

labels :: Result -> Map [String] Int

numDiscarded :: Result -> Int

numShrinkFinal :: Result -> Int

numShrinkTries :: Result -> Int

numShrinks :: Result -> Int

numTests :: Result -> Int

output :: Result -> String

reason :: Result -> String

tables :: Result -> Map String (Map String Int)

theException :: Result -> Maybe AnException

usedSeed :: Result -> QCGen

usedSize :: Result -> Int

witnesses :: Result -> [Witness]

pattern Fn :: (a -> b) -> Fun a b

pattern Fn2 :: (a -> b -> c) -> Fun (a, b) c

pattern Fn3 :: (a -> b -> c -> d) -> Fun (a, b, c) d

(==>) :: Testable prop => Bool -> prop -> Property

(.&.) :: (Testable prop1, Testable prop2) => prop1 -> prop2 -> Property

choose :: Random a => (a, a) -> Gen a

getSize :: Gen Int

label :: Testable prop => String -> prop -> Property

(><) :: (Gen a -> Gen a) -> (Gen a -> Gen a) -> Gen a -> Gen a

monomorphic :: Name -> ExpQ

polyQuickCheck :: Name -> ExpQ

polyVerboseCheck :: Name -> ExpQ

applyArbitrary2 :: (Arbitrary a, Arbitrary b) => (a -> b -> r) -> Gen r

applyArbitrary3 :: (Arbitrary a, Arbitrary b, Arbitrary c) => (a -> b -> c -> r) -> Gen r

applyArbitrary4 :: (Arbitrary a, Arbitrary b, Arbitrary c, Arbitrary d) => (a -> b -> c -> d -> r) -> Gen r

arbitrary1 :: (Arbitrary1 f, Arbitrary a) => Gen (f a)

arbitrary2 :: (Arbitrary2 f, Arbitrary a, Arbitrary b) => Gen (f a b)

arbitraryASCIIChar :: Gen Char

arbitraryBoundedEnum :: (Bounded a, Enum a) => Gen a

arbitraryBoundedIntegral :: (Bounded a, Integral a) => Gen a

arbitraryBoundedRandom :: (Bounded a, Random a) => Gen a

arbitraryPrintableChar :: Gen Char

arbitrarySizedBoundedIntegral :: (Bounded a, Integral a) => Gen a

arbitrarySizedFractional :: Fractional a => Gen a

arbitrarySizedIntegral :: Integral a => Gen a

arbitrarySizedNatural :: Integral a => Gen a

arbitraryUnicodeChar :: Gen Char

coarbitraryEnum :: Enum a => a -> Gen b -> Gen b

coarbitraryIntegral :: Integral a => a -> Gen b -> Gen b

coarbitraryReal :: Real a => a -> Gen b -> Gen b

coarbitraryShow :: Show a => a -> Gen b -> Gen b

genericCoarbitrary :: (Generic a, GCoArbitrary (Rep a)) => a -> Gen b -> Gen b

genericShrink :: (Generic a, RecursivelyShrink (Rep a), GSubterms (Rep a) a) => a -> [a]

infiniteList :: Arbitrary a => Gen [a]

orderedList :: (Ord a, Arbitrary a) => Gen [a]

recursivelyShrink :: (Generic a, RecursivelyShrink (Rep a)) => a -> [a]

shrink1 :: (Arbitrary1 f, Arbitrary a) => f a -> [f a]

shrink2 :: (Arbitrary2 f, Arbitrary a, Arbitrary b) => f a b -> [f a b]

shrinkBoundedEnum :: (Bounded a, Enum a, Eq a) => a -> [a]

shrinkDecimal :: RealFrac a => a -> [a]

shrinkIntegral :: Integral a => a -> [a]

shrinkList :: (a -> [a]) -> [a] -> [[a]]

shrinkMap :: Arbitrary a => (a -> b) -> (b -> a) -> b -> [b]

shrinkMapBy :: (a -> b) -> (b -> a) -> (a -> [a]) -> b -> [b]

shrinkNothing :: a -> [a]

shrinkRealFrac :: RealFrac a => a -> [a]

subterms :: (Generic a, GSubterms (Rep a) a) => a -> [a]

vector :: Arbitrary a => Int -> Gen [a]

discard :: a

labelledExamples :: Testable prop => prop -> IO ()

labelledExamplesResult :: Testable prop => prop -> IO Result

labelledExamplesWith :: Testable prop => Args -> prop -> IO ()

labelledExamplesWithResult :: Testable prop => Args -> prop -> IO Result

applyFun :: Fun a b -> a -> b

applyFun2 :: Fun (a, b) c -> a -> b -> c

applyFun3 :: Fun (a, b, c) d -> a -> b -> c -> d

functionBoundedEnum :: (Eq a, Bounded a, Enum a) => (a -> b) -> a :-> b

functionIntegral :: Integral a => (a -> b) -> a :-> b

functionMap :: Function b => (a -> b) -> (b -> a) -> (a -> c) -> a :-> c

functionRealFrac :: RealFrac a => (a -> b) -> a :-> b

functionShow :: (Show a, Read a) => (a -> c) -> a :-> c

functionVoid :: (forall b. void -> b) -> void :-> c

chooseAny :: Random a => Gen a

chooseBoundedIntegral :: (Bounded a, Integral a) => (a, a) -> Gen a

chooseEnum :: Enum a => (a, a) -> Gen a

chooseInt :: (Int, Int) -> Gen Int

chooseInteger :: (Integer, Integer) -> Gen Integer

elements :: HasCallStack => [a] -> Gen a

frequency :: HasCallStack => [(Int, Gen a)] -> Gen a

generate :: Gen a -> IO a

growingElements :: HasCallStack => [a] -> Gen a

infiniteListOf :: Gen a -> Gen [a]

listOf :: Gen a -> Gen [a]

listOf1 :: Gen a -> Gen [a]

oneof :: HasCallStack => [Gen a] -> Gen a

resize :: HasCallStack => Int -> Gen a -> Gen a

sample :: Show a => Gen a -> IO ()

sample' :: Gen a -> IO [a]

scale :: (Int -> Int) -> Gen a -> Gen a

shuffle :: [a] -> Gen [a]

sized :: (Int -> Gen a) -> Gen a

sublistOf :: [a] -> Gen [a]

suchThat :: Gen a -> (a -> Bool) -> Gen a

suchThatMap :: Gen a -> (a -> Maybe b) -> Gen b

suchThatMaybe :: Gen a -> (a -> Bool) -> Gen (Maybe a)

variant :: Integral n => n -> Gen a -> Gen a

vectorOf :: Int -> Gen a -> Gen [a]

(.&&.) :: (Testable prop1, Testable prop2) => prop1 -> prop2 -> Property

(.||.) :: (Testable prop1, Testable prop2) => prop1 -> prop2 -> Property

(=/=) :: (Eq a, Show a) => a -> a -> Property

(===) :: (Eq a, Show a) => a -> a -> Property

again :: Testable prop => prop -> Property

castWitness :: Typeable a => Witness -> Maybe a

checkCoverage :: Testable prop => prop -> Property

checkCoverageWith :: Testable prop => Confidence -> prop -> Property

classify :: Testable prop => Bool -> String -> prop -> Property

coerceWitness :: Typeable a => Witness -> a

collect :: (Show a, Testable prop) => a -> prop -> Property

conjoin :: Testable prop => [prop] -> Property

counterexample :: Testable prop => String -> prop -> Property

cover :: Testable prop => Double -> Bool -> String -> prop -> Property

coverTable :: Testable prop => String -> [(String, Double)] -> prop -> Property

discardAfter :: Testable prop => Int -> prop -> Property

disjoin :: Testable prop => [prop] -> Property

expectFailure :: Testable prop => prop -> Property

forAll :: (Show a, Testable prop) => Gen a -> (a -> prop) -> Property

forAllBlind :: Testable prop => Gen a -> (a -> prop) -> Property

forAllShow :: Testable prop => Gen a -> (a -> String) -> (a -> prop) -> Property

forAllShrink :: (Show a, Testable prop) => Gen a -> (a -> [a]) -> (a -> prop) -> Property

forAllShrinkBlind :: Testable prop => Gen a -> (a -> [a]) -> (a -> prop) -> Property

forAllShrinkShow :: Testable prop => Gen a -> (a -> [a]) -> (a -> String) -> (a -> prop) -> Property

idempotentIOProperty :: Testable prop => IO prop -> Property

ioProperty :: Testable prop => IO prop -> Property

mapSize :: Testable prop => (Int -> Int) -> prop -> Property

noShrinking :: Testable prop => prop -> Property

once :: Testable prop => prop -> Property

printTestCase :: Testable prop => String -> prop -> Property

shrinking :: Testable prop => (a -> [a]) -> a -> (a -> prop) -> Property

stdConfidence :: Confidence

tabulate :: Testable prop => String -> [String] -> prop -> Property

total :: NFData a => a -> Property

verboseShrinking :: Testable prop => prop -> Property

whenFail :: Testable prop => IO () -> prop -> Property

whenFail' :: Testable prop => IO () -> prop -> Property

withDiscardRatio :: Testable prop => Int -> prop -> Property

withMaxShrinks :: Testable prop => Int -> prop -> Property

withMaxSize :: Testable prop => Int -> prop -> Property

withMaxSuccess :: Testable prop => Int -> prop -> Property

within :: Testable prop => Int -> prop -> Property

witness :: (Typeable a, Show a, Testable prop) => a -> prop -> Property

isSuccess :: Result -> Bool

recheck :: Testable prop => Result -> prop -> IO ()

Internal

If you are building a test suite, you don't need these functions.

They may be used by other tasty add-on packages (such as tasty-hspec).

newtype QC

Constructors

QC Property

optionSetToArgs :: OptionSet -> IO (Int, Args)

Convert tasty options into QuickCheck options.

This is a low-level function that was originally added for tasty-hspec but may be used by others.

The returned Int is kept only for backward compatibility purposes. It has no use in tasty-quickcheck.

Since: 0.9.1