Text.EDE.Filters

Description

The means to construct your own filters.

Prelude

The default filters available to a template are documented by the subsequent categories.

These filters cannot be overriden and attempting to supply your own filters to renderWith will cause the similarly named filters to disappear when they are merged with the prelude during evaluation. (See: union)

Boolean

! :: Bool -> Bool

See: not

&& :: Bool -> Bool -> Bool

|| :: Bool -> Bool -> Bool

Equality

== :: a -> a -> Bool

!= :: a -> a -> Bool

See: /=

Relational

> :: a -> a -> Bool

>= :: a -> a -> Bool

<= :: a -> a -> Bool

< :: a -> a -> Bool

Numeric

+ :: Scientific -> Scientific -> Scientific

- :: Scientific -> Scientific -> Scientific

* :: Scientific -> Scientific -> Scientific

abs :: Scientific -> Scientific

negate :: Scientific -> Scientific

signum :: Scientific -> Scientific

Fractional

ceiling :: Scientific -> Scientific

floor :: Scientific -> Scientific

round :: Scientific -> Scientific

truncate :: Scientific -> Scientific

Textual

dropLower :: Text -> Text

Drop preceding lowercase characters.

dropUpper :: Text -> Text

Drop preceding uppercase characters.

indentLines :: Text -> Text

prependLines :: Text -> Text

justifyLeft :: Text -> Text

Using whitespace as fill character.

justifyRight :: Text -> Text

Using whitespace as fill character.

center :: Text -> Text

Using whitespace as fill character.

replace :: Text -> Text

remove :: Text -> Text

Shortcut for: replace(pattern, "")

splitWords :: Text -> Text

strip :: Text -> Text

stripPrefix :: Text -> Text

stripSuffix :: Text -> Text

stripStart :: Text -> Text

stripEnd :: Text -> Text

takeWord :: Text -> Text

dropWord :: Text -> Text

lowerHead :: Text -> Text

upperHead :: Text -> Text

toCamel :: Text -> Text

toEllipsis :: Text -> Text

toEllipsisWith :: Text -> Text

toPascal :: Text -> Text

toSnake :: Text -> Text

toSpinal :: Text -> Text

toTitle :: Text -> Text

toTrain :: Text -> Text

toLower :: Text -> Text

toUpper :: Text -> Text

toOrdinal :: Scientific -> Text

Collection

length :: Collection -> Scientific

See: Text.length, Vector.length, HashMap.size

empty :: Collection -> Bool

See: Text.null, Vector.null, HashMap.null

reverse :: Collection -> Scientific

See: Text.reverse, Vector.reverse

head :: Collection -> Scientific

See: Text.head, Vector.head

last :: Collection -> Scientific

See: Text.last, Vector.last

tail :: Collection -> Scientific

See: Text.tail, Vector.tail

init :: Collection -> Scientific

See: Text.init, Vector.init

keys :: Collection -> Scientific

elems :: Collection -> Scientific

Polymorphic

show :: a -> Text

Constructing filters

data Term

A HOAS representation of (possibly partially applied) values in the environment.

Constructors

TVal !Value

TLam (Term -> Result Term)

Classes

class Quote a where

Minimal complete definition

Nothing

Methods

quote :: Id -> Int -> a -> Term

default quote :: ToJSON a => Id -> Int -> a -> Term

class Unquote a where

Minimal complete definition

Nothing

Methods

unquote :: Id -> Int -> Term -> Result a

default unquote :: FromJSON a => Id -> Int -> Term -> Result a

Restricted quoters

(@:) :: Quote a => Id -> a -> (Id, Term)

qapply :: Delta -> Term -> Term -> Result Term

Fully apply two Terms.

qpoly2 :: Quote a => Id -> (Value -> Value -> a) -> (Id, Term)

Quote a binary function which takes the most general binding value.

qnum1 :: Id -> (Scientific -> Scientific) -> (Id, Term)

Quote an unary numeric function.

qnum2 :: Quote a => Id -> (Scientific -> Scientific -> a) -> (Id, Term)

Quote a binary numeric function.

qcol1 :: (Quote a, Quote b, Quote c) => Id -> (Text -> a) -> (HashMap Text Value -> b) -> (Vector Value -> c) -> (Id, Term)

Quote a comprehensive set of unary functions to create a binding that supports all collection types.

Errors

typeErr :: Id -> Int -> AnsiDoc -> AnsiDoc -> Result a

argumentErr :: Pretty a => Id -> Int -> a -> Result b