Text.EDE.Internal.AST

Description

Warning: this is an internal module, and does not have a stable API or name. Functions in this module may not check or enforce preconditions expected by public modules. Use at your own risk!

AST smart constructors.

Documentation

newtype Fix f

Constructors

Fix (f (Fix f))

cofreeFix :: Functor f => a -> Fix f -> Cofree f a

var :: Id -> Var

eapp :: a -> [Exp a] -> Exp a

efun :: Id -> Exp a -> Exp a

efilter :: Exp a -> (Id, [Exp a]) -> Exp a

elet :: Maybe (Id, Exp a) -> Exp a -> Exp a

eset :: Id -> Exp a -> Exp a -> Exp a

eblock :: Id -> Exp a -> Exp a

eOverrideBlock :: Id -> Exp a -> Exp a -> Exp a

ecase :: Exp a -> [Alt (Exp a)] -> Maybe (Exp a) -> Exp a

eif :: (Exp a, Exp a) -> [(Exp a, Exp a)] -> Maybe (Exp a) -> Exp a

eempty :: Exp a -> Exp a -> Maybe (Exp a) -> Exp a

true :: Exp a -> Alt (Exp a)

false :: Exp a -> Alt (Exp a)

wild :: Exp a -> Alt (Exp a)

blank :: Fix ExpF