Text.EDE.Internal.Eval

Description

Warning: this is an internal module, and does not have a stable API or name. Functions in this module may not check or enforce preconditions expected by public modules. Use at your own risk!

Documentation

data Env

Constructors

Env

Fields

_extends :: Bool

_templates :: HashMap Id (Exp Delta)

_quoted :: HashMap Id Term

_values :: HashMap Id Value

_blocks :: HashMap Id (Exp Delta)

type Context = ReaderT Env Result

render :: HashMap Id (Exp Delta) -> HashMap Id Term -> Exp Delta -> HashMap Id Value -> Result Builder

eval :: Exp Delta -> Context Term

bind :: (HashMap Text Value -> HashMap Text Value) -> Context a -> Context a

bindBlock :: (HashMap Id (Exp Delta) -> HashMap Id (Exp Delta)) -> Context a -> Context a

setExtended :: Bool -> Context a -> Context a

variable :: Delta -> Var -> Context Value

predicate :: Exp Delta -> Context Term

A variable can be tested for truthiness, but a non-whnf expr cannot.

binding :: Delta -> Term -> Term -> Context Term

build :: Delta -> Value -> Context Builder

throwError :: Delta -> AnsiDoc -> Context a