Text.EDE.Internal.Filters

Description

Warning: this is an internal module, and does not have a stable API or name. Functions in this module may not check or enforce preconditions expected by public modules. Use at your own risk!

Documentation

stdlib :: HashMap Text Term

(@:) :: Quote a => Id -> a -> (Id, Term)

qpoly2 :: Quote a => Id -> (Value -> Value -> a) -> (Id, Term)

Quote a binary function which takes the most general binding value.

qnum1 :: Id -> (Scientific -> Scientific) -> (Id, Term)

Quote an unary numeric function.

qnum2 :: Quote a => Id -> (Scientific -> Scientific -> a) -> (Id, Term)

Quote a binary numeric function.

qlist1 :: (Quote a, Quote b) => Id -> (Text -> a) -> (Vector Value -> b) -> (Id, Term)

Quote a comprehensive set of unary functions to create a binding that supports list collection types.

qcol1 :: (Quote a, Quote b, Quote c) => Id -> (Text -> a) -> (HashMap Text Value -> b) -> (Vector Value -> c) -> (Id, Term)

Quote a comprehensive set of unary functions to create a binding that supports all collection types.

headT :: Text -> Value

lastT :: Text -> Value

tailT :: Text -> Value

initT :: Text -> Value

headV :: Vector Value -> Value

lastV :: Vector Value -> Value

tailV :: Vector Value -> Value

initV :: Vector Value -> Value

text :: (Text -> Text) -> Text -> Value

vec :: (Vector Value -> Value) -> Vector Value -> Value

safe :: b -> (a -> Bool) -> (a -> b) -> a -> b