Text.EDE.Internal.Parser

Description

Warning: this is an internal module, and does not have a stable API or name. Functions in this module may not check or enforce preconditions expected by public modules. Use at your own risk!

Documentation

data Env

Constructors

Env

Fields

_settings :: !Syntax

_includes :: HashMap Text (NonEmpty Delta)

_extends :: HashMap Text (NonEmpty Delta)

_blocks :: HashMap Text (Exp Delta)

settings :: Lens' Env Syntax

includes :: Lens' Env (HashMap Text (NonEmpty Delta))

extends :: Lens' Env (HashMap Text (NonEmpty Delta))

blocks :: Lens' Env (HashMap Text (Exp Delta))

type Parser m = (Monad m, MonadFail m, MonadState Env m, TokenParsing m, DeltaParsing m, LookAheadParsing m, Errable m)

newtype EDE a

Constructors

EDE

Fields

runEDE :: Parser a

runParser :: Syntax -> Text -> ByteString -> Result (Exp Delta, HashMap Text (NonEmpty Delta), HashMap Text (NonEmpty Delta), HashMap Text (Exp Delta))

pragma :: Parser m => m ()

document :: Parser m => m (Exp Delta)

inline :: Parser m => m (Exp Delta)

fragment :: Parser m => m (Exp Delta)

statement :: Parser m => m (Exp Delta)

block :: Parser m => String -> m a -> m a

multiLine :: Parser m => m b -> m a -> m a

singleLine :: Parser m => m b -> m a -> m a

blankLine :: Parser m => m (Exp Delta)

ifelif :: Parser m => m (Exp Delta)

cases :: Parser m => m (Exp Delta)

loop :: Parser m => m (Exp Delta)

include :: Parser m => m (Exp Delta)

extends' :: Parser m => m (Exp Delta)

binding :: Parser m => m (Exp Delta)

set :: Parser m => m (Exp Delta)

block' :: Parser m => m (Exp Delta)

raw :: Parser m => m (Exp Delta)

comment :: Parser m => m (Exp Delta)

else' :: Parser m => m (Maybe (Exp Delta))

exit :: Parser m => String -> m ()

pattern :: Parser m => m Pat

term :: Parser m => m (Exp Delta)

term0 :: Parser m => m (Exp Delta)

filter' :: Parser m => m (Id, [Exp Delta])

collection :: Parser m => m (Exp Delta)

literal :: Parser m => m Value

number :: Parser m => m Scientific

bool :: Parser m => m Bool

object :: Parser m => m (HashMap Text Value)

array :: Parser m => m (Vector Value)

operator :: Parser m => Text -> m Delta

keyword :: Parser m => String -> m Delta

variable :: (Monad m, TokenParsing m) => m Var

identifier :: (Monad m, TokenParsing m) => m Id

spaces :: (Monad m, TokenParsing m) => m ()

manyTill1 :: Alternative m => m a -> m b -> m [a]

manyEndBy1 :: Alternative m => m a -> m a -> m [a]

chainl1' :: Alternative m => m a -> m b -> m (a -> b -> a) -> m a

ann :: (DeltaParsing m, Functor f) => m (f (Fix f)) -> m (Cofree f Delta)

pack :: Functor f => f String -> f Value

triml :: Parser m => m a -> m a

trimr :: Parser m => m a -> m a

pragmak :: Parser m => String -> m ()

pragmal :: Parser m => m String

pragmar :: Parser m => m String

commentl :: MonadState Env m => m String

commentr :: MonadState Env m => m String

inlinel :: Parser m => m String

inliner :: Parser m => m String

blockl :: Parser m => m String

blockr :: Parser m => m String

left :: MonadState s m => Getter s Delim -> m String

right :: MonadState s m => Getter s Delim -> m String