Text.EDE.Internal.Quoting

Description

Warning: this is an internal module, and does not have a stable API or name. Functions in this module may not check or enforce preconditions expected by public modules. Use at your own risk!

Documentation

data Term

A HOAS representation of (possibly partially applied) values in the environment.

Constructors

TVal !Value

TLam (Term -> Result Term)

qapply :: Delta -> Term -> Term -> Result Term

Fully apply two Terms.

qprim :: (ToJSON a, Quote a) => a -> Term

Quote a primitive Value from the top-level.

class Unquote a where

Minimal complete definition

Nothing

Methods

unquote :: Id -> Int -> Term -> Result a

default unquote :: FromJSON a => Id -> Int -> Term -> Result a

class Quote a where

Minimal complete definition

Nothing

Methods

quote :: Id -> Int -> a -> Term

default quote :: ToJSON a => Id -> Int -> a -> Term

typeErr :: Id -> Int -> AnsiDoc -> AnsiDoc -> Result a

argumentErr :: Pretty a => Id -> Int -> a -> Result b