Text.EDE.Internal.Types

Description

Warning: this is an internal module, and does not have a stable API or name. Functions in this module may not check or enforce preconditions expected by public modules. Use at your own risk!

Documentation

type AnsiDoc = Doc AnsiStyle

class AnsiPretty a where

Methods

apretty :: a -> AnsiDoc

newtype PP a

Convenience wrapper for Pretty instances.

Constructors

PP

Fields

unPP :: a

pp :: AnsiPretty (PP a) => a -> AnsiDoc

(</>) :: Doc ann -> Doc ann -> Doc ann

bold :: AnsiDoc -> AnsiDoc

red :: AnsiDoc -> AnsiDoc

data Result a

The result of running parsing or rendering steps.

Constructors

Success a

Failure AnsiDoc

_Failure :: forall a. Prism' (Result a) AnsiDoc

_Success :: forall a a. Prism (Result a) (Result a) a a

eitherResult :: Result a -> Either String a

Convert a Result to an Either with the Left case holding a formatted error message, and Right being the successful result over which Result is paramterised.

result

Arguments

:: (AnsiDoc -> b)
Function to apply to the Failure case.

-> (a -> b)
Function to apply to the Success case.

-> Result a
The Result to map over.

-> b

Perform a case analysis on a Result.

success :: Monad m => a -> m (Result a)

Convenience for returning a successful Result.

failure :: Monad m => AnsiDoc -> m (Result a)

Convenience for returning an error Result.

type Delim = (String, String)

data Syntax

Constructors

Syntax

Fields

_delimPragma :: !Delim

_delimInline :: !Delim

_delimComment :: !Delim

_delimBlock :: !Delim

class HasSyntax c where

Minimal complete definition

syntax

Methods

syntax :: Lens' c Syntax

delimBlock :: Lens' c Delim

delimComment :: Lens' c Delim

delimInline :: Lens' c Delim

delimPragma :: Lens' c Delim

type Resolver m = Syntax -> Id -> Delta -> m (Result Template)

A function to resolve the target of an include expression.

data Template

A parsed and compiled template.

Constructors

Template

Fields

_tmplName :: !Text

_tmplExp :: !(Exp Delta)

_tmplIncl :: HashMap Id (Exp Delta)

_tmplExtends :: HashMap Id (Exp Delta)

_tmplBlocks :: HashMap Id (Exp Delta)

type Id = Text

newtype Var

Constructors

Var (NonEmpty Id)

data Collection where

Constructors

Col :: Foldable f => Int -> f (Maybe Text, Value) -> Collection

data Pat

Constructors

PWild

PVar !Var

PLit !Value

type Alt a = (Pat, a)

data ExpF a

Constructors

ELit !Value

EVar !Var

EFun !Id

EApp !a !a

ELet !Id !a !a

EBlock !Id !a

EOverrideBlock !Id !a !a

ECase !a [Alt a]

ELoop !Id !a !a

EIncl !Text

EExt !Text

type Exp = Cofree ExpF

fromValue :: Value -> Maybe (HashMap Text Value)

Unwrap a Value to an Object safely.

See Aeson's documentation for more details.

fromPairs :: [Pair] -> HashMap Text Value

Create an Object from a list of name/value Pairs.

See Aeson's documentation for more details.