Text.Megaparsec.Unicode

Description

Utility functions for working with Unicode.

Since: 9.7.0

Documentation

stringLength :: Traversable t => t Char -> Int

Calculate length of a string taking into account the fact that certain Chars may span more than 1 column.

Since: 9.7.0

charLength :: Char -> Int

Return length of an individual Char.

Since: 9.7.0

isWideChar :: Char -> Bool

Determine whether the given Char is “wide”, that is, whether it spans 2 columns instead of one.

Since: 9.7.0