Cardano.Node.Configuration.LedgerDB

Documentation

newtype DeprecatedOptions

Some options that existed in the TopLevel were now moved to a subsection. We use this field to propagate the results from parsing those into the monadic part of the node so that we can emit warnings.

Constructors

DeprecatedOptions [String]

data LedgerDbConfiguration

Constructors

LedgerDbConfiguration NumOfDiskSnapshots SnapshotInterval QueryBatchSize LedgerDbSelectorFlag DeprecatedOptions

data LedgerDbSelectorFlag

Choose the LedgerDB Backend

As of UTxO-HD, the LedgerDB now uses either an in-memory backend or LMDB to keep track of differences in the UTxO set.

V2InMemory: uses more memory than the minimum requirements but is somewhat faster.

V1LMDB: uses less memory but is somewhat slower.

V1InMemory: Not intended for production. It is an in-memory reproduction of the LMDB implementation.

Constructors

V1LMDB

Fields

FlushFrequency

The frequency at which changes are flushed to the disk.

(Maybe FilePath)

Path for the live tables.

(Maybe Gigabytes)

A map size can be specified, this is the maximum disk space the LMDB database can fill. If not provided, the default of 16GB will be used.

(Maybe Int)

An override to the max number of readers.

V1InMemory FlushFrequency

V2InMemory

data Gigabytes

A number of gigabytes.

noDeprecatedOptions :: DeprecatedOptions

selectorToArgs :: LedgerDbSelectorFlag -> Complete LedgerDbFlavorArgs IO