Cardano.Node.Configuration.Logging

Documentation

data LoggingLayer

The LoggingLayer interface that we can expose. We want to do this since we want to be able to mock out any function tied to logging.

The good side of this is that _each function has it's own effects_ and that is ideal for tracking the functions effects and constraining the user (programmer) of those function to use specific effects in them. https://github.com/input-output-hk/cardano-sl/blob/develop/util/src/Pos/Util/Log/LogSafe.hs

Constructors

LoggingLayer

Fields

llBasicTrace :: forall m. MonadIO m => Trace m Text

llLogDebug :: forall m a. (MonadIO m, Show a) => Trace m a -> a -> m ()

llLogInfo :: forall m a. (MonadIO m, Show a) => Trace m a -> a -> m ()

llLogNotice :: forall m a. (MonadIO m, Show a) => Trace m a -> a -> m ()

llLogWarning :: forall m a. (MonadIO m, Show a) => Trace m a -> a -> m ()

llLogError :: forall m a. (MonadIO m, Show a) => Trace m a -> a -> m ()

llAppendName :: forall m a. Show a => LoggerName -> Trace m a -> Trace m a

llBracketMonadIO :: forall a t. Show a => Trace IO a -> Severity -> Text -> IO t -> IO t

llBracketMonadM :: forall m a t. (MonadCatch m, MonadIO m, Show a) => Trace m a -> Severity -> Text -> m t -> m t

llBracketMonadX :: forall m a t. (MonadIO m, Show a) => Trace m a -> Severity -> Text -> m t -> m t

llBracketStmIO :: forall a t. Show a => Trace IO a -> Severity -> Text -> STM t -> IO t

llBracketStmLogIO :: forall a t. Show a => Trace IO a -> Severity -> Text -> STM (t, [(LOMeta, LOContent a)]) -> IO t

llConfiguration :: Configuration

llAddBackend :: Backend Text -> BackendKind -> IO ()

llSwitchboard :: Switchboard Text

llEKGDirect :: Maybe EKGDirect

data EKGDirect

Constructors

EKGDirect

Fields

ekgServer :: Server

ekgGauges :: MVar (Map Text Gauge)

ekgLabels :: MVar (Map Text Label)

ekgCounters :: MVar (Map Text Counter)

createLoggingLayer :: Text -> NodeConfiguration -> SomeConsensusProtocol -> ExceptT ConfigError IO LoggingLayer

Create logging feature for `cardano-node`

nodeBasicInfo :: NodeConfiguration -> SomeConsensusProtocol -> UTCTime -> IO [LogObject Text]

shutdownLoggingLayer :: LoggingLayer -> IO ()

traceCounter :: Text -> Trace IO Text -> Int -> IO ()

type Trace (m :: Type -> Type) a = Tracer m (LoggerName, LogObject a)

data Configuration

type LoggerName = Text

data Severity

Constructors

Debug

Info

Notice

Warning

Error

Critical

Alert

Emergency

mkLOMeta :: MonadIO m => Severity -> PrivacyAnnotation -> m LOMeta

data LOMeta

Constructors

LOMeta

Fields

tstamp :: !UTCTime

tid :: !Text

hostname :: !Text

severity :: !Severity

privacy :: !PrivacyAnnotation

data LOContent a

Constructors

LogMessage a

LogError !Text

LogValue !Text !Measurable

LogStructuredText Object Text

LogStructured Object

ObserveOpen !CounterState

ObserveDiff !CounterState

ObserveClose !CounterState

AggregatedMessage [(Text, Aggregated)]

MonitoringEffect !MonitorAction

Command !CommandValue

KillPill