Cardano.Node.Configuration.POM

Documentation

data NodeConfiguration

Constructors

NodeConfiguration

Fields

ncSocketConfig ∷ !SocketConfig

ncConfigFile ∷ !ConfigYamlFilePath

Filepath of the configuration yaml file. This file determines all the configuration settings required for the cardano node (logging, tracing, protocol, slot length etc)

ncTopologyFile ∷ !TopologyFile

ncDatabaseFile ∷ !DbFile

ncProtocolFiles ∷ !ProtocolFilepaths

ncValidateDB ∷ !Bool

ncShutdownConfig ∷ !ShutdownConfig

ncProtocolConfig ∷ !NodeProtocolConfiguration

ncDiffusionMode ∷ !DiffusionMode

ncSnapshotInterval ∷ !SnapshotInterval

ncTestEnableDevelopmentNetworkProtocols ∷ !Bool

During the development and integration of new network protocols (node-to-node and node-to-client) we wish to be able to test them but not have everybody use them by default on the mainnet. Avoiding enabling them by default makes it practical to include such not-yet-ready protocol versions into released versions of the node without the danger that node operators on the mainnet will start using them prematurely, before the testing is complete.

The flag defaults to False

This flag should be set to True when testing the new protocol versions.

ncMaxConcurrencyBulkSync ∷ !(Maybe MaxConcurrencyBulkSync)

ncMaxConcurrencyDeadline ∷ !(Maybe MaxConcurrencyDeadline)

ncLoggingSwitch ∷ !Bool

ncLogMetrics ∷ !Bool

ncTraceConfig ∷ !TraceOptions

ncTraceForwardSocket ∷ !(Maybe (SocketPath, ForwarderMode))

ncMaybeMempoolCapacityOverride ∷ !(Maybe MempoolCapacityBytesOverride)

ncProtocolIdleTimeout ∷ DiffTime

Protocol idleness timeout, see daProtocolIdleTimeout.

ncTimeWaitTimeout ∷ DiffTime

Wait time timeout, see daTimeWaitTimeout.

ncAcceptedConnectionsLimit ∷ !AcceptedConnectionsLimit

Node AcceptedConnectionsLimit

ncTargetNumberOfRootPeers ∷ Int

ncTargetNumberOfKnownPeers ∷ Int

ncTargetNumberOfEstablishedPeers ∷ Int

ncTargetNumberOfActivePeers ∷ Int

ncEnableP2P ∷ SomeNetworkP2PMode

data NetworkP2PMode

Constructors

EnabledP2PMode

DisabledP2PMode

data SomeNetworkP2PMode where

Constructors

SomeNetworkP2PMode ∷ ∀ p2p. NetworkP2PMode p2p → SomeNetworkP2PMode

data PartialNodeConfiguration

Constructors

PartialNodeConfiguration

Fields

pncSocketConfig ∷ !(Last SocketConfig)

pncConfigFile ∷ !(Last ConfigYamlFilePath)

Filepath of the configuration yaml file. This file determines all the configuration settings required for the cardano node (logging, tracing, protocol, slot length etc)

pncTopologyFile ∷ !(Last TopologyFile)

pncDatabaseFile ∷ !(Last DbFile)

pncProtocolFiles ∷ !(Last ProtocolFilepaths)

pncValidateDB ∷ !(Last Bool)

pncShutdownConfig ∷ !(Last ShutdownConfig)

pncProtocolConfig ∷ !(Last NodeProtocolConfiguration)

pncDiffusionMode ∷ !(Last DiffusionMode)

pncSnapshotInterval ∷ !(Last SnapshotInterval)

pncTestEnableDevelopmentNetworkProtocols ∷ !(Last Bool)

pncMaxConcurrencyBulkSync ∷ !(Last MaxConcurrencyBulkSync)

pncMaxConcurrencyDeadline ∷ !(Last MaxConcurrencyDeadline)

pncLoggingSwitch ∷ !(Last Bool)

pncLogMetrics ∷ !(Last Bool)

pncTraceConfig ∷ !(Last PartialTraceOptions)

pncTraceForwardSocket ∷ !(Last (SocketPath, ForwarderMode))

pncMaybeMempoolCapacityOverride ∷ !(Last MempoolCapacityBytesOverride)

pncProtocolIdleTimeout ∷ !(Last DiffTime)

pncTimeWaitTimeout ∷ !(Last DiffTime)

pncAcceptedConnectionsLimit ∷ !(Last AcceptedConnectionsLimit)

pncTargetNumberOfRootPeers ∷ !(Last Int)

pncTargetNumberOfKnownPeers ∷ !(Last Int)

pncTargetNumberOfEstablishedPeers ∷ !(Last Int)

pncTargetNumberOfActivePeers ∷ !(Last Int)

pncEnableP2P ∷ !(Last NetworkP2PMode)

defaultPartialNodeConfiguration ∷ PartialNodeConfiguration

Default configuration is mainnet

lastOption ∷ Parser a → Parser (Last a)

makeNodeConfiguration ∷ PartialNodeConfiguration → Either String NodeConfiguration

parseNodeConfigurationFP ∷ Maybe ConfigYamlFilePath → IO PartialNodeConfiguration

pncProtocol ∷ PartialNodeConfiguration → Either Text Protocol

ncProtocol ∷ NodeConfiguration → Protocol