Cardano.Node.Configuration.POM

Documentation

data NodeConfiguration

Constructors

NodeConfiguration

Fields

ncSocketConfig :: !SocketConfig

ncConfigFile :: !ConfigYamlFilePath

Filepath of the configuration yaml file. This file determines all the configuration settings required for the cardano node (logging, tracing, protocol, slot length etc)

ncTopologyFile :: !TopologyFile

ncDatabaseFile :: !NodeDatabasePaths

ncProtocolFiles :: !ProtocolFilepaths

ncValidateDB :: !Bool

ncShutdownConfig :: !ShutdownConfig

ncStartAsNonProducingNode :: !Bool

ncProtocolConfig :: !NodeProtocolConfiguration

ncDiffusionMode :: !DiffusionMode

ncExperimentalProtocolsEnabled :: !Bool

During the development and integration of new network protocols (node-to-node and node-to-client) we wish to be able to test them but not have everybody use them by default on the mainnet. Avoiding enabling them by default makes it practical to include such not-yet-ready protocol versions into released versions of the node without the danger that node operators on the mainnet will start using them prematurely, before the testing is complete.

The flag defaults to False

This flag should be set to True when testing the new protocol versions.

ncMaxConcurrencyBulkSync :: !(Maybe MaxConcurrencyBulkSync)

ncMaxConcurrencyDeadline :: !(Maybe MaxConcurrencyDeadline)

ncLoggingSwitch :: !Bool

ncLogMetrics :: !Bool

ncTraceConfig :: !TraceOptions

ncTraceForwardSocket :: !(Maybe (HowToConnect, ForwarderMode))

ncMaybeMempoolCapacityOverride :: !(Maybe MempoolCapacityBytesOverride)

ncLedgerDbConfig :: !LedgerDbConfiguration

ncProtocolIdleTimeout :: DiffTime

Protocol idleness timeout, see daProtocolIdleTimeout.

ncTimeWaitTimeout :: DiffTime

Wait time timeout, see daTimeWaitTimeout.

ncEgressPollInterval :: DiffTime

ncChainSyncIdleTimeout :: TimeoutOverride

Timeout override for ChainSync, see ChainSyncTimeout

ncMempoolTimeoutSoft :: DiffTime

If the mempool takes longer than this to validate a tx, then it discards the tx instead of adding it.

ncMempoolTimeoutHard :: DiffTime

If the mempool takes longer than this to validate a tx, then it disconnects from the peer.

WARNING: if this is less than mempoolTimeoutSoft, then mempoolTimeoutSoft is irrelevant. If it's equal or just barely larger, then the soft/hard distinction will likely be unreliable.

ncMempoolTimeoutCapacity :: DiffTime

If the mempool takes longer than this cumulatively to validate when each entered the mempool, then the mempool is at capacity, ie it's full, ie no tx can be added.

ncAcceptedConnectionsLimit :: !AcceptedConnectionsLimit

Node AcceptedConnectionsLimit

ncDeadlineTargetOfRootPeers :: !Int

ncDeadlineTargetOfKnownPeers :: !Int

ncDeadlineTargetOfEstablishedPeers :: !Int

ncDeadlineTargetOfActivePeers :: !Int

ncDeadlineTargetOfKnownBigLedgerPeers :: !Int

ncDeadlineTargetOfEstablishedBigLedgerPeers :: !Int

ncDeadlineTargetOfActiveBigLedgerPeers :: !Int

ncSyncTargetOfRootPeers :: !Int

ncSyncTargetOfKnownPeers :: !Int

ncSyncTargetOfEstablishedPeers :: !Int

ncSyncTargetOfActivePeers :: !Int

ncSyncTargetOfKnownBigLedgerPeers :: !Int

ncSyncTargetOfEstablishedBigLedgerPeers :: !Int

ncSyncTargetOfActiveBigLedgerPeers :: !Int

ncConsensusMode :: !ConsensusMode

ncMinBigLedgerPeersForTrustedState :: NumberOfBigLedgerPeers

ncPeerSharing :: PeerSharing

ncGenesisConfig :: GenesisConfig

ncResponderCoreAffinityPolicy :: ResponderCoreAffinityPolicy

data ResponderCoreAffinityPolicy

We expose the ForkPolicy as a NodeConfiguration field.

NoResponderCoreAffinity corresponds to noBindForkPolicy

ResponderCoreAffinity corresponds to responderForkPolicy with a randomIO generated salt and getNumCapabilities.

Constructors

NoResponderCoreAffinity

ResponderCoreAffinity

data PartialNodeConfiguration

Constructors

PartialNodeConfiguration

Fields

pncSocketConfig :: !(Last SocketConfig)

pncConfigFile :: !(Last ConfigYamlFilePath)

Filepath of the configuration yaml file. This file determines all the configuration settings required for the cardano node (logging, tracing, protocol, slot length etc)

pncTopologyFile :: !(Last TopologyFile)

pncDatabaseFile :: !(Last NodeDatabasePaths)

pncProtocolFiles :: !(Last ProtocolFilepaths)

pncProtocolFiles can only be supplied with command line arguments.

pncValidateDB :: !(Last Bool)

pncShutdownConfig :: !(Last ShutdownConfig)

pncStartAsNonProducingNode :: !(Last Bool)

pncProtocolConfig :: !(Last NodeProtocolConfiguration)

pncDiffusionMode :: !(Last DiffusionMode)

pncExperimentalProtocolsEnabled :: !(Last Bool)

pncMaxConcurrencyBulkSync :: !(Last MaxConcurrencyBulkSync)

pncMaxConcurrencyDeadline :: !(Last MaxConcurrencyDeadline)

pncLoggingSwitch :: !(Last Bool)

pncLogMetrics :: !(Last Bool)

pncTraceConfig :: !(Last PartialTraceOptions)

pncTraceForwardSocket :: !(Last (HowToConnect, ForwarderMode))

pncMaybeMempoolCapacityOverride :: !(Last MempoolCapacityBytesOverride)

pncLedgerDbConfig :: !(Last LedgerDbConfiguration)

pncProtocolIdleTimeout :: !(Last DiffTime)

pncTimeWaitTimeout :: !(Last DiffTime)

pncEgressPollInterval :: !(Last DiffTime)

pncChainSyncIdleTimeout :: !(Last DiffTime)

pncMempoolTimeoutSoft :: !(Last DiffTime)

pncMempoolTimeoutHard :: !(Last DiffTime)

pncMempoolTimeoutCapacity :: !(Last DiffTime)

pncAcceptedConnectionsLimit :: !(Last AcceptedConnectionsLimit)

pncDeadlineTargetOfRootPeers :: !(Last Int)

pncDeadlineTargetOfKnownPeers :: !(Last Int)

pncDeadlineTargetOfEstablishedPeers :: !(Last Int)

pncDeadlineTargetOfActivePeers :: !(Last Int)

pncDeadlineTargetOfKnownBigLedgerPeers :: !(Last Int)

pncDeadlineTargetOfEstablishedBigLedgerPeers :: !(Last Int)

pncDeadlineTargetOfActiveBigLedgerPeers :: !(Last Int)

pncSyncTargetOfRootPeers :: !(Last Int)

pncSyncTargetOfKnownPeers :: !(Last Int)

pncSyncTargetOfEstablishedPeers :: !(Last Int)

pncSyncTargetOfActivePeers :: !(Last Int)

pncSyncTargetOfKnownBigLedgerPeers :: !(Last Int)

pncSyncTargetOfEstablishedBigLedgerPeers :: !(Last Int)

pncSyncTargetOfActiveBigLedgerPeers :: !(Last Int)

pncMinBigLedgerPeersForTrustedState :: !(Last NumberOfBigLedgerPeers)

pncConsensusMode :: !(Last ConsensusMode)

pncPeerSharing :: !(Last PeerSharing)

pncGenesisConfigFlags :: !(Last GenesisConfigFlags)

pncResponderCoreAffinityPolicy :: !(Last ResponderCoreAffinityPolicy)

data TimeoutOverride

Isomorphic to a `Maybe DiffTime`, but expresses what Nothing means, in this case that we want to NOT override the default timeout.

Constructors

NoTimeoutOverride

TimeoutOverride DiffTime

defaultPartialNodeConfiguration :: PartialNodeConfiguration

Default configuration is mainnet

lastOption :: Parser a -> Parser (Last a)

makeNodeConfiguration :: PartialNodeConfiguration -> Either String NodeConfiguration

parseNodeConfigurationFP :: Maybe ConfigYamlFilePath -> IO PartialNodeConfiguration

pncProtocol :: PartialNodeConfiguration -> Either Text Protocol

ncProtocol :: NodeConfiguration -> Protocol

getForkPolicy :: Hashable peerAddr => ResponderCoreAffinityPolicy -> IO (ForkPolicy peerAddr)

Convert NCForkPolicy to a ForkPolicy