Cardano.Node.Configuration.Socket

Documentation

data SocketConfig

Constructors

SocketConfig

Fields

ncNodeIPv4Addr :: !(Last NodeHostIPv4Address)

ncNodeIPv6Addr :: !(Last NodeHostIPv6Address)

ncNodePortNumber :: !(Last PortNumber)

ncSocketPath :: !(Last SocketPath)

gatherConfiguredSockets :: SocketConfig -> ExceptT SocketConfigError IO (Maybe SocketOrSocketInfo, Maybe SocketOrSocketInfo, Maybe LocalSocketOrSocketInfo)

Gather from the various sources of configuration which sockets we will use for the public node-to-node and the local node-to-client IPC. It returns SocketOrSocketInfo for ipv4, ipv6 and local socket.

We get such configuration from:

node config file

node cli

systemd socket activation

data SocketOrSocketInfo' socket info

Since we support systemd socket activation, we have to handle being given actual already-constructed sockets, or the info needed to make new sockets later.

Constructors

ActualSocket socket

SocketInfo info

type SocketOrSocketInfo = SocketOrSocketInfo' Socket SockAddr

type LocalSocketOrSocketInfo = SocketOrSocketInfo' LocalSocket LocalAddress

getSocketOrSocketInfoAddr :: SocketOrSocketInfo -> IO SockAddr

data SocketConfigError

Errors for the current module.

Constructors

NoPublicSocketGiven

NoLocalSocketGiven

ClashingPublicIpv4SocketGiven

ClashingPublicIpv6SocketGiven

ClashingLocalSocketGiven

LocalSocketError SocketPath IOException

GetAddrInfoError (Maybe NodeHostIPAddress) (Maybe PortNumber) IOException

renderSocketConfigError :: SocketConfigError -> String