Cardano.Node.Configuration.Topology

Documentation

newtype TopologyError

Constructors

NodeIdNotFoundInToplogyFile FilePath

data NetworkTopology

Constructors

MockNodeTopology ![NodeSetup]

RealNodeTopology ![RemoteAddress]

newtype NodeHostIPAddress

Constructors

NodeHostIPAddress

Fields

unNodeHostIPAddress :: IP

newtype NodeHostIPv4Address

Constructors

NodeHostIPv4Address

Fields

unNodeHostIPv4Address :: IPv4

newtype NodeHostIPv6Address

Constructors

NodeHostIPv6Address

Fields

unNodeHostIPv6Address :: IPv6

data NodeSetup

Constructors

NodeSetup

Fields

nodeId :: !Word64

nodeIPv4Address :: !(Maybe NodeIPv4Address)

nodeIPv6Address :: !(Maybe NodeIPv6Address)

producers :: ![RemoteAddress]

data RemoteAddress

Domain name with port number

Constructors

RemoteAddress

Fields

raAddress :: !Text

Either a dns address or an ip address.

raPort :: !PortNumber

Port number of the destination.

raValency :: !Int

If a DNS address is given valency governs to how many resolved IP addresses should we maintain active (hot) connection; if an IP address is given valency is used as a Boolean value, 0 means to ignore the address;

nodeAddressToSockAddr :: NodeIPAddress -> SockAddr

readTopologyFile :: NodeConfiguration -> IO (Either Text NetworkTopology)

Read the NetworkTopology configuration from the specified file. While running a real protocol, this gives your node its own address and other remote peers it will attempt to connect to.

readTopologyFileOrError :: NodeConfiguration -> IO NetworkTopology

remoteAddressToNodeAddress :: RemoteAddress -> Maybe (Either NodeIPAddress (NodeDnsAddress, Int))

Parse raAddress field as an IP address; if it parses and the valency is non zero return corresponding NodeAddress.