Cardano.Node.Configuration.TopologyP2P

Documentation

newtype TopologyError

Constructors

NodeIdNotFoundInToplogyFile FilePath

data NetworkTopology adr

Describes the P2P topology of a node. Whenever the node actually runs, the type parameter adr should be RelayAccessPoint. However, we might want to use and serialize this type with adr being NodeId, or another placeholder type, if we want the user to be able to edit the topology without knowing the actual addresses of the nodes: those might only be knowable at runtime.

Constructors

RealNodeTopology

Fields

ntLocalRootPeersGroups :: !(LocalRootPeersGroups adr)

ntPublicRootPeers :: ![PublicRootPeers adr]

ntUseLedgerPeers :: !UseLedgerPeers

ntUseBootstrapPeers :: !UseBootstrapPeers

ntPeerSnapshotPath :: !(Maybe PeerSnapshotFile)

newtype PublicRootPeers adr

Constructors

PublicRootPeers

Fields

publicRoots :: RootConfig adr

data LocalRootPeersGroup adr

A local root peers group. Local roots are treated by the outbound governor in a special way. The node will make sure that a node has the requested number (valency/hotValency) of connections to the local root peer group. warmValency value is the value of warm/established connections that the node will attempt to maintain. By default this value will be equal to hotValency.

Constructors

LocalRootPeersGroup

Fields

localRoots :: RootConfig adr

hotValency :: HotValency

warmValency :: WarmValency

trustable :: PeerTrustable

trustable configures whether the root should be trusted in fallback state.

rootDiffusionMode :: DiffusionMode

diffusion mode; used for local root peers.

newtype LocalRootPeersGroups adr

Constructors

LocalRootPeersGroups

Fields

groups :: [LocalRootPeersGroup adr]

data RootConfig adr

Each root peer consists of a list of access points and a shared PeerAdvertise field.

Constructors

RootConfig

Fields

rootAccessPoints :: [adr]

a list of relay access points, each of which is either an ip address or domain name and a port number.

rootAdvertise :: PeerAdvertise

advertise configures whether the root should be advertised through peer sharing.

newtype NodeHostIPAddress

Constructors

NodeHostIPAddress

Fields

unNodeHostIPAddress :: IP

newtype NodeHostIPv4Address

Constructors

NodeHostIPv4Address

Fields

unNodeHostIPv4Address :: IPv4

newtype NodeHostIPv6Address

Constructors

NodeHostIPv6Address

Fields

unNodeHostIPv6Address :: IPv6

data NodeSetup adr

Constructors

NodeSetup

Fields

nodeId :: !Word64

nodeIPv4Address :: !(Maybe NodeIPv4Address)

nodeIPv6Address :: !(Maybe NodeIPv6Address)

producers :: ![RootConfig adr]

useLedger :: !UseLedgerPeers

nodeAddressToSockAddr :: NodeIPAddress -> SockAddr

readTopologyFile :: forall adr. FromJSON adr => NodeConfiguration -> Tracer IO (StartupTrace blk) -> IO (Either Text (NetworkTopology adr))

Read the NetworkTopology configuration from the specified file.

readPeerSnapshotFile :: PeerSnapshotFile -> IO (Either Text LedgerPeerSnapshot)

readTopologyFileOrError :: forall adr. FromJSON adr => NodeConfiguration -> Tracer IO (StartupTrace blk) -> IO (NetworkTopology adr)

rootConfigToRelayAccessPoint :: forall adr. RootConfig adr -> [(adr, PeerAdvertise)]

Transforms a RootConfig into a pair of RelayAccessPoint and its corresponding PeerAdvertise value.

Re-exports

data DiffusionMode

The flag which indicates whether the node runs only initiator or both initiator or responder node.

This data structure has two proposes:

instruct the diffusion layer if it should listen on incoming connections;

it is communicated via NodeToNodeVersionData during handshake negotiation. In non-p2p mode we always send InitiatorOnlyDiffusionMode, in p2p mode we send exactly what the diffusion is given. In non-p2p mode every connection outbound port is ephemeral, the remote side cannot connect to it, however in p2p mode the outbound port is actually the port on which the node is listening (if it runs in InitiatorAndResponderDiffusionMode).

Constructors

InitiatorOnlyDiffusionMode

InitiatorAndResponderDiffusionMode

data PeerAdvertise

Should this peer be advertised to other peers asking for known peers? For certain peers specified by configuration it would be an appropriate policy to keep them private.

Constructors

DoNotAdvertisePeer

DoAdvertisePeer

data PeerTrustable

Is this Peer trustable as a bootstrap peer?

This trustability flag is used on local root peers (pre-genesis) to distinguish which locally configured peer is considered safe to trust for bootstrap purposes

Constructors

IsTrustable

IsNotTrustable

data RelayAccessPoint

A relay can have either an IP address and a port number or a domain with a port number

Constructors

RelayAccessDomain !Domain !PortNumber

RelayAccessSRVDomain !Domain
SRV domain, prefixed (as defined in CIP#0155)

RelayAccessAddress !IP !PortNumber

data UseBootstrapPeers

UseBootstrapPeers is read from the topology file. It's value might changed by the user while the node is running.

Constructors

DontUseBootstrapPeers

UseBootstrapPeers [RelayAccessPoint]

data UseLedgerPeers

Only use the ledger after the given slot number.

Constructors

DontUseLedgerPeers

UseLedgerPeers AfterSlot