Cardano.Node.Configuration.TopologyP2P

Documentation

newtype TopologyError

Constructors

NodeIdNotFoundInToplogyFile FilePath

data NetworkTopology

Constructors

RealNodeTopology

Fields

ntLocalRootPeersGroups :: !LocalRootPeersGroups

ntPublicRootPeers :: ![PublicRootPeers]

ntUseLedgerPeers :: !UseLedgerPeers

ntUseBootstrapPeers :: !UseBootstrapPeers

ntPeerSnapshotPath :: !(Maybe PeerSnapshotFile)

newtype PublicRootPeers

Constructors

PublicRootPeers

Fields

publicRoots :: RootConfig

data LocalRootPeersGroup

A local root peers group. Local roots are treated by the outbound governor in a special way. The node will make sure that a node has the requested number (valency/hotValency) of connections to the local root peer group. warmValency value is the value of warm/established connections that the node will attempt to maintain. By default this value will be equal to hotValency.

Constructors

LocalRootPeersGroup

Fields

localRoots :: RootConfig

hotValency :: HotValency

warmValency :: WarmValency

trustable :: PeerTrustable

trustable configures whether the root should be trusted in fallback state.

rootDiffusionMode :: DiffusionMode

diffusion mode; used for local root peers.

newtype LocalRootPeersGroups

Constructors

LocalRootPeersGroups

Fields

groups :: [LocalRootPeersGroup]

data RootConfig

Each root peer consists of a list of access points and a shared PeerAdvertise field.

Constructors

RootConfig

Fields

rootAccessPoints :: [RelayAccessPoint]

a list of relay access points, each of which is either an ip address or domain name and a port number.

rootAdvertise :: PeerAdvertise

advertise configures whether the root should be advertised through peer sharing.

newtype NodeHostIPAddress

Constructors

NodeHostIPAddress

Fields

unNodeHostIPAddress :: IP

newtype NodeHostIPv4Address

Constructors

NodeHostIPv4Address

Fields

unNodeHostIPv4Address :: IPv4

newtype NodeHostIPv6Address

Constructors

NodeHostIPv6Address

Fields

unNodeHostIPv6Address :: IPv6

data NodeSetup

Constructors

NodeSetup

Fields

nodeId :: !Word64

nodeIPv4Address :: !(Maybe NodeIPv4Address)

nodeIPv6Address :: !(Maybe NodeIPv6Address)

producers :: ![RootConfig]

useLedger :: !UseLedgerPeers

data PeerAdvertise

Should this peer be advertised to other peers asking for known peers? For certain peers specified by configuration it would be an appropriate policy to keep them private.

Constructors

DoNotAdvertisePeer

DoAdvertisePeer

nodeAddressToSockAddr :: NodeIPAddress -> SockAddr

readTopologyFile :: NodeConfiguration -> Tracer IO (StartupTrace blk) -> IO (Either Text NetworkTopology)

Read the NetworkTopology configuration from the specified file.

readPeerSnapshotFile :: PeerSnapshotFile -> IO LedgerPeerSnapshot

readTopologyFileOrError :: NodeConfiguration -> Tracer IO (StartupTrace blk) -> IO NetworkTopology

rootConfigToRelayAccessPoint :: RootConfig -> [(RelayAccessPoint, PeerAdvertise)]

Transforms a RootConfig into a pair of RelayAccessPoint and its corresponding PeerAdvertise value.