Cardano.Node.Handlers.Shutdown

Documentation

data ShutdownOn

Constructors

ASlot !SlotNo

ABlock !BlockNo

NoShutdown

parseShutdownOn ∷ Parser ShutdownOn

Generalised shutdown handling

data ShutdownConfig

Constructors

ShutdownConfig

Fields

scIPC ∷ !(Maybe Fd)

scOnSyncLimit ∷ !(Maybe ShutdownOn)

withShutdownHandling

Arguments

∷ ShutdownConfig

→ Tracer IO ShutdownTrace

→ IO ()
Action to potentially shutdown via file descriptor

→ IO ()

We provide an optional cross-platform method to politely request shut down. The parent process passes us the file descriptor number of the read end of a pipe, via the CLI with --shutdown-ipc FD

data ShutdownTrace

Constructors

ShutdownRequested
Received shutdown request

AbnormalShutdown
Non-isEOFError shutdown request

ShutdownUnexpectedInput Text
Received shutdown request but found unexpected input in --shutdown-ipc FD:

RequestingShutdown Text
Ringing the node shutdown doorbell for reason

ShutdownArmedAt ShutdownOn
Will terminate upon reaching a ChainDB sync limit

Watch ChainDB for passing a configured slot sync limit threshold,

maybeSpawnOnSlotSyncedShutdownHandler ∷ HasHeader (Header blk) ⇒ ShutdownConfig → Tracer IO ShutdownTrace → ResourceRegistry IO → ChainDB IO blk → IO ()

Spawn a thread that would cause node to shutdown upon ChainDB reaching the configuration-defined slot.

Orphan instances

FromJSON BlockNo

Instance details

Methods

parseJSON ∷ Value → Parser BlockNo

parseJSONList ∷ Value → Parser [BlockNo]

ToJSON BlockNo

Instance details

Methods

toJSON ∷ BlockNo → Value

toEncoding ∷ BlockNo → Encoding

toJSONList ∷ [BlockNo] → Value

toEncodingList ∷ [BlockNo] → Encoding