Cardano.Node.Parsers

Documentation

nodeCLIParser :: Parser PartialNodeConfiguration

parseConfigFile :: Parser FilePath

parserHelpHeader :: String -> Parser a -> Doc

Produce just the brief help header for a given CLI option parser, without the options.

parserHelpOptions :: Parser a -> Doc

Produce just the options help for a given CLI option parser, without the header.

renderHelpDoc :: Int -> Doc -> String

Render the help pretty document.

parseHostPort :: String -> Either String HowToConnect