Cardano.Node.Protocol.Byron

Documentation

mkSomeConsensusProtocolByron :: NodeByronProtocolConfiguration -> Maybe ProtocolFilepaths -> ExceptT ByronProtocolInstantiationError IO SomeConsensusProtocol

Make SomeConsensusProtocol using the Byron instance.

This lets us handle multiple protocols in a generic way.

This also serves a purpose as a sanity check that we have all the necessary type class instances available.

Errors

data ByronProtocolInstantiationError

Constructors

CanonicalDecodeFailure !FilePath !Text

GenesisHashMismatch !GenesisHash !GenesisHash

DelegationCertificateFilepathNotSpecified

GenesisConfigurationError !FilePath !ConfigurationError

GenesisReadError !FilePath !GenesisDataError

CredentialsError !ByronLeaderCredentialsError

SigningKeyDeserialiseFailure !FilePath

SigningKeyFilepathNotSpecified

Reusable parts

readGenesis :: GenesisFile -> Maybe GenesisHash -> RequiresNetworkMagic -> ExceptT ByronProtocolInstantiationError IO Config

readLeaderCredentials :: Config -> Maybe ProtocolFilepaths -> ExceptT ByronProtocolInstantiationError IO (Maybe ByronLeaderCredentials)