Cardano.Node.Protocol.Shelley

Documentation

mkSomeConsensusProtocolShelley ∷ NodeShelleyProtocolConfiguration → Maybe ProtocolFilepaths → ExceptT ShelleyProtocolInstantiationError IO SomeConsensusProtocol

Make SomeConsensusProtocol using the Shelley instance.

This lets us handle multiple protocols in a generic way.

This also serves a purpose as a sanity check that we have all the necessary type class instances available.

Errors

data ShelleyProtocolInstantiationError

Constructors

GenesisReadError GenesisReadError

GenesisValidationError GenesisValidationError

PraosLeaderCredentialsError PraosLeaderCredentialsError

data GenesisReadError

Constructors

GenesisReadFileError !FilePath !IOException

GenesisHashMismatch !GenesisHash !GenesisHash

GenesisDecodeError !FilePath !String

newtype GenesisValidationError

Constructors

GenesisValidationErrors [ValidationErr]

data PraosLeaderCredentialsError

Constructors

CredentialsReadError !FilePath !IOException

EnvelopeParseError !FilePath !String

FileError !(FileError TextEnvelopeError)

OCertNotSpecified

VRFKeyNotSpecified

KESKeyNotSpecified

MismatchedKesKey FilePath FilePath

Reusable parts

readGenesis ∷ GenesisFile → Maybe GenesisHash → ExceptT GenesisReadError IO (ShelleyGenesis StandardShelley, GenesisHash)

readGenesisAny ∷ FromJSON genesis ⇒ GenesisFile → Maybe GenesisHash → ExceptT GenesisReadError IO (genesis, GenesisHash)

readLeaderCredentials ∷ Maybe ProtocolFilepaths → ExceptT PraosLeaderCredentialsError IO [ShelleyLeaderCredentials StandardCrypto]

genesisHashToPraosNonce ∷ GenesisHash → Nonce

validateGenesis ∷ ShelleyGenesis StandardShelley → ExceptT GenesisValidationError IO ()