Cardano.Node.Protocol.Shelley

Documentation

mkSomeConsensusProtocolShelley :: NodeShelleyProtocolConfiguration -> Maybe ProtocolFilepaths -> ExceptT ShelleyProtocolInstantiationError IO SomeConsensusProtocol

Make SomeConsensusProtocol using the Shelley instance.

This lets us handle multiple protocols in a generic way.

This also serves a purpose as a sanity check that we have all the necessary type class instances available.

Errors

data ShelleyProtocolInstantiationError

Constructors

GenesisReadError GenesisReadError

GenesisValidationError GenesisValidationError

PraosLeaderCredentialsError PraosLeaderCredentialsError

data GenesisReadError

Constructors

GenesisReadFileError !FilePath !IOException

GenesisHashMismatch !GenesisHash !GenesisHash

GenesisDecodeError !FilePath !String

newtype GenesisValidationError

Constructors

GenesisValidationErrors [ValidationErr]

data PraosLeaderCredentialsError

Constructors

CredentialsReadError !FilePath !IOException

EnvelopeParseError !FilePath !String

FileError !(FileError TextEnvelopeError)

OCertNotSpecified

VRFKeyNotSpecified

KESKeyNotSpecified

MismatchedKesKey FilePath FilePath

Reusable parts

readGenesis :: GenesisFile -> Maybe GenesisHash -> ExceptT GenesisReadError IO (ShelleyGenesis, GenesisHash)

readGenesisAny :: FromJSON genesis => GenesisFile -> Maybe GenesisHash -> ExceptT GenesisReadError IO (genesis, GenesisHash)

Read a genesis file using a memory-safe strategy.

Genesis files are used in testing and benchmarking to create testnets with large datasets. To be able to read these files in memory constrained environments we use a conditional loading strategy: if file is below ~10 megabytes it is entirely read into memory, otherwise the ByteString is created using mmap that uses the virtual memory subsystem to do on-demand loading. With current usage and how benchmakring is done only Shelley and Conway are affected (Shelley's initialFunds and Conway's "delegs" and "initialDReps").

readLeaderCredentials :: Maybe ProtocolFilepaths -> ExceptT PraosLeaderCredentialsError IO [ShelleyLeaderCredentials StandardCrypto]

genesisHashToPraosNonce :: GenesisHash -> Nonce

validateGenesis :: ShelleyGenesis -> ExceptT GenesisValidationError IO ()

checkExpectedGenesisHash

Arguments

:: ByteString
genesis bytes

-> Maybe GenesisHash
expected hash, check for hash match, if provided

-> ExceptT GenesisReadError IO GenesisHash