Cardano.Node.Queries

Documentation

class ConvertTxId blk where

Convert a transaction ID to raw bytes.

Methods

txIdToRawBytes ∷ TxId (GenTx blk) → ByteString

KES

newtype MaxKESEvolutions

The maximum number of evolutions that a KES key can undergo before it is considered expired.

Constructors

MaxKESEvolutions Word64

newtype OperationalCertStartKESPeriod

The start KES period of the configured operational certificate.

Constructors

OperationalCertStartKESPeriod Period

class HasKESInfo blk where

Minimal complete definition

Nothing

Methods

getKESInfo ∷ Proxy blk → ForgeStateUpdateError blk → Maybe KESInfo

data KESMetricsData

KES-related data to be traced as metrics.

Constructors

NoKESMetricsData
The current protocol does not support KES.

TPraosKESMetricsData

Fields

!Period

The current KES period of the hot key, relative to the start KES period of the operational certificate.

!MaxKESEvolutions

The configured max KES evolutions.

!OperationalCertStartKESPeriod

The start KES period of the configured operational certificate.

class HasKESMetricsData blk where

Minimal complete definition

Nothing

Methods

getKESMetricsData ∷ Proxy blk → ForgeStateInfo blk → KESMetricsData

General ledger

class LedgerQueries blk where

Methods

ledgerUtxoSize ∷ LedgerState blk → Int

ledgerDelegMapSize ∷ LedgerState blk → Int