Cardano.Node.Tracing.Consistency

Documentation

getAllNamespaces :: [([Text], [Text])]

Returns a list of all namespaces from all tracers

checkNodeTraceConfiguration :: FilePath -> IO NSWarnings

Check the configuration in the given file. If there is no configuration in the file check the standard configuration An empty return list means, everything is well

checkNodeTraceConfiguration' :: TraceConfig -> NSWarnings

Check the configuration in the given file. If there is no configuration in the file check the standard configuration An empty return list means, everything is well