Cardano.Node.Tracing.Formatting

Orphan instances

LogFormatting SlotNo

Instance details

Methods

forMachine :: DetailLevel -> SlotNo -> Object

forHuman :: SlotNo -> Text

asMetrics :: SlotNo -> [Metric]

LogFormatting Void
A bit of a weird one, but needed because some of the very general consensus interfaces are sometimes instantiated to Void, when there are no cases needed.

Instance details

Methods

forMachine :: DetailLevel -> Void -> Object

forHuman :: Void -> Text

asMetrics :: Void -> [Metric]

LogFormatting ()

Instance details

Methods

forMachine :: DetailLevel -> () -> Object

forHuman :: () -> Text

asMetrics :: () -> [Metric]

ConvertRawHash blk => ConvertRawHash (Header blk)
Derives ConvertRawHash for Header blk from ConvertRawHash blk. Safe because HeaderHash (Header blk) = HeaderHash blk.

Instance details

Methods

toRawHash :: proxy (Header blk) -> HeaderHash (Header blk) -> ByteString

fromRawHash :: proxy (Header blk) -> ByteString -> HeaderHash (Header blk)

toShortRawHash :: proxy (Header blk) -> HeaderHash (Header blk) -> ShortByteString

fromShortRawHash :: proxy (Header blk) -> ShortByteString -> HeaderHash (Header blk)

hashSize :: proxy (Header blk) -> Word32

ConvertRawHash blk => LogFormatting (RealPoint blk)

Instance details

Methods

forMachine :: DetailLevel -> RealPoint blk -> Object

forHuman :: RealPoint blk -> Text

asMetrics :: RealPoint blk -> [Metric]

ConvertRawHash blk => LogFormatting (Anchor blk)

Instance details

Methods

forMachine :: DetailLevel -> Anchor blk -> Object

forHuman :: Anchor blk -> Text

asMetrics :: Anchor blk -> [Metric]

(ConvertRawHash blk, HasHeader blk) => LogFormatting (AnchoredFragment blk)

Instance details

Methods

forMachine :: DetailLevel -> AnchoredFragment blk -> Object

forHuman :: AnchoredFragment blk -> Text

asMetrics :: AnchoredFragment blk -> [Metric]

ConvertRawHash blk => LogFormatting (Point blk)

Instance details

Methods

forMachine :: DetailLevel -> Point blk -> Object

forHuman :: Point blk -> Text

asMetrics :: Point blk -> [Metric]