Cardano.Node.Tracing.Peers

Documentation

newtype NodePeers

This type contains an information about current peers of the node. It will be asked by external applications as a DataPoint.

Constructors

NodePeers [PeerInfoPP]

traceNodePeers :: Trace IO NodePeers -> [PeerT blk] -> IO ()

Orphan instances

MetaTrace NodePeers

Instance details

Methods

namespaceFor :: NodePeers -> Namespace NodePeers

severityFor :: Namespace NodePeers -> Maybe NodePeers -> Maybe SeverityS

privacyFor :: Namespace NodePeers -> Maybe NodePeers -> Maybe Privacy

detailsFor :: Namespace NodePeers -> Maybe NodePeers -> Maybe DetailLevel

documentFor :: Namespace NodePeers -> Maybe Text

metricsDocFor :: Namespace NodePeers -> [(Text, Text)]

allNamespaces :: [Namespace NodePeers]