Cardano.Node.Types

Configuration

class AdjustFilePaths a where

Methods

adjustFilePaths :: (FilePath -> FilePath) -> a -> a

data ConfigError

Errors for the cardano-config module.

Constructors

ConfigErrorFileNotFound FilePath

ConfigErrorNoEKG

newtype ConfigYamlFilePath

Filepath of the configuration yaml file. This file determines all the configuration settings required for the cardano node (logging, tracing, protocol, slot length etc)

Constructors

ConfigYamlFilePath

Fields

unConfigPath :: FilePath

newtype DbFile

Constructors

DbFile

Fields

unDB :: FilePath

newtype GenesisFile

Constructors

GenesisFile

Fields

unGenesisFile :: FilePath

newtype PeerSnapshotFile

Path containing a serialized ledger peer snapshot for use by diffusion layer to facilitate bootstrapping a node in Genesis consensus mode

Constructors

PeerSnapshotFile

Fields

unPeerSnapshotFile :: FilePath

newtype CheckpointsFile

Constructors

CheckpointsFile

Fields

unCheckpointsFile :: FilePath

data ProtocolFilepaths

Constructors

ProtocolFilepaths

Fields

byronCertFile :: !(Maybe FilePath)

byronKeyFile :: !(Maybe FilePath)

shelleyKESFile :: !(Maybe FilePath)

shelleyVRFFile :: !(Maybe FilePath)

shelleyCertFile :: !(Maybe FilePath)

shelleyBulkCredsFile :: !(Maybe FilePath)

hasProtocolFile :: ProtocolFilepaths -> Bool

If any of the ProtocolFilepath is given PeerSharing option will be disabled by default.

newtype GenesisHash

Constructors

GenesisHash (Hash Blake2b_256 ByteString)

newtype CheckpointsHash

Constructors

CheckpointsHash (Hash Blake2b_256 ByteString)

newtype MaxConcurrencyBulkSync

Constructors

MaxConcurrencyBulkSync

Fields

unMaxConcurrencyBulkSync :: Word

newtype MaxConcurrencyDeadline

Constructors

MaxConcurrencyDeadline

Fields

unMaxConcurrencyDeadline :: Word

Networking

newtype TopologyFile

Constructors

TopologyFile

Fields

unTopology :: FilePath

newtype NodeConsensusMode

Newtype wrapper which provides FromJSON instance for ConsensusMode.

Constructors

NodeConsensusMode

Fields

getConsensusMode :: ConsensusMode

newtype NodeDiffusionMode

Newtype wrapper which provides ToJSON and FromJSON instances for DiffusionMode.

Constructors

NodeDiffusionMode

Fields

getDiffusionMode :: DiffusionMode

Consensus protocol configuration

data NodeByronProtocolConfiguration

Constructors

NodeByronProtocolConfiguration

Fields

npcByronGenesisFile :: !GenesisFile

npcByronGenesisFileHash :: !(Maybe GenesisHash)

npcByronReqNetworkMagic :: !RequiresNetworkMagic

npcByronPbftSignatureThresh :: !(Maybe Double)

npcByronSupportedProtocolVersionMajor :: !Word16

These declare the version of the protocol that the node is prepared to run. This is usually the version of the protocol in use on the chain now, but during protocol updates this version will be the one that we declare that we are ready to move to. This is the endorsement mechanism for determining when enough block producers are ready to move to the next version.

npcByronSupportedProtocolVersionMinor :: !Word16

npcByronSupportedProtocolVersionAlt :: !Word8

data NodeHardForkProtocolConfiguration

Configuration relating to a hard forks themselves, not the specific eras.

Constructors

NodeHardForkProtocolConfiguration

Fields

npcExperimentalHardForksEnabled :: Bool

During the development and integration of new eras we wish to be able to test the hard fork transition into the new era, but we do not wish to generally have the node advertise that it understands the new era. Avoiding advertising new development eras until they are ready makes it practical to include new not-yet-ready eras into the main release version of the node without the danger that operators on the mainnet will prematurely advertise that their nodes are capable of crossing the next hard fork.

It should always remain at the default of false for nodes running on the mainnet.

This flag should be set to true for nodes taking part in testnets for testing the new era.

npcTestShelleyHardForkAtEpoch :: Maybe EpochNo

For testing purposes we support specifying that the hard fork happens at an exact epoch number (ie the first epoch of the new era).

Obviously if this is used, all the nodes in the test cluster must be configured the same, or they will disagree.

npcTestShelleyHardForkAtVersion :: Maybe Word

For testing purposes we support specifying that the hard fork happens at a given major protocol version. For example this can be used to cause the Shelley hard fork to occur at the transition from protocol version 0 to version 1 (rather than the default of from 1 to 2) which can make the test setup simpler.

Obviously if this is used, all the nodes in the test cluster must be configured the same, or they will disagree.

npcTestAllegraHardForkAtEpoch :: Maybe EpochNo

For testing purposes we support specifying that the hard fork happens at an exact epoch number (ie the first epoch of the new era).

Obviously if this is used, all the nodes in the test cluster must be configured the same, or they will disagree.

npcTestAllegraHardForkAtVersion :: Maybe Word

For testing purposes we support specifying that the hard fork happens at a given major protocol version.

Obviously if this is used, all the nodes in the test cluster must be configured the same, or they will disagree.

npcTestMaryHardForkAtEpoch :: Maybe EpochNo

For testing purposes we support specifying that the hard fork happens at an exact epoch number (ie the first epoch of the new era).

Obviously if this is used, all the nodes in the test cluster must be configured the same, or they will disagree.

npcTestMaryHardForkAtVersion :: Maybe Word

For testing purposes we support specifying that the hard fork happens at a given major protocol version.

Obviously if this is used, all the nodes in the test cluster must be configured the same, or they will disagree.

npcTestAlonzoHardForkAtEpoch :: Maybe EpochNo

For testing purposes we support specifying that the hard fork happens at an exact epoch number (ie the first epoch of the new era).

Obviously if this is used, all the nodes in the test cluster must be configured the same, or they will disagree.

npcTestAlonzoHardForkAtVersion :: Maybe Word

For testing purposes we support specifying that the hard fork happens at a given major protocol version.

Obviously if this is used, all the nodes in the test cluster must be configured the same, or they will disagree.

npcTestBabbageHardForkAtEpoch :: Maybe EpochNo

npcTestBabbageHardForkAtVersion :: Maybe Word

For testing purposes we support specifying that the hard fork happens at a given major protocol version.

Obviously if this is used, all the nodes in the test cluster must be configured the same, or they will disagree.

npcTestConwayHardForkAtEpoch :: Maybe EpochNo

npcTestConwayHardForkAtVersion :: Maybe Word

npcTestDijkstraHardForkAtEpoch :: Maybe EpochNo

npcTestDijkstraHardForkAtVersion :: Maybe Word

npcTestStartingEra :: NodeHardForkProtocolConfiguration -> Maybe AnyShelleyBasedEra

Find the starting era for the test network, if it was configured.

Starting eras have zero defined as a forking epoch. So here we're taking the last zeroed configuration value. Returns Nothing if no HardForkAt option is present, or all of them have non-zero value - meaning we are starting from the very first era: Byron.

In mainnet config, the starting era is not configured, so this function will return Nothing.

Introduced in https://github.com/IntersectMBO/cardano-node/pull/5896 as a part of the fix of reading Plutus V2 cost model. Can be removed when era-sensitive AlonzoGenesis decoding gets removed.

data NodeProtocolConfiguration

Constructors

NodeProtocolConfigurationCardano NodeByronProtocolConfiguration NodeShelleyProtocolConfiguration NodeAlonzoProtocolConfiguration NodeConwayProtocolConfiguration (Maybe NodeDijkstraProtocolConfiguration) NodeHardForkProtocolConfiguration NodeCheckpointsConfiguration

data NodeShelleyProtocolConfiguration

Constructors

NodeShelleyProtocolConfiguration

Fields

npcShelleyGenesisFile :: !GenesisFile

npcShelleyGenesisFileHash :: !(Maybe GenesisHash)

data NodeAlonzoProtocolConfiguration

Constructors

NodeAlonzoProtocolConfiguration

Fields

npcAlonzoGenesisFile :: !GenesisFile

npcAlonzoGenesisFileHash :: !(Maybe GenesisHash)

data NodeConwayProtocolConfiguration

Constructors

NodeConwayProtocolConfiguration

Fields

npcConwayGenesisFile :: !GenesisFile

If no conway genesis file is provided, we want to enforce a maximum protocol version of 8 to avoid a permanent hard fork.

npcConwayGenesisFileHash :: !(Maybe GenesisHash)

data NodeDijkstraProtocolConfiguration

Constructors

NodeDijkstraProtocolConfiguration

Fields

npcDijkstraGenesisFile :: !GenesisFile

npcDijkstraGenesisFileHash :: !(Maybe GenesisHash)

data NodeCheckpointsConfiguration

See CheckpointsMap.

Constructors

NodeCheckpointsConfiguration

Fields

npcCheckpointsFile :: !(Maybe CheckpointsFile)

npcCheckpointsFileHash :: !(Maybe CheckpointsHash)

data VRFPrivateKeyFilePermissionError

Constructors

OtherPermissionsExist FilePath

GroupPermissionsExist FilePath

GenericPermissionsExist FilePath

renderVRFPrivateKeyFilePermissionError :: VRFPrivateKeyFilePermissionError -> Text