Cardano.Tracing.Metrics

Documentation

data ForgingStats

This structure stores counters of blockchain-related events, per individual forge thread. These counters are driven by traces.

Constructors

ForgingStats

Fields

fsTxsProcessedNum :: !(IORef Int)

Transactions removed from mempool.

fsState :: !(TVar (Map ThreadId (TVar ForgeThreadStats)))

fsBlocksUncoupled :: !(TVar Int64)

Blocks forged since last restart not on the current chain

data ForgeThreadStats

Per-forging-thread statistics.

Constructors

ForgeThreadStats

Fields

ftsNodeCannotForgeNum :: !Int

ftsNodeIsLeaderNum :: !Int

ftsBlocksForgedNum :: !Int

ftsSlotsMissedNum :: !Int

Potentially missed slots. Note that this is not the same as the number of missed blocks, since this includes all occurrences of not reaching a leadership check decision, whether or not leadership was possible or not.

Also note that when the aggregate total for this metric is reported in the multi-pool case, it can be much larger than the actual number of slots occurring since node start, for it is a sum total for all threads.

ftsLastSlot :: !Int

mapForgingCurrentThreadStats :: ForgingStats -> (ForgeThreadStats -> (ForgeThreadStats, a)) -> IO a

mapForgingCurrentThreadStats_ :: ForgingStats -> (ForgeThreadStats -> ForgeThreadStats) -> IO ()

mapForgingStatsTxsProcessed :: ForgingStats -> (Int -> Int) -> IO Int

mkForgingStats :: IO ForgingStats

threadStatsProjection :: ForgingStats -> (ForgeThreadStats -> a) -> IO [a]