Cardano.Tracing.OrphanInstances.Common

ToObject and helpers

class ToObject a where

Minimal complete definition

Nothing

Methods

toObject ∷ TracingVerbosity → a → Object

textTransformer ∷ a → Object → Text

data TracingVerbosity

Constructors

MinimalVerbosity

NormalVerbosity

MaximalVerbosity

mkObject ∷ ToObject a ⇒ [(Text, a)] → HashMap Text a

emptyObject ∷ ToObject a ⇒ HashMap Text a

class ToJSON a

toJSON ∷ ToJSON a ⇒ a → Value

(.=) ∷ (KeyValue kv, ToJSON v) ⇒ Key → v → kv

Transformable and helpers

class (Monad m, HasPrivacyAnnotation b, HasSeverityAnnotation b) ⇒ Transformable a (m ∷ Type → Type) b where

Minimal complete definition

Nothing

Methods

trTransformer ∷ TracingVerbosity → Trace m a → Tracer m b

trStructured ∷ ∀ b (m ∷ Type → Type) a. (ToObject b, MonadIO m, HasPrivacyAnnotation b, HasSeverityAnnotation b) ⇒ TracingVerbosity → Trace m a → Tracer m b

trStructuredText ∷ ∀ b (m ∷ Type → Type) a. (ToObject b, MonadIO m, HasTextFormatter b, HasPrivacyAnnotation b, HasSeverityAnnotation b) ⇒ TracingVerbosity → Trace m a → Tracer m b

class HasTextFormatter a where

Minimal complete definition

Nothing

Methods

formatText ∷ a → Object → Text

Severity and Privacy

class HasSeverityAnnotation a where

Minimal complete definition

Nothing

Methods

getSeverityAnnotation ∷ a → Severity

data Severity

Constructors

Debug

Info

Notice

Warning

Error

Critical

Alert

Emergency

class HasPrivacyAnnotation a where

Minimal complete definition

Nothing

Methods

getPrivacyAnnotation ∷ a → PrivacyAnnotation

data PrivacyAnnotation

Constructors

Confidential

Public

Tracer and related

data Tracer (m ∷ Type → Type) a

data LogObject a

Constructors

LogObject

Fields

loName ∷ LoggerName

loMeta ∷ !LOMeta

loContent ∷ !(LOContent a)

data LOContent a

Constructors

LogMessage a

LogError !Text

LogValue !Text !Measurable

LogStructuredText Object Text

LogStructured Object

ObserveOpen !CounterState

ObserveDiff !CounterState

ObserveClose !CounterState

AggregatedMessage [(Text, Aggregated)]

MonitoringEffect !MonitorAction

Command !CommandValue

KillPill

mkLOMeta ∷ MonadIO m ⇒ Severity → PrivacyAnnotation → m LOMeta

Orphan instances

FromJSON TracingVerbosity

Instance details

Methods

parseJSON ∷ Value → Parser TracingVerbosity

parseJSONList ∷ Value → Parser [TracingVerbosity]

FromJSON PortNumber

Instance details

Methods

parseJSON ∷ Value → Parser PortNumber

parseJSONList ∷ Value → Parser [PortNumber]

ToJSON ByronHash

Instance details

Methods

toJSON ∷ ByronHash → Value

toEncoding ∷ ByronHash → Encoding

toJSONList ∷ [ByronHash] → Value

toEncodingList ∷ [ByronHash] → Encoding

Show TracingVerbosity

Instance details

Methods

showsPrec ∷ Int → TracingVerbosity → ShowS

show ∷ TracingVerbosity → String

showList ∷ [TracingVerbosity] → ShowS

HasPrivacyAnnotation ResourceStats

Instance details

Methods

getPrivacyAnnotation ∷ ResourceStats → PrivacyAnnotation

HasSeverityAnnotation ResourceStats

Instance details

Methods

getSeverityAnnotation ∷ ResourceStats → Severity

ToObject Void
A bit of a weird one, but needed because some of the very general consensus interfaces are sometimes instantiated to Void, when there are no cases needed.

Instance details

Methods

toObject ∷ TracingVerbosity → Void → Object

textTransformer ∷ Void → Object → Text

ToObject ResourceStats

Instance details

Methods

toObject ∷ TracingVerbosity → ResourceStats → Object

textTransformer ∷ ResourceStats → Object → Text

Transformable Text IO ResourceStats

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO ResourceStats

ToJSON (HeaderHash blk) ⇒ ToJSON (Tip blk)

Instance details

Methods

toJSON ∷ Tip blk → Value

toEncoding ∷ Tip blk → Encoding

toJSONList ∷ [Tip blk] → Value

toEncodingList ∷ [Tip blk] → Encoding

ToJSON (OneEraHash xs)

Instance details

Methods

toJSON ∷ OneEraHash xs → Value

toEncoding ∷ OneEraHash xs → Encoding

toJSONList ∷ [OneEraHash xs] → Value

toEncodingList ∷ [OneEraHash xs] → Encoding