Cardano.Tracing.OrphanInstances.Consensus

Orphan instances

ToJSON SanityCheckIssue

Instance details

Methods

toJSON :: SanityCheckIssue -> Value

toEncoding :: SanityCheckIssue -> Encoding

toJSONList :: [SanityCheckIssue] -> Value

toEncodingList :: [SanityCheckIssue] -> Encoding

omitField :: SanityCheckIssue -> Bool

HasPrivacyAnnotation ConsensusStartupException

Instance details

Methods

getPrivacyAnnotation :: ConsensusStartupException -> PrivacyAnnotation

HasPrivacyAnnotation SanityCheckIssue

Instance details

Methods

getPrivacyAnnotation :: SanityCheckIssue -> PrivacyAnnotation

HasPrivacyAnnotation ()

Instance details

Methods

getPrivacyAnnotation :: () -> PrivacyAnnotation

HasSeverityAnnotation ConsensusStartupException

Instance details

Methods

getSeverityAnnotation :: ConsensusStartupException -> Severity

HasSeverityAnnotation SanityCheckIssue

Instance details

Methods

getSeverityAnnotation :: SanityCheckIssue -> Severity

HasSeverityAnnotation ()

Instance details

Methods

getSeverityAnnotation :: () -> Severity

HasTextFormatter ConsensusStartupException

Instance details

Methods

formatText :: ConsensusStartupException -> Object -> Text

HasTextFormatter ()

Instance details

Methods

formatText :: () -> Object -> Text

ToObject ConsensusStartupException

Instance details

Methods

toObject :: TracingVerbosity -> ConsensusStartupException -> Object

textTransformer :: ConsensusStartupException -> Object -> Text

ToObject SanityCheckIssue

Instance details

Methods

toObject :: TracingVerbosity -> SanityCheckIssue -> Object

textTransformer :: SanityCheckIssue -> Object -> Text

ToObject MempoolSize

Instance details

Methods

toObject :: TracingVerbosity -> MempoolSize -> Object

textTransformer :: MempoolSize -> Object -> Text

ToObject BftValidationErr
instances of ToObject

NOTE: this list is sorted by the unqualified name of the outermost type.

Instance details

Methods

toObject :: TracingVerbosity -> BftValidationErr -> Object

textTransformer :: BftValidationErr -> Object -> Text

ToObject DiskSnapshot

Instance details

Methods

toObject :: TracingVerbosity -> DiskSnapshot -> Object

textTransformer :: DiskSnapshot -> Object -> Text

Transformable Text IO ConsensusStartupException

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO ConsensusStartupException

Transformable Text IO SanityCheckIssue

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO SanityCheckIssue

Transformable Text IO ()

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO ()

(ToObject (ApplyTxErr blk), ToObject (GenTx blk), ToJSON (GenTxId blk), LedgerSupportsMempool blk, ConvertRawHash blk) => Transformable Text IO (TraceEventMempool blk)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceEventMempool blk)

(ConvertRawHash blk, LedgerSupportsProtocol blk) => Transformable Text IO (TraceChainSyncClientEvent blk)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceChainSyncClientEvent blk)

ToObject peer => Transformable Text IO (TraceEventDbf peer)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceEventDbf peer)

ConvertRawHash blk => Transformable Text IO (TraceChainSyncServerEvent blk)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceChainSyncServerEvent blk)

Transformable Text IO (TraceLocalTxSubmissionServerEvent blk)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLocalTxSubmissionServerEvent blk)

(HasPrivacyAnnotation (TraceAddBlockEvent blk), HasSeverityAnnotation (TraceAddBlockEvent blk), LedgerSupportsProtocol blk, ToObject (TraceAddBlockEvent blk)) => Transformable Text IO (TraceAddBlockEvent blk)
instances of Transformable

NOTE: this list is sorted by the unqualified name of the outermost type.

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceAddBlockEvent blk)

(ConvertRawHash blk, LedgerSupportsProtocol blk, InspectLedger blk, ToObject (Header blk), ToObject (LedgerEvent blk), ToObject (SelectView (BlockProtocol blk))) => Transformable Text IO (TraceEvent blk)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceEvent blk)

ToObject selection => Transformable Text IO (TraceGsmEvent selection)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceGsmEvent selection)

(tx ~ GenTx blk, HasTxId tx, RunNode blk, ToObject (LedgerError blk), ToObject (OtherHeaderEnvelopeError blk), ToObject (ValidationErr (BlockProtocol blk)), ToObject (CannotForge blk), ToObject (ForgeStateUpdateError blk), LedgerSupportsMempool blk) => Transformable Text IO (TraceForgeEvent blk)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceForgeEvent blk)

(HasPrivacyAnnotation a, HasSeverityAnnotation a, ToObject a) => Transformable Text IO (TraceLabelCreds a)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelCreds a)

(ToObject peer, ConvertRawHash blk) => Transformable Text IO (TraceLabelPeer peer (TraceBlockFetchServerEvent blk))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer peer (TraceBlockFetchServerEvent blk))

(ToObject peer, ToObject (TraceChainSyncClientEvent blk)) => Transformable Text IO (TraceLabelPeer peer (TraceChainSyncClientEvent blk))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer peer (TraceChainSyncClientEvent blk))

(ToObject peer, ConvertRawHash blk) => Transformable Text IO (TraceLabelPeer peer (TraceEventCsj peer blk))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer peer (TraceEventCsj peer blk))

(ToObject peer, ToObject (TraceChainSyncServerEvent blk)) => Transformable Text IO (TraceLabelPeer peer (TraceChainSyncServerEvent blk))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer peer (TraceChainSyncServerEvent blk))

(ToObject peer, ConvertRawHash blk, GetHeader blk) => Transformable Text IO (TraceGDDEvent peer blk)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceGDDEvent peer blk)

ToJSON t => ToJSON (Enclosing' t)

Instance details

Methods

toJSON :: Enclosing' t -> Value

toEncoding :: Enclosing' t -> Encoding

toJSONList :: [Enclosing' t] -> Value

toEncodingList :: [Enclosing' t] -> Encoding

omitField :: Enclosing' t -> Bool

Generic (Enclosing' t)

Instance details

Associated Types

type Rep (Enclosing' t) :: Type -> Type

Methods

from :: Enclosing' t -> Rep (Enclosing' t) x

to :: Rep (Enclosing' t) x -> Enclosing' t

HasPrivacyAnnotation (TraceEventMempool blk)

Instance details

Methods

getPrivacyAnnotation :: TraceEventMempool blk -> PrivacyAnnotation

HasPrivacyAnnotation (TraceBlockFetchServerEvent blk)

Instance details

Methods

getPrivacyAnnotation :: TraceBlockFetchServerEvent blk -> PrivacyAnnotation

HasPrivacyAnnotation (TraceChainSyncClientEvent blk)

Instance details

Methods

getPrivacyAnnotation :: TraceChainSyncClientEvent blk -> PrivacyAnnotation

HasPrivacyAnnotation (TraceEventDbf peer)

Instance details

Methods

getPrivacyAnnotation :: TraceEventDbf peer -> PrivacyAnnotation

HasPrivacyAnnotation (TraceChainSyncServerEvent blk)

Instance details

Methods

getPrivacyAnnotation :: TraceChainSyncServerEvent blk -> PrivacyAnnotation

HasPrivacyAnnotation (TraceLocalTxSubmissionServerEvent blk)

Instance details

Methods

getPrivacyAnnotation :: TraceLocalTxSubmissionServerEvent blk -> PrivacyAnnotation

HasPrivacyAnnotation (TraceEvent blk)

Instance details

Methods

getPrivacyAnnotation :: TraceEvent blk -> PrivacyAnnotation

HasPrivacyAnnotation (TraceGsmEvent selection)

Instance details

Methods

getPrivacyAnnotation :: TraceGsmEvent selection -> PrivacyAnnotation

HasPrivacyAnnotation (TraceForgeEvent blk)

Instance details

Methods

getPrivacyAnnotation :: TraceForgeEvent blk -> PrivacyAnnotation

HasPrivacyAnnotation a => HasPrivacyAnnotation (TraceLabelCreds a)

Instance details

Methods

getPrivacyAnnotation :: TraceLabelCreds a -> PrivacyAnnotation

HasSeverityAnnotation (LedgerEvent blk)

Instance details

Methods

getSeverityAnnotation :: LedgerEvent blk -> Severity

HasSeverityAnnotation (TraceEventMempool blk)

Instance details

Methods

getSeverityAnnotation :: TraceEventMempool blk -> Severity

HasSeverityAnnotation (TraceBlockFetchServerEvent blk)

Instance details

Methods

getSeverityAnnotation :: TraceBlockFetchServerEvent blk -> Severity

HasSeverityAnnotation (TraceChainSyncClientEvent blk)

Instance details

Methods

getSeverityAnnotation :: TraceChainSyncClientEvent blk -> Severity

HasSeverityAnnotation (TraceEventDbf peer)

Instance details

Methods

getSeverityAnnotation :: TraceEventDbf peer -> Severity

HasSeverityAnnotation (TraceChainSyncServerEvent blk)

Instance details

Methods

getSeverityAnnotation :: TraceChainSyncServerEvent blk -> Severity

HasSeverityAnnotation (TraceLocalTxSubmissionServerEvent blk)

Instance details

Methods

getSeverityAnnotation :: TraceLocalTxSubmissionServerEvent blk -> Severity

HasSeverityAnnotation (TraceEvent blk)

Instance details

Methods

getSeverityAnnotation :: TraceEvent blk -> Severity

HasSeverityAnnotation (TraceGsmEvent selection)

Instance details

Methods

getSeverityAnnotation :: TraceGsmEvent selection -> Severity

HasSeverityAnnotation (TraceForgeEvent blk)

Instance details

Methods

getSeverityAnnotation :: TraceForgeEvent blk -> Severity

HasSeverityAnnotation a => HasSeverityAnnotation (TraceLabelCreds a)

Instance details

Methods

getSeverityAnnotation :: TraceLabelCreds a -> Severity

HasTextFormatter (TraceEventDbf peer)

Instance details

Methods

formatText :: TraceEventDbf peer -> Object -> Text

LedgerSupportsProtocol blk => HasTextFormatter (TraceAddBlockEvent blk)

Instance details

Methods

formatText :: TraceAddBlockEvent blk -> Object -> Text

(ConvertRawHash blk, LedgerSupportsProtocol blk, InspectLedger blk) => HasTextFormatter (TraceEvent blk)

Instance details

Methods

formatText :: TraceEvent blk -> Object -> Text

(tx ~ GenTx blk, ConvertRawHash blk, HasTxId tx, LedgerSupportsMempool blk, LedgerSupportsProtocol blk, LedgerSupportsMempool blk, Show (TxId tx), Show (ForgeStateUpdateError blk), Show (CannotForge blk), LedgerSupportsMempool blk) => HasTextFormatter (TraceForgeEvent blk)

Instance details

Methods

formatText :: TraceForgeEvent blk -> Object -> Text

ConvertRawHash blk => ToObject (RealPoint blk)

Instance details

Methods

toObject :: TracingVerbosity -> RealPoint blk -> Object

textTransformer :: RealPoint blk -> Object -> Text

(ConvertRawHash blk, GetHeader blk) => ToObject (DensityBounds blk)

Instance details

Methods

toObject :: TracingVerbosity -> DensityBounds blk -> Object

textTransformer :: DensityBounds blk -> Object -> Text

(StandardHash blk, ToObject (OtherHeaderEnvelopeError blk)) => ToObject (HeaderEnvelopeError blk)

Instance details

Methods

toObject :: TracingVerbosity -> HeaderEnvelopeError blk -> Object

textTransformer :: HeaderEnvelopeError blk -> Object -> Text

(StandardHash blk, ToObject (ValidationErr (BlockProtocol blk)), ToObject (OtherHeaderEnvelopeError blk)) => ToObject (HeaderError blk)

Instance details

Methods

toObject :: TracingVerbosity -> HeaderError blk -> Object

textTransformer :: HeaderError blk -> Object -> Text

(StandardHash blk, ToObject (LedgerError blk), ToObject (OtherHeaderEnvelopeError blk), ToObject (ValidationErr (BlockProtocol blk))) => ToObject (ExtValidationError blk)

Instance details

Methods

toObject :: TracingVerbosity -> ExtValidationError blk -> Object

textTransformer :: ExtValidationError blk -> Object -> Text

(ToObject (LedgerUpdate blk), ToObject (LedgerWarning blk)) => ToObject (LedgerEvent blk)

Instance details

Methods

toObject :: TracingVerbosity -> LedgerEvent blk -> Object

textTransformer :: LedgerEvent blk -> Object -> Text

(ToObject (ApplyTxErr blk), ToObject (GenTx blk), ToJSON (GenTxId blk), LedgerSupportsMempool blk, ConvertRawHash blk) => ToObject (TraceEventMempool blk)

Instance details

Methods

toObject :: TracingVerbosity -> TraceEventMempool blk -> Object

textTransformer :: TraceEventMempool blk -> Object -> Text

ConvertRawHash blk => ToObject (TraceBlockFetchServerEvent blk)

Instance details

Methods

toObject :: TracingVerbosity -> TraceBlockFetchServerEvent blk -> Object

textTransformer :: TraceBlockFetchServerEvent blk -> Object -> Text

(ConvertRawHash blk, LedgerSupportsProtocol blk) => ToObject (TraceChainSyncClientEvent blk)

Instance details

Methods

toObject :: TracingVerbosity -> TraceChainSyncClientEvent blk -> Object

textTransformer :: TraceChainSyncClientEvent blk -> Object -> Text

(LedgerSupportsProtocol blk, ConvertRawHash blk) => ToObject (Instruction blk)

Instance details

Methods

toObject :: TracingVerbosity -> Instruction blk -> Object

textTransformer :: Instruction blk -> Object -> Text

(LedgerSupportsProtocol blk, ConvertRawHash blk) => ToObject (JumpInstruction blk)

Instance details

Methods

toObject :: TracingVerbosity -> JumpInstruction blk -> Object

textTransformer :: JumpInstruction blk -> Object -> Text

ToObject peer => ToObject (TraceEventDbf peer)

Instance details

Methods

toObject :: TracingVerbosity -> TraceEventDbf peer -> Object

textTransformer :: TraceEventDbf peer -> Object -> Text

(LedgerSupportsProtocol blk, ConvertRawHash blk) => ToObject (JumpInfo blk)

Instance details

Methods

toObject :: TracingVerbosity -> JumpInfo blk -> Object

textTransformer :: JumpInfo blk -> Object -> Text

ConvertRawHash blk => ToObject (TraceChainSyncServerEvent blk)

Instance details

Methods

toObject :: TracingVerbosity -> TraceChainSyncServerEvent blk -> Object

textTransformer :: TraceChainSyncServerEvent blk -> Object -> Text

ToObject (TraceLocalTxSubmissionServerEvent blk)

Instance details

Methods

toObject :: TracingVerbosity -> TraceLocalTxSubmissionServerEvent blk -> Object

textTransformer :: TraceLocalTxSubmissionServerEvent blk -> Object -> Text

Show (PBftVerKeyHash c) => ToObject (PBftCannotForge c)

Instance details

Methods

toObject :: TracingVerbosity -> PBftCannotForge c -> Object

textTransformer :: PBftCannotForge c -> Object -> Text

Show (PBftVerKeyHash c) => ToObject (PBftValidationErr c)

Instance details

Methods

toObject :: TracingVerbosity -> PBftValidationErr c -> Object

textTransformer :: PBftValidationErr c -> Object -> Text

(ConvertRawHash blk, LedgerSupportsProtocol blk, ToObject (Header blk), ToObject (LedgerEvent blk), ToObject (SelectView (BlockProtocol blk))) => ToObject (TraceEvent blk)

Instance details

Methods

toObject :: TracingVerbosity -> TraceEvent blk -> Object

textTransformer :: TraceEvent blk -> Object -> Text

ToObject selection => ToObject (TraceGsmEvent selection)

Instance details

Methods

toObject :: TracingVerbosity -> TraceGsmEvent selection -> Object

textTransformer :: TraceGsmEvent selection -> Object -> Text

(RunNode blk, ToObject (LedgerError blk), ToObject (OtherHeaderEnvelopeError blk), ToObject (ValidationErr (BlockProtocol blk)), ToObject (CannotForge blk), ToObject (ForgeStateUpdateError blk)) => ToObject (TraceForgeEvent blk)

Instance details

Methods

toObject :: TracingVerbosity -> TraceForgeEvent blk -> Object

textTransformer :: TraceForgeEvent blk -> Object -> Text

ToObject a => ToObject (TraceLabelCreds a)

Instance details

Methods

toObject :: TracingVerbosity -> TraceLabelCreds a -> Object

textTransformer :: TraceLabelCreds a -> Object -> Text

ConvertRawHash blk => ConvertRawHash (Header blk)

Instance details

Methods

toRawHash :: proxy (Header blk) -> HeaderHash (Header blk) -> ByteString

fromRawHash :: proxy (Header blk) -> ByteString -> HeaderHash (Header blk)

toShortRawHash :: proxy (Header blk) -> HeaderHash (Header blk) -> ShortByteString

fromShortRawHash :: proxy (Header blk) -> ShortByteString -> HeaderHash (Header blk)

hashSize :: proxy (Header blk) -> Word32

Condense t => Condense (Enclosing' t)

Instance details

Methods

condense :: Enclosing' t -> String

HasPrivacyAnnotation (TraceGDDEvent peer blk)

Instance details

Methods

getPrivacyAnnotation :: TraceGDDEvent peer blk -> PrivacyAnnotation

HasPrivacyAnnotation (TraceEventCsj peer blk)

Instance details

Methods

getPrivacyAnnotation :: TraceEventCsj peer blk -> PrivacyAnnotation

HasSeverityAnnotation (TraceGDDEvent peer blk)

Instance details

Methods

getSeverityAnnotation :: TraceGDDEvent peer blk -> Severity

HasSeverityAnnotation (TraceEventCsj peer blk)

Instance details

Methods

getSeverityAnnotation :: TraceEventCsj peer blk -> Severity

HasTextFormatter (TraceLabelPeer peer (TraceBlockFetchServerEvent blk))

Instance details

Methods

formatText :: TraceLabelPeer peer (TraceBlockFetchServerEvent blk) -> Object -> Text

(BlockSupportsProtocol blk, Show peer, Show (Header blk)) => HasTextFormatter (TraceLabelPeer peer (TraceChainSyncClientEvent blk))

Instance details

Methods

formatText :: TraceLabelPeer peer (TraceChainSyncClientEvent blk) -> Object -> Text

(StandardHash blk, Show peer) => HasTextFormatter (TraceLabelPeer peer (TraceChainSyncServerEvent blk))

Instance details

Methods

formatText :: TraceLabelPeer peer (TraceChainSyncServerEvent blk) -> Object -> Text

(ToObject peer, ConvertRawHash blk, GetHeader blk) => ToObject (TraceGDDEvent peer blk)

Instance details

Methods

toObject :: TracingVerbosity -> TraceGDDEvent peer blk -> Object

textTransformer :: TraceGDDEvent peer blk -> Object -> Text

(ToObject peer, ConvertRawHash blk) => ToObject (TraceEventCsj peer blk)

Instance details

Methods

toObject :: TracingVerbosity -> TraceEventCsj peer blk -> Object

textTransformer :: TraceEventCsj peer blk -> Object -> Text

ConvertRawHash blk => ToObject (TraceChunkValidation blk ChunkNo)

Instance details

Methods

toObject :: TracingVerbosity -> TraceChunkValidation blk ChunkNo -> Object

textTransformer :: TraceChunkValidation blk ChunkNo -> Object -> Text

ConvertRawHash blk => ToObject (Tip blk)

Instance details

Methods

toObject :: TracingVerbosity -> Tip blk -> Object

textTransformer :: Tip blk -> Object -> Text