Cardano.Tracing.OrphanInstances.Network

Orphan instances

ToJSON IP

Instance details

Methods

toJSON ∷ IP → Value

toEncoding ∷ IP → Encoding

toJSONList ∷ [IP] → Value

toEncodingList ∷ [IP] → Encoding

ToJSON SockAddr

Instance details

Methods

toJSON ∷ SockAddr → Value

toEncoding ∷ SockAddr → Encoding

toJSONList ∷ [SockAddr] → Value

toEncodingList ∷ [SockAddr] → Encoding

ToJSON MiniProtocolNum

Instance details

Methods

toJSON ∷ MiniProtocolNum → Value

toEncoding ∷ MiniProtocolNum → Encoding

toJSONList ∷ [MiniProtocolNum] → Value

toEncodingList ∷ [MiniProtocolNum] → Encoding

ToJSON ReconnectDelay

Instance details

Methods

toJSON ∷ ReconnectDelay → Value

toEncoding ∷ ReconnectDelay → Encoding

toJSONList ∷ [ReconnectDelay] → Value

toEncodingList ∷ [ReconnectDelay] → Encoding

ToJSON PeerSelectionTargets

Instance details

Methods

toJSON ∷ PeerSelectionTargets → Value

toEncoding ∷ PeerSelectionTargets → Encoding

toJSONList ∷ [PeerSelectionTargets] → Value

toEncodingList ∷ [PeerSelectionTargets] → Encoding

ToJSON PeerStatus

Instance details

Methods

toJSON ∷ PeerStatus → Value

toEncoding ∷ PeerStatus → Encoding

toJSONList ∷ [PeerStatus] → Value

toEncodingList ∷ [PeerStatus] → Encoding

ToJSON NodeToClientVersion

Instance details

Methods

toJSON ∷ NodeToClientVersion → Value

toEncoding ∷ NodeToClientVersion → Encoding

toJSONList ∷ [NodeToClientVersion] → Value

toEncodingList ∷ [NodeToClientVersion] → Encoding

ToJSON NodeToClientVersionData

Instance details

Methods

toJSON ∷ NodeToClientVersionData → Value

toEncoding ∷ NodeToClientVersionData → Encoding

toJSONList ∷ [NodeToClientVersionData] → Value

toEncodingList ∷ [NodeToClientVersionData] → Encoding

ToJSON NodeToNodeVersion

Instance details

Methods

toJSON ∷ NodeToNodeVersion → Value

toEncoding ∷ NodeToNodeVersion → Encoding

toJSONList ∷ [NodeToNodeVersion] → Value

toEncodingList ∷ [NodeToNodeVersion] → Encoding

ToJSON NodeToNodeVersionData

Instance details

Methods

toJSON ∷ NodeToNodeVersionData → Value

toEncoding ∷ NodeToNodeVersionData → Encoding

toJSONList ∷ [NodeToNodeVersionData] → Value

toEncodingList ∷ [NodeToNodeVersionData] → Encoding

ToJSON ProtocolLimitFailure

Instance details

Methods

toJSON ∷ ProtocolLimitFailure → Value

toEncoding ∷ ProtocolLimitFailure → Encoding

toJSONList ∷ [ProtocolLimitFailure] → Value

toEncodingList ∷ [ProtocolLimitFailure] → Encoding

ToJSON AbstractState

Instance details

Methods

toJSON ∷ AbstractState → Value

toEncoding ∷ AbstractState → Encoding

toJSONList ∷ [AbstractState] → Value

toEncodingList ∷ [AbstractState] → Encoding

ToJSON ConnectionManagerCounters

Instance details

Methods

toJSON ∷ ConnectionManagerCounters → Value

toEncoding ∷ ConnectionManagerCounters → Encoding

toJSONList ∷ [ConnectionManagerCounters] → Value

toEncodingList ∷ [ConnectionManagerCounters] → Encoding

ToJSON RemoteSt

Instance details

Methods

toJSON ∷ RemoteSt → Value

toEncoding ∷ RemoteSt → Encoding

toJSONList ∷ [RemoteSt] → Value

toEncodingList ∷ [RemoteSt] → Encoding

ToJSON LocalAddress

Instance details

Methods

toJSON ∷ LocalAddress → Value

toEncoding ∷ LocalAddress → Encoding

toJSONList ∷ [LocalAddress] → Value

toEncodingList ∷ [LocalAddress] → Encoding

ToJSONKey SockAddr

Instance details

Methods

toJSONKey ∷ ToJSONKeyFunction SockAddr

toJSONKeyList ∷ ToJSONKeyFunction [SockAddr]

ToJSONKey RelayAccessPoint

Instance details

Methods

toJSONKey ∷ ToJSONKeyFunction RelayAccessPoint

toJSONKeyList ∷ ToJSONKeyFunction [RelayAccessPoint]

ToJSONKey LocalAddress

Instance details

Methods

toJSONKey ∷ ToJSONKeyFunction LocalAddress

toJSONKeyList ∷ ToJSONKeyFunction [LocalAddress]

HasPrivacyAnnotation PeerSelectionCounters

Instance details

Methods

getPrivacyAnnotation ∷ PeerSelectionCounters → PrivacyAnnotation

HasPrivacyAnnotation TraceLedgerPeers

Instance details

Methods

getPrivacyAnnotation ∷ TraceLedgerPeers → PrivacyAnnotation

HasPrivacyAnnotation TracePublicRootPeers

Instance details

Methods

getPrivacyAnnotation ∷ TracePublicRootPeers → PrivacyAnnotation

HasPrivacyAnnotation AcceptConnectionsPolicyTrace

Instance details

Methods

getPrivacyAnnotation ∷ AcceptConnectionsPolicyTrace → PrivacyAnnotation

HasSeverityAnnotation PeerSelectionCounters

Instance details

Methods

getSeverityAnnotation ∷ PeerSelectionCounters → Severity

HasSeverityAnnotation TraceLedgerPeers

Instance details

Methods

getSeverityAnnotation ∷ TraceLedgerPeers → Severity

HasSeverityAnnotation TracePublicRootPeers

Instance details

Methods

getSeverityAnnotation ∷ TracePublicRootPeers → Severity

HasSeverityAnnotation AcceptConnectionsPolicyTrace

Instance details

Methods

getSeverityAnnotation ∷ AcceptConnectionsPolicyTrace → Severity

HasTextFormatter PeerSelectionCounters

Instance details

Methods

formatText ∷ PeerSelectionCounters → Object → Text

HasTextFormatter TraceLedgerPeers

Instance details

Methods

formatText ∷ TraceLedgerPeers → Object → Text

HasTextFormatter TracePublicRootPeers

Instance details

Methods

formatText ∷ TracePublicRootPeers → Object → Text

HasTextFormatter AcceptConnectionsPolicyTrace

Instance details

Methods

formatText ∷ AcceptConnectionsPolicyTrace → Object → Text

ToObject SlotNo

Instance details

Methods

toObject ∷ TracingVerbosity → SlotNo → Object

textTransformer ∷ SlotNo → Object → Text

ToObject LocalConnectionId

Instance details

Methods

toObject ∷ TracingVerbosity → LocalConnectionId → Object

textTransformer ∷ LocalConnectionId → Object → Text

ToObject RemoteAddress

Instance details

Methods

toObject ∷ TracingVerbosity → RemoteAddress → Object

textTransformer ∷ RemoteAddress → Object → Text

ToObject RemoteConnectionId

Instance details

Methods

toObject ∷ TracingVerbosity → RemoteConnectionId → Object

textTransformer ∷ RemoteConnectionId → Object → Text

ToObject PeerSelectionCounters

Instance details

Methods

toObject ∷ TracingVerbosity → PeerSelectionCounters → Object

textTransformer ∷ PeerSelectionCounters → Object → Text

ToObject TraceLedgerPeers

Instance details

Methods

toObject ∷ TracingVerbosity → TraceLedgerPeers → Object

textTransformer ∷ TraceLedgerPeers → Object → Text

ToObject TracePublicRootPeers

Instance details

Methods

toObject ∷ TracingVerbosity → TracePublicRootPeers → Object

textTransformer ∷ TracePublicRootPeers → Object → Text

ToObject AcceptConnectionsPolicyTrace

Instance details

Methods

toObject ∷ TracingVerbosity → AcceptConnectionsPolicyTrace → Object

textTransformer ∷ AcceptConnectionsPolicyTrace → Object → Text

ToObject LocalAddress

Instance details

Methods

toObject ∷ TracingVerbosity → LocalAddress → Object

textTransformer ∷ LocalAddress → Object → Text

Transformable Text IO PeerSelectionCounters

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO PeerSelectionCounters

Transformable Text IO TraceLedgerPeers

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO TraceLedgerPeers

Transformable Text IO TracePublicRootPeers

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO TracePublicRootPeers

Transformable Text IO AcceptConnectionsPolicyTrace

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO AcceptConnectionsPolicyTrace

Transformable Text IO (Identity (SubscriptionTrace LocalAddress))

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (Identity (SubscriptionTrace LocalAddress))

Show addr ⇒ Transformable Text IO (TraceKeepAliveClient addr)

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (TraceKeepAliveClient addr)

Transformable Text IO (DebugPeerSelection SockAddr)

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (DebugPeerSelection SockAddr)

Transformable Text IO (TracePeerSelection SockAddr)

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (TracePeerSelection SockAddr)

(Show addr, ToJSON addr, ToObject addr) ⇒ Transformable Text IO (AbstractTransitionTrace addr)

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (AbstractTransitionTrace addr)

(ToJSON addr, Show addr) ⇒ Transformable Text IO (InboundGovernorTrace addr)

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (InboundGovernorTrace addr)

(Show addr, ToJSON addr) ⇒ Transformable Text IO (RemoteTransitionTrace addr)

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (RemoteTransitionTrace addr)

(Show addr, ToObject addr, ToJSON addr) ⇒ Transformable Text IO (ServerTrace addr)

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (ServerTrace addr)

Transformable Text IO (WithDomainName DnsTrace)

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (WithDomainName DnsTrace)

Transformable Text IO (WithDomainName (SubscriptionTrace SockAddr))

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (WithDomainName (SubscriptionTrace SockAddr))

Transformable Text IO (WithIPList (SubscriptionTrace SockAddr))

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (WithIPList (SubscriptionTrace SockAddr))

(StandardHash header, Show peer, ToObject peer) ⇒ Transformable Text IO [TraceLabelPeer peer (FetchDecision [Point header])]

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO [TraceLabelPeer peer (FetchDecision [Point header])]

ToObject localPeer ⇒ Transformable Text IO (TraceLabelPeer localPeer (TraceSendRecv (ChainSync (Serialised blk) (Point blk) (Tip blk))))

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (TraceLabelPeer localPeer (TraceSendRecv (ChainSync (Serialised blk) (Point blk) (Tip blk))))

(ShowQuery (BlockQuery blk), ToObject localPeer) ⇒ Transformable Text IO (TraceLabelPeer localPeer (TraceSendRecv (LocalStateQuery blk (Point blk) (Query blk))))

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (TraceLabelPeer localPeer (TraceSendRecv (LocalStateQuery blk (Point blk) (Query blk))))

(applyTxErr ~ ApplyTxErr blk, ToObject localPeer) ⇒ Transformable Text IO (TraceLabelPeer localPeer (TraceSendRecv (LocalTxMonitor (GenTxId blk) (GenTx blk) SlotNo)))

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (TraceLabelPeer localPeer (TraceSendRecv (LocalTxMonitor (GenTxId blk) (GenTx blk) SlotNo)))

(applyTxErr ~ ApplyTxErr blk, ToObject localPeer) ⇒ Transformable Text IO (TraceLabelPeer localPeer (TraceSendRecv (LocalTxSubmission (GenTx blk) applyTxErr)))

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (TraceLabelPeer localPeer (TraceSendRecv (LocalTxSubmission (GenTx blk) applyTxErr)))

(HasHeader header, ConvertRawHash header, ToObject peer) ⇒ Transformable Text IO (TraceLabelPeer peer (TraceFetchClientState header))

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (TraceLabelPeer peer (TraceFetchClientState header))

ToObject peer ⇒ Transformable Text IO (TraceLabelPeer peer (TraceTxSubmissionInbound (GenTxId blk) (GenTx blk)))

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (TraceLabelPeer peer (TraceTxSubmissionInbound (GenTxId blk) (GenTx blk)))

(ToObject peer, Show (TxId (GenTx blk)), Show (GenTx blk)) ⇒ Transformable Text IO (TraceLabelPeer peer (TraceTxSubmissionOutbound (GenTxId blk) (GenTx blk)))

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (TraceLabelPeer peer (TraceTxSubmissionOutbound (GenTxId blk) (GenTx blk)))

(ToObject peer, ToObject (AnyMessageAndAgency (TraceTxSubmissionInbound (GenTxId blk) (GenTx blk)))) ⇒ Transformable Text IO (TraceLabelPeer peer (TraceSendRecv (TraceTxSubmissionInbound (GenTxId blk) (GenTx blk))))

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (TraceLabelPeer peer (TraceSendRecv (TraceTxSubmissionInbound (GenTxId blk) (GenTx blk))))

(ToObject peer, ConvertTxId blk, RunNode blk, HasTxs blk) ⇒ Transformable Text IO (TraceLabelPeer peer (TraceSendRecv (BlockFetch blk (Point blk))))

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (TraceLabelPeer peer (TraceSendRecv (BlockFetch blk (Point blk))))

ToObject peer ⇒ Transformable Text IO (TraceLabelPeer peer (TraceSendRecv (ChainSync (Header blk) (Point blk) (Tip blk))))

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (TraceLabelPeer peer (TraceSendRecv (ChainSync (Header blk) (Point blk) (Tip blk))))

(ToObject peer, Show (TxId (GenTx blk)), Show (GenTx blk)) ⇒ Transformable Text IO (TraceLabelPeer peer (TraceSendRecv (TxSubmission2 (GenTxId blk) (GenTx blk))))

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (TraceLabelPeer peer (TraceSendRecv (TxSubmission2 (GenTxId blk) (GenTx blk))))

(Show peer, ToObject peer) ⇒ Transformable Text IO (WithMuxBearer peer MuxTrace)

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (WithMuxBearer peer MuxTrace)

Transformable Text IO (DiffusionTracer RemoteAddress LocalAddress)
instances of Transformable

NOTE: this list is sorted by the unqualified name of the outermost type.

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (DiffusionTracer RemoteAddress LocalAddress)

Transformable Text IO (HandshakeTr RemoteAddress NodeToNodeVersion)

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (HandshakeTr RemoteAddress NodeToNodeVersion)

Show lAddr ⇒ Transformable Text IO (PeerSelectionActionsTrace SockAddr lAddr)

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (PeerSelectionActionsTrace SockAddr lAddr)

Show exception ⇒ Transformable Text IO (TraceLocalRootPeers RemoteAddress exception)

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (TraceLocalRootPeers RemoteAddress exception)

Transformable Text IO (TraceTxSubmissionInbound txid tx)

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (TraceTxSubmissionInbound txid tx)

(Show tx, Show txid) ⇒ Transformable Text IO (TraceTxSubmissionOutbound txid tx)

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (TraceTxSubmissionOutbound txid tx)

(Show addr, Show versionNumber, Show agreedOptions, ToObject addr, ToJSON addr, ToJSON versionNumber, ToJSON agreedOptions) ⇒ Transformable Text IO (ConnectionManagerTrace addr (ConnectionHandlerTrace versionNumber agreedOptions))

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (ConnectionManagerTrace addr (ConnectionHandlerTrace versionNumber agreedOptions))

Show addr ⇒ Transformable Text IO (WithAddr addr ErrorPolicyTrace)

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (WithAddr addr ErrorPolicyTrace)

Transformable Text IO (HandshakeTr LocalAddress NodeToClientVersion)

Instance details

Methods

trTransformer ∷ TracingVerbosity → Trace IO Text → Tracer IO (HandshakeTr LocalAddress NodeToClientVersion)

(ToJSONKey peerAddr, ToJSON peerAddr, Ord peerAddr, Show peerAddr) ⇒ ToJSON (LocalRootPeers peerAddr)

Instance details

Methods

toJSON ∷ LocalRootPeers peerAddr → Value

toEncoding ∷ LocalRootPeers peerAddr → Encoding

toJSONList ∷ [LocalRootPeers peerAddr] → Value

toEncodingList ∷ [LocalRootPeers peerAddr] → Encoding

ToJSON peerAddr ⇒ ToJSON (ConnectionId peerAddr)

Instance details

Methods

toJSON ∷ ConnectionId peerAddr → Value

toEncoding ∷ ConnectionId peerAddr → Encoding

toJSONList ∷ [ConnectionId peerAddr] → Value

toEncodingList ∷ [ConnectionId peerAddr] → Encoding

ToJSON state ⇒ ToJSON (MaybeUnknown state)

Instance details

Methods

toJSON ∷ MaybeUnknown state → Value

toEncoding ∷ MaybeUnknown state → Encoding

toJSONList ∷ [MaybeUnknown state] → Value

toEncodingList ∷ [MaybeUnknown state] → Encoding

ToJSON addr ⇒ ToJSON (OperationResult addr)

Instance details

Methods

toJSON ∷ OperationResult addr → Value

toEncoding ∷ OperationResult addr → Encoding

toJSONList ∷ [OperationResult addr] → Value

toEncodingList ∷ [OperationResult addr] → Encoding

Show vNumber ⇒ ToJSON (HandshakeException vNumber)

Instance details

Methods

toJSON ∷ HandshakeException vNumber → Value

toEncoding ∷ HandshakeException vNumber → Encoding

toJSONList ∷ [HandshakeException vNumber] → Value

toEncodingList ∷ [HandshakeException vNumber] → Encoding

Show vNumber ⇒ ToJSON (HandshakeProtocolError vNumber)

Instance details

Methods

toJSON ∷ HandshakeProtocolError vNumber → Value

toEncoding ∷ HandshakeProtocolError vNumber → Encoding

toJSONList ∷ [HandshakeProtocolError vNumber] → Value

toEncodingList ∷ [HandshakeProtocolError vNumber] → Encoding

Show vNumber ⇒ ToJSON (RefuseReason vNumber)

Instance details

Methods

toJSON ∷ RefuseReason vNumber → Value

toEncoding ∷ RefuseReason vNumber → Encoding

toJSONList ∷ [RefuseReason vNumber] → Value

toEncodingList ∷ [RefuseReason vNumber] → Encoding

ToJSON addr ⇒ ToJSONKey (ConnectionId addr)

Instance details

Methods

toJSONKey ∷ ToJSONKeyFunction (ConnectionId addr)

toJSONKeyList ∷ ToJSONKeyFunction [ConnectionId addr]

HasPrivacyAnnotation (Identity (SubscriptionTrace LocalAddress))

Instance details

Methods

getPrivacyAnnotation ∷ Identity (SubscriptionTrace LocalAddress) → PrivacyAnnotation

HasPrivacyAnnotation (TraceFetchClientState header)

Instance details

Methods

getPrivacyAnnotation ∷ TraceFetchClientState header → PrivacyAnnotation

HasPrivacyAnnotation (TraceKeepAliveClient remotePeer)

Instance details

Methods

getPrivacyAnnotation ∷ TraceKeepAliveClient remotePeer → PrivacyAnnotation

HasPrivacyAnnotation (DebugPeerSelection addr)

Instance details

Methods

getPrivacyAnnotation ∷ DebugPeerSelection addr → PrivacyAnnotation

HasPrivacyAnnotation (TracePeerSelection addr)

Instance details

Methods

getPrivacyAnnotation ∷ TracePeerSelection addr → PrivacyAnnotation

HasPrivacyAnnotation (AbstractTransitionTrace addr)

Instance details

Methods

getPrivacyAnnotation ∷ AbstractTransitionTrace addr → PrivacyAnnotation

HasPrivacyAnnotation (TraceSendRecv a)

Instance details

Methods

getPrivacyAnnotation ∷ TraceSendRecv a → PrivacyAnnotation

HasPrivacyAnnotation (InboundGovernorTrace addr)

Instance details

Methods

getPrivacyAnnotation ∷ InboundGovernorTrace addr → PrivacyAnnotation

HasPrivacyAnnotation (RemoteTransitionTrace addr)

Instance details

Methods

getPrivacyAnnotation ∷ RemoteTransitionTrace addr → PrivacyAnnotation

HasPrivacyAnnotation (ServerTrace addr)

Instance details

Methods

getPrivacyAnnotation ∷ ServerTrace addr → PrivacyAnnotation

HasPrivacyAnnotation (WithDomainName DnsTrace)

Instance details

Methods

getPrivacyAnnotation ∷ WithDomainName DnsTrace → PrivacyAnnotation

HasPrivacyAnnotation (WithDomainName (SubscriptionTrace SockAddr))

Instance details

Methods

getPrivacyAnnotation ∷ WithDomainName (SubscriptionTrace SockAddr) → PrivacyAnnotation

HasPrivacyAnnotation (WithIPList (SubscriptionTrace SockAddr))

Instance details

Methods

getPrivacyAnnotation ∷ WithIPList (SubscriptionTrace SockAddr) → PrivacyAnnotation

HasPrivacyAnnotation [TraceLabelPeer peer (FetchDecision [Point header])]

Instance details

Methods

getPrivacyAnnotation ∷ [TraceLabelPeer peer (FetchDecision [Point header])] → PrivacyAnnotation

HasSeverityAnnotation (Identity (SubscriptionTrace LocalAddress))

Instance details

Methods

getSeverityAnnotation ∷ Identity (SubscriptionTrace LocalAddress) → Severity

HasSeverityAnnotation (TraceFetchClientState header)

Instance details

Methods

getSeverityAnnotation ∷ TraceFetchClientState header → Severity

HasSeverityAnnotation (TraceKeepAliveClient remotePeer)

Instance details

Methods

getSeverityAnnotation ∷ TraceKeepAliveClient remotePeer → Severity

HasSeverityAnnotation (DebugPeerSelection addr)

Instance details

Methods

getSeverityAnnotation ∷ DebugPeerSelection addr → Severity

HasSeverityAnnotation (TracePeerSelection addr)

Instance details

Methods

getSeverityAnnotation ∷ TracePeerSelection addr → Severity

HasSeverityAnnotation (AbstractTransitionTrace addr)

Instance details

Methods

getSeverityAnnotation ∷ AbstractTransitionTrace addr → Severity

HasSeverityAnnotation (TraceSendRecv a)

Instance details

Methods

getSeverityAnnotation ∷ TraceSendRecv a → Severity

HasSeverityAnnotation (InboundGovernorTrace addr)

Instance details

Methods

getSeverityAnnotation ∷ InboundGovernorTrace addr → Severity

HasSeverityAnnotation (RemoteTransitionTrace addr)

Instance details

Methods

getSeverityAnnotation ∷ RemoteTransitionTrace addr → Severity

HasSeverityAnnotation (ServerTrace addr)

Instance details

Methods

getSeverityAnnotation ∷ ServerTrace addr → Severity

HasSeverityAnnotation (WithDomainName DnsTrace)

Instance details

Methods

getSeverityAnnotation ∷ WithDomainName DnsTrace → Severity

HasSeverityAnnotation (WithDomainName (SubscriptionTrace SockAddr))

Instance details

Methods

getSeverityAnnotation ∷ WithDomainName (SubscriptionTrace SockAddr) → Severity

HasSeverityAnnotation (WithIPList (SubscriptionTrace SockAddr))

Instance details

Methods

getSeverityAnnotation ∷ WithIPList (SubscriptionTrace SockAddr) → Severity

HasSeverityAnnotation [TraceLabelPeer peer (FetchDecision [Point header])]

Instance details

Methods

getSeverityAnnotation ∷ [TraceLabelPeer peer (FetchDecision [Point header])] → Severity

HasTextFormatter (Identity (SubscriptionTrace LocalAddress))

Instance details

Methods

formatText ∷ Identity (SubscriptionTrace LocalAddress) → Object → Text

Show addr ⇒ HasTextFormatter (TraceKeepAliveClient addr)

Instance details

Methods

formatText ∷ TraceKeepAliveClient addr → Object → Text

HasTextFormatter (DebugPeerSelection SockAddr)

Instance details

Methods

formatText ∷ DebugPeerSelection SockAddr → Object → Text

HasTextFormatter (TracePeerSelection SockAddr)

Instance details

Methods

formatText ∷ TracePeerSelection SockAddr → Object → Text

Show addr ⇒ HasTextFormatter (AbstractTransitionTrace addr)

Instance details

Methods

formatText ∷ AbstractTransitionTrace addr → Object → Text

Show addr ⇒ HasTextFormatter (InboundGovernorTrace addr)

Instance details

Methods

formatText ∷ InboundGovernorTrace addr → Object → Text

Show addr ⇒ HasTextFormatter (RemoteTransitionTrace addr)

Instance details

Methods

formatText ∷ RemoteTransitionTrace addr → Object → Text

Show addr ⇒ HasTextFormatter (ServerTrace addr)

Instance details

Methods

formatText ∷ ServerTrace addr → Object → Text

HasTextFormatter (WithDomainName DnsTrace)

Instance details

Methods

formatText ∷ WithDomainName DnsTrace → Object → Text

HasTextFormatter (WithDomainName (SubscriptionTrace SockAddr))

Instance details

Methods

formatText ∷ WithDomainName (SubscriptionTrace SockAddr) → Object → Text

HasTextFormatter (WithIPList (SubscriptionTrace SockAddr))

Instance details

Methods

formatText ∷ WithIPList (SubscriptionTrace SockAddr) → Object → Text

(StandardHash header, Show peer) ⇒ HasTextFormatter [TraceLabelPeer peer (FetchDecision [Point header])]

Instance details

Methods

formatText ∷ [TraceLabelPeer peer (FetchDecision [Point header])] → Object → Text

ToObject (Identity (SubscriptionTrace LocalAddress))

Instance details

Methods

toObject ∷ TracingVerbosity → Identity (SubscriptionTrace LocalAddress) → Object

textTransformer ∷ Identity (SubscriptionTrace LocalAddress) → Object → Text

(HasHeader header, ConvertRawHash header) ⇒ ToObject (TraceFetchClientState header)

Instance details

Methods

toObject ∷ TracingVerbosity → TraceFetchClientState header → Object

textTransformer ∷ TraceFetchClientState header → Object → Text

ToObject (FetchDecision [Point header])

Instance details

Methods

toObject ∷ TracingVerbosity → FetchDecision [Point header] → Object

textTransformer ∷ FetchDecision [Point header] → Object → Text

Show remotePeer ⇒ ToObject (TraceKeepAliveClient remotePeer)

Instance details

Methods

toObject ∷ TracingVerbosity → TraceKeepAliveClient remotePeer → Object

textTransformer ∷ TraceKeepAliveClient remotePeer → Object → Text

ToObject (DebugPeerSelection SockAddr)

Instance details

Methods

toObject ∷ TracingVerbosity → DebugPeerSelection SockAddr → Object

textTransformer ∷ DebugPeerSelection SockAddr → Object → Text

ToObject (TracePeerSelection SockAddr)

Instance details

Methods

toObject ∷ TracingVerbosity → TracePeerSelection SockAddr → Object

textTransformer ∷ TracePeerSelection SockAddr → Object → Text

ConvertRawHash blk ⇒ ToObject (Point blk)

Instance details

Methods

toObject ∷ TracingVerbosity → Point blk → Object

textTransformer ∷ Point blk → Object → Text

(Show addr, ToObject addr, ToJSON addr) ⇒ ToObject (AbstractTransitionTrace addr)

Instance details

Methods

toObject ∷ TracingVerbosity → AbstractTransitionTrace addr → Object

textTransformer ∷ AbstractTransitionTrace addr → Object → Text

ToObject (AnyMessageAndAgency ps) ⇒ ToObject (TraceSendRecv ps)

Instance details

Methods

toObject ∷ TracingVerbosity → TraceSendRecv ps → Object

textTransformer ∷ TraceSendRecv ps → Object → Text

(ToJSON addr, Show addr) ⇒ ToObject (InboundGovernorTrace addr)

Instance details

Methods

toObject ∷ TracingVerbosity → InboundGovernorTrace addr → Object

textTransformer ∷ InboundGovernorTrace addr → Object → Text

ToJSON addr ⇒ ToObject (RemoteTransitionTrace addr)

Instance details

Methods

toObject ∷ TracingVerbosity → RemoteTransitionTrace addr → Object

textTransformer ∷ RemoteTransitionTrace addr → Object → Text

(Show addr, ToObject addr, ToJSON addr) ⇒ ToObject (ServerTrace addr)

Instance details

Methods

toObject ∷ TracingVerbosity → ServerTrace addr → Object

textTransformer ∷ ServerTrace addr → Object → Text

ToObject (WithDomainName DnsTrace)

Instance details

Methods

toObject ∷ TracingVerbosity → WithDomainName DnsTrace → Object

textTransformer ∷ WithDomainName DnsTrace → Object → Text

ToObject (WithDomainName (SubscriptionTrace SockAddr))

Instance details

Methods

toObject ∷ TracingVerbosity → WithDomainName (SubscriptionTrace SockAddr) → Object

textTransformer ∷ WithDomainName (SubscriptionTrace SockAddr) → Object → Text

ToObject (WithIPList (SubscriptionTrace SockAddr))

Instance details

Methods

toObject ∷ TracingVerbosity → WithIPList (SubscriptionTrace SockAddr) → Object

textTransformer ∷ WithIPList (SubscriptionTrace SockAddr) → Object → Text

(ConvertTxId blk, RunNode blk, HasTxs blk) ⇒ ToObject (AnyMessageAndAgency (BlockFetch blk (Point blk)))
instances of ToObject

NOTE: this list is sorted by the unqualified name of the outermost type.

Instance details

Methods

toObject ∷ TracingVerbosity → AnyMessageAndAgency (BlockFetch blk (Point blk)) → Object

textTransformer ∷ AnyMessageAndAgency (BlockFetch blk (Point blk)) → Object → Text

ToObject (AnyMessageAndAgency (ChainSync blk pt tip))

Instance details

Methods

toObject ∷ TracingVerbosity → AnyMessageAndAgency (ChainSync blk pt tip) → Object

textTransformer ∷ AnyMessageAndAgency (ChainSync blk pt tip) → Object → Text

(∀ result. Show (query result)) ⇒ ToObject (AnyMessageAndAgency (LocalStateQuery blk pt query))

Instance details

Methods

toObject ∷ TracingVerbosity → AnyMessageAndAgency (LocalStateQuery blk pt query) → Object

textTransformer ∷ AnyMessageAndAgency (LocalStateQuery blk pt query) → Object → Text

ToObject (AnyMessageAndAgency (LocalTxMonitor txid tx slotno))

Instance details

Methods

toObject ∷ TracingVerbosity → AnyMessageAndAgency (LocalTxMonitor txid tx slotno) → Object

textTransformer ∷ AnyMessageAndAgency (LocalTxMonitor txid tx slotno) → Object → Text

ToObject (AnyMessageAndAgency (LocalTxSubmission tx err))

Instance details

Methods

toObject ∷ TracingVerbosity → AnyMessageAndAgency (LocalTxSubmission tx err) → Object

textTransformer ∷ AnyMessageAndAgency (LocalTxSubmission tx err) → Object → Text

(Show txid, Show tx) ⇒ ToObject (AnyMessageAndAgency (TxSubmission2 txid tx))

Instance details

Methods

toObject ∷ TracingVerbosity → AnyMessageAndAgency (TxSubmission2 txid tx) → Object

textTransformer ∷ AnyMessageAndAgency (TxSubmission2 txid tx) → Object → Text

ToObject peer ⇒ ToObject [TraceLabelPeer peer (FetchDecision [Point header])]

Instance details

Methods

toObject ∷ TracingVerbosity → [TraceLabelPeer peer (FetchDecision [Point header])] → Object

textTransformer ∷ [TraceLabelPeer peer (FetchDecision [Point header])] → Object → Text

HasPrivacyAnnotation a ⇒ HasPrivacyAnnotation (TraceLabelPeer peer a)

Instance details

Methods

getPrivacyAnnotation ∷ TraceLabelPeer peer a → PrivacyAnnotation

HasPrivacyAnnotation (WithMuxBearer peer MuxTrace)

Instance details

Methods

getPrivacyAnnotation ∷ WithMuxBearer peer MuxTrace → PrivacyAnnotation

HasPrivacyAnnotation (DiffusionTracer ntnAddr ntcAddr)

Instance details

Methods

getPrivacyAnnotation ∷ DiffusionTracer ntnAddr ntcAddr → PrivacyAnnotation

HasPrivacyAnnotation (HandshakeTr RemoteAddress NodeToNodeVersion)

Instance details

Methods

getPrivacyAnnotation ∷ HandshakeTr RemoteAddress NodeToNodeVersion → PrivacyAnnotation

HasPrivacyAnnotation (PeerSelectionActionsTrace SockAddr lAddr)

Instance details

Methods

getPrivacyAnnotation ∷ PeerSelectionActionsTrace SockAddr lAddr → PrivacyAnnotation

HasPrivacyAnnotation (TraceLocalRootPeers RemoteAddress exception)

Instance details

Methods

getPrivacyAnnotation ∷ TraceLocalRootPeers RemoteAddress exception → PrivacyAnnotation

HasPrivacyAnnotation (TraceTxSubmissionInbound txid tx)

Instance details

Methods

getPrivacyAnnotation ∷ TraceTxSubmissionInbound txid tx → PrivacyAnnotation

HasPrivacyAnnotation (TraceTxSubmissionOutbound txid tx)

Instance details

Methods

getPrivacyAnnotation ∷ TraceTxSubmissionOutbound txid tx → PrivacyAnnotation

HasPrivacyAnnotation (ConnectionManagerTrace addr connTrace)

Instance details

Methods

getPrivacyAnnotation ∷ ConnectionManagerTrace addr connTrace → PrivacyAnnotation

HasPrivacyAnnotation (WithAddr addr ErrorPolicyTrace)

Instance details

Methods

getPrivacyAnnotation ∷ WithAddr addr ErrorPolicyTrace → PrivacyAnnotation

HasSeverityAnnotation a ⇒ HasSeverityAnnotation (TraceLabelPeer peer a)

Instance details

Methods

getSeverityAnnotation ∷ TraceLabelPeer peer a → Severity

HasSeverityAnnotation (WithMuxBearer peer MuxTrace)

Instance details

Methods

getSeverityAnnotation ∷ WithMuxBearer peer MuxTrace → Severity

HasSeverityAnnotation (DiffusionTracer ntnAddr ntcAddr)

Instance details

Methods

getSeverityAnnotation ∷ DiffusionTracer ntnAddr ntcAddr → Severity

HasSeverityAnnotation (HandshakeTr RemoteAddress NodeToNodeVersion)

Instance details

Methods

getSeverityAnnotation ∷ HandshakeTr RemoteAddress NodeToNodeVersion → Severity

HasSeverityAnnotation (PeerSelectionActionsTrace SockAddr lAddr)

Instance details

Methods

getSeverityAnnotation ∷ PeerSelectionActionsTrace SockAddr lAddr → Severity

HasSeverityAnnotation (TraceLocalRootPeers RemoteAddress exception)

Instance details

Methods

getSeverityAnnotation ∷ TraceLocalRootPeers RemoteAddress exception → Severity

HasSeverityAnnotation (TraceTxSubmissionInbound txid tx)

Instance details

Methods

getSeverityAnnotation ∷ TraceTxSubmissionInbound txid tx → Severity

HasSeverityAnnotation (TraceTxSubmissionOutbound txid tx)

Instance details

Methods

getSeverityAnnotation ∷ TraceTxSubmissionOutbound txid tx → Severity

HasSeverityAnnotation (ConnectionManagerTrace addr (ConnectionHandlerTrace versionNumber agreedOptions))

Instance details

Methods

getSeverityAnnotation ∷ ConnectionManagerTrace addr (ConnectionHandlerTrace versionNumber agreedOptions) → Severity

HasSeverityAnnotation (WithAddr addr ErrorPolicyTrace)

Instance details

Methods

getSeverityAnnotation ∷ WithAddr addr ErrorPolicyTrace → Severity

(Show header, StandardHash header, Show peer) ⇒ HasTextFormatter (TraceLabelPeer peer (TraceFetchClientState header))

Instance details

Methods

formatText ∷ TraceLabelPeer peer (TraceFetchClientState header) → Object → Text

(Show peer, StandardHash blk, Show (Header blk)) ⇒ HasTextFormatter (TraceLabelPeer peer (TraceSendRecv (ChainSync (Header blk) (Point blk) (Tip blk))))

Instance details

Methods

formatText ∷ TraceLabelPeer peer (TraceSendRecv (ChainSync (Header blk) (Point blk) (Tip blk))) → Object → Text

Show peer ⇒ HasTextFormatter (WithMuxBearer peer MuxTrace)

Instance details

Methods

formatText ∷ WithMuxBearer peer MuxTrace → Object → Text

HasTextFormatter (DiffusionTracer RemoteAddress LocalAddress)

Instance details

Methods

formatText ∷ DiffusionTracer RemoteAddress LocalAddress → Object → Text

HasTextFormatter (HandshakeTr RemoteAddress NodeToNodeVersion)

Instance details

Methods

formatText ∷ HandshakeTr RemoteAddress NodeToNodeVersion → Object → Text

Show lAddr ⇒ HasTextFormatter (PeerSelectionActionsTrace SockAddr lAddr)

Instance details

Methods

formatText ∷ PeerSelectionActionsTrace SockAddr lAddr → Object → Text

Show exception ⇒ HasTextFormatter (TraceLocalRootPeers RemoteAddress exception)

Instance details

Methods

formatText ∷ TraceLocalRootPeers RemoteAddress exception → Object → Text

HasTextFormatter (TraceTxSubmissionInbound txid tx)

Instance details

Methods

formatText ∷ TraceTxSubmissionInbound txid tx → Object → Text

(Show tx, Show txid) ⇒ HasTextFormatter (TraceTxSubmissionOutbound txid tx)

Instance details

Methods

formatText ∷ TraceTxSubmissionOutbound txid tx → Object → Text

(Show addr, Show versionNumber, Show agreedOptions) ⇒ HasTextFormatter (ConnectionManagerTrace addr (ConnectionHandlerTrace versionNumber agreedOptions))

Instance details

Methods

formatText ∷ ConnectionManagerTrace addr (ConnectionHandlerTrace versionNumber agreedOptions) → Object → Text

Show addr ⇒ HasTextFormatter (WithAddr addr ErrorPolicyTrace)

Instance details

Methods

formatText ∷ WithAddr addr ErrorPolicyTrace → Object → Text

(ToObject peer, ToObject a) ⇒ ToObject (TraceLabelPeer peer a)

Instance details

Methods

toObject ∷ TracingVerbosity → TraceLabelPeer peer a → Object

textTransformer ∷ TraceLabelPeer peer a → Object → Text

ToObject peer ⇒ ToObject (WithMuxBearer peer MuxTrace)

Instance details

Methods

toObject ∷ TracingVerbosity → WithMuxBearer peer MuxTrace → Object

textTransformer ∷ WithMuxBearer peer MuxTrace → Object → Text

(Show ntnAddr, Show ntcAddr) ⇒ ToObject (DiffusionTracer ntnAddr ntcAddr)

Instance details

Methods

toObject ∷ TracingVerbosity → DiffusionTracer ntnAddr ntcAddr → Object

textTransformer ∷ DiffusionTracer ntnAddr ntcAddr → Object → Text

ToObject (HandshakeTr RemoteAddress NodeToNodeVersion)

Instance details

Methods

toObject ∷ TracingVerbosity → HandshakeTr RemoteAddress NodeToNodeVersion → Object

textTransformer ∷ HandshakeTr RemoteAddress NodeToNodeVersion → Object → Text

Show lAddr ⇒ ToObject (PeerSelectionActionsTrace SockAddr lAddr)

Instance details

Methods

toObject ∷ TracingVerbosity → PeerSelectionActionsTrace SockAddr lAddr → Object

textTransformer ∷ PeerSelectionActionsTrace SockAddr lAddr → Object → Text

Show exception ⇒ ToObject (TraceLocalRootPeers RemoteAddress exception)

Instance details

Methods

toObject ∷ TracingVerbosity → TraceLocalRootPeers RemoteAddress exception → Object

textTransformer ∷ TraceLocalRootPeers RemoteAddress exception → Object → Text

ToObject (TraceTxSubmissionInbound txid tx)

Instance details

Methods

toObject ∷ TracingVerbosity → TraceTxSubmissionInbound txid tx → Object

textTransformer ∷ TraceTxSubmissionInbound txid tx → Object → Text

(Show txid, Show tx) ⇒ ToObject (TraceTxSubmissionOutbound txid tx)

Instance details

Methods

toObject ∷ TracingVerbosity → TraceTxSubmissionOutbound txid tx → Object

textTransformer ∷ TraceTxSubmissionOutbound txid tx → Object → Text

(Show versionNumber, ToJSON versionNumber, ToJSON agreedOptions) ⇒ ToObject (ConnectionHandlerTrace versionNumber agreedOptions)

Instance details

Methods

toObject ∷ TracingVerbosity → ConnectionHandlerTrace versionNumber agreedOptions → Object

textTransformer ∷ ConnectionHandlerTrace versionNumber agreedOptions → Object → Text

(Show addr, Show versionNumber, Show agreedOptions, ToObject addr, ToJSON addr, ToJSON versionNumber, ToJSON agreedOptions) ⇒ ToObject (ConnectionManagerTrace addr (ConnectionHandlerTrace versionNumber agreedOptions))

Instance details

Methods

toObject ∷ TracingVerbosity → ConnectionManagerTrace addr (ConnectionHandlerTrace versionNumber agreedOptions) → Object

textTransformer ∷ ConnectionManagerTrace addr (ConnectionHandlerTrace versionNumber agreedOptions) → Object → Text

Show addr ⇒ ToObject (WithAddr addr ErrorPolicyTrace)

Instance details

Methods

toObject ∷ TracingVerbosity → WithAddr addr ErrorPolicyTrace → Object

textTransformer ∷ WithAddr addr ErrorPolicyTrace → Object → Text

(Show (ClientHasAgency st), Show (ServerHasAgency st)) ⇒ ToJSON (PeerHasAgency pr st)

Instance details

Methods

toJSON ∷ PeerHasAgency pr st → Value

toEncoding ∷ PeerHasAgency pr st → Encoding

toJSONList ∷ [PeerHasAgency pr st] → Value

toEncodingList ∷ [PeerHasAgency pr st] → Encoding

HasPrivacyAnnotation (HandshakeTr LocalAddress NodeToClientVersion)

Instance details

Methods

getPrivacyAnnotation ∷ HandshakeTr LocalAddress NodeToClientVersion → PrivacyAnnotation

HasSeverityAnnotation (HandshakeTr LocalAddress NodeToClientVersion)

Instance details

Methods

getSeverityAnnotation ∷ HandshakeTr LocalAddress NodeToClientVersion → Severity

HasTextFormatter (HandshakeTr LocalAddress NodeToClientVersion)

Instance details

Methods

formatText ∷ HandshakeTr LocalAddress NodeToClientVersion → Object → Text

ToObject (HandshakeTr LocalAddress NodeToClientVersion)

Instance details

Methods

toObject ∷ TracingVerbosity → HandshakeTr LocalAddress NodeToClientVersion → Object

textTransformer ∷ HandshakeTr LocalAddress NodeToClientVersion → Object → Text