Cardano.Tracing.OrphanInstances.Network

Documentation

newtype Verbose a

Constructors

Verbose a

newtype FetchDecisionToJSON point

Constructors

FetchDecisionToJSON (FetchDecision [point])

Orphan instances

ToJSON Time

Instance details

Methods

toJSON :: Time -> Value

toEncoding :: Time -> Encoding

toJSONList :: [Time] -> Value

toEncodingList :: [Time] -> Encoding

omitField :: Time -> Bool

ToJSON IP

Instance details

Methods

toJSON :: IP -> Value

toEncoding :: IP -> Encoding

toJSONList :: [IP] -> Value

toEncodingList :: [IP] -> Encoding

omitField :: IP -> Bool

HasPrivacyAnnotation TraceLedgerPeers

Instance details

Methods

getPrivacyAnnotation :: TraceLedgerPeers -> PrivacyAnnotation

HasPrivacyAnnotation DNSTrace

Instance details

Methods

getPrivacyAnnotation :: DNSTrace -> PrivacyAnnotation

HasPrivacyAnnotation TracePublicRootPeers

Instance details

Methods

getPrivacyAnnotation :: TracePublicRootPeers -> PrivacyAnnotation

HasPrivacyAnnotation CardanoDebugPeerSelection

Instance details

Methods

getPrivacyAnnotation :: CardanoDebugPeerSelection -> PrivacyAnnotation

HasPrivacyAnnotation CardanoTraceLocalRootPeers

Instance details

Methods

getPrivacyAnnotation :: CardanoTraceLocalRootPeers -> PrivacyAnnotation

HasPrivacyAnnotation CardanoTracePeerSelection

Instance details

Methods

getPrivacyAnnotation :: CardanoTracePeerSelection -> PrivacyAnnotation

HasPrivacyAnnotation TraceChurnMode

Instance details

Methods

getPrivacyAnnotation :: TraceChurnMode -> PrivacyAnnotation

HasPrivacyAnnotation AcceptConnectionsPolicyTrace

Instance details

Methods

getPrivacyAnnotation :: AcceptConnectionsPolicyTrace -> PrivacyAnnotation

HasSeverityAnnotation TraceLedgerPeers

Instance details

Methods

getSeverityAnnotation :: TraceLedgerPeers -> Severity

HasSeverityAnnotation DNSTrace

Instance details

Methods

getSeverityAnnotation :: DNSTrace -> Severity

HasSeverityAnnotation TracePublicRootPeers

Instance details

Methods

getSeverityAnnotation :: TracePublicRootPeers -> Severity

HasSeverityAnnotation CardanoDebugPeerSelection

Instance details

Methods

getSeverityAnnotation :: CardanoDebugPeerSelection -> Severity

HasSeverityAnnotation CardanoTraceLocalRootPeers

Instance details

Methods

getSeverityAnnotation :: CardanoTraceLocalRootPeers -> Severity

HasSeverityAnnotation CardanoTracePeerSelection

Instance details

Methods

getSeverityAnnotation :: CardanoTracePeerSelection -> Severity

HasSeverityAnnotation TraceChurnMode

Instance details

Methods

getSeverityAnnotation :: TraceChurnMode -> Severity

HasSeverityAnnotation AcceptConnectionsPolicyTrace

Instance details

Methods

getSeverityAnnotation :: AcceptConnectionsPolicyTrace -> Severity

HasTextFormatter TraceLedgerPeers

Instance details

Methods

formatText :: TraceLedgerPeers -> Object -> Text

HasTextFormatter DNSTrace

Instance details

Methods

formatText :: DNSTrace -> Object -> Text

HasTextFormatter TracePublicRootPeers

Instance details

Methods

formatText :: TracePublicRootPeers -> Object -> Text

HasTextFormatter CardanoDebugPeerSelection

Instance details

Methods

formatText :: CardanoDebugPeerSelection -> Object -> Text

HasTextFormatter CardanoTraceLocalRootPeers

Instance details

Methods

formatText :: CardanoTraceLocalRootPeers -> Object -> Text

HasTextFormatter CardanoTracePeerSelection

Instance details

Methods

formatText :: CardanoTracePeerSelection -> Object -> Text

HasTextFormatter TraceChurnMode

Instance details

Methods

formatText :: TraceChurnMode -> Object -> Text

HasTextFormatter AcceptConnectionsPolicyTrace

Instance details

Methods

formatText :: AcceptConnectionsPolicyTrace -> Object -> Text

ToObject SlotNo

Instance details

Methods

toObject :: TracingVerbosity -> SlotNo -> Object

textTransformer :: SlotNo -> Object -> Text

ToObject LocalConnectionId

Instance details

Methods

toObject :: TracingVerbosity -> LocalConnectionId -> Object

textTransformer :: LocalConnectionId -> Object -> Text

ToObject RemoteAddress

Instance details

Methods

toObject :: TracingVerbosity -> RemoteAddress -> Object

textTransformer :: RemoteAddress -> Object -> Text

ToObject RemoteConnectionId

Instance details

Methods

toObject :: TracingVerbosity -> RemoteConnectionId -> Object

textTransformer :: RemoteConnectionId -> Object -> Text

ToObject TraceLedgerPeers

Instance details

Methods

toObject :: TracingVerbosity -> TraceLedgerPeers -> Object

textTransformer :: TraceLedgerPeers -> Object -> Text

ToObject DNSTrace

Instance details

Methods

toObject :: TracingVerbosity -> DNSTrace -> Object

textTransformer :: DNSTrace -> Object -> Text

ToObject TracePublicRootPeers

Instance details

Methods

toObject :: TracingVerbosity -> TracePublicRootPeers -> Object

textTransformer :: TracePublicRootPeers -> Object -> Text

ToObject CardanoDebugPeerSelection

Instance details

Methods

toObject :: TracingVerbosity -> CardanoDebugPeerSelection -> Object

textTransformer :: CardanoDebugPeerSelection -> Object -> Text

ToObject CardanoPeerSelectionCounters

Instance details

Methods

toObject :: TracingVerbosity -> CardanoPeerSelectionCounters -> Object

textTransformer :: CardanoPeerSelectionCounters -> Object -> Text

ToObject CardanoTraceLocalRootPeers

Instance details

Methods

toObject :: TracingVerbosity -> CardanoTraceLocalRootPeers -> Object

textTransformer :: CardanoTraceLocalRootPeers -> Object -> Text

ToObject CardanoTracePeerSelection

Instance details

Methods

toObject :: TracingVerbosity -> CardanoTracePeerSelection -> Object

textTransformer :: CardanoTracePeerSelection -> Object -> Text

ToObject TraceChurnMode

Instance details

Methods

toObject :: TracingVerbosity -> TraceChurnMode -> Object

textTransformer :: TraceChurnMode -> Object -> Text

ToObject ConnStateId

Instance details

Methods

toObject :: TracingVerbosity -> ConnStateId -> Object

textTransformer :: ConnStateId -> Object -> Text

ToObject AcceptConnectionsPolicyTrace

Instance details

Methods

toObject :: TracingVerbosity -> AcceptConnectionsPolicyTrace -> Object

textTransformer :: AcceptConnectionsPolicyTrace -> Object -> Text

ToObject LocalAddress

Instance details

Methods

toObject :: TracingVerbosity -> LocalAddress -> Object

textTransformer :: LocalAddress -> Object -> Text

Transformable Text IO TraceLedgerPeers

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO TraceLedgerPeers

Transformable Text IO DNSTrace

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO DNSTrace

Transformable Text IO TracePublicRootPeers

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO TracePublicRootPeers

Transformable Text IO CardanoDebugPeerSelection

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO CardanoDebugPeerSelection

Transformable Text IO CardanoPeerSelectionCounters

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO CardanoPeerSelectionCounters

Transformable Text IO CardanoTraceLocalRootPeers

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO CardanoTraceLocalRootPeers

Transformable Text IO CardanoTracePeerSelection

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO CardanoTracePeerSelection

Transformable Text IO TraceChurnMode

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO TraceChurnMode

Transformable Text IO AcceptConnectionsPolicyTrace

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO AcceptConnectionsPolicyTrace

Show addr => Transformable Text IO (TraceKeepAliveClient addr)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceKeepAliveClient addr)

(Show addr, ToJSON addr, ToObject addr) => Transformable Text IO (AbstractTransitionTrace addr)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (AbstractTransitionTrace addr)

(Show addr, ToJSON addr) => Transformable Text IO (RemoteTransitionTrace addr)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (RemoteTransitionTrace addr)

(ToJSON addr, Show addr, ToJSONKey addr) => Transformable Text IO (Trace addr)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace0 IO Text -> Tracer IO (Trace addr)

(Show addr, ToObject addr, ToJSON addr) => Transformable Text IO (Trace addr)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace0 IO Text -> Tracer IO (Trace addr)

(StandardHash header, Show peer, ToJSON peer, ConvertRawHash header) => Transformable Text IO [TraceLabelPeer peer (FetchDecision [Point header])]

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO [TraceLabelPeer peer (FetchDecision [Point header])]

ToObject localPeer => Transformable Text IO (TraceLabelPeer localPeer (TraceSendRecv (ChainSync (Serialised blk) (Point blk) (Tip blk))))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer localPeer (TraceSendRecv (ChainSync (Serialised blk) (Point blk) (Tip blk))))

ToObject localPeer => Transformable Text IO (TraceLabelPeer localPeer (TraceSendRecv KeepAlive))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer localPeer (TraceSendRecv KeepAlive))

(forall (fp :: QueryFootprint). ShowQuery (BlockQuery blk fp), ToObject localPeer) => Transformable Text IO (TraceLabelPeer localPeer (TraceSendRecv (LocalStateQuery blk (Point blk) (Query blk))))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer localPeer (TraceSendRecv (LocalStateQuery blk (Point blk) (Query blk))))

(applyTxErr ~ ApplyTxErr blk, ToObject localPeer) => Transformable Text IO (TraceLabelPeer localPeer (TraceSendRecv (LocalTxMonitor (GenTxId blk) (GenTx blk) SlotNo)))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer localPeer (TraceSendRecv (LocalTxMonitor (GenTxId blk) (GenTx blk) SlotNo)))

(applyTxErr ~ ApplyTxErr blk, ToObject localPeer) => Transformable Text IO (TraceLabelPeer localPeer (TraceSendRecv (LocalTxSubmission (GenTx blk) applyTxErr)))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer localPeer (TraceSendRecv (LocalTxSubmission (GenTx blk) applyTxErr)))

(ToObject localPeer, ToJSON addr) => Transformable Text IO (TraceLabelPeer localPeer (TraceSendRecv (PeerSharing addr)))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer localPeer (TraceSendRecv (PeerSharing addr)))

(HasPrivacyAnnotation (TraceSendRecv (LocalStateQuery blk (Point blk) (Query blk)) f), HasSeverityAnnotation (TraceSendRecv (LocalStateQuery blk (Point blk) (Query blk)) f), forall (fp :: QueryFootprint). ShowQuery (BlockQuery blk fp), ToObject localPeer) => Transformable Text IO (TraceLabelPeer localPeer (TraceSendRecv (LocalStateQuery blk (Point blk) (Query blk)) f))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer localPeer (TraceSendRecv (LocalStateQuery blk (Point blk) (Query blk)) f))

(HasHeader header, ConvertRawHash header, ToObject peer) => Transformable Text IO (TraceLabelPeer peer (TraceFetchClientState header))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer peer (TraceFetchClientState header))

ToObject peer => Transformable Text IO (TraceLabelPeer peer (TraceTxSubmissionInbound (GenTxId blk) (GenTx blk)))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer peer (TraceTxSubmissionInbound (GenTxId blk) (GenTx blk)))

(ToObject peer, Show (TxId (GenTx blk)), Show (GenTx blk)) => Transformable Text IO (TraceLabelPeer peer (TraceTxSubmissionOutbound (GenTxId blk) (GenTx blk)))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer peer (TraceTxSubmissionOutbound (GenTxId blk) (GenTx blk)))

(ToObject peer, ToObject (AnyMessage (TraceTxSubmissionInbound (GenTxId blk) (GenTx blk)))) => Transformable Text IO (TraceLabelPeer peer (TraceSendRecv (TraceTxSubmissionInbound (GenTxId blk) (GenTx blk))))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer peer (TraceSendRecv (TraceTxSubmissionInbound (GenTxId blk) (GenTx blk))))

(ToObject peer, ConvertTxId blk, RunNode blk, HasTxs blk) => Transformable Text IO (TraceLabelPeer peer (TraceSendRecv (BlockFetch blk (Point blk))))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer peer (TraceSendRecv (BlockFetch blk (Point blk))))

ToObject peer => Transformable Text IO (TraceLabelPeer peer (TraceSendRecv (ChainSync (Header blk) (Point blk) (Tip blk))))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer peer (TraceSendRecv (ChainSync (Header blk) (Point blk) (Tip blk))))

(ToObject peer, Show (TxId (GenTx blk)), Show (GenTx blk)) => Transformable Text IO (TraceLabelPeer peer (TraceSendRecv (TxSubmission2 (GenTxId blk) (GenTx blk))))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer peer (TraceSendRecv (TxSubmission2 (GenTxId blk) (GenTx blk))))

(Show peer, Show tr, HasPrivacyAnnotation (WithBearer peer tr), HasSeverityAnnotation (WithBearer peer tr), ToObject (WithBearer peer tr)) => Transformable Text IO (WithBearer peer tr)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (WithBearer peer tr)

(StandardHash header, Show peer, ToJSON peer, ConvertRawHash header) => Transformable Text IO (TraceDecisionEvent peer header)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceDecisionEvent peer header)

Transformable Text IO (DiffusionTracer RemoteAddress LocalAddress)
instances of Transformable

NOTE: this list is sorted by the unqualified name of the outermost type.

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (DiffusionTracer RemoteAddress LocalAddress)

Show lAddr => Transformable Text IO (PeerSelectionActionsTrace SockAddr lAddr)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (PeerSelectionActionsTrace SockAddr lAddr)

Transformable Text IO (TraceTxSubmissionInbound txid tx)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceTxSubmissionInbound txid tx)

(Show tx, Show txid) => Transformable Text IO (TraceTxSubmissionOutbound txid tx)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceTxSubmissionOutbound txid tx)

(Show addr, Show versionNumber, Show agreedOptions, ToObject addr, ToJSON addr, ToJSON versionNumber, ToJSON agreedOptions) => Transformable Text IO (Trace addr (ConnectionHandlerTrace versionNumber agreedOptions))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace0 IO Text -> Tracer IO (Trace addr (ConnectionHandlerTrace versionNumber agreedOptions))

HasPrivacyAnnotation (TraceFetchClientState header)

Instance details

Methods

getPrivacyAnnotation :: TraceFetchClientState header -> PrivacyAnnotation

HasPrivacyAnnotation (TraceKeepAliveClient remotePeer)

Instance details

Methods

getPrivacyAnnotation :: TraceKeepAliveClient remotePeer -> PrivacyAnnotation

HasPrivacyAnnotation (PeerSelectionCounters extraCounters)

Instance details

Methods

getPrivacyAnnotation :: PeerSelectionCounters extraCounters -> PrivacyAnnotation

HasPrivacyAnnotation (AbstractTransitionTrace addr)

Instance details

Methods

getPrivacyAnnotation :: AbstractTransitionTrace addr -> PrivacyAnnotation

HasPrivacyAnnotation (TraceSendRecv a)

Instance details

Methods

getPrivacyAnnotation :: TraceSendRecv a -> PrivacyAnnotation

HasPrivacyAnnotation (RemoteTransitionTrace addr)

Instance details

Methods

getPrivacyAnnotation :: RemoteTransitionTrace addr -> PrivacyAnnotation

HasPrivacyAnnotation (Trace addr)

Instance details

Methods

getPrivacyAnnotation :: Trace addr -> PrivacyAnnotation

HasPrivacyAnnotation (Trace addr)

Instance details

Methods

getPrivacyAnnotation :: Trace addr -> PrivacyAnnotation

HasPrivacyAnnotation [TraceLabelPeer peer (FetchDecision [Point header])]

Instance details

Methods

getPrivacyAnnotation :: [TraceLabelPeer peer (FetchDecision [Point header])] -> PrivacyAnnotation

HasSeverityAnnotation (TraceFetchClientState header)

Instance details

Methods

getSeverityAnnotation :: TraceFetchClientState header -> Severity

HasSeverityAnnotation (TraceKeepAliveClient remotePeer)

Instance details

Methods

getSeverityAnnotation :: TraceKeepAliveClient remotePeer -> Severity

HasSeverityAnnotation (PeerSelectionCounters extraCounters)

Instance details

Methods

getSeverityAnnotation :: PeerSelectionCounters extraCounters -> Severity

HasSeverityAnnotation (AbstractTransitionTrace addr)

Instance details

Methods

getSeverityAnnotation :: AbstractTransitionTrace addr -> Severity

HasSeverityAnnotation (TraceSendRecv a)

Instance details

Methods

getSeverityAnnotation :: TraceSendRecv a -> Severity

HasSeverityAnnotation (RemoteTransitionTrace addr)

Instance details

Methods

getSeverityAnnotation :: RemoteTransitionTrace addr -> Severity

HasSeverityAnnotation (Trace addr)

Instance details

Methods

getSeverityAnnotation :: Trace addr -> Severity

HasSeverityAnnotation (Trace addr)

Instance details

Methods

getSeverityAnnotation :: Trace addr -> Severity

HasSeverityAnnotation [TraceLabelPeer peer (FetchDecision [Point header])]

Instance details

Methods

getSeverityAnnotation :: [TraceLabelPeer peer (FetchDecision [Point header])] -> Severity

Show addr => HasTextFormatter (TraceKeepAliveClient addr)

Instance details

Methods

formatText :: TraceKeepAliveClient addr -> Object -> Text

Show extraCounters => HasTextFormatter (PeerSelectionCounters extraCounters)

Instance details

Methods

formatText :: PeerSelectionCounters extraCounters -> Object -> Text

Show addr => HasTextFormatter (AbstractTransitionTrace addr)

Instance details

Methods

formatText :: AbstractTransitionTrace addr -> Object -> Text

Show addr => HasTextFormatter (RemoteTransitionTrace addr)

Instance details

Methods

formatText :: RemoteTransitionTrace addr -> Object -> Text

Show addr => HasTextFormatter (Trace addr)

Instance details

Methods

formatText :: Trace addr -> Object -> Text

Show addr => HasTextFormatter (Trace addr)

Instance details

Methods

formatText :: Trace addr -> Object -> Text

(StandardHash header, Show peer) => HasTextFormatter [TraceLabelPeer peer (FetchDecision [Point header])]

Instance details

Methods

formatText :: [TraceLabelPeer peer (FetchDecision [Point header])] -> Object -> Text

(HasHeader header, ConvertRawHash header) => ToObject (TraceFetchClientState header)

Instance details

Methods

toObject :: TracingVerbosity -> TraceFetchClientState header -> Object

textTransformer :: TraceFetchClientState header -> Object -> Text

Show remotePeer => ToObject (TraceKeepAliveClient remotePeer)

Instance details

Methods

toObject :: TracingVerbosity -> TraceKeepAliveClient remotePeer -> Object

textTransformer :: TraceKeepAliveClient remotePeer -> Object -> Text

ConvertRawHash blk => ToObject (Anchor blk)

Instance details

Methods

toObject :: TracingVerbosity -> Anchor blk -> Object

textTransformer :: Anchor blk -> Object -> Text

(ConvertRawHash blk, HasHeader blk) => ToObject (AnchoredFragment blk)

Instance details

Methods

toObject :: TracingVerbosity -> AnchoredFragment blk -> Object

textTransformer :: AnchoredFragment blk -> Object -> Text

(Show addr, ToObject addr, ToJSON addr) => ToObject (AbstractTransitionTrace addr)

Instance details

Methods

toObject :: TracingVerbosity -> AbstractTransitionTrace addr -> Object

textTransformer :: AbstractTransitionTrace addr -> Object -> Text

ToObject (AnyMessage ps) => ToObject (TraceSendRecv ps)

Instance details

Methods

toObject :: TracingVerbosity -> TraceSendRecv ps -> Object

textTransformer :: TraceSendRecv ps -> Object -> Text

ToJSON addr => ToObject (RemoteTransitionTrace addr)

Instance details

Methods

toObject :: TracingVerbosity -> RemoteTransitionTrace addr -> Object

textTransformer :: RemoteTransitionTrace addr -> Object -> Text

(ToJSON addr, Show addr, ToJSONKey addr) => ToObject (Trace addr)

Instance details

Methods

toObject :: TracingVerbosity -> Trace addr -> Object

textTransformer :: Trace addr -> Object -> Text

(Show addr, ToObject addr, ToJSON addr) => ToObject (Trace addr)

Instance details

Methods

toObject :: TracingVerbosity -> Trace addr -> Object

textTransformer :: Trace addr -> Object -> Text

(ConvertTxId blk, RunNode blk, HasTxs blk) => ToObject (AnyMessage (BlockFetch blk (Point blk)))
instances of ToObject

NOTE: this list is sorted by the unqualified name of the outermost type.

Instance details

Methods

toObject :: TracingVerbosity -> AnyMessage (BlockFetch blk (Point blk)) -> Object

textTransformer :: AnyMessage (BlockFetch blk (Point blk)) -> Object -> Text

ToObject (AnyMessage (ChainSync blk pt tip))

Instance details

Methods

toObject :: TracingVerbosity -> AnyMessage (ChainSync blk pt tip) -> Object

textTransformer :: AnyMessage (ChainSync blk pt tip) -> Object -> Text

ToObject (AnyMessage KeepAlive)

Instance details

Methods

toObject :: TracingVerbosity -> AnyMessage KeepAlive -> Object

textTransformer :: AnyMessage KeepAlive -> Object -> Text

(forall result. Show (query result)) => ToObject (AnyMessage (LocalStateQuery blk pt query))

Instance details

Methods

toObject :: TracingVerbosity -> AnyMessage (LocalStateQuery blk pt query) -> Object

textTransformer :: AnyMessage (LocalStateQuery blk pt query) -> Object -> Text

ToObject (AnyMessage (LocalTxMonitor txid tx slotno))

Instance details

Methods

toObject :: TracingVerbosity -> AnyMessage (LocalTxMonitor txid tx slotno) -> Object

textTransformer :: AnyMessage (LocalTxMonitor txid tx slotno) -> Object -> Text

ToObject (AnyMessage (LocalTxSubmission tx err))

Instance details

Methods

toObject :: TracingVerbosity -> AnyMessage (LocalTxSubmission tx err) -> Object

textTransformer :: AnyMessage (LocalTxSubmission tx err) -> Object -> Text

ToJSON peerAddr => ToObject (AnyMessage (PeerSharing peerAddr))

Instance details

Methods

toObject :: TracingVerbosity -> AnyMessage (PeerSharing peerAddr) -> Object

textTransformer :: AnyMessage (PeerSharing peerAddr) -> Object -> Text

(Show txid, Show tx) => ToObject (AnyMessage (TxSubmission2 txid tx))

Instance details

Methods

toObject :: TracingVerbosity -> AnyMessage (TxSubmission2 txid tx) -> Object

textTransformer :: AnyMessage (TxSubmission2 txid tx) -> Object -> Text

(ToJSON peer, ConvertRawHash header) => ToObject [TraceLabelPeer peer (FetchDecision [Point header])]

Instance details

Methods

toObject :: TracingVerbosity -> [TraceLabelPeer peer (FetchDecision [Point header])] -> Object

textTransformer :: [TraceLabelPeer peer (FetchDecision [Point header])] -> Object -> Text

ConvertRawHash header => ToJSON (Point header)

Instance details

Methods

toJSON :: Point header -> Value

toEncoding :: Point header -> Encoding

toJSONList :: [Point header] -> Value

toEncodingList :: [Point header] -> Encoding

omitField :: Point header -> Bool

HasPrivacyAnnotation a => HasPrivacyAnnotation (TraceLabelPeer peer a)

Instance details

Methods

getPrivacyAnnotation :: TraceLabelPeer peer a -> PrivacyAnnotation

HasPrivacyAnnotation (WithBearer peer ChannelTrace)

Instance details

Methods

getPrivacyAnnotation :: WithBearer peer ChannelTrace -> PrivacyAnnotation

HasPrivacyAnnotation (WithBearer peer Trace)

Instance details

Methods

getPrivacyAnnotation :: WithBearer peer Trace -> PrivacyAnnotation

HasPrivacyAnnotation (WithBearer peer BearerTrace)

Instance details

Methods

getPrivacyAnnotation :: WithBearer peer BearerTrace -> PrivacyAnnotation

HasPrivacyAnnotation (TraceDecisionEvent peer header)

Instance details

Methods

getPrivacyAnnotation :: TraceDecisionEvent peer header -> PrivacyAnnotation

HasPrivacyAnnotation (DiffusionTracer ntnAddr ntcAddr)

Instance details

Methods

getPrivacyAnnotation :: DiffusionTracer ntnAddr ntcAddr -> PrivacyAnnotation

HasPrivacyAnnotation (HandshakeTr RemoteAddress NodeToNodeVersion)

Instance details

Methods

getPrivacyAnnotation :: HandshakeTr RemoteAddress NodeToNodeVersion -> PrivacyAnnotation

HasPrivacyAnnotation (PeerSelectionActionsTrace SockAddr lAddr)

Instance details

Methods

getPrivacyAnnotation :: PeerSelectionActionsTrace SockAddr lAddr -> PrivacyAnnotation

HasPrivacyAnnotation (TraceTxSubmissionInbound txid tx)

Instance details

Methods

getPrivacyAnnotation :: TraceTxSubmissionInbound txid tx -> PrivacyAnnotation

HasPrivacyAnnotation (TraceTxSubmissionOutbound txid tx)

Instance details

Methods

getPrivacyAnnotation :: TraceTxSubmissionOutbound txid tx -> PrivacyAnnotation

HasPrivacyAnnotation (Trace addr connTrace)

Instance details

Methods

getPrivacyAnnotation :: Trace addr connTrace -> PrivacyAnnotation

HasPrivacyAnnotation (TraceSendRecv (LocalStateQuery blk (Point blk) (Query blk)) f)

Instance details

Methods

getPrivacyAnnotation :: TraceSendRecv (LocalStateQuery blk (Point blk) (Query blk)) f -> PrivacyAnnotation

HasSeverityAnnotation a => HasSeverityAnnotation (TraceLabelPeer peer a)

Instance details

Methods

getSeverityAnnotation :: TraceLabelPeer peer a -> Severity

HasSeverityAnnotation (WithBearer peer ChannelTrace)

Instance details

Methods

getSeverityAnnotation :: WithBearer peer ChannelTrace -> Severity

HasSeverityAnnotation (WithBearer peer Trace)

Instance details

Methods

getSeverityAnnotation :: WithBearer peer Trace -> Severity

HasSeverityAnnotation (WithBearer peer BearerTrace)

Instance details

Methods

getSeverityAnnotation :: WithBearer peer BearerTrace -> Severity

HasSeverityAnnotation (TraceDecisionEvent peer header)

Instance details

Methods

getSeverityAnnotation :: TraceDecisionEvent peer header -> Severity

HasSeverityAnnotation (DiffusionTracer ntnAddr ntcAddr)

Instance details

Methods

getSeverityAnnotation :: DiffusionTracer ntnAddr ntcAddr -> Severity

HasSeverityAnnotation (HandshakeTr RemoteAddress NodeToNodeVersion)

Instance details

Methods

getSeverityAnnotation :: HandshakeTr RemoteAddress NodeToNodeVersion -> Severity

HasSeverityAnnotation (PeerSelectionActionsTrace SockAddr lAddr)

Instance details

Methods

getSeverityAnnotation :: PeerSelectionActionsTrace SockAddr lAddr -> Severity

HasSeverityAnnotation (TraceTxSubmissionInbound txid tx)

Instance details

Methods

getSeverityAnnotation :: TraceTxSubmissionInbound txid tx -> Severity

HasSeverityAnnotation (TraceTxSubmissionOutbound txid tx)

Instance details

Methods

getSeverityAnnotation :: TraceTxSubmissionOutbound txid tx -> Severity

HasSeverityAnnotation (Trace addr (ConnectionHandlerTrace versionNumber agreedOptions))

Instance details

Methods

getSeverityAnnotation :: Trace addr (ConnectionHandlerTrace versionNumber agreedOptions) -> Severity

HasSeverityAnnotation (TraceSendRecv (LocalStateQuery blk (Point blk) (Query blk)) f)

Instance details

Methods

getSeverityAnnotation :: TraceSendRecv (LocalStateQuery blk (Point blk) (Query blk)) f -> Severity

(Show header, StandardHash header, Show peer) => HasTextFormatter (TraceLabelPeer peer (TraceFetchClientState header))

Instance details

Methods

formatText :: TraceLabelPeer peer (TraceFetchClientState header) -> Object -> Text

(Show peer, StandardHash blk, Show (Header blk)) => HasTextFormatter (TraceLabelPeer peer (TraceSendRecv (ChainSync (Header blk) (Point blk) (Tip blk))))

Instance details

Methods

formatText :: TraceLabelPeer peer (TraceSendRecv (ChainSync (Header blk) (Point blk) (Tip blk))) -> Object -> Text

(Show peer, Show tr) => HasTextFormatter (WithBearer peer tr)

Instance details

Methods

formatText :: WithBearer peer tr -> Object -> Text

(StandardHash header, Show peer) => HasTextFormatter (TraceDecisionEvent peer header)

Instance details

Methods

formatText :: TraceDecisionEvent peer header -> Object -> Text

HasTextFormatter (DiffusionTracer RemoteAddress LocalAddress)

Instance details

Methods

formatText :: DiffusionTracer RemoteAddress LocalAddress -> Object -> Text

Show lAddr => HasTextFormatter (PeerSelectionActionsTrace SockAddr lAddr)

Instance details

Methods

formatText :: PeerSelectionActionsTrace SockAddr lAddr -> Object -> Text

HasTextFormatter (TraceTxSubmissionInbound txid tx)

Instance details

Methods

formatText :: TraceTxSubmissionInbound txid tx -> Object -> Text

(Show tx, Show txid) => HasTextFormatter (TraceTxSubmissionOutbound txid tx)

Instance details

Methods

formatText :: TraceTxSubmissionOutbound txid tx -> Object -> Text

(Show addr, Show versionNumber, Show agreedOptions) => HasTextFormatter (Trace addr (ConnectionHandlerTrace versionNumber agreedOptions))

Instance details

Methods

formatText :: Trace addr (ConnectionHandlerTrace versionNumber agreedOptions) -> Object -> Text

(ToObject peer, ToObject a) => ToObject (TraceLabelPeer peer a)

Instance details

Methods

toObject :: TracingVerbosity -> TraceLabelPeer peer a -> Object

textTransformer :: TraceLabelPeer peer a -> Object -> Text

(Typeable tr, ToObject peer, Show tr) => ToObject (WithBearer peer tr)

Instance details

Methods

toObject :: TracingVerbosity -> WithBearer peer tr -> Object

textTransformer :: WithBearer peer tr -> Object -> Text

(ToJSON peer, ConvertRawHash header) => ToObject (TraceDecisionEvent peer header)

Instance details

Methods

toObject :: TracingVerbosity -> TraceDecisionEvent peer header -> Object

textTransformer :: TraceDecisionEvent peer header -> Object -> Text

(Show ntnAddr, Show ntcAddr) => ToObject (DiffusionTracer ntnAddr ntcAddr)

Instance details

Methods

toObject :: TracingVerbosity -> DiffusionTracer ntnAddr ntcAddr -> Object

textTransformer :: DiffusionTracer ntnAddr ntcAddr -> Object -> Text

Show lAddr => ToObject (PeerSelectionActionsTrace SockAddr lAddr)

Instance details

Methods

toObject :: TracingVerbosity -> PeerSelectionActionsTrace SockAddr lAddr -> Object

textTransformer :: PeerSelectionActionsTrace SockAddr lAddr -> Object -> Text

ToObject (TraceTxSubmissionInbound txid tx)

Instance details

Methods

toObject :: TracingVerbosity -> TraceTxSubmissionInbound txid tx -> Object

textTransformer :: TraceTxSubmissionInbound txid tx -> Object -> Text

(Show txid, Show tx) => ToObject (TraceTxSubmissionOutbound txid tx)

Instance details

Methods

toObject :: TracingVerbosity -> TraceTxSubmissionOutbound txid tx -> Object

textTransformer :: TraceTxSubmissionOutbound txid tx -> Object -> Text

ConvertRawHash blk => ToObject (Point blk)

Instance details

Methods

toObject :: TracingVerbosity -> Point blk -> Object

textTransformer :: Point blk -> Object -> Text

(Show versionNumber, ToJSON versionNumber, ToJSON agreedOptions) => ToObject (ConnectionHandlerTrace versionNumber agreedOptions)

Instance details

Methods

toObject :: TracingVerbosity -> ConnectionHandlerTrace versionNumber agreedOptions -> Object

textTransformer :: ConnectionHandlerTrace versionNumber agreedOptions -> Object -> Text

(Show addr, Show versionNumber, Show agreedOptions, ToObject addr, ToJSON addr, ToJSON versionNumber, ToJSON agreedOptions) => ToObject (Trace addr (ConnectionHandlerTrace versionNumber agreedOptions))

Instance details

Methods

toObject :: TracingVerbosity -> Trace addr (ConnectionHandlerTrace versionNumber agreedOptions) -> Object

textTransformer :: Trace addr (ConnectionHandlerTrace versionNumber agreedOptions) -> Object -> Text

ToObject (AnyMessage ps f) => ToObject (TraceSendRecv ps f)

Instance details

Methods

toObject :: TracingVerbosity -> TraceSendRecv ps f -> Object

textTransformer :: TraceSendRecv ps f -> Object -> Text

(forall result. Show (query result)) => ToObject (AnyMessage (LocalStateQuery blk pt query) f)

Instance details

Methods

toObject :: TracingVerbosity -> AnyMessage (LocalStateQuery blk pt query) f -> Object

textTransformer :: AnyMessage (LocalStateQuery blk pt query) f -> Object -> Text

HasPrivacyAnnotation (HandshakeTr LocalAddress NodeToClientVersion)

Instance details

Methods

getPrivacyAnnotation :: HandshakeTr LocalAddress NodeToClientVersion -> PrivacyAnnotation

HasSeverityAnnotation (HandshakeTr LocalAddress NodeToClientVersion)

Instance details

Methods

getSeverityAnnotation :: HandshakeTr LocalAddress NodeToClientVersion -> Severity