Cardano.Tracing.OrphanInstances.Network

Documentation

newtype Verbose a

Constructors

Verbose a

newtype FetchDecisionToJSON point

Constructors

FetchDecisionToJSON (FetchDecision [point])

Orphan instances

FromJSON AssociationMode

Instance details

Methods

parseJSON :: Value -> Parser AssociationMode

parseJSONList :: Value -> Parser [AssociationMode]

omittedField :: Maybe AssociationMode

FromJSON HotValency

Instance details

Methods

parseJSON :: Value -> Parser HotValency

parseJSONList :: Value -> Parser [HotValency]

omittedField :: Maybe HotValency

FromJSON WarmValency

Instance details

Methods

parseJSON :: Value -> Parser WarmValency

parseJSONList :: Value -> Parser [WarmValency]

omittedField :: Maybe WarmValency

FromJSON NodeToClientVersion

Instance details

Methods

parseJSON :: Value -> Parser NodeToClientVersion

parseJSONList :: Value -> Parser [NodeToClientVersion]

omittedField :: Maybe NodeToClientVersion

FromJSON NodeToNodeVersion

Instance details

Methods

parseJSON :: Value -> Parser NodeToNodeVersion

parseJSONList :: Value -> Parser [NodeToNodeVersion]

omittedField :: Maybe NodeToNodeVersion

FromJSON UseBootstrapPeers

Instance details

Methods

parseJSON :: Value -> Parser UseBootstrapPeers

parseJSONList :: Value -> Parser [UseBootstrapPeers]

omittedField :: Maybe UseBootstrapPeers

FromJSON LedgerStateJudgement

Instance details

Methods

parseJSON :: Value -> Parser LedgerStateJudgement

parseJSONList :: Value -> Parser [LedgerStateJudgement]

omittedField :: Maybe LedgerStateJudgement

FromJSON UseLedgerPeers

Instance details

Methods

parseJSON :: Value -> Parser UseLedgerPeers

parseJSONList :: Value -> Parser [UseLedgerPeers]

omittedField :: Maybe UseLedgerPeers

FromJSON PeerSharing

Instance details

Methods

parseJSON :: Value -> Parser PeerSharing

parseJSONList :: Value -> Parser [PeerSharing]

omittedField :: Maybe PeerSharing

FromJSON PeerTrustable

Instance details

Methods

parseJSON :: Value -> Parser PeerTrustable

parseJSONList :: Value -> Parser [PeerTrustable]

omittedField :: Maybe PeerTrustable

ToJSON IP

Instance details

Methods

toJSON :: IP -> Value

toEncoding :: IP -> Encoding

toJSONList :: [IP] -> Value

toEncodingList :: [IP] -> Encoding

omitField :: IP -> Bool

ToJSON SockAddr

Instance details

Methods

toJSON :: SockAddr -> Value

toEncoding :: SockAddr -> Encoding

toJSONList :: [SockAddr] -> Value

toEncodingList :: [SockAddr] -> Encoding

omitField :: SockAddr -> Bool

ToJSON MiniProtocolNum

Instance details

Methods

toJSON :: MiniProtocolNum -> Value

toEncoding :: MiniProtocolNum -> Encoding

toJSONList :: [MiniProtocolNum] -> Value

toEncodingList :: [MiniProtocolNum] -> Encoding

omitField :: MiniProtocolNum -> Bool

ToJSON PeerGSV

Instance details

Methods

toJSON :: PeerGSV -> Value

toEncoding :: PeerGSV -> Encoding

toJSONList :: [PeerGSV] -> Value

toEncodingList :: [PeerGSV] -> Encoding

omitField :: PeerGSV -> Bool

ToJSON RepromoteDelay

Instance details

Methods

toJSON :: RepromoteDelay -> Value

toEncoding :: RepromoteDelay -> Encoding

toJSONList :: [RepromoteDelay] -> Value

toEncodingList :: [RepromoteDelay] -> Encoding

omitField :: RepromoteDelay -> Bool

ToJSON AssociationMode

Instance details

Methods

toJSON :: AssociationMode -> Value

toEncoding :: AssociationMode -> Encoding

toJSONList :: [AssociationMode] -> Value

toEncodingList :: [AssociationMode] -> Encoding

omitField :: AssociationMode -> Bool

ToJSON PeerSelectionTargets

Instance details

Methods

toJSON :: PeerSelectionTargets -> Value

toEncoding :: PeerSelectionTargets -> Encoding

toJSONList :: [PeerSelectionTargets] -> Value

toEncodingList :: [PeerSelectionTargets] -> Encoding

omitField :: PeerSelectionTargets -> Bool

ToJSON KnownPeerInfo

Instance details

Methods

toJSON :: KnownPeerInfo -> Value

toEncoding :: KnownPeerInfo -> Encoding

toJSONList :: [KnownPeerInfo] -> Value

toEncodingList :: [KnownPeerInfo] -> Encoding

omitField :: KnownPeerInfo -> Bool

ToJSON HotValency

Instance details

Methods

toJSON :: HotValency -> Value

toEncoding :: HotValency -> Encoding

toJSONList :: [HotValency] -> Value

toEncodingList :: [HotValency] -> Encoding

omitField :: HotValency -> Bool

ToJSON LocalRootConfig

Instance details

Methods

toJSON :: LocalRootConfig -> Value

toEncoding :: LocalRootConfig -> Encoding

toJSONList :: [LocalRootConfig] -> Value

toEncodingList :: [LocalRootConfig] -> Encoding

omitField :: LocalRootConfig -> Bool

ToJSON WarmValency

Instance details

Methods

toJSON :: WarmValency -> Value

toEncoding :: WarmValency -> Encoding

toJSONList :: [WarmValency] -> Value

toEncodingList :: [WarmValency] -> Encoding

omitField :: WarmValency -> Bool

ToJSON PeerStatus

Instance details

Methods

toJSON :: PeerStatus -> Value

toEncoding :: PeerStatus -> Encoding

toJSONList :: [PeerStatus] -> Value

toEncodingList :: [PeerStatus] -> Encoding

omitField :: PeerStatus -> Bool

ToJSON NodeToClientVersion

Instance details

Methods

toJSON :: NodeToClientVersion -> Value

toEncoding :: NodeToClientVersion -> Encoding

toJSONList :: [NodeToClientVersion] -> Value

toEncodingList :: [NodeToClientVersion] -> Encoding

omitField :: NodeToClientVersion -> Bool

ToJSON NodeToClientVersionData

Instance details

Methods

toJSON :: NodeToClientVersionData -> Value

toEncoding :: NodeToClientVersionData -> Encoding

toJSONList :: [NodeToClientVersionData] -> Value

toEncodingList :: [NodeToClientVersionData] -> Encoding

omitField :: NodeToClientVersionData -> Bool

ToJSON DiffusionMode

Instance details

Methods

toJSON :: DiffusionMode -> Value

toEncoding :: DiffusionMode -> Encoding

toJSONList :: [DiffusionMode] -> Value

toEncodingList :: [DiffusionMode] -> Encoding

omitField :: DiffusionMode -> Bool

ToJSON NodeToNodeVersion

Instance details

Methods

toJSON :: NodeToNodeVersion -> Value

toEncoding :: NodeToNodeVersion -> Encoding

toJSONList :: [NodeToNodeVersion] -> Value

toEncodingList :: [NodeToNodeVersion] -> Encoding

omitField :: NodeToNodeVersion -> Bool

ToJSON NodeToNodeVersionData

Instance details

Methods

toJSON :: NodeToNodeVersionData -> Value

toEncoding :: NodeToNodeVersionData -> Encoding

toJSONList :: [NodeToNodeVersionData] -> Value

toEncodingList :: [NodeToNodeVersionData] -> Encoding

omitField :: NodeToNodeVersionData -> Bool

ToJSON UseBootstrapPeers

Instance details

Methods

toJSON :: UseBootstrapPeers -> Value

toEncoding :: UseBootstrapPeers -> Encoding

toJSONList :: [UseBootstrapPeers] -> Value

toEncodingList :: [UseBootstrapPeers] -> Encoding

omitField :: UseBootstrapPeers -> Bool

ToJSON LedgerStateJudgement

Instance details

Methods

toJSON :: LedgerStateJudgement -> Value

toEncoding :: LedgerStateJudgement -> Encoding

toJSONList :: [LedgerStateJudgement] -> Value

toEncodingList :: [LedgerStateJudgement] -> Encoding

omitField :: LedgerStateJudgement -> Bool

ToJSON UseLedgerPeers

Instance details

Methods

toJSON :: UseLedgerPeers -> Value

toEncoding :: UseLedgerPeers -> Encoding

toJSONList :: [UseLedgerPeers] -> Value

toEncodingList :: [UseLedgerPeers] -> Encoding

omitField :: UseLedgerPeers -> Bool

ToJSON PeerSharing

Instance details

Methods

toJSON :: PeerSharing -> Value

toEncoding :: PeerSharing -> Encoding

toJSONList :: [PeerSharing] -> Value

toEncodingList :: [PeerSharing] -> Encoding

omitField :: PeerSharing -> Bool

ToJSON PeerTrustable

Instance details

Methods

toJSON :: PeerTrustable -> Value

toEncoding :: PeerTrustable -> Encoding

toJSONList :: [PeerTrustable] -> Value

toEncodingList :: [PeerTrustable] -> Encoding

omitField :: PeerTrustable -> Bool

ToJSON ProtocolLimitFailure

Instance details

Methods

toJSON :: ProtocolLimitFailure -> Value

toEncoding :: ProtocolLimitFailure -> Encoding

toJSONList :: [ProtocolLimitFailure] -> Value

toEncodingList :: [ProtocolLimitFailure] -> Encoding

omitField :: ProtocolLimitFailure -> Bool

ToJSON ConnStateId

Instance details

Methods

toJSON :: ConnStateId -> Value

toEncoding :: ConnStateId -> Encoding

toJSONList :: [ConnStateId] -> Value

toEncodingList :: [ConnStateId] -> Encoding

omitField :: ConnStateId -> Bool

ToJSON AbstractState

Instance details

Methods

toJSON :: AbstractState -> Value

toEncoding :: AbstractState -> Encoding

toJSONList :: [AbstractState] -> Value

toEncodingList :: [AbstractState] -> Encoding

omitField :: AbstractState -> Bool

ToJSON ConnectionManagerCounters

Instance details

Methods

toJSON :: ConnectionManagerCounters -> Value

toEncoding :: ConnectionManagerCounters -> Encoding

toJSONList :: [ConnectionManagerCounters] -> Value

toEncodingList :: [ConnectionManagerCounters] -> Encoding

omitField :: ConnectionManagerCounters -> Bool

ToJSON RemoteSt

Instance details

Methods

toJSON :: RemoteSt -> Value

toEncoding :: RemoteSt -> Encoding

toJSONList :: [RemoteSt] -> Value

toEncodingList :: [RemoteSt] -> Encoding

omitField :: RemoteSt -> Bool

ToJSON LocalAddress

Instance details

Methods

toJSON :: LocalAddress -> Value

toEncoding :: LocalAddress -> Encoding

toJSONList :: [LocalAddress] -> Value

toEncodingList :: [LocalAddress] -> Encoding

omitField :: LocalAddress -> Bool

ToJSON Time

Instance details

Methods

toJSON :: Time -> Value

toEncoding :: Time -> Encoding

toJSONList :: [Time] -> Value

toEncodingList :: [Time] -> Encoding

omitField :: Time -> Bool

ToJSONKey SockAddr

Instance details

Methods

toJSONKey :: ToJSONKeyFunction SockAddr

toJSONKeyList :: ToJSONKeyFunction [SockAddr]

ToJSONKey DomainAccessPoint

Instance details

Methods

toJSONKey :: ToJSONKeyFunction DomainAccessPoint

toJSONKeyList :: ToJSONKeyFunction [DomainAccessPoint]

ToJSONKey RelayAccessPoint

Instance details

Methods

toJSONKey :: ToJSONKeyFunction RelayAccessPoint

toJSONKeyList :: ToJSONKeyFunction [RelayAccessPoint]

ToJSONKey LocalAddress

Instance details

Methods

toJSONKey :: ToJSONKeyFunction LocalAddress

toJSONKeyList :: ToJSONKeyFunction [LocalAddress]

HasPrivacyAnnotation PeerSelectionCounters

Instance details

Methods

getPrivacyAnnotation :: PeerSelectionCounters -> PrivacyAnnotation

HasPrivacyAnnotation TraceLedgerPeers

Instance details

Methods

getPrivacyAnnotation :: TraceLedgerPeers -> PrivacyAnnotation

HasPrivacyAnnotation TracePublicRootPeers

Instance details

Methods

getPrivacyAnnotation :: TracePublicRootPeers -> PrivacyAnnotation

HasPrivacyAnnotation AcceptConnectionsPolicyTrace

Instance details

Methods

getPrivacyAnnotation :: AcceptConnectionsPolicyTrace -> PrivacyAnnotation

HasSeverityAnnotation PeerSelectionCounters

Instance details

Methods

getSeverityAnnotation :: PeerSelectionCounters -> Severity

HasSeverityAnnotation TraceLedgerPeers

Instance details

Methods

getSeverityAnnotation :: TraceLedgerPeers -> Severity

HasSeverityAnnotation TracePublicRootPeers

Instance details

Methods

getSeverityAnnotation :: TracePublicRootPeers -> Severity

HasSeverityAnnotation AcceptConnectionsPolicyTrace

Instance details

Methods

getSeverityAnnotation :: AcceptConnectionsPolicyTrace -> Severity

HasTextFormatter PeerSelectionCounters

Instance details

Methods

formatText :: PeerSelectionCounters -> Object -> Text

HasTextFormatter TraceLedgerPeers

Instance details

Methods

formatText :: TraceLedgerPeers -> Object -> Text

HasTextFormatter TracePublicRootPeers

Instance details

Methods

formatText :: TracePublicRootPeers -> Object -> Text

HasTextFormatter AcceptConnectionsPolicyTrace

Instance details

Methods

formatText :: AcceptConnectionsPolicyTrace -> Object -> Text

ToObject SlotNo

Instance details

Methods

toObject :: TracingVerbosity -> SlotNo -> Object

textTransformer :: SlotNo -> Object -> Text

ToObject LocalConnectionId

Instance details

Methods

toObject :: TracingVerbosity -> LocalConnectionId -> Object

textTransformer :: LocalConnectionId -> Object -> Text

ToObject RemoteAddress

Instance details

Methods

toObject :: TracingVerbosity -> RemoteAddress -> Object

textTransformer :: RemoteAddress -> Object -> Text

ToObject RemoteConnectionId

Instance details

Methods

toObject :: TracingVerbosity -> RemoteConnectionId -> Object

textTransformer :: RemoteConnectionId -> Object -> Text

ToObject PeerSelectionCounters

Instance details

Methods

toObject :: TracingVerbosity -> PeerSelectionCounters -> Object

textTransformer :: PeerSelectionCounters -> Object -> Text

ToObject TraceLedgerPeers

Instance details

Methods

toObject :: TracingVerbosity -> TraceLedgerPeers -> Object

textTransformer :: TraceLedgerPeers -> Object -> Text

ToObject TracePublicRootPeers

Instance details

Methods

toObject :: TracingVerbosity -> TracePublicRootPeers -> Object

textTransformer :: TracePublicRootPeers -> Object -> Text

ToObject ConnStateId

Instance details

Methods

toObject :: TracingVerbosity -> ConnStateId -> Object

textTransformer :: ConnStateId -> Object -> Text

ToObject AcceptConnectionsPolicyTrace

Instance details

Methods

toObject :: TracingVerbosity -> AcceptConnectionsPolicyTrace -> Object

textTransformer :: AcceptConnectionsPolicyTrace -> Object -> Text

ToObject LocalAddress

Instance details

Methods

toObject :: TracingVerbosity -> LocalAddress -> Object

textTransformer :: LocalAddress -> Object -> Text

Transformable Text IO PeerSelectionCounters

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO PeerSelectionCounters

Transformable Text IO TraceLedgerPeers

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO TraceLedgerPeers

Transformable Text IO TracePublicRootPeers

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO TracePublicRootPeers

Transformable Text IO AcceptConnectionsPolicyTrace

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO AcceptConnectionsPolicyTrace

Transformable Text IO (Identity (SubscriptionTrace LocalAddress))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (Identity (SubscriptionTrace LocalAddress))

Show addr => Transformable Text IO (TraceKeepAliveClient addr)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceKeepAliveClient addr)

Transformable Text IO (DebugPeerSelection SockAddr)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (DebugPeerSelection SockAddr)

Transformable Text IO (TracePeerSelection SockAddr)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TracePeerSelection SockAddr)

(Show addr, ToJSON addr, ToObject addr) => Transformable Text IO (AbstractTransitionTrace addr)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (AbstractTransitionTrace addr)

(Show addr, ToJSON addr) => Transformable Text IO (RemoteTransitionTrace addr)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (RemoteTransitionTrace addr)

(ToJSON addr, Show addr) => Transformable Text IO (Trace addr)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace0 IO Text -> Tracer IO (Trace addr)

(Show addr, ToObject addr, ToJSON addr) => Transformable Text IO (Trace addr)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace0 IO Text -> Tracer IO (Trace addr)

Transformable Text IO (WithDomainName DnsTrace)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (WithDomainName DnsTrace)

Transformable Text IO (WithDomainName (SubscriptionTrace SockAddr))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (WithDomainName (SubscriptionTrace SockAddr))

Transformable Text IO (WithIPList (SubscriptionTrace SockAddr))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (WithIPList (SubscriptionTrace SockAddr))

(StandardHash header, Show peer, ToJSON peer, ConvertRawHash header) => Transformable Text IO [TraceLabelPeer peer (FetchDecision [Point header])]

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO [TraceLabelPeer peer (FetchDecision [Point header])]

ToObject localPeer => Transformable Text IO (TraceLabelPeer localPeer (TraceSendRecv (ChainSync (Serialised blk) (Point blk) (Tip blk))))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer localPeer (TraceSendRecv (ChainSync (Serialised blk) (Point blk) (Tip blk))))

ToObject localPeer => Transformable Text IO (TraceLabelPeer localPeer (TraceSendRecv KeepAlive))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer localPeer (TraceSendRecv KeepAlive))

(ShowQuery (BlockQuery blk), ToObject localPeer) => Transformable Text IO (TraceLabelPeer localPeer (TraceSendRecv (LocalStateQuery blk (Point blk) (Query blk))))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer localPeer (TraceSendRecv (LocalStateQuery blk (Point blk) (Query blk))))

(applyTxErr ~ ApplyTxErr blk, ToObject localPeer) => Transformable Text IO (TraceLabelPeer localPeer (TraceSendRecv (LocalTxMonitor (GenTxId blk) (GenTx blk) SlotNo)))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer localPeer (TraceSendRecv (LocalTxMonitor (GenTxId blk) (GenTx blk) SlotNo)))

(applyTxErr ~ ApplyTxErr blk, ToObject localPeer) => Transformable Text IO (TraceLabelPeer localPeer (TraceSendRecv (LocalTxSubmission (GenTx blk) applyTxErr)))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer localPeer (TraceSendRecv (LocalTxSubmission (GenTx blk) applyTxErr)))

(HasPrivacyAnnotation (TraceSendRecv (LocalStateQuery blk (Point blk) (Query blk)) f), HasSeverityAnnotation (TraceSendRecv (LocalStateQuery blk (Point blk) (Query blk)) f), ShowQuery (BlockQuery blk), ToObject localPeer) => Transformable Text IO (TraceLabelPeer localPeer (TraceSendRecv (LocalStateQuery blk (Point blk) (Query blk)) f))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer localPeer (TraceSendRecv (LocalStateQuery blk (Point blk) (Query blk)) f))

(HasHeader header, ConvertRawHash header, ToObject peer) => Transformable Text IO (TraceLabelPeer peer (TraceFetchClientState header))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer peer (TraceFetchClientState header))

ToObject peer => Transformable Text IO (TraceLabelPeer peer (TraceTxSubmissionInbound (GenTxId blk) (GenTx blk)))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer peer (TraceTxSubmissionInbound (GenTxId blk) (GenTx blk)))

(ToObject peer, Show (TxId (GenTx blk)), Show (GenTx blk)) => Transformable Text IO (TraceLabelPeer peer (TraceTxSubmissionOutbound (GenTxId blk) (GenTx blk)))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer peer (TraceTxSubmissionOutbound (GenTxId blk) (GenTx blk)))

(ToObject peer, ToObject (AnyMessage (TraceTxSubmissionInbound (GenTxId blk) (GenTx blk)))) => Transformable Text IO (TraceLabelPeer peer (TraceSendRecv (TraceTxSubmissionInbound (GenTxId blk) (GenTx blk))))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer peer (TraceSendRecv (TraceTxSubmissionInbound (GenTxId blk) (GenTx blk))))

(ToObject peer, ConvertTxId blk, RunNode blk, HasTxs blk) => Transformable Text IO (TraceLabelPeer peer (TraceSendRecv (BlockFetch blk (Point blk))))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer peer (TraceSendRecv (BlockFetch blk (Point blk))))

ToObject peer => Transformable Text IO (TraceLabelPeer peer (TraceSendRecv (ChainSync (Header blk) (Point blk) (Tip blk))))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer peer (TraceSendRecv (ChainSync (Header blk) (Point blk) (Tip blk))))

(ToObject peer, Show (TxId (GenTx blk)), Show (GenTx blk)) => Transformable Text IO (TraceLabelPeer peer (TraceSendRecv (TxSubmission2 (GenTxId blk) (GenTx blk))))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLabelPeer peer (TraceSendRecv (TxSubmission2 (GenTxId blk) (GenTx blk))))

(Show peer, ToObject peer) => Transformable Text IO (WithBearer peer Trace)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace0 IO Text -> Tracer IO (WithBearer peer Trace)

(StandardHash header, Show peer, ToJSON peer, ConvertRawHash header) => Transformable Text IO (TraceDecisionEvent peer header)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceDecisionEvent peer header)

Transformable Text IO (DiffusionTracer RemoteAddress LocalAddress)
instances of Transformable

NOTE: this list is sorted by the unqualified name of the outermost type.

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (DiffusionTracer RemoteAddress LocalAddress)

Transformable Text IO (HandshakeTr RemoteAddress NodeToNodeVersion)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (HandshakeTr RemoteAddress NodeToNodeVersion)

Show lAddr => Transformable Text IO (PeerSelectionActionsTrace SockAddr lAddr)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (PeerSelectionActionsTrace SockAddr lAddr)

Show exception => Transformable Text IO (TraceLocalRootPeers RemoteAddress exception)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceLocalRootPeers RemoteAddress exception)

Transformable Text IO (TraceTxSubmissionInbound txid tx)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceTxSubmissionInbound txid tx)

(Show tx, Show txid) => Transformable Text IO (TraceTxSubmissionOutbound txid tx)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (TraceTxSubmissionOutbound txid tx)

(Show addr, Show versionNumber, Show agreedOptions, ToObject addr, ToJSON addr, ToJSON versionNumber, ToJSON agreedOptions) => Transformable Text IO (Trace addr (ConnectionHandlerTrace versionNumber agreedOptions))

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace0 IO Text -> Tracer IO (Trace addr (ConnectionHandlerTrace versionNumber agreedOptions))

Show addr => Transformable Text IO (WithAddr addr ErrorPolicyTrace)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (WithAddr addr ErrorPolicyTrace)

Transformable Text IO (HandshakeTr LocalAddress NodeToClientVersion)

Instance details

Methods

trTransformer :: TracingVerbosity -> Trace IO Text -> Tracer IO (HandshakeTr LocalAddress NodeToClientVersion)

ToJSON peerAddr => ToJSON (PublicRootPeers peerAddr)

Instance details

Methods

toJSON :: PublicRootPeers peerAddr -> Value

toEncoding :: PublicRootPeers peerAddr -> Encoding

toJSONList :: [PublicRootPeers peerAddr] -> Value

toEncodingList :: [PublicRootPeers peerAddr] -> Encoding

omitField :: PublicRootPeers peerAddr -> Bool

(ToJSONKey peerAddr, ToJSON peerAddr, Ord peerAddr, Show peerAddr) => ToJSON (LocalRootPeers peerAddr)

Instance details

Methods

toJSON :: LocalRootPeers peerAddr -> Value

toEncoding :: LocalRootPeers peerAddr -> Encoding

toJSONList :: [LocalRootPeers peerAddr] -> Value

toEncodingList :: [LocalRootPeers peerAddr] -> Encoding

omitField :: LocalRootPeers peerAddr -> Bool

ToJSON peerAddr => ToJSON (ConnectionId peerAddr)

Instance details

Methods

toJSON :: ConnectionId peerAddr -> Value

toEncoding :: ConnectionId peerAddr -> Encoding

toJSONList :: [ConnectionId peerAddr] -> Value

toEncodingList :: [ConnectionId peerAddr] -> Encoding

omitField :: ConnectionId peerAddr -> Bool

ToJSON addr => ToJSON (LocalAddr addr)

Instance details

Methods

toJSON :: LocalAddr addr -> Value

toEncoding :: LocalAddr addr -> Encoding

toJSONList :: [LocalAddr addr] -> Value

toEncodingList :: [LocalAddr addr] -> Encoding

omitField :: LocalAddr addr -> Bool

ToJSON state => ToJSON (MaybeUnknown state)

Instance details

Methods

toJSON :: MaybeUnknown state -> Value

toEncoding :: MaybeUnknown state -> Encoding

toJSONList :: [MaybeUnknown state] -> Value

toEncodingList :: [MaybeUnknown state] -> Encoding

omitField :: MaybeUnknown state -> Bool

ToJSON addr => ToJSON (OperationResult addr)

Instance details

Methods

toJSON :: OperationResult addr -> Value

toEncoding :: OperationResult addr -> Encoding

toJSONList :: [OperationResult addr] -> Value

toEncodingList :: [OperationResult addr] -> Encoding

omitField :: OperationResult addr -> Bool

Show vNumber => ToJSON (HandshakeException vNumber)

Instance details

Methods

toJSON :: HandshakeException vNumber -> Value

toEncoding :: HandshakeException vNumber -> Encoding

toJSONList :: [HandshakeException vNumber] -> Value

toEncodingList :: [HandshakeException vNumber] -> Encoding

omitField :: HandshakeException vNumber -> Bool

Show vNumber => ToJSON (HandshakeProtocolError vNumber)

Instance details

Methods

toJSON :: HandshakeProtocolError vNumber -> Value

toEncoding :: HandshakeProtocolError vNumber -> Encoding

toJSONList :: [HandshakeProtocolError vNumber] -> Value

toEncodingList :: [HandshakeProtocolError vNumber] -> Encoding

omitField :: HandshakeProtocolError vNumber -> Bool

Show vNumber => ToJSON (RefuseReason vNumber)

Instance details

Methods

toJSON :: RefuseReason vNumber -> Value

toEncoding :: RefuseReason vNumber -> Encoding

toJSONList :: [RefuseReason vNumber] -> Value

toEncodingList :: [RefuseReason vNumber] -> Encoding

omitField :: RefuseReason vNumber -> Bool

ToJSON addr => ToJSON (PeerSharingResult addr)

Instance details

Methods

toJSON :: PeerSharingResult addr -> Value

toEncoding :: PeerSharingResult addr -> Encoding

toJSONList :: [PeerSharingResult addr] -> Value

toEncodingList :: [PeerSharingResult addr] -> Encoding

omitField :: PeerSharingResult addr -> Bool

ToJSON addr => ToJSONKey (ConnectionId addr)

Instance details

Methods

toJSONKey :: ToJSONKeyFunction (ConnectionId addr)

toJSONKeyList :: ToJSONKeyFunction [ConnectionId addr]

HasPrivacyAnnotation (Identity (SubscriptionTrace LocalAddress))

Instance details

Methods

getPrivacyAnnotation :: Identity (SubscriptionTrace LocalAddress) -> PrivacyAnnotation

HasPrivacyAnnotation (TraceFetchClientState header)

Instance details

Methods

getPrivacyAnnotation :: TraceFetchClientState header -> PrivacyAnnotation

HasPrivacyAnnotation (TraceKeepAliveClient remotePeer)

Instance details

Methods

getPrivacyAnnotation :: TraceKeepAliveClient remotePeer -> PrivacyAnnotation

HasPrivacyAnnotation (DebugPeerSelection addr)

Instance details

Methods

getPrivacyAnnotation :: DebugPeerSelection addr -> PrivacyAnnotation

HasPrivacyAnnotation (TracePeerSelection addr)

Instance details

Methods

getPrivacyAnnotation :: TracePeerSelection addr -> PrivacyAnnotation

HasPrivacyAnnotation (AbstractTransitionTrace addr)

Instance details

Methods

getPrivacyAnnotation :: AbstractTransitionTrace addr -> PrivacyAnnotation

HasPrivacyAnnotation (TraceSendRecv a)

Instance details

Methods

getPrivacyAnnotation :: TraceSendRecv a -> PrivacyAnnotation

HasPrivacyAnnotation (RemoteTransitionTrace addr)

Instance details

Methods

getPrivacyAnnotation :: RemoteTransitionTrace addr -> PrivacyAnnotation

HasPrivacyAnnotation (Trace addr)

Instance details

Methods

getPrivacyAnnotation :: Trace addr -> PrivacyAnnotation

HasPrivacyAnnotation (Trace addr)

Instance details

Methods

getPrivacyAnnotation :: Trace addr -> PrivacyAnnotation

HasPrivacyAnnotation (WithDomainName DnsTrace)

Instance details

Methods

getPrivacyAnnotation :: WithDomainName DnsTrace -> PrivacyAnnotation

HasPrivacyAnnotation (WithDomainName (SubscriptionTrace SockAddr))

Instance details

Methods

getPrivacyAnnotation :: WithDomainName (SubscriptionTrace SockAddr) -> PrivacyAnnotation

HasPrivacyAnnotation (WithIPList (SubscriptionTrace SockAddr))

Instance details

Methods

getPrivacyAnnotation :: WithIPList (SubscriptionTrace SockAddr) -> PrivacyAnnotation

HasPrivacyAnnotation [TraceLabelPeer peer (FetchDecision [Point header])]

Instance details

Methods

getPrivacyAnnotation :: [TraceLabelPeer peer (FetchDecision [Point header])] -> PrivacyAnnotation

HasSeverityAnnotation (Identity (SubscriptionTrace LocalAddress))

Instance details

Methods

getSeverityAnnotation :: Identity (SubscriptionTrace LocalAddress) -> Severity

HasSeverityAnnotation (TraceFetchClientState header)

Instance details

Methods

getSeverityAnnotation :: TraceFetchClientState header -> Severity

HasSeverityAnnotation (TraceKeepAliveClient remotePeer)

Instance details

Methods

getSeverityAnnotation :: TraceKeepAliveClient remotePeer -> Severity

HasSeverityAnnotation (DebugPeerSelection addr)

Instance details

Methods

getSeverityAnnotation :: DebugPeerSelection addr -> Severity

HasSeverityAnnotation (TracePeerSelection addr)

Instance details

Methods

getSeverityAnnotation :: TracePeerSelection addr -> Severity

HasSeverityAnnotation (AbstractTransitionTrace addr)

Instance details

Methods

getSeverityAnnotation :: AbstractTransitionTrace addr -> Severity

HasSeverityAnnotation (TraceSendRecv a)

Instance details

Methods

getSeverityAnnotation :: TraceSendRecv a -> Severity

HasSeverityAnnotation (RemoteTransitionTrace addr)

Instance details

Methods

getSeverityAnnotation :: RemoteTransitionTrace addr -> Severity

HasSeverityAnnotation (Trace addr)

Instance details

Methods

getSeverityAnnotation :: Trace addr -> Severity

HasSeverityAnnotation (Trace addr)

Instance details

Methods

getSeverityAnnotation :: Trace addr -> Severity

HasSeverityAnnotation (WithDomainName DnsTrace)

Instance details

Methods

getSeverityAnnotation :: WithDomainName DnsTrace -> Severity

HasSeverityAnnotation (WithDomainName (SubscriptionTrace SockAddr))

Instance details

Methods

getSeverityAnnotation :: WithDomainName (SubscriptionTrace SockAddr) -> Severity

HasSeverityAnnotation (WithIPList (SubscriptionTrace SockAddr))

Instance details

Methods

getSeverityAnnotation :: WithIPList (SubscriptionTrace SockAddr) -> Severity

HasSeverityAnnotation [TraceLabelPeer peer (FetchDecision [Point header])]

Instance details

Methods

getSeverityAnnotation :: [TraceLabelPeer peer (FetchDecision [Point header])] -> Severity

HasTextFormatter (Identity (SubscriptionTrace LocalAddress))

Instance details

Methods

formatText :: Identity (SubscriptionTrace LocalAddress) -> Object -> Text

Show addr => HasTextFormatter (TraceKeepAliveClient addr)

Instance details

Methods

formatText :: TraceKeepAliveClient addr -> Object -> Text

HasTextFormatter (DebugPeerSelection SockAddr)

Instance details

Methods

formatText :: DebugPeerSelection SockAddr -> Object -> Text

HasTextFormatter (TracePeerSelection SockAddr)

Instance details

Methods

formatText :: TracePeerSelection SockAddr -> Object -> Text

Show addr => HasTextFormatter (AbstractTransitionTrace addr)

Instance details

Methods

formatText :: AbstractTransitionTrace addr -> Object -> Text

Show addr => HasTextFormatter (RemoteTransitionTrace addr)

Instance details

Methods

formatText :: RemoteTransitionTrace addr -> Object -> Text

Show addr => HasTextFormatter (Trace addr)

Instance details

Methods

formatText :: Trace addr -> Object -> Text

Show addr => HasTextFormatter (Trace addr)

Instance details

Methods

formatText :: Trace addr -> Object -> Text

HasTextFormatter (WithDomainName DnsTrace)

Instance details

Methods

formatText :: WithDomainName DnsTrace -> Object -> Text

HasTextFormatter (WithDomainName (SubscriptionTrace SockAddr))

Instance details

Methods

formatText :: WithDomainName (SubscriptionTrace SockAddr) -> Object -> Text

HasTextFormatter (WithIPList (SubscriptionTrace SockAddr))

Instance details

Methods

formatText :: WithIPList (SubscriptionTrace SockAddr) -> Object -> Text

(StandardHash header, Show peer) => HasTextFormatter [TraceLabelPeer peer (FetchDecision [Point header])]

Instance details

Methods

formatText :: [TraceLabelPeer peer (FetchDecision [Point header])] -> Object -> Text

ToObject (Identity (SubscriptionTrace LocalAddress))

Instance details

Methods

toObject :: TracingVerbosity -> Identity (SubscriptionTrace LocalAddress) -> Object

textTransformer :: Identity (SubscriptionTrace LocalAddress) -> Object -> Text

(HasHeader header, ConvertRawHash header) => ToObject (TraceFetchClientState header)

Instance details

Methods

toObject :: TracingVerbosity -> TraceFetchClientState header -> Object

textTransformer :: TraceFetchClientState header -> Object -> Text

Show remotePeer => ToObject (TraceKeepAliveClient remotePeer)

Instance details

Methods

toObject :: TracingVerbosity -> TraceKeepAliveClient remotePeer -> Object

textTransformer :: TraceKeepAliveClient remotePeer -> Object -> Text

ToObject (DebugPeerSelection SockAddr)

Instance details

Methods

toObject :: TracingVerbosity -> DebugPeerSelection SockAddr -> Object

textTransformer :: DebugPeerSelection SockAddr -> Object -> Text

ToObject (TracePeerSelection SockAddr)

Instance details

Methods

toObject :: TracingVerbosity -> TracePeerSelection SockAddr -> Object

textTransformer :: TracePeerSelection SockAddr -> Object -> Text

ConvertRawHash blk => ToObject (Anchor blk)

Instance details

Methods

toObject :: TracingVerbosity -> Anchor blk -> Object

textTransformer :: Anchor blk -> Object -> Text

(ConvertRawHash blk, HasHeader blk) => ToObject (AnchoredFragment blk)

Instance details

Methods

toObject :: TracingVerbosity -> AnchoredFragment blk -> Object

textTransformer :: AnchoredFragment blk -> Object -> Text

(Show addr, ToObject addr, ToJSON addr) => ToObject (AbstractTransitionTrace addr)

Instance details

Methods

toObject :: TracingVerbosity -> AbstractTransitionTrace addr -> Object

textTransformer :: AbstractTransitionTrace addr -> Object -> Text

ToObject (AnyMessage ps) => ToObject (TraceSendRecv ps)

Instance details

Methods

toObject :: TracingVerbosity -> TraceSendRecv ps -> Object

textTransformer :: TraceSendRecv ps -> Object -> Text

ToJSON addr => ToObject (RemoteTransitionTrace addr)

Instance details

Methods

toObject :: TracingVerbosity -> RemoteTransitionTrace addr -> Object

textTransformer :: RemoteTransitionTrace addr -> Object -> Text

(ToJSON addr, Show addr) => ToObject (Trace addr)

Instance details

Methods

toObject :: TracingVerbosity -> Trace addr -> Object

textTransformer :: Trace addr -> Object -> Text

(Show addr, ToObject addr, ToJSON addr) => ToObject (Trace addr)

Instance details

Methods

toObject :: TracingVerbosity -> Trace addr -> Object

textTransformer :: Trace addr -> Object -> Text

ToObject (WithDomainName DnsTrace)

Instance details

Methods

toObject :: TracingVerbosity -> WithDomainName DnsTrace -> Object

textTransformer :: WithDomainName DnsTrace -> Object -> Text

ToObject (WithDomainName (SubscriptionTrace SockAddr))

Instance details

Methods

toObject :: TracingVerbosity -> WithDomainName (SubscriptionTrace SockAddr) -> Object

textTransformer :: WithDomainName (SubscriptionTrace SockAddr) -> Object -> Text

ToObject (WithIPList (SubscriptionTrace SockAddr))

Instance details

Methods

toObject :: TracingVerbosity -> WithIPList (SubscriptionTrace SockAddr) -> Object

textTransformer :: WithIPList (SubscriptionTrace SockAddr) -> Object -> Text

(ConvertTxId blk, RunNode blk, HasTxs blk) => ToObject (AnyMessage (BlockFetch blk (Point blk)))
instances of ToObject

NOTE: this list is sorted by the unqualified name of the outermost type.

Instance details

Methods

toObject :: TracingVerbosity -> AnyMessage (BlockFetch blk (Point blk)) -> Object

textTransformer :: AnyMessage (BlockFetch blk (Point blk)) -> Object -> Text

ToObject (AnyMessage (ChainSync blk pt tip))

Instance details

Methods

toObject :: TracingVerbosity -> AnyMessage (ChainSync blk pt tip) -> Object

textTransformer :: AnyMessage (ChainSync blk pt tip) -> Object -> Text

ToObject (AnyMessage KeepAlive)

Instance details

Methods

toObject :: TracingVerbosity -> AnyMessage KeepAlive -> Object

textTransformer :: AnyMessage KeepAlive -> Object -> Text

(forall result. Show (query result)) => ToObject (AnyMessage (LocalStateQuery blk pt query))

Instance details

Methods

toObject :: TracingVerbosity -> AnyMessage (LocalStateQuery blk pt query) -> Object

textTransformer :: AnyMessage (LocalStateQuery blk pt query) -> Object -> Text

ToObject (AnyMessage (LocalTxMonitor txid tx slotno))

Instance details

Methods

toObject :: TracingVerbosity -> AnyMessage (LocalTxMonitor txid tx slotno) -> Object

textTransformer :: AnyMessage (LocalTxMonitor txid tx slotno) -> Object -> Text

ToObject (AnyMessage (LocalTxSubmission tx err))

Instance details

Methods

toObject :: TracingVerbosity -> AnyMessage (LocalTxSubmission tx err) -> Object

textTransformer :: AnyMessage (LocalTxSubmission tx err) -> Object -> Text

(Show txid, Show tx) => ToObject (AnyMessage (TxSubmission2 txid tx))

Instance details

Methods

toObject :: TracingVerbosity -> AnyMessage (TxSubmission2 txid tx) -> Object

textTransformer :: AnyMessage (TxSubmission2 txid tx) -> Object -> Text

(ToJSON peer, ConvertRawHash header) => ToObject [TraceLabelPeer peer (FetchDecision [Point header])]

Instance details

Methods

toObject :: TracingVerbosity -> [TraceLabelPeer peer (FetchDecision [Point header])] -> Object

textTransformer :: [TraceLabelPeer peer (FetchDecision [Point header])] -> Object -> Text

(ToJSON peer, ToJSON point) => ToJSON (TraceLabelPeer peer (FetchDecision [point]))

Instance details

Methods

toJSON :: TraceLabelPeer peer (FetchDecision [point]) -> Value

toEncoding :: TraceLabelPeer peer (FetchDecision [point]) -> Encoding

toJSONList :: [TraceLabelPeer peer (FetchDecision [point])] -> Value

toEncodingList :: [TraceLabelPeer peer (FetchDecision [point])] -> Encoding

omitField :: TraceLabelPeer peer (FetchDecision [point]) -> Bool

ConvertRawHash header => ToJSON (Point header)

Instance details

Methods

toJSON :: Point header -> Value

toEncoding :: Point header -> Encoding

toJSONList :: [Point header] -> Value

toEncodingList :: [Point header] -> Encoding

omitField :: Point header -> Bool

HasPrivacyAnnotation a => HasPrivacyAnnotation (TraceLabelPeer peer a)

Instance details

Methods

getPrivacyAnnotation :: TraceLabelPeer peer a -> PrivacyAnnotation

HasPrivacyAnnotation (WithBearer peer Trace)

Instance details

Methods

getPrivacyAnnotation :: WithBearer peer Trace -> PrivacyAnnotation

HasPrivacyAnnotation (TraceDecisionEvent peer header)

Instance details

Methods

getPrivacyAnnotation :: TraceDecisionEvent peer header -> PrivacyAnnotation

HasPrivacyAnnotation (DiffusionTracer ntnAddr ntcAddr)

Instance details

Methods

getPrivacyAnnotation :: DiffusionTracer ntnAddr ntcAddr -> PrivacyAnnotation

HasPrivacyAnnotation (HandshakeTr RemoteAddress NodeToNodeVersion)

Instance details

Methods

getPrivacyAnnotation :: HandshakeTr RemoteAddress NodeToNodeVersion -> PrivacyAnnotation

HasPrivacyAnnotation (PeerSelectionActionsTrace SockAddr lAddr)

Instance details

Methods

getPrivacyAnnotation :: PeerSelectionActionsTrace SockAddr lAddr -> PrivacyAnnotation

HasPrivacyAnnotation (TraceLocalRootPeers RemoteAddress exception)

Instance details

Methods

getPrivacyAnnotation :: TraceLocalRootPeers RemoteAddress exception -> PrivacyAnnotation

HasPrivacyAnnotation (TraceTxSubmissionInbound txid tx)

Instance details

Methods

getPrivacyAnnotation :: TraceTxSubmissionInbound txid tx -> PrivacyAnnotation

HasPrivacyAnnotation (TraceTxSubmissionOutbound txid tx)

Instance details

Methods

getPrivacyAnnotation :: TraceTxSubmissionOutbound txid tx -> PrivacyAnnotation

HasPrivacyAnnotation (Trace addr connTrace)

Instance details

Methods

getPrivacyAnnotation :: Trace addr connTrace -> PrivacyAnnotation

HasPrivacyAnnotation (TraceSendRecv (LocalStateQuery blk (Point blk) (Query blk)) f)

Instance details

Methods

getPrivacyAnnotation :: TraceSendRecv (LocalStateQuery blk (Point blk) (Query blk)) f -> PrivacyAnnotation

HasPrivacyAnnotation (WithAddr addr ErrorPolicyTrace)

Instance details

Methods

getPrivacyAnnotation :: WithAddr addr ErrorPolicyTrace -> PrivacyAnnotation

HasSeverityAnnotation a => HasSeverityAnnotation (TraceLabelPeer peer a)

Instance details

Methods

getSeverityAnnotation :: TraceLabelPeer peer a -> Severity

HasSeverityAnnotation (WithBearer peer Trace)

Instance details

Methods

getSeverityAnnotation :: WithBearer peer Trace -> Severity

HasSeverityAnnotation (TraceDecisionEvent peer header)

Instance details

Methods

getSeverityAnnotation :: TraceDecisionEvent peer header -> Severity

HasSeverityAnnotation (DiffusionTracer ntnAddr ntcAddr)

Instance details

Methods

getSeverityAnnotation :: DiffusionTracer ntnAddr ntcAddr -> Severity

HasSeverityAnnotation (HandshakeTr RemoteAddress NodeToNodeVersion)

Instance details

Methods

getSeverityAnnotation :: HandshakeTr RemoteAddress NodeToNodeVersion -> Severity

HasSeverityAnnotation (PeerSelectionActionsTrace SockAddr lAddr)

Instance details

Methods

getSeverityAnnotation :: PeerSelectionActionsTrace SockAddr lAddr -> Severity

HasSeverityAnnotation (TraceLocalRootPeers RemoteAddress exception)

Instance details

Methods

getSeverityAnnotation :: TraceLocalRootPeers RemoteAddress exception -> Severity

HasSeverityAnnotation (TraceTxSubmissionInbound txid tx)

Instance details

Methods

getSeverityAnnotation :: TraceTxSubmissionInbound txid tx -> Severity

HasSeverityAnnotation (TraceTxSubmissionOutbound txid tx)

Instance details

Methods

getSeverityAnnotation :: TraceTxSubmissionOutbound txid tx -> Severity

HasSeverityAnnotation (Trace addr (ConnectionHandlerTrace versionNumber agreedOptions))

Instance details

Methods

getSeverityAnnotation :: Trace addr (ConnectionHandlerTrace versionNumber agreedOptions) -> Severity

HasSeverityAnnotation (TraceSendRecv (LocalStateQuery blk (Point blk) (Query blk)) f)

Instance details

Methods

getSeverityAnnotation :: TraceSendRecv (LocalStateQuery blk (Point blk) (Query blk)) f -> Severity

HasSeverityAnnotation (WithAddr addr ErrorPolicyTrace)

Instance details

Methods

getSeverityAnnotation :: WithAddr addr ErrorPolicyTrace -> Severity

(Show header, StandardHash header, Show peer) => HasTextFormatter (TraceLabelPeer peer (TraceFetchClientState header))

Instance details

Methods

formatText :: TraceLabelPeer peer (TraceFetchClientState header) -> Object -> Text

(Show peer, StandardHash blk, Show (Header blk)) => HasTextFormatter (TraceLabelPeer peer (TraceSendRecv (ChainSync (Header blk) (Point blk) (Tip blk))))

Instance details

Methods

formatText :: TraceLabelPeer peer (TraceSendRecv (ChainSync (Header blk) (Point blk) (Tip blk))) -> Object -> Text

Show peer => HasTextFormatter (WithBearer peer Trace)

Instance details

Methods

formatText :: WithBearer peer Trace -> Object -> Text

(StandardHash header, Show peer) => HasTextFormatter (TraceDecisionEvent peer header)

Instance details

Methods

formatText :: TraceDecisionEvent peer header -> Object -> Text

HasTextFormatter (DiffusionTracer RemoteAddress LocalAddress)

Instance details

Methods

formatText :: DiffusionTracer RemoteAddress LocalAddress -> Object -> Text

HasTextFormatter (HandshakeTr RemoteAddress NodeToNodeVersion)

Instance details

Methods

formatText :: HandshakeTr RemoteAddress NodeToNodeVersion -> Object -> Text

Show lAddr => HasTextFormatter (PeerSelectionActionsTrace SockAddr lAddr)

Instance details

Methods

formatText :: PeerSelectionActionsTrace SockAddr lAddr -> Object -> Text

Show exception => HasTextFormatter (TraceLocalRootPeers RemoteAddress exception)

Instance details

Methods

formatText :: TraceLocalRootPeers RemoteAddress exception -> Object -> Text

HasTextFormatter (TraceTxSubmissionInbound txid tx)

Instance details

Methods

formatText :: TraceTxSubmissionInbound txid tx -> Object -> Text

(Show tx, Show txid) => HasTextFormatter (TraceTxSubmissionOutbound txid tx)

Instance details

Methods

formatText :: TraceTxSubmissionOutbound txid tx -> Object -> Text

(Show addr, Show versionNumber, Show agreedOptions) => HasTextFormatter (Trace addr (ConnectionHandlerTrace versionNumber agreedOptions))

Instance details

Methods

formatText :: Trace addr (ConnectionHandlerTrace versionNumber agreedOptions) -> Object -> Text

Show addr => HasTextFormatter (WithAddr addr ErrorPolicyTrace)

Instance details

Methods

formatText :: WithAddr addr ErrorPolicyTrace -> Object -> Text

(ToObject peer, ToObject a) => ToObject (TraceLabelPeer peer a)

Instance details

Methods

toObject :: TracingVerbosity -> TraceLabelPeer peer a -> Object

textTransformer :: TraceLabelPeer peer a -> Object -> Text

ToObject peer => ToObject (WithBearer peer Trace)

Instance details

Methods

toObject :: TracingVerbosity -> WithBearer peer Trace -> Object

textTransformer :: WithBearer peer Trace -> Object -> Text

(ToJSON peer, ConvertRawHash header) => ToObject (TraceDecisionEvent peer header)

Instance details

Methods

toObject :: TracingVerbosity -> TraceDecisionEvent peer header -> Object

textTransformer :: TraceDecisionEvent peer header -> Object -> Text

(Show ntnAddr, Show ntcAddr) => ToObject (DiffusionTracer ntnAddr ntcAddr)

Instance details

Methods

toObject :: TracingVerbosity -> DiffusionTracer ntnAddr ntcAddr -> Object

textTransformer :: DiffusionTracer ntnAddr ntcAddr -> Object -> Text

ToObject (HandshakeTr RemoteAddress NodeToNodeVersion)

Instance details

Methods

toObject :: TracingVerbosity -> HandshakeTr RemoteAddress NodeToNodeVersion -> Object

textTransformer :: HandshakeTr RemoteAddress NodeToNodeVersion -> Object -> Text

Show lAddr => ToObject (PeerSelectionActionsTrace SockAddr lAddr)

Instance details

Methods

toObject :: TracingVerbosity -> PeerSelectionActionsTrace SockAddr lAddr -> Object

textTransformer :: PeerSelectionActionsTrace SockAddr lAddr -> Object -> Text

Show exception => ToObject (TraceLocalRootPeers RemoteAddress exception)

Instance details

Methods

toObject :: TracingVerbosity -> TraceLocalRootPeers RemoteAddress exception -> Object

textTransformer :: TraceLocalRootPeers RemoteAddress exception -> Object -> Text

ToObject (TraceTxSubmissionInbound txid tx)

Instance details

Methods

toObject :: TracingVerbosity -> TraceTxSubmissionInbound txid tx -> Object

textTransformer :: TraceTxSubmissionInbound txid tx -> Object -> Text

(Show txid, Show tx) => ToObject (TraceTxSubmissionOutbound txid tx)

Instance details

Methods

toObject :: TracingVerbosity -> TraceTxSubmissionOutbound txid tx -> Object

textTransformer :: TraceTxSubmissionOutbound txid tx -> Object -> Text

ConvertRawHash blk => ToObject (Point blk)

Instance details

Methods

toObject :: TracingVerbosity -> Point blk -> Object

textTransformer :: Point blk -> Object -> Text

(Show versionNumber, ToJSON versionNumber, ToJSON agreedOptions) => ToObject (ConnectionHandlerTrace versionNumber agreedOptions)

Instance details

Methods

toObject :: TracingVerbosity -> ConnectionHandlerTrace versionNumber agreedOptions -> Object

textTransformer :: ConnectionHandlerTrace versionNumber agreedOptions -> Object -> Text

(Show addr, Show versionNumber, Show agreedOptions, ToObject addr, ToJSON addr, ToJSON versionNumber, ToJSON agreedOptions) => ToObject (Trace addr (ConnectionHandlerTrace versionNumber agreedOptions))

Instance details

Methods

toObject :: TracingVerbosity -> Trace addr (ConnectionHandlerTrace versionNumber agreedOptions) -> Object

textTransformer :: Trace addr (ConnectionHandlerTrace versionNumber agreedOptions) -> Object -> Text

ToObject (AnyMessage ps f) => ToObject (TraceSendRecv ps f)

Instance details

Methods

toObject :: TracingVerbosity -> TraceSendRecv ps f -> Object

textTransformer :: TraceSendRecv ps f -> Object -> Text

Show addr => ToObject (WithAddr addr ErrorPolicyTrace)

Instance details

Methods

toObject :: TracingVerbosity -> WithAddr addr ErrorPolicyTrace -> Object

textTransformer :: WithAddr addr ErrorPolicyTrace -> Object -> Text

(forall result. Show (query result)) => ToObject (AnyMessage (LocalStateQuery blk pt query) f)

Instance details

Methods

toObject :: TracingVerbosity -> AnyMessage (LocalStateQuery blk pt query) f -> Object

textTransformer :: AnyMessage (LocalStateQuery blk pt query) f -> Object -> Text

HasPrivacyAnnotation (HandshakeTr LocalAddress NodeToClientVersion)

Instance details

Methods

getPrivacyAnnotation :: HandshakeTr LocalAddress NodeToClientVersion -> PrivacyAnnotation

HasSeverityAnnotation (HandshakeTr LocalAddress NodeToClientVersion)

Instance details

Methods

getSeverityAnnotation :: HandshakeTr LocalAddress NodeToClientVersion -> Severity

HasTextFormatter (HandshakeTr LocalAddress NodeToClientVersion)

Instance details

Methods

formatText :: HandshakeTr LocalAddress NodeToClientVersion -> Object -> Text

ToObject (HandshakeTr LocalAddress NodeToClientVersion)

Instance details

Methods

toObject :: TracingVerbosity -> HandshakeTr LocalAddress NodeToClientVersion -> Object

textTransformer :: HandshakeTr LocalAddress NodeToClientVersion -> Object -> Text