Control.Tracer.Transformers.WithThreadAndTime

Description

Observing events with annotations of thread id and time.

transformer

data WithThreadAndTime a

Add some operational context, time and thread

note that this could, for example, be an instance of ToJSON or Generic or similar to project it into a more general framework

Constructors

WithThreadAndTime

Fields

occurredAt :: !SystemTime

withinThread :: !ThreadId

event :: !a

threadAndTimeTracer :: MonadIO m => Tracer m (WithThreadAndTime a) -> Tracer m a

Add the time and thread to a trace observation