Options.Applicative.Help.Pretty

Documentation

type Doc = Doc Ann

(.$.) :: Doc -> Doc -> Doc

(</>) :: Doc -> Doc -> Doc

groupOrNestLine :: Doc -> Doc

Render flattened text on this line, or start a new line before rendering any text.

This will also nest subsequent lines in the group.

altSep :: Doc -> Doc -> Doc

Separate items in an alternative with a pipe.

If the first document and the pipe don't fit on the line, then mandatorily flow the next entry onto the following line.

The (//) softbreak ensures that if the document does fit on the line, there is at least a space, but it's possible for y to still appear on the next line.

hangAtIfOver :: Int -> Int -> Doc -> Doc

Printer hacks to get nice indentation for long commands and subcommands.

If we're starting this section over the desired width (usually 1/3 of the ribbon), then we will make a line break, indent all of the usage, and go.

The ifAtRoot is an interesting clause. If this whole operation is put under a group then the linebreak will disappear; then item d will therefore not be at the starting column, and it won't be indented more.

data Ann

Constructors

AnnTrace Int String

AnnStyle SetStyle

enclose

Arguments

:: Doc
L

-> Doc
R

-> Doc
x

-> Doc
LxR

Traced version of enclose.

parens :: Doc -> Doc

Traced version of parens.

brackets :: Doc -> Doc

Traced version of brackets.

hang :: Int -> Doc -> Doc

Traced version of hang.

indent :: Int -> Doc -> Doc

Traced version of indent.

nest :: Int -> Doc -> Doc

Traced version of nest.

text :: String -> Doc

plain :: Doc -> Doc

deunderline :: Doc -> Doc

underline :: Doc -> Doc

debold :: Doc -> Doc

bold :: Doc -> Doc

ondullwhite :: Doc -> Doc

onwhite :: Doc -> Doc

ondullcyan :: Doc -> Doc

oncyan :: Doc -> Doc

ondullmagenta :: Doc -> Doc

onmagenta :: Doc -> Doc

ondullblue :: Doc -> Doc

onblue :: Doc -> Doc

ondullyellow :: Doc -> Doc

onyellow :: Doc -> Doc

ondullgreen :: Doc -> Doc

ongreen :: Doc -> Doc

ondullred :: Doc -> Doc

onred :: Doc -> Doc

ondullblack :: Doc -> Doc

onblack :: Doc -> Doc

dullwhite :: Doc -> Doc

white :: Doc -> Doc

dullcyan :: Doc -> Doc

cyan :: Doc -> Doc

dullmagenta :: Doc -> Doc

magenta :: Doc -> Doc

dullblue :: Doc -> Doc

blue :: Doc -> Doc

dullyellow :: Doc -> Doc

yellow :: Doc -> Doc

dullgreen :: Doc -> Doc

green :: Doc -> Doc

dullred :: Doc -> Doc

red :: Doc -> Doc

dullblack :: Doc -> Doc

black :: Doc -> Doc

(</>) :: Doc -> Doc -> Doc

(<$$>) :: Doc -> Doc -> Doc

(<//>) :: Doc -> Doc -> Doc

string :: String -> Doc

Traced version of string.

isEffectivelyEmpty :: Doc -> Bool

Determine if the document is empty when rendered

renderShowS :: SimpleDocStream ann -> ShowS