System.ThreadManager

Description

A thread manager including a time manager. The manager has responsibility to kill managed threads.

Documentation

data ThreadManager

Manager to manage the thread and the timer.

newThreadManager :: Manager -> IO ThreadManager

Starting a thread manager. Its action is initially set to 'return ()' and should be set by setAction. This allows that the action can include the manager itself.

stopAfter :: ThreadManager -> IO a -> (Maybe SomeException -> IO ()) -> IO a

Stopping the manager.

The action is run in the scope of an exception handler that catches all exceptions (including asynchronous ones); this allows the cleanup handler to cleanup in all circumstances. If an exception is caught, it is rethrown after the cleanup is complete.

data KilledByThreadManager

An exception used internally to kill a managed thread.

Constructors

KilledByThreadManager (Maybe SomeException)

Fork

forkManaged

Arguments

:: ThreadManager

-> String
Thread name

-> IO ()
Action

-> IO ()

Fork a managed thread.

This guarantees that the thread ID is added to the manager's queue before the thread starts, and is removed again when the thread terminates (normally or abnormally).

forkManagedFinally :: ThreadManager -> String -> IO () -> IO () -> IO ()

Fork a managed thread with a cleanup function.

forkManagedUnmask :: ThreadManager -> String -> ((forall x. IO x -> IO x) -> IO ()) -> IO ()

Like forkManaged, but run action with exceptions masked

forkManagedTimeout :: ThreadManager -> String -> (Handle -> IO ()) -> IO ()

Fork a managed thread with a handle created by a timeout manager.

forkManagedTimeoutFinally :: ThreadManager -> String -> (Handle -> IO ()) -> IO () -> IO ()

Fork a managed thread with a handle created by a timeout manager and with a cleanup function.

Synchronization

waitUntilAllGone :: ThreadManager -> IO ()

Wait until all managed thread are finished.

Re-exports

data Manager

A timeout manager

withHandle :: ThreadManager -> TimeoutAction -> (Handle -> IO a) -> IO (Maybe a)

data Handle

A handle used by a timeout manager.

tickle :: Handle -> IO ()

Setting the state to active. Manager turns active to inactive repeatedly.

pause :: Handle -> IO ()

Setting the state to paused. Manager does not change the value.

resume :: Handle -> IO ()

Setting the paused state to active. This is an alias to tickle.